gio
Enum DBusError

#[non_exhaustive]
pub enum DBusError {

    Failed,
    NoMemory,
    ServiceUnknown,
    NameHasNoOwner,
    NoReply,
    IoError,
    BadAddress,
    NotSupported,
    LimitsExceeded,
    AccessDenied,
    AuthFailed,
    NoServer,
    Timeout,
    NoNetwork,
    AddressInUse,
    Disconnected,
    InvalidArgs,
    FileNotFound,
    FileExists,
    UnknownMethod,
    TimedOut,
    MatchRuleNotFound,
    MatchRuleInvalid,
    SpawnExecFailed,
    SpawnForkFailed,
    SpawnChildExited,
    SpawnChildSignaled,
    SpawnFailed,
    SpawnSetupFailed,
    SpawnConfigInvalid,
    SpawnServiceInvalid,
    SpawnServiceNotFound,
    SpawnPermissionsInvalid,
    SpawnFileInvalid,
    SpawnNoMemory,
    UnixProcessIdUnknown,
    InvalidSignature,
    InvalidFileContent,
    SelinuxSecurityContextUnknown,
    AdtAuditDataUnknown,
    ObjectPathInUse,
    UnknownObject,
    UnknownInterface,
    UnknownProperty,
    PropertyReadOnly,

}

Error codes for the G_DBUS_ERROR error domain.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Failed

A generic error; “something went wrong” - see the error message for more.

NoMemory

There was not enough memory to complete an operation.

ServiceUnknown

The bus doesn’t know how to launch a service to supply the bus name you wanted.

NameHasNoOwner

The bus name you referenced doesn’t exist (i.e. no application owns it).

NoReply

No reply to a message expecting one, usually means a timeout occurred.

IoError

Something went wrong reading or writing to a socket, for example.

BadAddress

A D-Bus bus address was malformed.

NotSupported

Requested operation isn’t supported (like ENOSYS on UNIX).

LimitsExceeded

Some limited resource is exhausted.

AccessDenied

Security restrictions don’t allow doing what you’re trying to do.

AuthFailed

Authentication didn’t work.

NoServer

Unable to connect to server (probably caused by ECONNREFUSED on a socket).

Timeout

Certain timeout errors, possibly ETIMEDOUT on a socket. Note that NoReply is used for message reply timeouts. Warning: this is confusingly-named given that TimedOut also exists. We can’t fix it for compatibility reasons so just be careful.

NoNetwork

No network access (probably ENETUNREACH on a socket).

AddressInUse

Can’t bind a socket since its address is in use (i.e. EADDRINUSE).

Disconnected

The connection is disconnected and you’re trying to use it.

InvalidArgs

Invalid arguments passed to a method call.

FileNotFound

Missing file.

FileExists

Existing file and the operation you’re using does not silently overwrite.

UnknownMethod

Method name you invoked isn’t known by the object you invoked it on.

TimedOut

Certain timeout errors, e.g. while starting a service. Warning: this is confusingly-named given that Timeout also exists. We can’t fix it for compatibility reasons so just be careful.

MatchRuleNotFound

Tried to remove or modify a match rule that didn’t exist.

MatchRuleInvalid

The match rule isn’t syntactically valid.

SpawnExecFailed

While starting a new process, the exec() call failed.

SpawnForkFailed

While starting a new process, the fork() call failed.

SpawnChildExited

While starting a new process, the child exited with a status code.

SpawnChildSignaled

While starting a new process, the child exited on a signal.

SpawnFailed

While starting a new process, something went wrong.

SpawnSetupFailed

We failed to setup the environment correctly.

SpawnConfigInvalid

We failed to setup the config parser correctly.

SpawnServiceInvalid

Bus name was not valid.

SpawnServiceNotFound

Service file not found in system-services directory.

SpawnPermissionsInvalid

Permissions are incorrect on the setuid helper.

SpawnFileInvalid

Service file invalid (Name, User or Exec missing).

SpawnNoMemory

Tried to get a UNIX process ID and it wasn’t available.

UnixProcessIdUnknown

Tried to get a UNIX process ID and it wasn’t available.

InvalidSignature

A type signature is not valid.

InvalidFileContent

A file contains invalid syntax or is otherwise broken.

SelinuxSecurityContextUnknown

Asked for SELinux security context and it wasn’t available.

AdtAuditDataUnknown

Asked for ADT audit data and it wasn’t available.

ObjectPathInUse

There’s already an object with the requested object path.

UnknownObject

Object you invoked a method on isn’t known. Since 2.42

UnknownInterface

Interface you invoked a method on isn’t known by the object. Since 2.42

UnknownProperty

Property you tried to access isn’t known by the object. Since 2.42

PropertyReadOnly

Property you tried to set is read-only. Since 2.42

Implementations

impl DBusError

pub fn encode_gerror(error: &Error) -> GString

pub fn remote_error(error: &Error) -> Option<GString>

pub fn is_remote_error(error: &Error) -> bool

pub fn new_for_dbus_error( dbus_error_name: &str, dbus_error_message: &str, ) -> Error

pub fn register_error( error_domain: Quark, error_code: i32, dbus_error_name: &str, ) -> bool

pub fn strip_remote_error(error: &mut Error) -> bool

pub fn unregister_error( error_domain: Quark, error_code: i32, dbus_error_name: &str, ) -> bool