Enum gio::IOErrorEnum

#[non_exhaustive]pub enum IOErrorEnum {
    Failed,
    NotFound,
    Exists,
    IsDirectory,
    NotDirectory,
    NotEmpty,
    NotRegularFile,
    NotSymbolicLink,
    NotMountableFile,
    FilenameTooLong,
    InvalidFilename,
    TooManyLinks,
    NoSpace,
    InvalidArgument,
    PermissionDenied,
    NotSupported,
    NotMounted,
    AlreadyMounted,
    Closed,
    Cancelled,
    Pending,
    ReadOnly,
    CantCreateBackup,
    WrongEtag,
    TimedOut,
    WouldRecurse,
    Busy,
    WouldBlock,
    HostNotFound,
    WouldMerge,
    FailedHandled,
    TooManyOpenFiles,
    NotInitialized,
    AddressInUse,
    PartialInput,
    InvalidData,
    DbusError,
    HostUnreachable,
    NetworkUnreachable,
    ConnectionRefused,
    ProxyFailed,
    ProxyAuthFailed,
    ProxyNeedAuth,
    ProxyNotAllowed,
    BrokenPipe,
    NotConnected,
    MessageTooLarge,
    // some variants omitted
}

Error codes returned by GIO functions.

Note that this domain may be extended in future GLib releases. In general, new error codes either only apply to new APIs, or else replace IOErrorEnum::Failed in cases that were not explicitly distinguished before. You should therefore avoid writing code like

if (g_error_matches (error, G_IO_ERROR, G_IO_ERROR_FAILED))
  {
    // Assume that this is EPRINTERONFIRE
    ...
  }

but should instead treat all unrecognized error codes the same as IOErrorEnum::Failed.

See also PollableReturn for a cheaper way of returning IOErrorEnum::WouldBlock to callers without allocating a glib::Error.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Failed

Generic error condition for when an operation fails and no more specific IOErrorEnum value is defined.

NotFound

File not found.

Exists

File already exists.

IsDirectory

File is a directory.

NotDirectory

File is not a directory.

NotEmpty

File is a directory that isn't empty.

NotRegularFile

File is not a regular file.

NotSymbolicLink

File is not a symbolic link.

NotMountableFile

File cannot be mounted.

FilenameTooLong

Filename is too many characters.

InvalidFilename

Filename is invalid or contains invalid characters.

TooManyLinks

File contains too many symbolic links.

NoSpace

No space left on drive.

InvalidArgument

Invalid argument.

PermissionDenied

Permission denied.

NotSupported

Operation (or one of its parameters) not supported

NotMounted

File isn't mounted.

AlreadyMounted

File is already mounted.

Closed

File was closed.

Cancelled

Operation was cancelled. See Cancellable.

Pending

Operations are still pending.

ReadOnly

File is read only.

CantCreateBackup

Backup couldn't be created.

WrongEtag

File's Entity Tag was incorrect.

TimedOut

Operation timed out.

WouldRecurse

Operation would be recursive.

Busy

File is busy.

WouldBlock

Operation would block.

HostNotFound

Host couldn't be found (remote operations).

WouldMerge

Operation would merge files.

FailedHandled

Operation failed and a helper program has already interacted with the user. Do not display any error dialog.

TooManyOpenFiles

The current process has too many files open and can't open any more. Duplicate descriptors do count toward this limit. Since 2.20

NotInitialized

The object has not been initialized. Since 2.22

AddressInUse

The requested address is already in use. Since 2.22

PartialInput

Need more input to finish operation. Since 2.24

InvalidData

The input data was invalid. Since 2.24

DbusError

A remote object generated an error that doesn't correspond to a locally registered glib::Error error domain. Use DBusError::get_remote_error to extract the D-Bus error name and DBusError::strip_remote_error to fix up the message so it matches what was received on the wire. Since 2.26.

HostUnreachable

Host unreachable. Since 2.26

NetworkUnreachable

Network unreachable. Since 2.26

ConnectionRefused

Connection refused. Since 2.26

ProxyFailed

Connection to proxy server failed. Since 2.26

ProxyAuthFailed

Proxy authentication failed. Since 2.26

ProxyNeedAuth

Proxy server needs authentication. Since 2.26

ProxyNotAllowed

Proxy connection is not allowed by ruleset. Since 2.26

BrokenPipe

Broken pipe. Since 2.36

NotConnected

Transport endpoint is not connected. Since 2.44

MessageTooLarge

Message too large. Since 2.48.