Function gio::bus_get_sync

pub fn bus_get_sync(
    bus_type: BusType,
    cancellable: Option<&impl IsA<Cancellable>>
) -> Result<DBusConnection, Error>

Synchronously connects to the message bus specified by bus_type. Note that the returned object may shared with other callers, e.g. if two separate parts of a process calls this function with the same bus_type, they will share the same object.

This is a synchronous failable function. See bus_get() and g_bus_get_finish() for the asynchronous version.

The returned object is a singleton, that is, shared with other callers of bus_get() and bus_get_sync() for bus_type. In the event that you need a private message bus connection, use dbus_address_get_for_bus_sync() and DBusConnection::for_address() with G_DBUS_CONNECTION_FLAGS_AUTHENTICATION_CLIENT and G_DBUS_CONNECTION_FLAGS_MESSAGE_BUS_CONNECTION flags.

Note that the returned DBusConnection object will (usually) have the exit-on-close property set to true.

bus_type

a BusType

cancellable

a Cancellable or None

Returns

a DBusConnection or None if error is set. Free with g_object_unref().