Function gio::dbus_address_get_stream

pub fn dbus_address_get_stream<P: IsA<Cancellable>, Q: FnOnce(Result<(IOStream, Option<GString>), Error>) + Send + 'static>(
    address: &str, 
    cancellable: Option<&P>, 
    callback: Q
)

Asynchronously connects to an endpoint specified by address and sets up the connection so it is in a state to run the client-side of the D-Bus authentication conversation. address must be in the D-Bus address format.

When the operation is finished, callback will be invoked. You can then call g_dbus_address_get_stream_finish() to get the result of the operation.

This is an asynchronous failable function. See g_dbus_address_get_stream_sync() for the synchronous version.

address

A valid D-Bus address.

cancellable

A crate::Cancellable or None.

callback

A GAsyncReadyCallback to call when the request is satisfied.