Function gio::resources_get_info

pub fn resources_get_info(
    path: &str,
    lookup_flags: ResourceLookupFlags
) -> Result<(usize, u32), Error>

Looks for a file at the specified path in the set of globally registered resources and if found returns information about it.

lookup_flags controls the behaviour of the lookup.

path

A pathname inside the resource

lookup_flags

A ResourceLookupFlags

Returns

true if the file was found. false if there were errors

size

a location to place the length of the contents of the file, or None if the length is not needed

flags

a location to place the GResourceFlags about the file, or None if the flags are not needed