gio::functions

Function bus_get_sync

pub fn bus_get_sync(
    bus_type: BusType,
    cancellable: Option<&impl IsA<Cancellable>>,
) -> Result<DBusConnection, Error>

Synchronously connects to the message bus specified by @bus_type. Note that the returned object may shared with other callers, e.g. if two separate parts of a process calls this function with the same @bus_type, they will share the same object.

This is a synchronous failable function. See g_bus_get() and g_bus_get_finish() for the asynchronous version.

The returned object is a singleton, that is, shared with other callers of g_bus_get() and g_bus_get_sync() for @bus_type. In the event that you need a private message bus connection, use g_dbus_address_get_for_bus_sync() and g_dbus_connection_new_for_address() with G_DBUS_CONNECTION_FLAGS_AUTHENTICATION_CLIENT and G_DBUS_CONNECTION_FLAGS_MESSAGE_BUS_CONNECTION flags.

Note that the returned #GDBusConnection object will (usually) have the #GDBusConnection:exit-on-close property set to true.

bus_type

a #GBusType

cancellable

a #GCancellable or None

Returns

a #GDBusConnection or None if @error is set. Free with g_object_unref().