gio::functions

Function content_types_get_registered

pub fn content_types_get_registered() -> Vec<GString>

Gets a list of strings containing all the registered content types known to the system. The list and its data should be freed using g_list_free_full (list, g_free).

Returns

list of the registered content types