gio::functions

Function io_modules_scan_all_in_directory

pub fn io_modules_scan_all_in_directory(dirname: impl AsRef<Path>)

Scans all the modules in the specified directory, ensuring that any extension point implemented by a module is registered.

This may not actually load and initialize all the types in each module, some modules may be lazily loaded and initialized when an extension point it implements is used with e.g. g_io_extension_point_get_extensions() or g_io_extension_point_get_extension_by_name().

If you need to guarantee that all types are loaded in all the modules, use g_io_modules_load_all_in_directory().

dirname

pathname for a directory containing modules to scan.