gio::functions

Function resources_enumerate_children

pub fn resources_enumerate_children(
    path: &str,
    lookup_flags: ResourceLookupFlags,
) -> Result<Vec<GString>, Error>

Returns all the names of children at the specified @path in the set of globally registered resources.

The return result is a NULL terminated list of strings which should be released with strfreev().

@lookup_flags controls the behaviour of the lookup.

path

A path name inside the resource

lookup_flags

A ResourceLookupFlags

Returns

an array of constant strings