gio::functions

Function resources_get_info

pub fn resources_get_info(
    path: &str,
    lookup_flags: ResourceLookupFlags,
) -> Result<(usize, u32), Error>

Looks for a file at the specified @path in the set of globally registered resources and if found returns information about it.

@lookup_flags controls the behaviour of the lookup.

path

A path name inside the resource

lookup_flags

A ResourceLookupFlags

Returns

TRUE if the file was found, FALSE if there were errors

size

a location to place the length of the contents of the file, or NULL if the length is not needed

flags

a location to place the ResourceFlags about the file, or NULL if the flags are not needed