Function gio::functions::resources_lookup_data

pub fn resources_lookup_data(
    path: &str, 
    lookup_flags: ResourceLookupFlags
) -> Result<Bytes, Error>

Looks for a file at the specified path in the set of globally registered resources and returns a glib::Bytes that lets you directly access the data in memory.

The data is always followed by a zero byte, so you can safely use the data as a C string. However, that byte is not included in the size of the GBytes.

For uncompressed resource files this is a pointer directly into the resource bundle, which is typically in some readonly data section in the program binary. For compressed files we allocate memory on the heap and automatically uncompress the data.

lookup_flags controls the behaviour of the lookup.

path

A pathname inside the resource

lookup_flags

A ResourceLookupFlags

Returns

glib::Bytes or None on error. Free the returned object with g_bytes_unref()