Function gio::functions::resources_open_stream

pub fn resources_open_stream(
    path: &str, 
    lookup_flags: ResourceLookupFlags
) -> Result<InputStream, Error>

Looks for a file at the specified path in the set of globally registered resources and returns a InputStream that lets you read the data.

lookup_flags controls the behaviour of the lookup.

path

A pathname inside the resource

lookup_flags

A ResourceLookupFlags

Returns

InputStream or None on error. Free the returned object with g_object_unref()