gio::functions

Function unix_is_mount_path_system_internal

pub fn unix_is_mount_path_system_internal(mount_path: impl AsRef<Path>) -> bool

Determines if @mount_path is considered an implementation of the OS.

This is primarily used for hiding mountable and mounted volumes that only are used in the OS and has little to no relevance to the casual user.

mount_path

a mount path, e.g. /media/disk or /usr

Returns

true if @mount_path is considered an implementation detail of the OS; false otherwise