gio::functions

Function unix_is_system_device_path

pub fn unix_is_system_device_path(device_path: impl AsRef<Path>) -> bool

Available on Unix only.

Determines if @device_path is considered a block device path which is only used in implementation of the OS.

This is primarily used for hiding mounted volumes that are intended as APIs for programs to read, and system administrators at a shell; rather than something that should, for example, appear in a GUI. For example, the Linux /proc filesystem.

The list of device paths considered ‘system’ ones may change over time.

device_path

a device path, e.g. /dev/loop0 or nfsd

Returns

true if @device_path is considered an implementation detail of the OS; false otherwise