Trait gio::prelude::ActionMapExt

pub trait ActionMapExt: 'static {
    fn add_action<P: IsA<Action>>(&self, action: &P);

    fn lookup_action(&self, action_name: &str) -> Option<Action>;

    fn remove_action(&self, action_name: &str);
}

Trait containing all ActionMap methods.

Implementors

crate::ActionMap, crate::Application, crate::SimpleActionGroup

Required methods

fn add_action<P: IsA<Action>>(&self, action: &P)

Adds an action to the self.

If the action map already contains an action with the same name as action then the old action is dropped from the action map.

The action map takes its own reference on action.

action

a crate::Action

fn lookup_action(&self, action_name: &str) -> Option<Action>

Looks up the action with the name action_name in self.

If no such action exists, returns None.

action_name

the name of an action

Returns

a crate::Action, or None

fn remove_action(&self, action_name: &str)

Removes the named action from the action map.

If no action of this name is in the map then nothing happens.

action_name

the name of the action

Loading content...