gio::prelude

Trait AppInfoExtManual

pub trait AppInfoExtManual: IsA<AppInfo> + 'static {
    // Provided methods
    fn launch_uris_async<P: IsA<AppLaunchContext>, Q: IsA<Cancellable>, R: FnOnce(Result<(), Error>) + 'static>(
        &self,
        uris: &[&str],
        context: Option<&P>,
        cancellable: Option<&Q>,
        callback: R,
    ) { ... }
    fn launch_uris_future<P: IsA<AppLaunchContext> + Clone + 'static>(
        &self,
        uris: &[&str],
        context: Option<&P>,
    ) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>> { ... }
}

Provided Methods

fn launch_uris_async<P: IsA<AppLaunchContext>, Q: IsA<Cancellable>, R: FnOnce(Result<(), Error>) + 'static>( &self, uris: &[&str], context: Option<&P>, cancellable: Option<&Q>, callback: R, )

Available on crate feature v2_60 only.

Async version of AppInfoExt::launch_uris().

The @callback is invoked immediately after the application launch, but it waits for activation in case of D-Bus–activated applications and also provides extended error information for sandboxed applications, see notes for AppInfo::launch_default_for_uri_async().

uris

a list of URIs to launch.

context

the launch context

cancellable

a Cancellable

callback

a [type@Gio.AsyncReadyCallback] to call when the request is done

fn launch_uris_future<P: IsA<AppLaunchContext> + Clone + 'static>( &self, uris: &[&str], context: Option<&P>, ) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

Available on crate feature v2_60 only.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.