gio::prelude

Trait BufferedOutputStreamExt

pub trait BufferedOutputStreamExt: IsA<BufferedOutputStream> + 'static {
    // Provided methods
    fn auto_grows(&self) -> bool { ... }
    fn buffer_size(&self) -> usize { ... }
    fn set_auto_grow(&self, auto_grow: bool) { ... }
    fn set_buffer_size(&self, size: usize) { ... }
    fn connect_auto_grow_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
    fn connect_buffer_size_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all BufferedOutputStream methods.

Implementors

BufferedOutputStream

Provided Methods

fn auto_grows(&self) -> bool

Checks if the buffer automatically grows as data is added.

Returns

TRUE if the @self’s buffer automatically grows, FALSE otherwise.

fn buffer_size(&self) -> usize

Gets the size of the buffer in the @self.

Returns

the current size of the buffer.

fn set_auto_grow(&self, auto_grow: bool)

Sets whether or not the @self’s buffer should automatically grow. If @auto_grow is true, then each write will just make the buffer larger, and you must manually flush the buffer to actually write out the data to the underlying stream.

auto_grow

a #gboolean.

fn set_buffer_size(&self, size: usize)

Sets the size of the internal buffer to @size.

size

a #gsize.

fn connect_auto_grow_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

fn connect_buffer_size_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.