Trait gio::prelude::ConverterExt

pub trait ConverterExt: 'static {
    fn reset(&self);
}

Trait containing all Converter methods.

Implementors

CharsetConverter, Converter, ZlibCompressor, ZlibDecompressor

Required Methods

fn reset(&self)

Resets all internal state in the converter, making it behave as if it was just created. If the converter has any internal state that would produce output then that output is lost.