Trait gio::prelude::CredentialsExt

pub trait CredentialsExt: 'static {
    fn get_unix_pid(&self) -> Result<i32, Error>;

    fn get_unix_user(&self) -> Result<(), Error>;

    fn is_same_user<P: IsA<Credentials>>(
        &self, 
        other_credentials: &P
    ) -> Result<(), Error>;

    fn set_unix_user(&self, uid: u32) -> Result<(), Error>;

    fn to_string(&self) -> GString;
}

Trait containing all Credentials methods.

Implementors

Credentials

Required methods

fn get_unix_pid(&self) -> Result<i32, Error>

Tries to get the UNIX process identifier from self. This method is only available on UNIX platforms.

This operation can fail if Credentials is not supported on the OS or if the native credentials type does not contain information about the UNIX process ID.

Returns

The UNIX process ID, or -1 if error is set.

fn get_unix_user(&self) -> Result<(), Error>

Tries to get the UNIX user identifier from self. This method is only available on UNIX platforms.

This operation can fail if Credentials is not supported on the OS or if the native credentials type does not contain information about the UNIX user.

Returns

The UNIX user identifier or -1 if error is set.

fn is_same_user<P: IsA<Credentials>>(
&self,
other_credentials: &P
) -> Result<(), Error>

Checks if self and other_credentials is the same user.

This operation can fail if Credentials is not supported on the the OS.

other_credentials

A Credentials.

Returns

true if self and other_credentials has the same user, false otherwise or if error is set.

fn set_unix_user(&self, uid: u32) -> Result<(), Error>

Tries to set the UNIX user identifier on self. This method is only available on UNIX platforms.

This operation can fail if Credentials is not supported on the OS or if the native credentials type does not contain information about the UNIX user. It can also fail if the OS does not allow the use of "spoofed" credentials.

uid

The UNIX user identifier to set.

Returns

true if uid was set, false if error is set.

fn to_string(&self) -> GString

Creates a human-readable textual representation of self that can be used in logging and debug messages. The format of the returned string may change in future GLib release.

Returns

A string that should be freed with g_free.

Loading content...