gio::prelude

Trait DBusObjectManagerClientExt

pub trait DBusObjectManagerClientExt: IsA<DBusObjectManagerClient> + 'static {
    // Provided methods
    fn connection(&self) -> DBusConnection { ... }
    fn flags(&self) -> DBusObjectManagerClientFlags { ... }
    fn name(&self) -> GString { ... }
    fn name_owner(&self) -> Option<GString> { ... }
    fn connect_interface_proxy_signal<F: Fn(&Self, &DBusObjectProxy, &DBusProxy, &str, &str, &Variant) + Send + Sync + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
    fn connect_name_owner_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all DBusObjectManagerClient methods.

Implementors

DBusObjectManagerClient

Provided Methods

fn connection(&self) -> DBusConnection

Gets the #GDBusConnection used by @self.

Returns

A #GDBusConnection object. Do not free, the object belongs to @self.

fn flags(&self) -> DBusObjectManagerClientFlags

Gets the flags that @self was constructed with.

Returns

Zero of more flags from the #GDBusObjectManagerClientFlags enumeration.

fn name(&self) -> GString

Gets the name that @self is for, or None if not a message bus connection.

Returns

A unique or well-known name. Do not free, the string belongs to @self.

fn name_owner(&self) -> Option<GString>

The unique name that owns the name that @self is for or None if no-one currently owns that name. You can connect to the #GObject::notify signal to track changes to the #GDBusObjectManagerClient:name-owner property.

Returns

The name owner or None if no name owner exists. Free with g_free().

fn connect_interface_proxy_signal<F: Fn(&Self, &DBusObjectProxy, &DBusProxy, &str, &str, &Variant) + Send + Sync + 'static>( &self, f: F, ) -> SignalHandlerId

Emitted when a D-Bus signal is received on @interface_proxy.

This signal exists purely as a convenience to avoid having to connect signals to all interface proxies managed by @manager.

This signal is emitted in the thread-default main context (see [glib::MainContext::push_thread_default()][crate::glib::MainContext::push_thread_default()]) that @manager was constructed in.

object_proxy

The #GDBusObjectProxy on which an interface is emitting a D-Bus signal.

interface_proxy

The #GDBusProxy that is emitting a D-Bus signal.

sender_name

The sender of the signal or NULL if the connection is not a bus connection.

signal_name

The signal name.

parameters

A #GVariant tuple with parameters for the signal.

fn connect_name_owner_notify<F: Fn(&Self) + Send + Sync + 'static>( &self, f: F, ) -> SignalHandlerId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.