gio::prelude

Trait FileEnumeratorExtManual

pub trait FileEnumeratorExtManual: IsA<FileEnumerator> {
    // Provided methods
    fn into_stream(
        self,
        num_files: i32,
        priority: Priority,
    ) -> FileEnumeratorStream { ... }
    fn close(
        &self,
        cancellable: Option<&impl IsA<Cancellable>>,
    ) -> (bool, Option<Error>) { ... }
}

Provided Methods

fn into_stream(self, num_files: i32, priority: Priority) -> FileEnumeratorStream

Converts the enumerator into a Stream.

fn close( &self, cancellable: Option<&impl IsA<Cancellable>>, ) -> (bool, Option<Error>)

Releases all resources used by this enumerator, making the enumerator return IOErrorEnum::Closed on all calls.

This will be automatically called when the last reference is dropped, but you might want to call this function to make sure resources are released as early as possible.

cancellable

optional #GCancellable object, None to ignore.

Returns

#TRUE on success or #FALSE on error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.