Trait gio::prelude::FilterOutputStreamExt

pub trait FilterOutputStreamExt: 'static {
    fn base_stream(&self) -> OutputStream;

    fn closes_base_stream(&self) -> bool;

    fn set_close_base_stream(&self, close_base: bool);

    fn connect_close_base_stream_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all FilterOutputStream methods.

Implementors

crate::BufferedOutputStream, crate::ConverterOutputStream, crate::DataOutputStream, crate::FilterOutputStream

Required methods

fn base_stream(&self) -> OutputStream

Gets the base stream for the filter stream.

Returns

a crate::OutputStream.

fn closes_base_stream(&self) -> bool

Returns whether the base stream will be closed when self is closed.

Returns

true if the base stream will be closed.

fn set_close_base_stream(&self, close_base: bool)

Sets whether the base stream will be closed when self is closed.

close_base

true to close the base stream.

fn connect_close_base_stream_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...