Trait gio::prelude::InetAddressMaskExt

pub trait InetAddressMaskExt: 'static {
    fn equal<P: IsA<InetAddressMask>>(&self, mask2: &P) -> bool;

    fn address(&self) -> InetAddress;

    fn family(&self) -> SocketFamily;

    fn length(&self) -> u32;

    fn matches<P: IsA<InetAddress>>(&self, address: &P) -> bool;

    fn to_str(&self) -> GString;

    fn set_address<P: IsA<InetAddress>>(&self, address: Option<&P>);

    fn set_length(&self, length: u32);

    fn connect_address_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_family_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_length_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all InetAddressMask methods.

Implementors

crate::InetAddressMask

Required methods

fn equal<P: IsA<InetAddressMask>>(&self, mask2: &P) -> bool

Tests if self and mask2 are the same mask.

mask2

another crate::InetAddressMask

Returns

whether self and mask2 are the same mask

fn address(&self) -> InetAddress

Gets self’s base address

Returns

self’s base address

fn family(&self) -> SocketFamily

Gets the crate::SocketFamily of self’s address

Returns

the crate::SocketFamily of self’s address

fn length(&self) -> u32

Gets self’s length

Returns

self’s length

fn matches<P: IsA<InetAddress>>(&self, address: &P) -> bool

Tests if address falls within the range described by self.

address

a crate::InetAddress

Returns

whether address falls within the range described by self.

fn to_str(&self) -> GString

fn set_address<P: IsA<InetAddress>>(&self, address: Option<&P>)

fn set_length(&self, length: u32)

fn connect_address_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_family_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_length_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...