Trait gio::prelude::ListModelExtManual

pub trait ListModelExtManual: Sealed + IsA<ListModel> + Sized {
    // Provided methods
    fn snapshot(&self) -> Vec<Object> { ... }
    fn iter<LT: IsA<Object>>(&self) -> ListModelIter<'_, LT> { ... }
}

Provided Methods

fn snapshot(&self) -> Vec<Object>

Get an immutable snapshot of the container inside the ListModel. Any modification done to the returned container Vec will not be reflected on the ListModel.

fn iter<LT: IsA<Object>>(&self) -> ListModelIter<'_, LT>

If T::static_type().is_a(self.item_type()) then it returns an iterator over the ListModel elements, else the types are not compatible and returns an Err(...).

Panics

Panics if T::static_type().is_a(self.item_type()) is not true.