Trait gio::prelude::ListStoreExtManual

pub trait ListStoreExtManual {
    fn insert_sorted<P: IsA<Object>, F: FnMut(&Object, &Object) -> Ordering>(
        &self, 
        item: &P, 
        compare_func: F
    ) -> u32;
}

Required methods

fn insert_sorted<P: IsA<Object>, F: FnMut(&Object, &Object) -> Ordering>(
&self,
item: &P,
compare_func: F
) -> u32

Inserts item into self at a position to be determined by the compare_func.

The list must already be sorted before calling this function or the result is undefined. Usually you would approach this by only ever inserting items by way of this function.

This function takes a ref on item.

Feature: v2_44

item

the new item

compare_func

pairwise comparison function for sorting

user_data

user data for compare_func

Returns

the position at which item was inserted

Loading content...