Trait gio::prelude::MemoryMonitorExt

pub trait MemoryMonitorExt: 'static {
    fn connect_low_memory_warning<F: Fn(&Self, MemoryMonitorWarningLevel) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all MemoryMonitor methods.

Implementors

crate::MemoryMonitor

Required methods

fn connect_low_memory_warning<F: Fn(&Self, MemoryMonitorWarningLevel) + 'static>(
&self,
f: F
) -> SignalHandlerId

This is supported on crate feature v2_64 only.

Emitted when the system is running low on free memory. The signal handler should then take the appropriate action depending on the warning level. See the crate::MemoryMonitorWarningLevel documentation for details.

level

the crate::MemoryMonitorWarningLevel warning level

Loading content...