Trait gio::prelude::MemoryOutputStreamExt

pub trait MemoryOutputStreamExt: 'static {
    fn get_data_size(&self) -> usize;

    fn steal_as_bytes(&self) -> Option<Bytes>;

    fn connect_property_data_size_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all MemoryOutputStream methods.

Implementors

MemoryOutputStream

Required methods

fn get_data_size(&self) -> usize

Returns the number of bytes from the start up to including the last byte written in the stream that has not been truncated away.

Returns

the number of bytes written to the stream

fn steal_as_bytes(&self) -> Option<Bytes>

Returns data from the self as a glib::Bytes. self must be closed before calling this function.

Returns

the stream's data

fn connect_property_data_size_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...