Trait gio::prelude::PollableInputStreamExt

pub trait PollableInputStreamExt: 'static {
    fn can_poll(&self) -> bool;

    fn is_readable(&self) -> bool;
}

Trait containing all PollableInputStream methods.

Implementors

crate::ConverterInputStream, crate::MemoryInputStream, crate::PollableInputStream, crate::UnixInputStream

Required methods

fn can_poll(&self) -> bool

Checks if self is actually pollable. Some classes may implement crate::PollableInputStream but have only certain instances of that class be pollable. If this method returns false, then the behavior of other crate::PollableInputStream methods is undefined.

For any given stream, the value returned by this method is constant; a stream cannot switch from pollable to non-pollable or vice versa.

Returns

true if self is pollable, false if not.

fn is_readable(&self) -> bool

Checks if self can be read.

Note that some stream types may not be able to implement this 100% reliably, and it is possible that a call to [crate::InputStream::read()] after this returns true would still block. To guarantee non-blocking behavior, you should always use [crate::PollableInputStream::read_nonblocking()], which will return a crate::IOErrorEnum::WouldBlock error rather than blocking.

Returns

true if self is readable, false if not. If an error has occurred on self, this will result in [crate::PollableInputStream::is_readable()] returning true, and the next attempt to read will return the error.

Loading content...