Trait gio::prelude::ProxyResolverExt

pub trait ProxyResolverExt: 'static {
    fn is_supported(&self) -> bool;

    fn lookup<P: IsA<Cancellable>>(
        &self, 
        uri: &str, 
        cancellable: Option<&P>
    ) -> Result<Vec<GString>, Error>;

    fn lookup_async<P: IsA<Cancellable>, Q: FnOnce(Result<Vec<GString>, Error>) + Send + 'static>(
        &self, 
        uri: &str, 
        cancellable: Option<&P>, 
        callback: Q
    );

    fn lookup_async_future(
        &self, 
        uri: &str
    ) -> Pin<Box_<dyn Future<Output = Result<Vec<GString>, Error>> + 'static>>;
}

Trait containing all ProxyResolver methods.

Implementors

crate::ProxyResolver

Required methods

fn is_supported(&self) -> bool

Checks if self can be used on this system. (This is used internally; [crate::ProxyResolver::get_default()] will only return a proxy resolver that returns true for this method.)

Returns

true if self is supported.

fn lookup<P: IsA<Cancellable>>(
&self,
uri: &str,
cancellable: Option<&P>
) -> Result<Vec<GString>, Error>

Looks into the system proxy configuration to determine what proxy, if any, to use to connect to uri. The returned proxy URIs are of the form <protocol>://[user[:password]@]host:port or direct://, where <protocol> could be http, rtsp, socks or other proxying protocol.

If you don’t know what network protocol is being used on the socket, you should use none as the URI protocol. In this case, the resolver might still return a generic proxy type (such as SOCKS), but would not return protocol-specific proxy types (such as http).

direct:// is used when no proxy is needed. Direct connection should not be attempted unless it is part of the returned array of proxies.

uri

a URI representing the destination to connect to

cancellable

a crate::Cancellable, or None

Returns

A NULL-terminated array of proxy URIs. Must be freed with g_strfreev().

fn lookup_async<P: IsA<Cancellable>, Q: FnOnce(Result<Vec<GString>, Error>) + Send + 'static>(
&self,
uri: &str,
cancellable: Option<&P>,
callback: Q
)

Asynchronous lookup of proxy. See [crate::ProxyResolver::lookup()] for more details.

uri

a URI representing the destination to connect to

cancellable

a crate::Cancellable, or None

callback

callback to call after resolution completes

user_data

data for callback

fn lookup_async_future(
&self,
uri: &str
) -> Pin<Box_<dyn Future<Output = Result<Vec<GString>, Error>> + 'static>>

Loading content...