Trait gio::prelude::SimpleActionGroupExt

pub trait SimpleActionGroupExt {
    fn insert<P: IsA<Action>>(&self, action: &P);

    fn lookup(&self, action_name: &str) -> Option<Action>;

    fn remove(&self, action_name: &str);
}

Trait containing all SimpleActionGroup methods.

Implementors

SimpleActionGroup

Required Methods

fn insert<P: IsA<Action>>(&self, action: &P)

Adds an action to the action group.

If the action group already contains an action with the same name as action then the old action is dropped from the group.

The action group takes its own reference on action.

Deprecated since 2.38

Use ActionMap::add_action

action

a Action

fn lookup(&self, action_name: &str) -> Option<Action>

Looks up the action with the name action_name in the group.

If no such action exists, returns None.

Deprecated since 2.38

Use ActionMap::lookup_action

action_name

the name of an action

Returns

a Action, or None

fn remove(&self, action_name: &str)

Removes the named action from the action group.

If no action of this name is in the group then nothing happens.

Deprecated since 2.38

Use ActionMap::remove_action

action_name

the name of the action