gio::prelude
Trait SocketAddressExt

pub trait SocketAddressExt: IsA<SocketAddress> + 'static {
    // Provided methods
    fn family(&self) -> SocketFamily { ... }
    fn native_size(&self) -> isize { ... }
    fn connect_family_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all SocketAddress methods.

Implementors

InetSocketAddress, SocketAddress, UnixSocketAddress

Provided Methods

fn family(&self) -> SocketFamily

Gets the socket family type of @self.

Returns

the socket family type of @self

fn native_size(&self) -> isize

Gets the size of @self’s native struct sockaddr. You can use this to allocate memory to pass to g_socket_address_to_native().

Returns

the size of the native struct sockaddr that @self represents

fn connect_family_notify<F: Fn(&Self) + Send + Sync + 'static>( &self, f: F, ) -> SignalHandlerId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.