Trait gio::prelude::SocketConnectableExt

pub trait SocketConnectableExt: 'static {
    fn enumerate(&self) -> SocketAddressEnumerator;

    fn proxy_enumerate(&self) -> SocketAddressEnumerator;

    fn to_string(&self) -> GString;
}

Trait containing all SocketConnectable methods.

Implementors

crate::InetSocketAddress, crate::NetworkAddress, crate::NetworkService, crate::ProxyAddress, crate::SocketAddress, crate::SocketConnectable, crate::UnixSocketAddress

Required methods

fn enumerate(&self) -> SocketAddressEnumerator

Creates a crate::SocketAddressEnumerator for self.

Returns

a new crate::SocketAddressEnumerator.

fn proxy_enumerate(&self) -> SocketAddressEnumerator

Creates a crate::SocketAddressEnumerator for self that will return a crate::ProxyAddress for each of its addresses that you must connect to via a proxy.

If self does not implement [crate::SocketConnectable::proxy_enumerate()], this will fall back to calling [crate::SocketConnectable::enumerate()].

Returns

a new crate::SocketAddressEnumerator.

fn to_string(&self) -> GString

Format a crate::SocketConnectable as a string. This is a human-readable format for use in debugging output, and is not a stable serialization format. It is not suitable for use in user interfaces as it exposes too much information for a user.

If the crate::SocketConnectable implementation does not support string formatting, the implementation’s type name will be returned as a fallback.

Returns

the formatted string

Loading content...