gio::prelude

Trait SocketControlMessageExt

pub trait SocketControlMessageExt: IsA<SocketControlMessage> + 'static {
    // Provided methods
    fn level(&self) -> i32 { ... }
    fn msg_type(&self) -> i32 { ... }
    fn size(&self) -> usize { ... }
}

Trait containing all SocketControlMessage methods.

Implementors

SocketControlMessage, UnixCredentialsMessage, UnixFDMessage

Provided Methods

fn level(&self) -> i32

Returns the “level” (i.e. the originating protocol) of the control message. This is often SOL_SOCKET.

Returns

an integer describing the level

fn msg_type(&self) -> i32

Returns the protocol specific type of the control message. For instance, for UNIX fd passing this would be SCM_RIGHTS.

Returns

an integer describing the type of control message

fn size(&self) -> usize

Returns the space required for the control message, not including headers or alignment.

Returns

The number of bytes required.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.