Trait gio::prelude::TlsCertificateExt

pub trait TlsCertificateExt: 'static {
    fn issuer(&self) -> Option<TlsCertificate>;

    fn is_same<P: IsA<TlsCertificate>>(&self, cert_two: &P) -> bool;

    fn verify<P: IsA<SocketConnectable>, Q: IsA<TlsCertificate>>(
        &self, 
        identity: Option<&P>, 
        trusted_ca: Option<&Q>
    ) -> TlsCertificateFlags;

    fn certificate(&self) -> Option<ByteArray>;

    fn certificate_pem(&self) -> Option<GString>;
}

Trait containing all TlsCertificate methods.

Implementors

crate::TlsCertificate

Required methods

fn issuer(&self) -> Option<TlsCertificate>

Gets the crate::TlsCertificate representing self’s issuer, if known

Returns

The certificate of self’s issuer, or None if self is self-signed or signed with an unknown certificate.

fn is_same<P: IsA<TlsCertificate>>(&self, cert_two: &P) -> bool

Check if two crate::TlsCertificate objects represent the same certificate. The raw DER byte data of the two certificates are checked for equality. This has the effect that two certificates may compare equal even if their [crate::TlsCertificate:issuer], [crate::TlsCertificate:private-key], or [crate::TlsCertificate:private-key-pem] properties differ.

cert_two

second certificate to compare

Returns

whether the same or not

fn verify<P: IsA<SocketConnectable>, Q: IsA<TlsCertificate>>(
&self,
identity: Option<&P>,
trusted_ca: Option<&Q>
) -> TlsCertificateFlags

This verifies self and returns a set of crate::TlsCertificateFlags indicating any problems found with it. This can be used to verify a certificate outside the context of making a connection, or to check a certificate against a CA that is not part of the system CA database.

If identity is not None, self’s name(s) will be compared against it, and [crate::TlsCertificateFlags::BadIdentity] will be set in the return value if it does not match. If identity is None, that bit will never be set in the return value.

If trusted_ca is not None, then self (or one of the certificates in its chain) must be signed by it, or else [crate::TlsCertificateFlags::UnknownCa] will be set in the return value. If trusted_ca is None, that bit will never be set in the return value.

(All other crate::TlsCertificateFlags values will always be set or unset as appropriate.)

identity

the expected peer identity

trusted_ca

the certificate of a trusted authority

Returns

the appropriate crate::TlsCertificateFlags

fn certificate(&self) -> Option<ByteArray>

fn certificate_pem(&self) -> Option<GString>

Loading content...