Trait gio::prelude::TlsClientConnectionExt

pub trait TlsClientConnectionExt: 'static {
    fn copy_session_state<P: IsA<TlsClientConnection>>(&self, source: &P);

    fn accepted_cas(&self) -> Vec<ByteArray>;

    fn server_identity(&self) -> Option<SocketConnectable>;

    fn uses_ssl3(&self) -> bool;

    fn validation_flags(&self) -> TlsCertificateFlags;

    fn set_server_identity<P: IsA<SocketConnectable>>(&self, identity: &P);

    fn set_use_ssl3(&self, use_ssl3: bool);

    fn set_validation_flags(&self, flags: TlsCertificateFlags);

    fn connect_accepted_cas_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_server_identity_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_use_ssl3_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_validation_flags_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all TlsClientConnection methods.

Implementors

crate::TlsClientConnection

Required methods

fn copy_session_state<P: IsA<TlsClientConnection>>(&self, source: &P)

Possibly copies session state from one connection to another, for use in TLS session resumption. This is not normally needed, but may be used when the same session needs to be used between different endpoints, as is required by some protocols, such as FTP over TLS. source should have already completed a handshake and, since TLS 1.3, it should have been used to read data at least once. self should not have completed a handshake.

It is not possible to know whether a call to this function will actually do anything. Because session resumption is normally used only for performance benefit, the TLS backend might not implement this function. Even if implemented, it may not actually succeed in allowing self to resume source’s TLS session, because the server may not have sent a session resumption token to source, or it may refuse to accept the token from self. There is no way to know whether a call to this function is actually successful.

Using this function is not required to benefit from session resumption. If the TLS backend supports session resumption, the session will be resumed automatically if it is possible to do so without weakening the privacy guarantees normally provided by TLS, without need to call this function. For example, with TLS 1.3, a session ticket will be automatically copied from any crate::TlsClientConnection that has previously received session tickets from the server, provided a ticket is available that has not previously been used for session resumption, since session ticket reuse would be a privacy weakness. Using this function causes the ticket to be copied without regard for privacy considerations.

source

a crate::TlsClientConnection

fn accepted_cas(&self) -> Vec<ByteArray>

Gets the list of distinguished names of the Certificate Authorities that the server will accept certificates from. This will be set during the TLS handshake if the server requests a certificate. Otherwise, it will be None.

Each item in the list is a crate::glib::ByteArray which contains the complete subject DN of the certificate authority.

Returns

the list of CA DNs. You should unref each element with [crate::glib::ByteArray::unref()] and then the free the list with [crate::glib::List::free()].

fn server_identity(&self) -> Option<SocketConnectable>

Gets self’s expected server identity

Returns

a crate::SocketConnectable describing the expected server identity, or None if the expected identity is not known.

fn uses_ssl3(&self) -> bool

SSL 3.0 is no longer supported. See [crate::TlsClientConnection::set_use_ssl3()] for details.

Deprecated since 2.56

SSL 3.0 is insecure.

Returns

false

fn validation_flags(&self) -> TlsCertificateFlags

Gets self’s validation flags

Returns

the validation flags

fn set_server_identity<P: IsA<SocketConnectable>>(&self, identity: &P)

Sets self’s expected server identity, which is used both to tell servers on virtual hosts which certificate to present, and also to let self know what name to look for in the certificate when performing [crate::TlsCertificateFlags::BadIdentity] validation, if enabled.

identity

a crate::SocketConnectable describing the expected server identity

fn set_use_ssl3(&self, use_ssl3: bool)

Since GLib 2.42.1, SSL 3.0 is no longer supported.

From GLib 2.42.1 through GLib 2.62, this function could be used to force use of TLS 1.0, the lowest-supported TLS protocol version at the time. In the past, this was needed to connect to broken TLS servers that exhibited protocol version intolerance. Such servers are no longer common, and using TLS 1.0 is no longer considered acceptable.

Since GLib 2.64, this function does nothing.

Deprecated since 2.56

SSL 3.0 is insecure.

use_ssl3

a gboolean, ignored

fn set_validation_flags(&self, flags: TlsCertificateFlags)

Sets self’s validation flags, to override the default set of checks performed when validating a server certificate. By default, [crate::TlsCertificateFlags::ValidateAll] is used.

flags

the crate::TlsCertificateFlags to use

fn connect_accepted_cas_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_server_identity_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_use_ssl3_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_validation_flags_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...