Trait gio::prelude::TlsFileDatabaseExt

pub trait TlsFileDatabaseExt:
    IsA<TlsFileDatabase>
    + Sealed
    + 'static {
    // Provided methods
    fn anchors(&self) -> Option<GString> { ... }
    fn set_anchors(&self, anchors: Option<&str>) { ... }
    fn connect_anchors_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all TlsFileDatabase methods.

Implementors

TlsFileDatabase

Provided Methods

fn anchors(&self) -> Option<GString>

The path to a file containing PEM encoded certificate authority root anchors. The certificates in this file will be treated as root authorities for the purpose of verifying other certificates via the g_tls_database_verify_chain() operation.

fn set_anchors(&self, anchors: Option<&str>)

The path to a file containing PEM encoded certificate authority root anchors. The certificates in this file will be treated as root authorities for the purpose of verifying other certificates via the g_tls_database_verify_chain() operation.

fn connect_anchors_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

Object Safety

This trait is not object safe.