Trait gio::prelude::TlsServerConnectionExt

pub trait TlsServerConnectionExt:
    IsA<TlsServerConnection>
    + Sealed
    + 'static {
    // Provided methods
    fn authentication_mode(&self) -> TlsAuthenticationMode { ... }
    fn set_authentication_mode(
        &self,
        authentication_mode: TlsAuthenticationMode,
    ) { ... }
    fn connect_authentication_mode_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all TlsServerConnection methods.

Implementors

TlsServerConnection

Provided Methods

fn authentication_mode(&self) -> TlsAuthenticationMode

The #GTlsAuthenticationMode for the server. This can be changed before calling g_tls_connection_handshake() if you want to rehandshake with a different mode from the initial handshake.

fn set_authentication_mode(&self, authentication_mode: TlsAuthenticationMode)

The #GTlsAuthenticationMode for the server. This can be changed before calling g_tls_connection_handshake() if you want to rehandshake with a different mode from the initial handshake.

fn connect_authentication_mode_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

Object Safety

This trait is not object safe.