Trait gio::prelude::UnixInputStreamExt

pub trait UnixInputStreamExt: 'static {
    fn closes_fd(&self) -> bool;
}

Trait containing all UnixInputStream methods.

Implementors

crate::UnixInputStream

Required methods

fn closes_fd(&self) -> bool

Returns whether the file descriptor of self will be closed when the stream is closed.

Returns

true if the file descriptor is closed when done

Loading content...