Trait gio::prelude::UnixOutputStreamExtManual

pub trait UnixOutputStreamExtManual: Sealed + IsA<UnixOutputStream> + Sized {
    // Provided method
    unsafe fn set_close_fd(&self, close_fd: bool) { ... }
}

Provided Methods

unsafe fn set_close_fd(&self, close_fd: bool)

Sets whether the fd of this stream will be closed when the stream is closed.

Safety

If you pass in false as the parameter, you may only close the fd if the all references to the stream have been dropped. If you pass in true, you must never call close.

Object Safety

This trait is not object safe.