Trait gio::prelude::UnixSocketAddressExtManual

pub trait UnixSocketAddressExtManual {
    fn get_path(&self) -> Option<UnixSocketAddressPath>;
}

Required methods

fn get_path(&self) -> Option<UnixSocketAddressPath>

Gets self's path, or for abstract sockets the "name".

Guaranteed to be zero-terminated, but an abstract socket may contain embedded zeros, and thus you should use UnixSocketAddressExt::get_path_len to get the true length of this string.

Returns

the path for self

Loading content...