Trait gio::prelude::VolumeExt

pub trait VolumeExt: 'static {

    fn can_eject(&self) -> bool;

    fn can_mount(&self) -> bool;

    fn eject_with_operation<P: IsA<MountOperation>, Q: IsA<Cancellable>, R: FnOnce(Result<(), Error>) + Send + 'static>(
        &self, 
        flags: MountUnmountFlags, 
        mount_operation: Option<&P>, 
        cancellable: Option<&Q>, 
        callback: R
    );

    fn eject_with_operation_future<P: IsA<MountOperation> + Clone + 'static>(
        &self, 
        flags: MountUnmountFlags, 
        mount_operation: Option<&P>
    ) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>;

    fn enumerate_identifiers(&self) -> Vec<GString>;

    fn activation_root(&self) -> Option<File>;

    fn drive(&self) -> Option<Drive>;

    fn icon(&self) -> Icon;

    fn get_identifier(&self, kind: &str) -> Option<GString>;

    fn get_mount(&self) -> Option<Mount>;

    fn name(&self) -> GString;

    fn sort_key(&self) -> Option<GString>;

    fn symbolic_icon(&self) -> Icon;

    fn uuid(&self) -> Option<GString>;

    fn mount<P: IsA<MountOperation>, Q: IsA<Cancellable>, R: FnOnce(Result<(), Error>) + Send + 'static>(
        &self, 
        flags: MountMountFlags, 
        mount_operation: Option<&P>, 
        cancellable: Option<&Q>, 
        callback: R
    );

    fn mount_future<P: IsA<MountOperation> + Clone + 'static>(
        &self, 
        flags: MountMountFlags, 
        mount_operation: Option<&P>
    ) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>;

    fn should_automount(&self) -> bool;

    fn connect_changed<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn connect_removed<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

}

Trait containing all Volume methods.

Implementors

Volume

Required methods

fn can_eject(&self) -> bool

Checks if a volume can be ejected.

Returns

true if the self can be ejected. false otherwise

fn can_mount(&self) -> bool

Checks if a volume can be mounted.

Returns

true if the self can be mounted. false otherwise

fn eject_with_operation<P: IsA<MountOperation>, Q: IsA<Cancellable>, R: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
flags: MountUnmountFlags,
mount_operation: Option<&P>,
cancellable: Option<&Q>,
callback: R
)

Ejects a volume. This is an asynchronous operation, and is finished by calling Volume::eject_with_operation_finish with the self and AsyncResult data returned in the callback.

flags

flags affecting the unmount if required for eject

mount_operation

a MountOperation or None to avoid user interaction

cancellable

optional Cancellable object, None to ignore

callback

a GAsyncReadyCallback, or None

user_data

user data passed to callback

fn eject_with_operation_future<P: IsA<MountOperation> + Clone + 'static>(
&self,
flags: MountUnmountFlags,
mount_operation: Option<&P>
) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

fn enumerate_identifiers(&self) -> Vec<GString>

Gets the kinds of [identifiers][volume-identifier] that self has. Use Volume::get_identifier to obtain the identifiers themselves.

Returns

a None-terminated array of strings containing kinds of identifiers. Use g_strfreev to free.

fn activation_root(&self) -> Option<File>

fn drive(&self) -> Option<Drive>

fn icon(&self) -> Icon

fn get_identifier(&self, kind: &str) -> Option<GString>

Gets the identifier of the given kind for self. See the [introduction][volume-identifier] for more information about volume identifiers.

kind

the kind of identifier to return

Returns

a newly allocated string containing the requested identifier, or None if the Volume doesn’t have this kind of identifier

fn get_mount(&self) -> Option<Mount>

Gets the mount for the self.

Returns

a Mount or None if self isn’t mounted. The returned object should be unreffed with gobject::ObjectExt::unref when no longer needed.

fn name(&self) -> GString

fn sort_key(&self) -> Option<GString>

fn symbolic_icon(&self) -> Icon

fn uuid(&self) -> Option<GString>

fn mount<P: IsA<MountOperation>, Q: IsA<Cancellable>, R: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
flags: MountMountFlags,
mount_operation: Option<&P>,
cancellable: Option<&Q>,
callback: R
)

Mounts a volume. This is an asynchronous operation, and is finished by calling Volume::mount_finish with the self and AsyncResult returned in the callback.

flags

flags affecting the operation

mount_operation

a MountOperation or None to avoid user interaction

cancellable

optional Cancellable object, None to ignore

callback

a GAsyncReadyCallback, or None

user_data

user data that gets passed to callback

fn mount_future<P: IsA<MountOperation> + Clone + 'static>(
&self,
flags: MountMountFlags,
mount_operation: Option<&P>
) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

fn should_automount(&self) -> bool

Returns whether the volume should be automatically mounted.

Returns

true if the volume should be automatically mounted

fn connect_changed<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

Emitted when the volume has been changed.

fn connect_removed<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is emitted when the Volume have been removed. If the recipient is holding references to the object they should release them so the object can be finalized.

Loading content...