Trait gio::prelude::VolumeMonitorExt

pub trait VolumeMonitorExt: 'static {

    fn connected_drives(&self) -> Vec<Drive>;

    fn mount_for_uuid(&self, uuid: &str) -> Option<Mount>;

    fn mounts(&self) -> Vec<Mount>;

    fn volume_for_uuid(&self, uuid: &str) -> Option<Volume>;

    fn volumes(&self) -> Vec<Volume>;

    fn connect_drive_changed<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drive_connected<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drive_disconnected<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drive_eject_button<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drive_stop_button<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_mount_added<F: Fn(&Self, &Mount) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_mount_changed<F: Fn(&Self, &Mount) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_mount_pre_unmount<F: Fn(&Self, &Mount) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_mount_removed<F: Fn(&Self, &Mount) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_volume_added<F: Fn(&Self, &Volume) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_volume_changed<F: Fn(&Self, &Volume) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_volume_removed<F: Fn(&Self, &Volume) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

}

Trait containing all VolumeMonitor methods.

Implementors

crate::VolumeMonitor

Required methods

fn connected_drives(&self) -> Vec<Drive>

Gets a list of drives connected to the system.

The returned list should be freed with [crate::glib::List::free()], after its elements have been unreffed with [crate::glib::object::ObjectExt::unref()].

Returns

a [crate::glib::List] of connected crate::Drive objects.

fn mount_for_uuid(&self, uuid: &str) -> Option<Mount>

Finds a crate::Mount object by its UUID (see [crate::Mount::get_uuid()])

uuid

the UUID to look for

Returns

a crate::Mount or None if no such mount is available. Free the returned object with [crate::glib::object::ObjectExt::unref()].

fn mounts(&self) -> Vec<Mount>

Gets a list of the mounts on the system.

The returned list should be freed with [crate::glib::List::free()], after its elements have been unreffed with [crate::glib::object::ObjectExt::unref()].

Returns

a [crate::glib::List] of crate::Mount objects.

fn volume_for_uuid(&self, uuid: &str) -> Option<Volume>

Finds a crate::Volume object by its UUID (see [crate::Volume::get_uuid()])

uuid

the UUID to look for

Returns

a crate::Volume or None if no such volume is available. Free the returned object with [crate::glib::object::ObjectExt::unref()].

fn volumes(&self) -> Vec<Volume>

Gets a list of the volumes on the system.

The returned list should be freed with [crate::glib::List::free()], after its elements have been unreffed with [crate::glib::object::ObjectExt::unref()].

Returns

a [crate::glib::List] of crate::Volume objects.

fn connect_drive_changed<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a drive changes.

drive

the drive that changed

fn connect_drive_connected<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a drive is connected to the system.

drive

a crate::Drive that was connected.

fn connect_drive_disconnected<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a drive is disconnected from the system.

drive

a crate::Drive that was disconnected.

fn connect_drive_eject_button<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when the eject button is pressed on drive.

drive

the drive where the eject button was pressed

fn connect_drive_stop_button<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when the stop button is pressed on drive.

drive

the drive where the stop button was pressed

fn connect_mount_added<F: Fn(&Self, &Mount) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mount is added.

mount

a crate::Mount that was added.

fn connect_mount_changed<F: Fn(&Self, &Mount) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mount changes.

mount

a crate::Mount that changed.

fn connect_mount_pre_unmount<F: Fn(&Self, &Mount) + 'static>(
&self,
f: F
) -> SignalHandlerId

May be emitted when a mount is about to be removed.

This signal depends on the backend and is only emitted if GIO was used to unmount.

mount

a crate::Mount that is being unmounted.

fn connect_mount_removed<F: Fn(&Self, &Mount) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mount is removed.

mount

a crate::Mount that was removed.

fn connect_volume_added<F: Fn(&Self, &Volume) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mountable volume is added to the system.

volume

a crate::Volume that was added.

fn connect_volume_changed<F: Fn(&Self, &Volume) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when mountable volume is changed.

volume

a crate::Volume that changed.

fn connect_volume_removed<F: Fn(&Self, &Volume) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mountable volume is removed from the system.

volume

a crate::Volume that was removed.

Loading content...