Trait gio::prelude::ZlibCompressorExt

pub trait ZlibCompressorExt: 'static {
    fn get_file_info(&self) -> Option<FileInfo>;

    fn set_file_info(&self, file_info: Option<&FileInfo>);

    fn get_property_format(&self) -> ZlibCompressorFormat;

    fn get_property_level(&self) -> i32;

    fn connect_property_file_info_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all ZlibCompressor methods.

Implementors

ZlibCompressor

Required methods

fn get_file_info(&self) -> Option<FileInfo>

Returns the ZlibCompressor:file-info property.

Returns

a FileInfo, or None

fn set_file_info(&self, file_info: Option<&FileInfo>)

Sets file_info in self. If non-None, and self's ZlibCompressor:format property is ZlibCompressorFormat::Gzip, it will be used to set the file name and modification time in the GZIP header of the compressed data.

Note: it is an error to call this function while a compression is in progress; it may only be called immediately after creation of self, or after resetting it with Converter::reset.

file_info

a FileInfo

fn get_property_format(&self) -> ZlibCompressorFormat

fn get_property_level(&self) -> i32

fn connect_property_file_info_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...