Struct gio::AppInfo

pub struct AppInfo(_, _);

AppInfo and AppLaunchContext are used for describing and launching applications installed on the system.

As of GLib 2.20, URIs will always be converted to POSIX paths (using File::get_path) when using AppInfo::launch even if the application requested an URI and not a POSIX path. For example for an desktop-file based application with Exec key totem %U and a single URI, sftp://foo/file.avi, then /home/user/.gvfs/sftp on foo/file.avi will be passed. This will only work if a set of suitable GIO extensions (such as gvfs 2.26 compiled with FUSE support), is available and operational; if this is not the case, the URI will be passed unmodified to the application. Some URIs, such as mailto:, of course cannot be mapped to a POSIX path (in gvfs there's no FUSE mount for it); such URIs will be passed unmodified to the application.

Specifically for gvfs 2.26 and later, the POSIX URI will be mapped back to the GIO URI in the File constructors (since gvfs implements the Vfs extension point). As such, if the application needs to examine the URI, it needs to use File::get_uri or similar on File. In other words, an application cannot assume that the URI passed to e.g. File::new_for_commandline_arg is equal to the result of File::get_uri. The following snippet illustrates this:

GFile *f;
char *uri;

file = g_file_new_for_commandline_arg (uri_from_commandline);

uri = g_file_get_uri (file);
strcmp (uri, uri_from_commandline) == 0;
g_free (uri);

if (g_file_has_uri_scheme (file, "cdda"))
  {
    // do something special with uri
  }
g_object_unref (file);

This code will work when both cdda://sr0/Track 1.wav and /home/user/.gvfs/cdda on sr0/Track 1.wav is passed to the application. It should be noted that it's generally not safe for applications to rely on the format of a particular URIs. Different launcher applications (e.g. file managers) may have different ideas of what a given URI means.

Implements

AppInfoExt

Methods

impl AppInfo

fn create_from_commandline<'a, P: Into<Option<&'a str>>>(
commandline: &str,
application_name: P,
flags: AppInfoCreateFlags
) -> Result<AppInfo, Error>

Creates a new AppInfo from the given information.

Note that for commandline, the quoting rules of the Exec key of the freedesktop.org Desktop Entry Specification are applied. For example, if the commandline contains percent-encoded URIs, the percent-character must be doubled in order to prevent it from being swallowed by Exec key unquoting. See the specification for exact quoting rules.

commandline

the commandline to use

application_name

the application name, or None to use commandline

flags

flags that can specify details of the created AppInfo

Returns

new AppInfo for given command.

fn get_all() -> Vec<AppInfo>

Gets a list of all of the applications currently registered on this system.

For desktop files, this includes applications that have NoDisplay=true set or are excluded from display by means of OnlyShowIn or NotShowIn. See AppInfo::should_show. The returned list does not include applications which have the Hidden key set.

Returns

a newly allocated glib::List of references to GAppInfos.

fn get_all_for_type(content_type: &str) -> Vec<AppInfo>

Gets a list of all GAppInfos for a given content type, including the recommended and fallback GAppInfos. See AppInfo::get_recommended_for_type and AppInfo::get_fallback_for_type.

content_type

the content type to find a AppInfo for

Returns

glib::List of GAppInfos for given content_type or None on error.

fn get_default_for_type(
content_type: &str,
must_support_uris: bool
) -> Option<AppInfo>

Gets the default AppInfo for a given content type.

content_type

the content type to find a AppInfo for

must_support_uris

if true, the AppInfo is expected to support URIs

Returns

AppInfo for given content_type or None on error.

fn get_default_for_uri_scheme(uri_scheme: &str) -> Option<AppInfo>

Gets the default application for handling URIs with the given URI scheme. A URI scheme is the initial part of the URI, up to but not including the ':', e.g. "http", "ftp" or "sip".

uri_scheme

a string containing a URI scheme.

Returns

AppInfo for given uri_scheme or None on error.

fn get_fallback_for_type(content_type: &str) -> Vec<AppInfo>

Gets a list of fallback GAppInfos for a given content type, i.e. those applications which claim to support the given content type by MIME type subclassing and not directly.

content_type

the content type to find a AppInfo for

Returns

glib::List of GAppInfos for given content_type or None on error.

fn get_recommended_for_type(content_type: &str) -> Vec<AppInfo>

Gets a list of recommended GAppInfos for a given content type, i.e. those applications which claim to support the given content type exactly, and not by MIME type subclassing. Note that the first application of the list is the last used one, i.e. the last one for which AppInfo::set_as_last_used_for_type has been called.

content_type

the content type to find a AppInfo for

Returns

glib::List of GAppInfos for given content_type or None on error.

fn launch_default_for_uri<'a, P: Into<Option<&'a AppLaunchContext>>>(
uri: &str,
launch_context: P
) -> Result<(), Error>

Utility function that launches the default application registered to handle the specified uri. Synchronous I/O is done on the uri to detect the type of the file if required.

uri

the uri to show

launch_context

an optional AppLaunchContext

Returns

true on success, false on error.

fn reset_type_associations(content_type: &str)

Removes all changes to the type associations done by AppInfo::set_as_default_for_type, AppInfo::set_as_default_for_extension, AppInfo::add_supports_type or AppInfo::remove_supports_type.

content_type

a content type