Struct gio::DBusConnection

pub struct DBusConnection(_);

The crate::DBusConnection type is used for D-Bus connections to remote peers such as a message buses. It is a low-level API that offers a lot of flexibility. For instance, it lets you establish a connection over any transport that can by represented as a crate::IOStream.

This class is rarely used directly in D-Bus clients. If you are writing a D-Bus client, it is often easier to use the g_bus_own_name(), g_bus_watch_name() or [crate::DBusProxy::new_for_bus()] APIs.

As an exception to the usual GLib rule that a particular object must not be used by two threads at the same time, crate::DBusConnection’s methods may be called from any thread. This is so that g_bus_get() and g_bus_get_sync() can safely return the same crate::DBusConnection when called from any thread.

Most of the ways to obtain a crate::DBusConnection automatically initialize it (i.e. connect to D-Bus): for instance, Self::new() and g_bus_get(), and the synchronous versions of those methods, give you an initialized connection. Language bindings for GIO should use [crate::Initable::new()] or [crate::AsyncInitable::new_async()], which also initialize the connection.

If you construct an uninitialized crate::DBusConnection, such as via crate::glib::object::Object::new(), you must initialize it via [crate::Initable::init()] or [crate::AsyncInitable::init_async()] before using its methods or properties. Calling methods or accessing properties on a crate::DBusConnection that has not completed initialization successfully is considered to be invalid, and leads to undefined behaviour. In particular, if initialization fails with a crate::glib::Error, the only valid thing you can do with that crate::DBusConnection is to free it with [crate::glib::object::ObjectExt::unref()].

An example D-Bus server # {gdbus-server}

Here is an example for a D-Bus server: gdbus-example-server.c

An example for exporting a subtree # {gdbus-subtree-server}

Here is an example for exporting a subtree: gdbus-example-subtree.c

An example for file descriptor passing # {gdbus-unix-fd-client}

Here is an example for passing UNIX file descriptors: gdbus-unix-fd-client.c

An example for exporting a GObject # {gdbus-export}

Here is an example for exporting a crate::glib::object::Object: gdbus-example-export.c

Implements

glib::object::ObjectExt

Implementations

impl DBusConnection

pub fn register_object<MethodCall, SetProperty, GetProperty>(
&self,
object_path: &str,
interface_info: &DBusInterfaceInfo,
method_call: MethodCall,
get_property: GetProperty,
set_property: SetProperty
) -> Result<RegistrationId, Error> where
MethodCall: Fn(DBusConnection, &str, &str, &str, &str, Variant, DBusMethodInvocation) + Send + Sync + 'static,
GetProperty: Fn(DBusConnection, &str, &str, &str, &str) -> Variant + Send + Sync + 'static,
SetProperty: Fn(DBusConnection, &str, &str, &str, &str, Variant) -> bool + Send + Sync + 'static,

Registers callbacks for exported objects at object_path with the D-Bus interface that is described in interface_info.

Calls to functions in vtable (and user_data_free_func) will happen in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this method from.

Note that all crate::glib::Variant values passed to functions in vtable will match the signature given in interface_info - if a remote caller passes incorrect values, the org.freedesktop.DBus.Error.InvalidArgs is returned to the remote caller.

Additionally, if the remote caller attempts to invoke methods or access properties not mentioned in interface_info the org.freedesktop.DBus.Error.UnknownMethod resp. org.freedesktop.DBus.Error.InvalidArgs errors are returned to the caller.

It is considered a programming error if the GDBusInterfaceGetPropertyFunc function in vtable returns a crate::glib::Variant of incorrect type.

If an existing callback is already registered at object_path and interface_name, then error is set to crate::IOErrorEnum::Exists.

GDBus automatically implements the standard D-Bus interfaces org.freedesktop.DBus.Properties, org.freedesktop.DBus.Introspectable and org.freedesktop.Peer, so you don’t have to implement those for the objects you export. You can implement org.freedesktop.DBus.Properties yourself, e.g. to handle getting and setting of properties asynchronously.

Note that the reference count on interface_info will be incremented by 1 (unless allocated statically, e.g. if the reference count is -1, see [crate::DBusInterfaceInfo::ref()]) for as long as the object is exported. Also note that vtable will be copied.

See this [server][gdbus-server] for an example of how to use this method.

object_path

the object path to register at

interface_info

introspection data for the interface

vtable

a [crate::DBusInterfaceVTable] to call into or None

user_data

data to pass to functions in vtable

user_data_free_func

function to call when the object path is unregistered

Returns

0 if error is set, otherwise a registration id (never 0) that can be used with Self::unregister_object()

pub fn unregister_object(
&self,
registration_id: RegistrationId
) -> Result<(), BoolError>

Unregisters an object.

registration_id

a registration id obtained from Self::register_object()

Returns

true if the object was unregistered, false otherwise

pub fn export_action_group<P: IsA<ActionGroup>>(
&self,
object_path: &str,
action_group: &P
) -> Result<ActionGroupExportId, Error>

Exports action_group on self at object_path.

The implemented D-Bus API should be considered private. It is subject to change in the future.

A given object path can only have one action group exported on it. If this constraint is violated, the export will fail and 0 will be returned (with error set accordingly).

You can unexport the action group using Self::unexport_action_group() with the return value of this function.

The thread default main context is taken at the time of this call. All incoming action activations and state change requests are reported from this context. Any changes on the action group that cause it to emit signals must also come from this same context. Since incoming action activations and state change requests are rather likely to cause changes on the action group, this effectively limits a given action group to being exported from only one main context.

object_path

a D-Bus object path

action_group

a crate::ActionGroup

Returns

the ID of the export (never zero), or 0 in case of failure

pub fn unexport_action_group(&self, export_id: ActionGroupExportId)

Reverses the effect of a previous call to Self::export_action_group().

It is an error to call this function with an ID that wasn’t returned from Self::export_action_group() or to call it with the same ID more than once.

export_id

the ID from Self::export_action_group()

pub fn export_menu_model<P: IsA<MenuModel>>(
&self,
object_path: &str,
menu: &P
) -> Result<MenuModelExportId, Error>

Exports menu on self at object_path.

The implemented D-Bus API should be considered private. It is subject to change in the future.

An object path can only have one menu model exported on it. If this constraint is violated, the export will fail and 0 will be returned (with error set accordingly).

You can unexport the menu model using Self::unexport_menu_model() with the return value of this function.

object_path

a D-Bus object path

menu

a crate::MenuModel

Returns

the ID of the export (never zero), or 0 in case of failure

pub fn unexport_menu_model(&self, export_id: MenuModelExportId)

Reverses the effect of a previous call to Self::export_menu_model().

It is an error to call this function with an ID that wasn’t returned from Self::export_menu_model() or to call it with the same ID more than once.

export_id

the ID from Self::export_menu_model()

pub fn add_filter<P: Fn(&DBusConnection, &DBusMessage, bool) -> Option<DBusMessage> + 'static>(
&self,
filter_function: P
) -> FilterId

Adds a message filter. Filters are handlers that are run on all incoming and outgoing messages, prior to standard dispatch. Filters are run in the order that they were added. The same handler can be added as a filter more than once, in which case it will be run more than once. Filters added during a filter callback won’t be run on the message being processed. Filter functions are allowed to modify and even drop messages.

Note that filters are run in a dedicated message handling thread so they can’t block and, generally, can’t do anything but signal a worker thread. Also note that filters are rarely needed - use API such as Self::send_message_with_reply(), Self::signal_subscribe() or Self::call() instead.

If a filter consumes an incoming message the message is not dispatched anywhere else - not even the standard dispatch machinery (that API such as Self::signal_subscribe() and Self::send_message_with_reply() relies on) will see the message. Similarly, if a filter consumes an outgoing message, the message will not be sent to the other peer.

If user_data_free_func is non-None, it will be called (in the thread-default main context of the thread you are calling this method from) at some point after user_data is no longer needed. (It is not guaranteed to be called synchronously when the filter is removed, and may be called after self has been destroyed.)

filter_function

a filter function

user_data

user data to pass to filter_function

user_data_free_func

function to free user_data with when filter is removed or None

Returns

a filter identifier that can be used with Self::remove_filter()

pub fn remove_filter(&self, filter_id: FilterId)

Removes a filter.

Note that since filters run in a different thread, there is a race condition where it is possible that the filter will be running even after calling Self::remove_filter(), so you cannot just free data that the filter might be using. Instead, you should pass a GDestroyNotify to Self::add_filter(), which will be called when it is guaranteed that the data is no longer needed.

filter_id

an identifier obtained from Self::add_filter()

pub fn signal_subscribe<P: Fn(&DBusConnection, &str, &str, &str, &str, &Variant) + 'static>(
&self,
sender: Option<&str>,
interface_name: Option<&str>,
member: Option<&str>,
object_path: Option<&str>,
arg0: Option<&str>,
flags: DBusSignalFlags,
callback: P
) -> SignalSubscriptionId

Subscribes to signals on self and invokes callback with a whenever the signal is received. Note that callback will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this method from.

If self is not a message bus connection, sender must be None.

If sender is a well-known name note that callback is invoked with the unique name for the owner of sender, not the well-known name as one would expect. This is because the message bus rewrites the name. As such, to avoid certain race conditions, users should be tracking the name owner of the well-known name and use that when processing the received signal.

If one of [crate::DBusSignalFlags::MatchArg0Namespace] or [crate::DBusSignalFlags::MatchArg0Path] are given, arg0 is interpreted as part of a namespace or path. The first argument of a signal is matched against that part as specified by D-Bus.

If user_data_free_func is non-None, it will be called (in the thread-default main context of the thread you are calling this method from) at some point after user_data is no longer needed. (It is not guaranteed to be called synchronously when the signal is unsubscribed from, and may be called after self has been destroyed.)

As callback is potentially invoked in a different thread from where it’s emitted, it’s possible for this to happen after Self::signal_unsubscribe() has been called in another thread. Due to this, user_data should have a strong reference which is freed with user_data_free_func, rather than pointing to data whose lifecycle is tied to the signal subscription. For example, if a crate::glib::object::Object is used to store the subscription ID from Self::signal_subscribe(), a strong reference to that crate::glib::object::Object must be passed to user_data, and [crate::glib::object::ObjectExt::unref()] passed to user_data_free_func. You are responsible for breaking the resulting reference count cycle by explicitly unsubscribing from the signal when dropping the last external reference to the crate::glib::object::Object. Alternatively, a weak reference may be used.

It is guaranteed that if you unsubscribe from a signal using Self::signal_unsubscribe() from the same thread which made the corresponding Self::signal_subscribe() call, callback will not be invoked after Self::signal_unsubscribe() returns.

The returned subscription identifier is an opaque value which is guaranteed to never be zero.

This function can never fail.

sender

sender name to match on (unique or well-known name) or None to listen from all senders

interface_name

D-Bus interface name to match on or None to match on all interfaces

member

D-Bus signal name to match on or None to match on all signals

object_path

object path to match on or None to match on all object paths

arg0

contents of first string argument to match on or None to match on all kinds of arguments

flags

crate::DBusSignalFlags describing how arg0 is used in subscribing to the signal

callback

callback to invoke when there is a signal matching the requested data

user_data

user data to pass to callback

user_data_free_func

function to free user_data with when subscription is removed or None

Returns

a subscription identifier that can be used with Self::signal_unsubscribe()

pub fn signal_unsubscribe(&self, subscription_id: SignalSubscriptionId)

Unsubscribes from signals.

Note that there may still be D-Bus traffic to process (relating to this signal subscription) in the current thread-default crate::glib::MainContext after this function has returned. You should continue to iterate the crate::glib::MainContext until the GDestroyNotify function passed to Self::signal_subscribe() is called, in order to avoid memory leaks through callbacks queued on the crate::glib::MainContext after it’s stopped being iterated.

subscription_id

a subscription id obtained from Self::signal_subscribe()

impl DBusConnection

pub fn for_address_sync<P: IsA<Cancellable>>(
address: &str,
flags: DBusConnectionFlags,
observer: Option<&DBusAuthObserver>,
cancellable: Option<&P>
) -> Result<DBusConnection, Error>

Synchronously connects and sets up a D-Bus client connection for exchanging D-Bus messages with an endpoint specified by address which must be in the D-Bus address format.

This constructor can only be used to initiate client-side connections - use Self::new_sync() if you need to act as the server. In particular, flags cannot contain the [crate::DBusConnectionFlags::AuthenticationServer], [crate::DBusConnectionFlags::AuthenticationAllowAnonymous] or G_DBUS_CONNECTION_FLAGS_AUTHENTICATION_REQUIRE_SAME_USER flags.

This is a synchronous failable constructor. See [Self::new_for_address()] for the asynchronous version.

If observer is not None it may be used to control the authentication process.

address

a D-Bus address

flags

flags describing how to make the connection

observer

a crate::DBusAuthObserver or None

cancellable

a crate::Cancellable or None

Returns

a crate::DBusConnection or None if error is set. Free with [crate::glib::object::ObjectExt::unref()].

pub fn new_sync<P: IsA<IOStream>, Q: IsA<Cancellable>>(
stream: &P,
guid: Option<&str>,
flags: DBusConnectionFlags,
observer: Option<&DBusAuthObserver>,
cancellable: Option<&Q>
) -> Result<DBusConnection, Error>

Synchronously sets up a D-Bus connection for exchanging D-Bus messages with the end represented by stream.

If stream is a crate::SocketConnection, then the corresponding crate::Socket will be put into non-blocking mode.

The D-Bus connection will interact with stream from a worker thread. As a result, the caller should not interact with stream after this method has been called, except by calling [crate::glib::object::ObjectExt::unref()] on it.

If observer is not None it may be used to control the authentication process.

This is a synchronous failable constructor. See Self::new() for the asynchronous version.

stream

a crate::IOStream

guid

the GUID to use if authenticating as a server or None

flags

flags describing how to make the connection

observer

a crate::DBusAuthObserver or None

cancellable

a crate::Cancellable or None

Returns

a crate::DBusConnection or None if error is set. Free with [crate::glib::object::ObjectExt::unref()].

pub fn call<P: IsA<Cancellable>, Q: FnOnce(Result<Variant, Error>) + Send + 'static>(
&self,
bus_name: Option<&str>,
object_path: &str,
interface_name: &str,
method_name: &str,
parameters: Option<&Variant>,
reply_type: Option<&VariantTy>,
flags: DBusCallFlags,
timeout_msec: i32,
cancellable: Option<&P>,
callback: Q
)

Asynchronously invokes the method_name method on the interface_name D-Bus interface on the remote object at object_path owned by bus_name.

If self is closed then the operation will fail with crate::IOErrorEnum::Closed. If cancellable is canceled, the operation will fail with crate::IOErrorEnum::Cancelled. If parameters contains a value not compatible with the D-Bus protocol, the operation fails with crate::IOErrorEnum::InvalidArgument.

If reply_type is non-None then the reply will be checked for having this type and an error will be raised if it does not match. Said another way, if you give a reply_type then any non-None return value will be of this type. Unless it’s G_VARIANT_TYPE_UNIT, the reply_type will be a tuple containing one or more values.

If the parameters crate::glib::Variant is floating, it is consumed. This allows convenient ‘inline’ use of [crate::glib::Variant::new()], e.g.:

 g_dbus_connection_call (connection,
                         "org.freedesktop.StringThings",
                         "/org/freedesktop/StringThings",
                         "org.freedesktop.StringThings",
                         "TwoStrings",
                         g_variant_new ("(ss)",
                                        "Thing One",
                                        "Thing Two"),
                         NULL,
                         G_DBUS_CALL_FLAGS_NONE,
                         -1,
                         NULL,
                         (GAsyncReadyCallback) two_strings_done,
                         NULL);

This is an asynchronous method. When the operation is finished, callback will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this method from. You can then call [Self::call_finish()] to get the result of the operation. See Self::call_sync() for the synchronous version of this function.

If callback is None then the D-Bus method call message will be sent with the [crate::DBusMessageFlags::NoReplyExpected] flag set.

bus_name

a unique or well-known bus name or None if self is not a message bus connection

object_path

path of remote object

interface_name

D-Bus interface to invoke method on

method_name

the name of the method to invoke

parameters

a crate::glib::Variant tuple with parameters for the method or None if not passing parameters

reply_type

the expected type of the reply (which will be a tuple), or None

flags

flags from the crate::DBusCallFlags enumeration

timeout_msec

the timeout in milliseconds, -1 to use the default timeout or G_MAXINT for no timeout

cancellable

a crate::Cancellable or None

callback

a GAsyncReadyCallback to call when the request is satisfied or None if you don’t care about the result of the method invocation

user_data

the data to pass to callback

pub fn call_future(
&self,
bus_name: Option<&str>,
object_path: &str,
interface_name: &str,
method_name: &str,
parameters: Option<&Variant>,
reply_type: Option<&VariantTy>,
flags: DBusCallFlags,
timeout_msec: i32
) -> Pin<Box_<dyn Future<Output = Result<Variant, Error>> + 'static>>

pub fn call_sync<P: IsA<Cancellable>>(
&self,
bus_name: Option<&str>,
object_path: &str,
interface_name: &str,
method_name: &str,
parameters: Option<&Variant>,
reply_type: Option<&VariantTy>,
flags: DBusCallFlags,
timeout_msec: i32,
cancellable: Option<&P>
) -> Result<Variant, Error>

Synchronously invokes the method_name method on the interface_name D-Bus interface on the remote object at object_path owned by bus_name.

If self is closed then the operation will fail with crate::IOErrorEnum::Closed. If cancellable is canceled, the operation will fail with crate::IOErrorEnum::Cancelled. If parameters contains a value not compatible with the D-Bus protocol, the operation fails with crate::IOErrorEnum::InvalidArgument.

If reply_type is non-None then the reply will be checked for having this type and an error will be raised if it does not match. Said another way, if you give a reply_type then any non-None return value will be of this type.

If the parameters crate::glib::Variant is floating, it is consumed. This allows convenient ‘inline’ use of [crate::glib::Variant::new()], e.g.:

 g_dbus_connection_call_sync (connection,
                              "org.freedesktop.StringThings",
                              "/org/freedesktop/StringThings",
                              "org.freedesktop.StringThings",
                              "TwoStrings",
                              g_variant_new ("(ss)",
                                             "Thing One",
                                             "Thing Two"),
                              NULL,
                              G_DBUS_CALL_FLAGS_NONE,
                              -1,
                              NULL,
                              &error);

The calling thread is blocked until a reply is received. See Self::call() for the asynchronous version of this method.

bus_name

a unique or well-known bus name or None if self is not a message bus connection

object_path

path of remote object

interface_name

D-Bus interface to invoke method on

method_name

the name of the method to invoke

parameters

a crate::glib::Variant tuple with parameters for the method or None if not passing parameters

reply_type

the expected type of the reply, or None

flags

flags from the crate::DBusCallFlags enumeration

timeout_msec

the timeout in milliseconds, -1 to use the default timeout or G_MAXINT for no timeout

cancellable

a crate::Cancellable or None

Returns

None if error is set. Otherwise a non-floating crate::glib::Variant tuple with return values. Free with [crate::glib::Variant::unref()].

pub fn call_with_unix_fd_list<P: IsA<UnixFDList>, Q: IsA<Cancellable>, R: FnOnce(Result<(Variant, UnixFDList), Error>) + Send + 'static>(
&self,
bus_name: Option<&str>,
object_path: &str,
interface_name: &str,
method_name: &str,
parameters: Option<&Variant>,
reply_type: Option<&VariantTy>,
flags: DBusCallFlags,
timeout_msec: i32,
fd_list: Option<&P>,
cancellable: Option<&Q>,
callback: R
)

This is supported on Unix only.

Like Self::call() but also takes a crate::UnixFDList object.

The file descriptors normally correspond to G_VARIANT_TYPE_HANDLE values in the body of the message. For example, if a message contains two file descriptors, fd_list would have length 2, and g_variant_new_handle (0) and g_variant_new_handle (1) would appear somewhere in the body of the message (not necessarily in that order!) to represent the file descriptors at indexes 0 and 1 respectively.

When designing D-Bus APIs that are intended to be interoperable, please note that non-GDBus implementations of D-Bus can usually only access file descriptors if they are referenced in this way by a value of type G_VARIANT_TYPE_HANDLE in the body of the message.

This method is only available on UNIX.

bus_name

a unique or well-known bus name or None if self is not a message bus connection

object_path

path of remote object

interface_name

D-Bus interface to invoke method on

method_name

the name of the method to invoke

parameters

a crate::glib::Variant tuple with parameters for the method or None if not passing parameters

reply_type

the expected type of the reply, or None

flags

flags from the crate::DBusCallFlags enumeration

timeout_msec

the timeout in milliseconds, -1 to use the default timeout or G_MAXINT for no timeout

fd_list

a crate::UnixFDList or None

cancellable

a crate::Cancellable or None

callback

a GAsyncReadyCallback to call when the request is satisfied or None if you don’t * care about the result of the method invocation

user_data

The data to pass to callback.

pub fn call_with_unix_fd_list_future<P: IsA<UnixFDList> + Clone + 'static>(
&self,
bus_name: Option<&str>,
object_path: &str,
interface_name: &str,
method_name: &str,
parameters: Option<&Variant>,
reply_type: Option<&VariantTy>,
flags: DBusCallFlags,
timeout_msec: i32,
fd_list: Option<&P>
) -> Pin<Box_<dyn Future<Output = Result<(Variant, UnixFDList), Error>> + 'static>>

This is supported on Unix only.

pub fn call_with_unix_fd_list_sync<P: IsA<UnixFDList>, Q: IsA<Cancellable>>(
&self,
bus_name: Option<&str>,
object_path: &str,
interface_name: &str,
method_name: &str,
parameters: Option<&Variant>,
reply_type: Option<&VariantTy>,
flags: DBusCallFlags,
timeout_msec: i32,
fd_list: Option<&P>,
cancellable: Option<&Q>
) -> Result<(Variant, UnixFDList), Error>

This is supported on Unix only.

Like Self::call_sync() but also takes and returns crate::UnixFDList objects. See Self::call_with_unix_fd_list() and [Self::call_with_unix_fd_list_finish()] for more details.

This method is only available on UNIX.

bus_name

a unique or well-known bus name or None if self is not a message bus connection

object_path

path of remote object

interface_name

D-Bus interface to invoke method on

method_name

the name of the method to invoke

parameters

a crate::glib::Variant tuple with parameters for the method or None if not passing parameters

reply_type

the expected type of the reply, or None

flags

flags from the crate::DBusCallFlags enumeration

timeout_msec

the timeout in milliseconds, -1 to use the default timeout or G_MAXINT for no timeout

fd_list

a crate::UnixFDList or None

out_fd_list

return location for a crate::UnixFDList or None

cancellable

a crate::Cancellable or None

Returns

None if error is set. Otherwise a non-floating crate::glib::Variant tuple with return values. Free with [crate::glib::Variant::unref()].

pub fn close<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
cancellable: Option<&P>,
callback: Q
)

Closes self. Note that this never causes the process to exit (this might only happen if the other end of a shared message bus connection disconnects, see [crate::DBusConnection:exit-on-close]).

Once the connection is closed, operations such as sending a message will return with the error crate::IOErrorEnum::Closed. Closing a connection will not automatically flush the connection so queued messages may be lost. Use Self::flush() if you need such guarantees.

If self is already closed, this method fails with crate::IOErrorEnum::Closed.

When self has been closed, the [crate::DBusConnection::closed] signal is emitted in the [thread-default main context][g-main-context-push-thread-default] of the thread that self was constructed in.

This is an asynchronous method. When the operation is finished, callback will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this method from. You can then call [Self::close_finish()] to get the result of the operation. See Self::close_sync() for the synchronous version.

cancellable

a crate::Cancellable or None

callback

a GAsyncReadyCallback to call when the request is satisfied or None if you don’t care about the result

user_data

The data to pass to callback

pub fn close_future(
&self
) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

pub fn close_sync<P: IsA<Cancellable>>(
&self,
cancellable: Option<&P>
) -> Result<(), Error>

Synchronously closes self. The calling thread is blocked until this is done. See Self::close() for the asynchronous version of this method and more details about what it does.

cancellable

a crate::Cancellable or None

Returns

true if the operation succeeded, false if error is set

pub fn emit_signal(
&self,
destination_bus_name: Option<&str>,
object_path: &str,
interface_name: &str,
signal_name: &str,
parameters: Option<&Variant>
) -> Result<(), Error>

Emits a signal.

If the parameters GVariant is floating, it is consumed.

This can only fail if parameters is not compatible with the D-Bus protocol (crate::IOErrorEnum::InvalidArgument), or if self has been closed (crate::IOErrorEnum::Closed).

destination_bus_name

the unique bus name for the destination for the signal or None to emit to all listeners

object_path

path of remote object

interface_name

D-Bus interface to emit a signal on

signal_name

the name of the signal to emit

parameters

a crate::glib::Variant tuple with parameters for the signal or None if not passing parameters

Returns

true unless error is set

pub fn flush<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
cancellable: Option<&P>,
callback: Q
)

Asynchronously flushes self, that is, writes all queued outgoing message to the transport and then flushes the transport (using crate::prelude::OutputStreamExt::flush_async()). This is useful in programs that wants to emit a D-Bus signal and then exit immediately. Without flushing the connection, there is no guaranteed that the message has been sent to the networking buffers in the OS kernel.

This is an asynchronous method. When the operation is finished, callback will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this method from. You can then call [Self::flush_finish()] to get the result of the operation. See Self::flush_sync() for the synchronous version.

cancellable

a crate::Cancellable or None

callback

a GAsyncReadyCallback to call when the request is satisfied or None if you don’t care about the result

user_data

The data to pass to callback

pub fn flush_future(
&self
) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

pub fn flush_sync<P: IsA<Cancellable>>(
&self,
cancellable: Option<&P>
) -> Result<(), Error>

Synchronously flushes self. The calling thread is blocked until this is done. See Self::flush() for the asynchronous version of this method and more details about what it does.

cancellable

a crate::Cancellable or None

Returns

true if the operation succeeded, false if error is set

pub fn capabilities(&self) -> DBusCapabilityFlags

Gets the capabilities negotiated with the remote peer

Returns

zero or more flags from the crate::DBusCapabilityFlags enumeration

pub fn exits_on_close(&self) -> bool

Gets whether the process is terminated when self is closed by the remote peer. See [crate::DBusConnection:exit-on-close] for more details.

Returns

whether the process is terminated when self is closed by the remote peer

pub fn flags(&self) -> DBusConnectionFlags

This is supported on crate feature v2_60 only.

Gets the flags used to construct this connection

Returns

zero or more flags from the crate::DBusConnectionFlags enumeration

pub fn guid(&self) -> GString

The GUID of the peer performing the role of server when authenticating. See [crate::DBusConnection:guid] for more details.

Returns

The GUID. Do not free this string, it is owned by self.

pub fn last_serial(&self) -> u32

Retrieves the last serial number assigned to a crate::DBusMessage on the current thread. This includes messages sent via both low-level API such as Self::send_message() as well as high-level API such as Self::emit_signal(), Self::call() or crate::prelude::DBusProxyExt::call().

Returns

the last used serial or zero when no message has been sent within the current thread

pub fn peer_credentials(&self) -> Option<Credentials>

Gets the credentials of the authenticated peer. This will always return None unless self acted as a server (e.g. [crate::DBusConnectionFlags::AuthenticationServer] was passed) when set up and the client passed credentials as part of the authentication process.

In a message bus setup, the message bus is always the server and each application is a client. So this method will always return None for message bus clients.

Returns

a crate::Credentials or None if not available. Do not free this object, it is owned by self.

pub fn stream(&self) -> IOStream

Gets the underlying stream used for IO.

While the crate::DBusConnection is active, it will interact with this stream from a worker thread, so it is not safe to interact with the stream directly.

Returns

the stream used for IO

pub fn unique_name(&self) -> Option<GString>

Gets the unique name of self as assigned by the message bus. This can also be used to figure out if self is a message bus connection.

Returns

the unique name or None if self is not a message bus connection. Do not free this string, it is owned by self.

pub fn is_closed(&self) -> bool

Gets whether self is closed.

Returns

true if the connection is closed, false otherwise

pub fn send_message(
&self,
message: &DBusMessage,
flags: DBusSendMessageFlags
) -> Result<u32, Error>

Asynchronously sends message to the peer represented by self.

Unless flags contain the [crate::DBusSendMessageFlags::PreserveSerial] flag, the serial number will be assigned by self and set on message via crate::DBusMessage::set_serial(). If out_serial is not None, then the serial number used will be written to this location prior to submitting the message to the underlying transport. While it has a volatile qualifier, this is a historical artifact and the argument passed to it should not be volatile.

If self is closed then the operation will fail with crate::IOErrorEnum::Closed. If message is not well-formed, the operation fails with crate::IOErrorEnum::InvalidArgument.

See this [server][gdbus-server] and [client][gdbus-unix-fd-client] for an example of how to use this low-level API to send and receive UNIX file descriptors.

Note that message must be unlocked, unless flags contain the [crate::DBusSendMessageFlags::PreserveSerial] flag.

message

a crate::DBusMessage

flags

flags affecting how the message is sent

out_serial

return location for serial number assigned to message when sending it or None

Returns

true if the message was well-formed and queued for transmission, false if error is set

pub fn send_message_with_reply<P: IsA<Cancellable>, Q: FnOnce(Result<DBusMessage, Error>) + Send + 'static>(
&self,
message: &DBusMessage,
flags: DBusSendMessageFlags,
timeout_msec: i32,
cancellable: Option<&P>,
callback: Q
) -> u32

Asynchronously sends message to the peer represented by self.

Unless flags contain the [crate::DBusSendMessageFlags::PreserveSerial] flag, the serial number will be assigned by self and set on message via crate::DBusMessage::set_serial(). If out_serial is not None, then the serial number used will be written to this location prior to submitting the message to the underlying transport. While it has a volatile qualifier, this is a historical artifact and the argument passed to it should not be volatile.

If self is closed then the operation will fail with crate::IOErrorEnum::Closed. If cancellable is canceled, the operation will fail with crate::IOErrorEnum::Cancelled. If message is not well-formed, the operation fails with crate::IOErrorEnum::InvalidArgument.

This is an asynchronous method. When the operation is finished, callback will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this method from. You can then call [Self::send_message_with_reply_finish()] to get the result of the operation. See Self::send_message_with_reply_sync() for the synchronous version.

Note that message must be unlocked, unless flags contain the [crate::DBusSendMessageFlags::PreserveSerial] flag.

See this [server][gdbus-server] and [client][gdbus-unix-fd-client] for an example of how to use this low-level API to send and receive UNIX file descriptors.

message

a crate::DBusMessage

flags

flags affecting how the message is sent

timeout_msec

the timeout in milliseconds, -1 to use the default timeout or G_MAXINT for no timeout

out_serial

return location for serial number assigned to message when sending it or None

cancellable

a crate::Cancellable or None

callback

a GAsyncReadyCallback to call when the request is satisfied or None if you don’t care about the result

user_data

The data to pass to callback

pub fn send_message_with_reply_future(
&self,
message: &DBusMessage,
flags: DBusSendMessageFlags,
timeout_msec: i32
) -> Pin<Box_<dyn Future<Output = Result<DBusMessage, Error>> + 'static>>

pub fn send_message_with_reply_sync<P: IsA<Cancellable>>(
&self,
message: &DBusMessage,
flags: DBusSendMessageFlags,
timeout_msec: i32,
cancellable: Option<&P>
) -> Result<(DBusMessage, u32), Error>

Synchronously sends message to the peer represented by self and blocks the calling thread until a reply is received or the timeout is reached. See Self::send_message_with_reply() for the asynchronous version of this method.

Unless flags contain the [crate::DBusSendMessageFlags::PreserveSerial] flag, the serial number will be assigned by self and set on message via crate::DBusMessage::set_serial(). If out_serial is not None, then the serial number used will be written to this location prior to submitting the message to the underlying transport. While it has a volatile qualifier, this is a historical artifact and the argument passed to it should not be volatile.

If self is closed then the operation will fail with crate::IOErrorEnum::Closed. If cancellable is canceled, the operation will fail with crate::IOErrorEnum::Cancelled. If message is not well-formed, the operation fails with crate::IOErrorEnum::InvalidArgument.

Note that error is only set if a local in-process error occurred. That is to say that the returned crate::DBusMessage object may be of type crate::DBusMessageType::Error. Use crate::DBusMessage::to_gerror() to transcode this to a crate::glib::Error.

See this [server][gdbus-server] and [client][gdbus-unix-fd-client] for an example of how to use this low-level API to send and receive UNIX file descriptors.

Note that message must be unlocked, unless flags contain the [crate::DBusSendMessageFlags::PreserveSerial] flag.

message

a crate::DBusMessage

flags

flags affecting how the message is sent.

timeout_msec

the timeout in milliseconds, -1 to use the default timeout or G_MAXINT for no timeout

out_serial

return location for serial number assigned to message when sending it or None

cancellable

a crate::Cancellable or None

Returns

a locked crate::DBusMessage that is the reply to message or None if error is set

pub fn set_exit_on_close(&self, exit_on_close: bool)

Sets whether the process should be terminated when self is closed by the remote peer. See [crate::DBusConnection:exit-on-close] for more details.

Note that this function should be used with care. Most modern UNIX desktops tie the notion of a user session with the session bus, and expect all of a user’s applications to quit when their bus connection goes away. If you are setting exit_on_close to false for the shared session bus connection, you should make sure that your application exits when the user session ends.

exit_on_close

whether the process should be terminated when self is closed by the remote peer

pub fn start_message_processing(&self)

If self was created with [crate::DBusConnectionFlags::DelayMessageProcessing], this method starts processing messages. Does nothing on if self wasn’t created with this flag or if the method has already been called.

pub fn get_property_flags(&self) -> DBusConnectionFlags

Flags from the crate::DBusConnectionFlags enumeration.

pub fn new<P: IsA<IOStream>, Q: IsA<Cancellable>, R: FnOnce(Result<DBusConnection, Error>) + Send + 'static>(
stream: &P,
guid: Option<&str>,
flags: DBusConnectionFlags,
observer: Option<&DBusAuthObserver>,
cancellable: Option<&Q>,
callback: R
)

Asynchronously sets up a D-Bus connection for exchanging D-Bus messages with the end represented by stream.

If stream is a crate::SocketConnection, then the corresponding crate::Socket will be put into non-blocking mode.

The D-Bus connection will interact with stream from a worker thread. As a result, the caller should not interact with stream after this method has been called, except by calling [crate::glib::object::ObjectExt::unref()] on it.

If observer is not None it may be used to control the authentication process.

When the operation is finished, callback will be invoked. You can then call [Self::new_finish()] to get the result of the operation.

This is an asynchronous failable constructor. See Self::new_sync() for the synchronous version.

stream

a crate::IOStream

guid

the GUID to use if authenticating as a server or None

flags

flags describing how to make the connection

observer

a crate::DBusAuthObserver or None

cancellable

a crate::Cancellable or None

callback

a GAsyncReadyCallback to call when the request is satisfied

user_data

the data to pass to callback

pub fn new_future<P: IsA<IOStream> + Clone + 'static>(
stream: &P,
guid: Option<&str>,
flags: DBusConnectionFlags,
observer: Option<&DBusAuthObserver>
) -> Pin<Box_<dyn Future<Output = Result<DBusConnection, Error>> + 'static>>

pub fn for_address<P: IsA<Cancellable>, Q: FnOnce(Result<DBusConnection, Error>) + Send + 'static>(
address: &str,
flags: DBusConnectionFlags,
observer: Option<&DBusAuthObserver>,
cancellable: Option<&P>,
callback: Q
)

Asynchronously connects and sets up a D-Bus client connection for exchanging D-Bus messages with an endpoint specified by address which must be in the D-Bus address format.

This constructor can only be used to initiate client-side connections - use Self::new() if you need to act as the server. In particular, flags cannot contain the [crate::DBusConnectionFlags::AuthenticationServer], [crate::DBusConnectionFlags::AuthenticationAllowAnonymous] or G_DBUS_CONNECTION_FLAGS_AUTHENTICATION_REQUIRE_SAME_USER flags.

When the operation is finished, callback will be invoked. You can then call [Self::new_for_address_finish()] to get the result of the operation.

If observer is not None it may be used to control the authentication process.

This is an asynchronous failable constructor. See [Self::new_for_address_sync()] for the synchronous version.

address

a D-Bus address

flags

flags describing how to make the connection

observer

a crate::DBusAuthObserver or None

cancellable

a crate::Cancellable or None

callback

a GAsyncReadyCallback to call when the request is satisfied

user_data

the data to pass to callback

pub fn for_address_future(
address: &str,
flags: DBusConnectionFlags,
observer: Option<&DBusAuthObserver>
) -> Pin<Box_<dyn Future<Output = Result<DBusConnection, Error>> + 'static>>

pub fn connect_closed<F: Fn(&DBusConnection, bool, Option<&Error>) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when the connection is closed.

The cause of this event can be

If Self::close() is called. In this case remote_peer_vanished is set to false and error is None.

If the remote peer closes the connection. In this case remote_peer_vanished is set to true and error is set.

If the remote peer sends invalid or malformed data. In this case remote_peer_vanished is set to false and error is set.

Upon receiving this signal, you should give up your reference to connection. You are guaranteed that this signal is emitted only once.

remote_peer_vanished

true if connection is closed because the remote peer closed its end of the connection

error

a crate::glib::Error with more details about the event or None

pub fn connect_capabilities_notify<F: Fn(&DBusConnection) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_closed_notify<F: Fn(&DBusConnection) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_exit_on_close_notify<F: Fn(&DBusConnection) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_unique_name_notify<F: Fn(&DBusConnection) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId