gio

Struct DataOutputStream

pub struct DataOutputStream { /* private fields */ }

Data output stream implements OutputStream and includes functions for writing data directly to an output stream.

Properties

byte-order

Determines the byte ordering that is used when writing multi-byte entities (such as integers) to the stream.

Readable | Writeable

FilterOutputStream

base-stream

The underlying base stream on which the I/O ops will be done.

Readable | Writeable | Construct Only

close-base-stream

Whether the base stream should be closed when the filter stream is closed.

Readable | Writeable | Construct Only

Implements

DataOutputStreamExt, FilterOutputStreamExt, OutputStreamExt, [trait@glib::ObjectExt], SeekableExt, OutputStreamExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl DataOutputStream

pub const NONE: Option<&'static DataOutputStream> = None

pub fn new(base_stream: &impl IsA<OutputStream>) -> DataOutputStream

Creates a new data output stream for @base_stream.

base_stream

a #GOutputStream.

Returns

#GDataOutputStream.

pub fn builder() -> DataOutputStreamBuilder

Creates a new builder-pattern struct instance to construct DataOutputStream objects.

This method returns an instance of DataOutputStreamBuilder which can be used to create DataOutputStream objects.