Struct gio::FileInfo

pub struct FileInfo(_, _);

Functionality for manipulating basic metadata for files. FileInfo implements methods for getting information that all files should contain, and allows for manipulation of extended attributes.

See [GFileAttribute][gio-GFileAttribute] for more information on how GIO handles file attributes.

To obtain a FileInfo for a File, use File::query_info (or its async variant). To obtain a FileInfo for a file input or output stream, use FileInputStreamExt::query_info or FileOutputStreamExt::query_info (or their async variants).

To change the actual attributes of a file, you should then set the attribute in the FileInfo and call File::set_attributes_from_info or File::set_attributes_async on a GFile.

However, not all attributes can be changed in the file. For instance, the actual size of a file cannot be changed via FileInfo::set_size. You may call File::query_settable_attributes and File::query_writable_namespaces to discover the settable attributes of a particular file at runtime.

FileAttributeMatcher allows for searching through a FileInfo for attributes.

Implements

glib::object::ObjectExt

Methods

impl FileInfo

pub fn new() -> FileInfo

Creates a new file info structure.

Returns

a FileInfo.

pub fn clear_status(&self)

Clears the status information from self.

pub fn copy_into(&self, dest_info: &FileInfo)

First clears all of the [GFileAttribute][gio-GFileAttribute] of dest_info, and then copies all of the file attributes from self to dest_info.

dest_info

destination to copy attributes to.

pub fn dup(&self) -> Option<FileInfo>

Duplicates a file info structure.

Returns

a duplicate FileInfo of self.

pub fn get_attribute_as_string(&self, attribute: &str) -> Option<GString>

Gets the value of a attribute, formated as a string. This escapes things as needed to make the string valid UTF-8.

attribute

a file attribute key.

Returns

a UTF-8 string associated with the given attribute, or None if the attribute wasn’t set. When you're done with the string it must be freed with g_free.

pub fn get_attribute_boolean(&self, attribute: &str) -> bool

Gets the value of a boolean attribute. If the attribute does not contain a boolean value, false will be returned.

attribute

a file attribute key.

Returns

the boolean value contained within the attribute.

pub fn get_attribute_byte_string(&self, attribute: &str) -> Option<GString>

Gets the value of a byte string attribute. If the attribute does not contain a byte string, None will be returned.

attribute

a file attribute key.

Returns

the contents of the attribute value as a byte string, or None otherwise.

pub fn get_attribute_int32(&self, attribute: &str) -> i32

Gets a signed 32-bit integer contained within the attribute. If the attribute does not contain a signed 32-bit integer, or is invalid, 0 will be returned.

attribute

a file attribute key.

Returns

a signed 32-bit integer from the attribute.

pub fn get_attribute_int64(&self, attribute: &str) -> i64

Gets a signed 64-bit integer contained within the attribute. If the attribute does not contain an signed 64-bit integer, or is invalid, 0 will be returned.

attribute

a file attribute key.

Returns

a signed 64-bit integer from the attribute.

pub fn get_attribute_object(&self, attribute: &str) -> Option<Object>

Gets the value of a gobject::Object attribute. If the attribute does not contain a gobject::Object, None will be returned.

attribute

a file attribute key.

Returns

a gobject::Object associated with the given attribute, or None otherwise.

pub fn get_attribute_status(&self, attribute: &str) -> FileAttributeStatus

Gets the attribute status for an attribute key.

attribute

a file attribute key

Returns

a FileAttributeStatus for the given attribute, or FileAttributeStatus::Unset if the key is invalid.

pub fn get_attribute_string(&self, attribute: &str) -> Option<GString>

Gets the value of a string attribute. If the attribute does not contain a string, None will be returned.

attribute

a file attribute key.

Returns

the contents of the attribute value as a UTF-8 string, or None otherwise.

pub fn get_attribute_stringv(&self, attribute: &str) -> Vec<GString>

Gets the value of a stringv attribute. If the attribute does not contain a stringv, None will be returned.

attribute

a file attribute key.

Returns

the contents of the attribute value as a stringv, or None otherwise. Do not free. These returned strings are UTF-8.

pub fn get_attribute_type(&self, attribute: &str) -> FileAttributeType

Gets the attribute type for an attribute key.

attribute

a file attribute key.

Returns

a FileAttributeType for the given attribute, or FileAttributeType::Invalid if the key is not set.

pub fn get_attribute_uint32(&self, attribute: &str) -> u32

Gets an unsigned 32-bit integer contained within the attribute. If the attribute does not contain an unsigned 32-bit integer, or is invalid, 0 will be returned.

attribute

a file attribute key.

Returns

an unsigned 32-bit integer from the attribute.

pub fn get_attribute_uint64(&self, attribute: &str) -> u64

Gets a unsigned 64-bit integer contained within the attribute. If the attribute does not contain an unsigned 64-bit integer, or is invalid, 0 will be returned.

attribute

a file attribute key.

Returns

a unsigned 64-bit integer from the attribute.

pub fn get_content_type(&self) -> Option<GString>

Gets the file's content type.

Returns

a string containing the file's content type.

pub fn get_deletion_date(&self) -> Option<DateTime>

Returns the glib::DateTime representing the deletion date of the file, as available in G_FILE_ATTRIBUTE_TRASH_DELETION_DATE. If the G_FILE_ATTRIBUTE_TRASH_DELETION_DATE attribute is unset, None is returned.

Returns

a glib::DateTime, or None.

pub fn get_display_name(&self) -> Option<GString>

Gets a display name for a file.

Returns

a string containing the display name.

pub fn get_edit_name(&self) -> Option<GString>

Gets the edit name for a file.

Returns

a string containing the edit name.

pub fn get_etag(&self) -> Option<GString>

Gets the [entity tag][gfile-etag] for a given FileInfo. See G_FILE_ATTRIBUTE_ETAG_VALUE.

Returns

a string containing the value of the "etag:value" attribute.

pub fn get_file_type(&self) -> FileType

Gets a file's type (whether it is a regular file, symlink, etc). This is different from the file's content type, see FileInfo::get_content_type.

Returns

a FileType for the given file.

pub fn get_icon(&self) -> Option<Icon>

Gets the icon for a file.

Returns

Icon for the given self.

pub fn get_is_backup(&self) -> bool

Checks if a file is a backup file.

Returns

true if file is a backup file, false otherwise.

pub fn get_is_hidden(&self) -> bool

Checks if a file is hidden.

Returns

true if the file is a hidden file, false otherwise.

pub fn get_is_symlink(&self) -> bool

Checks if a file is a symlink.

Returns

true if the given self is a symlink.

pub fn get_modification_time(&self) -> TimeVal

Gets the modification time of the current self and sets it in result.

Deprecated since 2.62

Use FileInfo::get_modification_date_time instead, as glib::TimeVal is deprecated due to the year 2038 problem.

result

a glib::TimeVal.

pub fn get_name(&self) -> Option<PathBuf>

Gets the name for a file.

Returns

a string containing the file name.

pub fn get_size(&self) -> i64

Gets the file's size.

Returns

a goffset containing the file's size.

pub fn get_sort_order(&self) -> i32

Gets the value of the sort_order attribute from the FileInfo. See G_FILE_ATTRIBUTE_STANDARD_SORT_ORDER.

Returns

a gint32 containing the value of the "standard::sort_order" attribute.

pub fn get_symbolic_icon(&self) -> Option<Icon>

Gets the symbolic icon for a file.

Returns

Icon for the given self.

pub fn get_symlink_target(&self) -> Option<GString>

Gets the symlink target for a given FileInfo.

Returns

a string containing the symlink target.

pub fn has_attribute(&self, attribute: &str) -> bool

Checks if a file info structure has an attribute named attribute.

attribute

a file attribute key.

Returns

true if Ginfo has an attribute named attribute, false otherwise.

pub fn has_namespace(&self, name_space: &str) -> bool

Checks if a file info structure has an attribute in the specified name_space.

name_space

a file attribute namespace.

Returns

true if Ginfo has an attribute in name_space, false otherwise.

pub fn list_attributes(&self, name_space: Option<&str>) -> Vec<GString>

Lists the file info structure's attributes.

name_space

a file attribute key's namespace, or None to list all attributes.

Returns

a null-terminated array of strings of all of the possible attribute types for the given name_space, or None on error.

pub fn remove_attribute(&self, attribute: &str)

Removes all cases of attribute from self if it exists.

attribute

a file attribute key.

pub fn set_attribute_boolean(&self, attribute: &str, attr_value: bool)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a boolean value.

pub fn set_attribute_byte_string(&self, attribute: &str, attr_value: &str)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a byte string.

pub fn set_attribute_int32(&self, attribute: &str, attr_value: i32)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a signed 32-bit integer

pub fn set_attribute_int64(&self, attribute: &str, attr_value: i64)

Sets the attribute to contain the given attr_value, if possible.

attribute

attribute name to set.

attr_value

int64 value to set attribute to.

pub fn set_attribute_mask(&self, mask: &FileAttributeMatcher)

Sets mask on self to match specific attribute types.

mask

a FileAttributeMatcher.

pub fn set_attribute_object<P: IsA<Object>>(
&self,
attribute: &str,
attr_value: &P
)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a gobject::Object.

pub fn set_attribute_status(
&self,
attribute: &str,
status: FileAttributeStatus
) -> bool

Sets the attribute status for an attribute key. This is only needed by external code that implement File::set_attributes_from_info or similar functions.

The attribute must exist in self for this to work. Otherwise false is returned and self is unchanged.

attribute

a file attribute key

status

a FileAttributeStatus

Returns

true if the status was changed, false if the key was not set.

pub fn set_attribute_string(&self, attribute: &str, attr_value: &str)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a UTF-8 string.

pub fn set_attribute_stringv(&self, attribute: &str, attr_value: &[&str])

Sets the attribute to contain the given attr_value, if possible.

Sinze: 2.22

attribute

a file attribute key

attr_value

a None terminated array of UTF-8 strings.

pub fn set_attribute_uint32(&self, attribute: &str, attr_value: u32)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

an unsigned 32-bit integer.

pub fn set_attribute_uint64(&self, attribute: &str, attr_value: u64)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

an unsigned 64-bit integer.

pub fn set_content_type(&self, content_type: &str)

Sets the content type attribute for a given FileInfo. See G_FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE.

content_type

a content type. See [GContentType][gio-GContentType]

pub fn set_display_name(&self, display_name: &str)

Sets the display name for the current FileInfo. See G_FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME.

display_name

a string containing a display name.

pub fn set_edit_name(&self, edit_name: &str)

Sets the edit name for the current file. See G_FILE_ATTRIBUTE_STANDARD_EDIT_NAME.

edit_name

a string containing an edit name.

pub fn set_file_type(&self, type_: FileType)

Sets the file type in a FileInfo to type_. See G_FILE_ATTRIBUTE_STANDARD_TYPE.

type_

a FileType.

pub fn set_icon<P: IsA<Icon>>(&self, icon: &P)

Sets the icon for a given FileInfo. See G_FILE_ATTRIBUTE_STANDARD_ICON.

icon

a Icon.

pub fn set_is_hidden(&self, is_hidden: bool)

Sets the "is_hidden" attribute in a FileInfo according to is_hidden. See G_FILE_ATTRIBUTE_STANDARD_IS_HIDDEN.

is_hidden

a gboolean.

pub fn set_is_symlink(&self, is_symlink: bool)

Sets the "is_symlink" attribute in a FileInfo according to is_symlink. See G_FILE_ATTRIBUTE_STANDARD_IS_SYMLINK.

is_symlink

a gboolean.

pub fn set_modification_time(&self, mtime: &mut TimeVal)

Sets the G_FILE_ATTRIBUTE_TIME_MODIFIED attribute in the file info to the given time value.

Deprecated since 2.62

Use FileInfo::set_modification_date_time instead, as glib::TimeVal is deprecated due to the year 2038 problem.

mtime

a glib::TimeVal.

pub fn set_name<P: AsRef<Path>>(&self, name: P)

Sets the name attribute for the current FileInfo. See G_FILE_ATTRIBUTE_STANDARD_NAME.

name

a string containing a name.

pub fn set_size(&self, size: i64)

Sets the G_FILE_ATTRIBUTE_STANDARD_SIZE attribute in the file info to the given size.

size

a goffset containing the file's size.

pub fn set_sort_order(&self, sort_order: i32)

Sets the sort order attribute in the file info structure. See G_FILE_ATTRIBUTE_STANDARD_SORT_ORDER.

sort_order

a sort order integer.

pub fn set_symbolic_icon<P: IsA<Icon>>(&self, icon: &P)

Sets the symbolic icon for a given FileInfo. See G_FILE_ATTRIBUTE_STANDARD_SYMBOLIC_ICON.

icon

a Icon.

pub fn set_symlink_target(&self, symlink_target: &str)

Sets the G_FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET attribute in the file info to the given symlink target.

symlink_target

a static string containing a path to a symlink target.

pub fn unset_attribute_mask(&self)

Unsets a mask set by FileInfo::set_attribute_mask, if one is set.