Struct gio::IOExtensionPoint

pub struct IOExtensionPoint(_);

An extension point provides a mechanism to extend the functionality of a library or application. Each extension point is identified by a name, and it may optionally require that any implementation must be of a certain type. IOExtensionPoint is an opaque data structure and can only be accessed using the following functions.

Implementations

impl IOExtensionPoint

pub fn builder(name: &str) -> IOExtensionPointBuilder<'_>

Create a new builder for an extension point.

pub fn lookup(name: &str) -> Option<Self>

Looks up an existing extension point.

name

the name of the extension point

Returns

the IOExtensionPoint, or None if there is no registered extension point with the given name.

pub fn extensions(&self) -> Vec<IOExtension>

Gets a list of all extensions that implement this extension point. The list is sorted by priority, beginning with the highest priority.

Returns

a GList of GIOExtensions. The list is owned by GIO and should not be modified.

pub fn extension_by_name(&self, name: &str) -> Option<IOExtension>

Finds a IOExtension for an extension point by name.

name

the name of the extension to get

Returns

the IOExtension for self that has the given name, or None if there is no extension with that name

pub fn required_type(&self) -> Type

Gets the required type for self.

Returns

the GType that all implementations must have, or G_TYPE_INVALID if the extension point has no required type

pub fn implement(
extension_point_name: &str,
type_: Type,
extension_name: &str,
priority: i32
) -> Option<IOExtension>

Registers type_ as extension for the extension point with name extension_point_name.

If type_ has already been registered as an extension for this extension point, the existing IOExtension object is returned.

extension_point_name

the name of the extension point

type_

the GType to register as extension

extension_name

the name for the extension

priority

the priority for the extension

Returns

a IOExtension object for GType