Struct gio::Icon

pub struct Icon(_, _);

Icon is a very minimal interface for icons. It provides functions for checking the equality of two icons, hashing of icons and serializing an icon to and from strings.

Icon does not provide the actual pixmap for the icon as this is out of GIO's scope, however implementations of Icon may contain the name of an icon (see ThemedIcon), or the path to an icon (see LoadableIcon).

To obtain a hash of a Icon, see Icon::hash.

To check if two GIcons are equal, see Icon::equal.

For serializing a Icon, use Icon::serialize and Icon::deserialize.

If you want to consume Icon (for example, in a toolkit) you must be prepared to handle at least the three following cases: LoadableIcon, ThemedIcon and EmblemedIcon. It may also make sense to have fast-paths for other cases (like handling GdkPixbuf directly, for example) but all compliant Icon implementations outside of GIO must implement LoadableIcon.

If your application or library provides one or more Icon implementations you need to ensure that your new implementation also implements LoadableIcon. Additionally, you must provide an implementation of Icon::serialize that gives a result that is understood by Icon::deserialize, yielding one of the built-in icon types.

Implements

IconExt

Methods

impl Icon

pub fn deserialize(value: &Variant) -> Option<Icon>

Deserializes a Icon previously serialized using Icon::serialize.

Feature: v2_38

value

a glib::Variant created with Icon::serialize

Returns

a Icon, or None when deserialization fails.

pub fn new_for_string(str: &str) -> Result<Icon, Error>

Generate a Icon instance from str. This function can fail if str is not valid - see Icon::to_string for discussion.

If your application or library provides one or more Icon implementations you need to ensure that each glib::Type is registered with the type system prior to calling Icon::new_for_string.

str

A string obtained via Icon::to_string.

Returns

An object implementing the Icon interface or None if error is set.