Struct gio::InetAddress

pub struct InetAddress { /* private fields */ }

GInetAddress represents an IPv4 or IPv6 internet address. Use ResolverExt::lookup_by_name() or ResolverExt::lookup_by_name_async() to look up the GInetAddress for a hostname. Use ResolverExt::lookup_by_address() or ResolverExt::lookup_by_address_async() to look up the hostname for a GInetAddress.

To actually connect to a remote host, you will need a InetSocketAddress (which includes a GInetAddress as well as a port number).

Properties

bytes

The raw address data.

Readable | Writeable | Construct Only

family

The address family (IPv4 or IPv6).

Readable | Writeable | Construct Only

is-any

Whether this is the “any” address for its family. See g_inet_address_get_is_any().

Readable

is-link-local

Whether this is a link-local address. See g_inet_address_get_is_link_local().

Readable

is-loopback

Whether this is the loopback address for its family. See g_inet_address_get_is_loopback().

Readable

is-mc-global

Whether this is a global multicast address. See g_inet_address_get_is_mc_global().

Readable

is-mc-link-local

Whether this is a link-local multicast address. See g_inet_address_get_is_mc_link_local().

Readable

is-mc-node-local

Whether this is a node-local multicast address. See g_inet_address_get_is_mc_node_local().

Readable

is-mc-org-local

Whether this is an organization-local multicast address. See g_inet_address_get_is_mc_org_local().

Readable

is-mc-site-local

Whether this is a site-local multicast address. See g_inet_address_get_is_mc_site_local().

Readable

is-multicast

Whether this is a multicast address. See g_inet_address_get_is_multicast().

Readable

is-site-local

Whether this is a site-local address. See g_inet_address_get_is_loopback().

Readable

Implements

InetAddressExt, [trait@glib::ObjectExt], InetAddressExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl InetAddress

pub fn from_bytes(inet_address_bytes: InetAddressBytes<'_>) -> Self

Creates a new #GInetAddress from the given @family and @bytes. @bytes should be 4 bytes for SocketFamily::Ipv4 and 16 bytes for SocketFamily::Ipv6.

bytes

raw address data

family

the address family of @bytes

Returns

a new #GInetAddress corresponding to @family and @bytes. Free the returned object with g_object_unref().

impl InetAddress

pub const NONE: Option<&'static InetAddress> = None

pub fn new_any(family: SocketFamily) -> InetAddress

Creates a #GInetAddress for the “any” address (unassigned/“don’t care”) for @family.

family

the address family

Returns

a new #GInetAddress corresponding to the “any” address for @family. Free the returned object with g_object_unref().

pub fn from_string(string: &str) -> Option<InetAddress>

Parses @string as an IP address and creates a new #GInetAddress.

string

a string representation of an IP address

Returns

a new #GInetAddress corresponding to @string, or None if @string could not be parsed. Free the returned object with g_object_unref().

pub fn new_loopback(family: SocketFamily) -> InetAddress

Creates a #GInetAddress for the loopback address for @family.

family

the address family

Returns

a new #GInetAddress corresponding to the loopback address for @family. Free the returned object with g_object_unref().