Struct gio::MenuItem

pub struct MenuItem(_, _);

MenuItem is an opaque structure type. You must access it using the functions below.

Implements

MenuItemExt, glib::object::ObjectExt

Methods

impl MenuItem

pub fn new<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
label: P,
detailed_action: Q
) -> MenuItem

Creates a new MenuItem.

If label is non-None it is used to set the "label" attribute of the new item.

If detailed_action is non-None it is used to set the "action" and possibly the "target" attribute of the new item. See MenuItemExt::set_detailed_action for more information.

label

the section label, or None

detailed_action

the detailed action string, or None

Returns

a new MenuItem

pub fn new_from_model<P: IsA<MenuModel>>(model: &P, item_index: i32) -> MenuItem

Creates a MenuItem as an exact copy of an existing menu item in a MenuModel.

item_index must be valid (ie: be sure to call MenuModelExt::get_n_items first).

Feature: v2_34

model

a MenuModel

item_index

the index of an item in model

Returns

a new MenuItem.

pub fn new_section<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
label: P,
section: &Q
) -> MenuItem

Creates a new MenuItem representing a section.

This is a convenience API around MenuItem::new and MenuItemExt::set_section.

The effect of having one menu appear as a section of another is exactly as it sounds: the items from section become a direct part of the menu that menu_item is added to.

Visual separation is typically displayed between two non-empty sections. If label is non-None then it will be encorporated into this visual indication. This allows for labeled subsections of a menu.

As a simple example, consider a typical "Edit" menu from a simple program. It probably contains an "Undo" and "Redo" item, followed by a separator, followed by "Cut", "Copy" and "Paste".

This would be accomplished by creating three Menu instances. The first would be populated with the "Undo" and "Redo" items, and the second with the "Cut", "Copy" and "Paste" items. The first and second menus would then be added as submenus of the third. In XML format, this would look something like the following:

<menu id='edit-menu'>
  <section>
    <item label='Undo'/>
    <item label='Redo'/>
  </section>
  <section>
    <item label='Cut'/>
    <item label='Copy'/>
    <item label='Paste'/>
  </section>
</menu>

The following example is exactly equivalent. It is more illustrative of the exact relationship between the menus and items (keeping in mind that the 'link' element defines a new menu that is linked to the containing one). The style of the second example is more verbose and difficult to read (and therefore not recommended except for the purpose of understanding what is really going on).

<menu id='edit-menu'>
  <item>
    <link name='section'>
      <item label='Undo'/>
      <item label='Redo'/>
    </link>
  </item>
  <item>
    <link name='section'>
      <item label='Cut'/>
      <item label='Copy'/>
      <item label='Paste'/>
    </link>
  </item>
</menu>

label

the section label, or None

section

a MenuModel with the items of the section

Returns

a new MenuItem

pub fn new_submenu<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
label: P,
submenu: &Q
) -> MenuItem

Creates a new MenuItem representing a submenu.

This is a convenience API around MenuItem::new and MenuItemExt::set_submenu.

label

the section label, or None

submenu

a MenuModel with the items of the submenu

Returns

a new MenuItem