gio

Struct RegistrationBuilder

pub struct RegistrationBuilder<'a> { /* private fields */ }

Build a registered DBus object, by handling different parts of DBus.

Implementations

impl<'a> RegistrationBuilder<'a>

pub fn method_call<F: Fn(DBusConnection, Option<&str>, &str, Option<&str>, &str, Variant, DBusMethodInvocation) + 'static>( self, f: F, ) -> Self

pub fn typed_method_call<T: DBusMethodCall>(self) -> MethodCallBuilder<'a, T>

Handle method calls on this object.

Return a builder for method calls which parses method names and parameters with the given DBusMethodCall and then allows to dispatch the parsed call either synchronously or asynchronously.

pub fn property<F: Fn(DBusConnection, Option<&str>, &str, &str, &str) -> Variant + 'static>( self, f: F, ) -> Self

pub fn set_property<F: Fn(DBusConnection, Option<&str>, &str, &str, &str, Variant) -> bool + 'static>( self, f: F, ) -> Self

pub fn build(self) -> Result<RegistrationId, Error>