Struct gio::SimpleAction

pub struct SimpleAction(_, _);

A SimpleAction is the obvious simple implementation of the Action interface. This is the easiest way to create an action for purposes of adding it to a SimpleActionGroup.

See also GtkAction.

Implements

SimpleActionExt, ObjectExt, ActionExt

Methods

impl SimpleAction

fn new<'a, P: Into<Option<&'a VariantTy>>>(
name: &str,
parameter_type: P
) -> SimpleAction

Creates a new action.

The created action is stateless. See SimpleAction::new_stateful.

name

the name of the action

parameter_type

the type of parameter to the activate function

Returns

a new SimpleAction

fn new_stateful<'a, P: Into<Option<&'a VariantTy>>>(
name: &str,
parameter_type: P,
state: &Variant
) -> SimpleAction

Creates a new stateful action.

state is the initial state of the action. All future state values must have the same glib::VariantType as the initial state.

If the state GVariant is floating, it is consumed.

name

the name of the action

parameter_type

the type of the parameter to the activate function

state

the initial state of the action

Returns

a new SimpleAction