Struct gio::Subprocess

pub struct Subprocess(_, _);

Subprocess allows the creation of and interaction with child processes.

Processes can be communicated with using standard GIO-style APIs (ie: InputStream, OutputStream). There are GIO-style APIs to wait for process termination (ie: cancellable and with an asynchronous variant).

There is an API to force a process to terminate, as well as a race-free API for sending UNIX signals to a subprocess.

One major advantage that GIO brings over the core GLib library is comprehensive API for asynchronous I/O, such OutputStreamExt::splice_async. This makes GSubprocess significantly more powerful and flexible than equivalent APIs in some other languages such as the subprocess.py included with Python. For example, using Subprocess one could create two child processes, reading standard output from the first, processing it, and writing to the input stream of the second, all without blocking the main loop.

A powerful Subprocess::communicate API is provided similar to the communicate() method of subprocess.py. This enables very easy interaction with a subprocess that has been opened with pipes.

Subprocess defaults to tight control over the file descriptors open in the child process, avoiding dangling-fd issues that are caused by a simple fork/exec. The only open file descriptors in the spawned process are ones that were explicitly specified by the Subprocess API (unless SubprocessFlags::InheritFds was specified).

Subprocess will quickly reap all child processes as they exit, avoiding "zombie processes" remaining around for long periods of time. Subprocess::wait can be used to wait for this to happen, but it will happen even without the call being explicitly made.

As a matter of principle, Subprocess has no API that accepts shell-style space-separated strings. It will, however, match the typical shell behaviour of searching the PATH for executables that do not contain a directory separator in their name.

Subprocess attempts to have a very simple API for most uses (ie: spawning a subprocess with arguments and support for most typical kinds of input and output redirection). See Subprocess::new. The SubprocessLauncher API is provided for more complicated cases (advanced types of redirection, environment variable manipulation, change of working directory, child setup functions, etc).

A typical use of Subprocess will involve calling Subprocess::new, followed by Subprocess::wait_async or Subprocess::wait. After the process exits, the status can be checked using functions such as Subprocess::get_if_exited (which are similar to the familiar WIFEXITED-style POSIX macros).

Implements

glib::object::ObjectExt

Methods

impl Subprocess

pub fn communicate_utf8_async<R: FnOnce(Result<(GString, GString), Error>) + Send + 'static>(
&self,
stdin_buf: Option<String>,
cancellable: Option<&Cancellable>,
callback: R
)

Asynchronous version of Subprocess::communicate_utf8. Complete invocation with Subprocess::communicate_utf8_finish.

stdin_buf

Input data, or None

cancellable

Cancellable

callback

Callback

user_data

User data

pub fn communicate_utf8_async_future(
&self,
stdin_buf: Option<String>
) -> Box<dyn Future<Output = Result<(GString, GString), Error>> + Unpin>

impl Subprocess

pub fn newv(
argv: &[&OsStr],
flags: SubprocessFlags
) -> Result<Subprocess, Error>

Create a new process with the given flags and argument list.

The argument list is expected to be None-terminated.

argv

commandline arguments for the subprocess

flags

flags that define the behaviour of the subprocess

Returns

A newly created Subprocess, or None on error (and error will be set)

pub fn communicate<P: IsA<Cancellable>>(
&self,
stdin_buf: Option<&Bytes>,
cancellable: Option<&P>
) -> Result<(Option<Bytes>, Option<Bytes>), Error>

Communicate with the subprocess until it terminates, and all input and output has been completed.

If stdin_buf is given, the subprocess must have been created with SubprocessFlags::StdinPipe. The given data is fed to the stdin of the subprocess and the pipe is closed (ie: EOF).

At the same time (as not to cause blocking when dealing with large amounts of data), if SubprocessFlags::StdoutPipe or SubprocessFlags::StderrPipe were used, reads from those streams. The data that was read is returned in stdout and/or the stderr.

If the subprocess was created with SubprocessFlags::StdoutPipe, stdout_buf will contain the data read from stdout. Otherwise, for subprocesses not created with SubprocessFlags::StdoutPipe, stdout_buf will be set to None. Similar provisions apply to stderr_buf and SubprocessFlags::StderrPipe.

As usual, any output variable may be given as None to ignore it.

If you desire the stdout and stderr data to be interleaved, create the subprocess with SubprocessFlags::StdoutPipe and SubprocessFlags::StderrMerge. The merged result will be returned in stdout_buf and stderr_buf will be set to None.

In case of any error (including cancellation), false will be returned with error set. Some or all of the stdin data may have been written. Any stdout or stderr data that has been read will be discarded. None of the out variables (aside from error) will have been set to anything in particular and should not be inspected.

In the case that true is returned, the subprocess has exited and the exit status inspection APIs (eg: Subprocess::get_if_exited, Subprocess::get_exit_status) may be used.

You should not attempt to use any of the subprocess pipes after starting this function, since they may be left in strange states, even if the operation was cancelled. You should especially not attempt to interact with the pipes while the operation is in progress (either from another thread or if using the asynchronous version).

stdin_buf

data to send to the stdin of the subprocess, or None

cancellable

a Cancellable

stdout_buf

data read from the subprocess stdout

stderr_buf

data read from the subprocess stderr

Returns

true if successful

pub fn communicate_async<P: IsA<Cancellable>, Q: FnOnce(Result<(Bytes, Bytes), Error>) + Send + 'static>(
&self,
stdin_buf: Option<&Bytes>,
cancellable: Option<&P>,
callback: Q
)

Asynchronous version of Subprocess::communicate. Complete invocation with Subprocess::communicate_finish.

stdin_buf

Input data, or None

cancellable

Cancellable

callback

Callback

user_data

User data

pub fn communicate_async_future(
&self,
stdin_buf: Option<&Bytes>
) -> Box_<dyn Future<Output = Result<(Bytes, Bytes), Error>> + Unpin>

pub fn communicate_utf8<P: IsA<Cancellable>>(
&self,
stdin_buf: Option<&str>,
cancellable: Option<&P>
) -> Result<(Option<GString>, Option<GString>), Error>

pub fn force_exit(&self)

pub fn get_exit_status(&self) -> i32

pub fn get_identifier(&self) -> Option<GString>

pub fn get_if_exited(&self) -> bool

pub fn get_if_signaled(&self) -> bool

pub fn get_status(&self) -> i32

pub fn get_stderr_pipe(&self) -> Option<InputStream>

pub fn get_stdin_pipe(&self) -> Option<OutputStream>

pub fn get_stdout_pipe(&self) -> Option<InputStream>

pub fn get_successful(&self) -> bool

pub fn get_term_sig(&self) -> i32

pub fn send_signal(&self, signal_num: i32)

pub fn wait<P: IsA<Cancellable>>(
&self,
cancellable: Option<&P>
) -> Result<(), Error>

pub fn wait_async<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
cancellable: Option<&P>,
callback: Q
)

pub fn wait_async_future(
&self
) -> Box_<dyn Future<Output = Result<(), Error>> + Unpin>

pub fn wait_check<P: IsA<Cancellable>>(
&self,
cancellable: Option<&P>
) -> Result<(), Error>

pub fn wait_check_async<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
cancellable: Option<&P>,
callback: Q
)

pub fn wait_check_async_future(
&self
) -> Box_<dyn Future<Output = Result<(), Error>> + Unpin>