gio
Struct TlsConnection

pub struct TlsConnection { /* private fields */ }

GTlsConnection is the base TLS connection class type, which wraps a IOStream and provides TLS encryption on top of it. Its subclasses, TlsClientConnection and TlsServerConnection, implement client-side and server-side TLS, respectively.

For DTLS (Datagram TLS) support, see Gio::DtlsConnection.

This is an Abstract Base Class, you cannot instantiate it.

Properties

advertised-protocols

The list of application-layer protocols that the connection advertises that it is willing to speak. See g_tls_connection_set_advertised_protocols().

Readable | Writeable

base-io-stream

The #GIOStream that the connection wraps. The connection holds a reference to this stream, and may run operations on the stream from other threads throughout its lifetime. Consequently, after the #GIOStream has been constructed, application code may only run its own operations on this stream when no #GIOStream operations are running.

Readable | Writeable | Construct Only

certificate

The connection’s certificate; see g_tls_connection_set_certificate().

Readable | Writeable

ciphersuite-name

The name of the TLS ciphersuite in use. See g_tls_connection_get_ciphersuite_name().

Readable

database

The certificate database to use when verifying this TLS connection. If no certificate database is set, then the default database will be used. See g_tls_backend_get_default_database().

When using a non-default database, #GTlsConnection must fall back to using the #GTlsDatabase to perform certificate verification using g_tls_database_verify_chain(), which means certificate verification will not be able to make use of TLS session context. This may be less secure. For example, if you create your own #GTlsDatabase that just wraps the default #GTlsDatabase, you might expect that you have not changed anything, but this is not true because you may have altered the behavior of #GTlsConnection by causing it to use g_tls_database_verify_chain(). See the documentation of g_tls_database_verify_chain() for more details on specific security checks that may not be performed. Accordingly, setting a non-default database is discouraged except for specialty applications with unusual security requirements.

Readable | Writeable

interaction

A #GTlsInteraction object to be used when the connection or certificate database need to interact with the user. This will be used to prompt the user for passwords where necessary.

Readable | Writeable

negotiated-protocol

The application-layer protocol negotiated during the TLS handshake. See g_tls_connection_get_negotiated_protocol().

Readable

peer-certificate

The connection’s peer’s certificate, after the TLS handshake has completed or failed. Note in particular that this is not yet set during the emission of #GTlsConnection::accept-certificate.

(You can watch for a #GObject::notify signal on this property to detect when a handshake has occurred.)

Readable

peer-certificate-errors

The errors noticed while verifying #GTlsConnection:peer-certificate. Normally this should be 0, but it may not be if #GTlsClientConnection:validation-flags is not TlsCertificateFlags::VALIDATE_ALL, or if #GTlsConnection::accept-certificate overrode the default behavior.

GLib guarantees that if certificate verification fails, at least one error will be set, but it does not guarantee that all possible errors will be set. Accordingly, you may not safely decide to ignore any particular type of error. For example, it would be incorrect to mask TlsCertificateFlags::EXPIRED if you want to allow expired certificates, because this could potentially be the only error flag set even if other problems exist with the certificate.

Readable

protocol-version

The TLS protocol version in use. See g_tls_connection_get_protocol_version().

Readable

rehandshake-mode

The rehandshaking mode. See g_tls_connection_set_rehandshake_mode().

Readable | Writeable | Construct

require-close-notify

Whether or not proper TLS close notification is required. See g_tls_connection_set_require_close_notify().

Readable | Writeable | Construct

use-system-certdb

Whether or not the system certificate database will be used to verify peer certificates. See g_tls_connection_set_use_system_certdb().

Readable | Writeable | Construct

IOStream

closed

Whether the stream is closed.

Readable

input-stream

The InputStream to read from.

Readable

output-stream

The OutputStream to write to.

Readable

Signals

accept-certificate

Emitted during the TLS handshake after the peer certificate has been received. You can examine @peer_cert’s certification path by calling g_tls_certificate_get_issuer() on it.

For a client-side connection, @peer_cert is the server’s certificate, and the signal will only be emitted if the certificate was not acceptable according to @conn’s #GTlsClientConnection:validation_flags. If you would like the certificate to be accepted despite @errors, return true from the signal handler. Otherwise, if no handler accepts the certificate, the handshake will fail with TlsError::BadCertificate.

GLib guarantees that if certificate verification fails, this signal will be emitted with at least one error will be set in @errors, but it does not guarantee that all possible errors will be set. Accordingly, you may not safely decide to ignore any particular type of error. For example, it would be incorrect to ignore TlsCertificateFlags::EXPIRED if you want to allow expired certificates, because this could potentially be the only error flag set even if other problems exist with the certificate.

For a server-side connection, @peer_cert is the certificate presented by the client, if this was requested via the server’s #GTlsServerConnection:authentication_mode. On the server side, the signal is always emitted when the client presents a certificate, and the certificate will only be accepted if a handler returns true.

Note that if this signal is emitted as part of asynchronous I/O in the main thread, then you should not attempt to interact with the user before returning from the signal handler. If you want to let the user decide whether or not to accept the certificate, you would have to return false from the signal handler on the first attempt, and then after the connection attempt returns a TlsError::BadCertificate, you can interact with the user, and if the user decides to accept the certificate, remember that fact, create a new connection, and return true from the signal handler the next time.

If you are doing I/O in another thread, you do not need to worry about this, and can simply block in the signal handler until the UI thread returns an answer.

Implements

TlsConnectionExt, IOStreamExt, [trait@glib::ObjectExt], TlsConnectionExtManual, IOStreamExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl TlsConnection

pub const NONE: Option<&'static TlsConnection> = None