gio
Struct TlsServerConnection

pub struct TlsServerConnection { /* private fields */ }

GTlsServerConnection is the server-side subclass of TlsConnection, representing a server-side TLS connection.

Properties

authentication-mode

The #GTlsAuthenticationMode for the server. This can be changed before calling g_tls_connection_handshake() if you want to rehandshake with a different mode from the initial handshake.

Readable | Writeable

TlsConnection

advertised-protocols

The list of application-layer protocols that the connection advertises that it is willing to speak. See g_tls_connection_set_advertised_protocols().

Readable | Writeable

base-io-stream

The #GIOStream that the connection wraps. The connection holds a reference to this stream, and may run operations on the stream from other threads throughout its lifetime. Consequently, after the #GIOStream has been constructed, application code may only run its own operations on this stream when no #GIOStream operations are running.

Readable | Writeable | Construct Only

certificate

The connection’s certificate; see g_tls_connection_set_certificate().

Readable | Writeable

ciphersuite-name

The name of the TLS ciphersuite in use. See g_tls_connection_get_ciphersuite_name().

Readable

database

The certificate database to use when verifying this TLS connection. If no certificate database is set, then the default database will be used. See g_tls_backend_get_default_database().

When using a non-default database, #GTlsConnection must fall back to using the #GTlsDatabase to perform certificate verification using g_tls_database_verify_chain(), which means certificate verification will not be able to make use of TLS session context. This may be less secure. For example, if you create your own #GTlsDatabase that just wraps the default #GTlsDatabase, you might expect that you have not changed anything, but this is not true because you may have altered the behavior of #GTlsConnection by causing it to use g_tls_database_verify_chain(). See the documentation of g_tls_database_verify_chain() for more details on specific security checks that may not be performed. Accordingly, setting a non-default database is discouraged except for specialty applications with unusual security requirements.

Readable | Writeable

interaction

A #GTlsInteraction object to be used when the connection or certificate database need to interact with the user. This will be used to prompt the user for passwords where necessary.

Readable | Writeable

negotiated-protocol

The application-layer protocol negotiated during the TLS handshake. See g_tls_connection_get_negotiated_protocol().

Readable

peer-certificate

The connection’s peer’s certificate, after the TLS handshake has completed or failed. Note in particular that this is not yet set during the emission of #GTlsConnection::accept-certificate.

(You can watch for a #GObject::notify signal on this property to detect when a handshake has occurred.)

Readable

peer-certificate-errors

The errors noticed while verifying #GTlsConnection:peer-certificate. Normally this should be 0, but it may not be if #GTlsClientConnection:validation-flags is not TlsCertificateFlags::VALIDATE_ALL, or if #GTlsConnection::accept-certificate overrode the default behavior.

GLib guarantees that if certificate verification fails, at least one error will be set, but it does not guarantee that all possible errors will be set. Accordingly, you may not safely decide to ignore any particular type of error. For example, it would be incorrect to mask TlsCertificateFlags::EXPIRED if you want to allow expired certificates, because this could potentially be the only error flag set even if other problems exist with the certificate.

Readable

protocol-version

The TLS protocol version in use. See g_tls_connection_get_protocol_version().

Readable

rehandshake-mode

The rehandshaking mode. See g_tls_connection_set_rehandshake_mode().

Readable | Writeable | Construct

require-close-notify

Whether or not proper TLS close notification is required. See g_tls_connection_set_require_close_notify().

Readable | Writeable | Construct

use-system-certdb

Whether or not the system certificate database will be used to verify peer certificates. See g_tls_connection_set_use_system_certdb().

Readable | Writeable | Construct

IOStream

closed

Whether the stream is closed.

Readable

input-stream

The InputStream to read from.

Readable

output-stream

The OutputStream to write to.

Readable

Implements

TlsServerConnectionExt, TlsConnectionExt, IOStreamExt, [trait@glib::ObjectExt], TlsConnectionExtManual, IOStreamExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl TlsServerConnection

pub const NONE: Option<&'static TlsServerConnection> = None

pub fn new( base_io_stream: &impl IsA<IOStream>, certificate: Option<&impl IsA<TlsCertificate>>, ) -> Result<TlsServerConnection, Error>

Creates a new #GTlsServerConnection wrapping @base_io_stream (which must have pollable input and output streams).

See the documentation for #GTlsConnection:base-io-stream for restrictions on when application code can run operations on the @base_io_stream after this function has returned.

base_io_stream

the #GIOStream to wrap

certificate

the default server certificate, or None

Returns

the new #GTlsServerConnection, or None on error