gio

Struct UnixMountEntry

pub struct UnixMountEntry { /* private fields */ }

Defines a Unix mount entry (e.g. /media/cdrom). This corresponds roughly to a mtab entry.

GLib type: Boxed type with copy-on-clone semantics.

Implementations

impl UnixMountEntry

pub fn mounts() -> (Vec<UnixMountEntry>, u64)

pub fn mounts_from_file( table_path: impl AsRef<Path>, ) -> (Vec<UnixMountEntry>, u64)

pub fn unix_mount_get_mount_path(&self) -> PathBuf

๐Deprecated: Use mount_path()

pub fn is_changed_since(time: u64) -> bool

impl UnixMountEntry

pub fn as_ptr(&self) -> *mut GUnixMountEntry

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut GUnixMountEntry) -> &Self

Borrows the underlying C value.

pub unsafe fn from_glib_ptr_borrow_mut( ptr: &mut *mut GUnixMountEntry, ) -> &mut Self

Borrows the underlying C value mutably.

impl UnixMountEntry

pub fn device_path(&self) -> PathBuf

Gets the device path for a Unix mount.

Returns

a string containing the device path

pub fn fs_type(&self) -> GString

Gets the filesystem type for the Unix mount.

Returns

a string containing the file system type

pub fn mount_path(&self) -> PathBuf

Gets the mount path for a Unix mount.

Returns

the mount path for @self

pub fn options(&self) -> Option<GString>

Gets a comma separated list of mount options for the Unix mount.

For example: rw,relatime,seclabel,data=ordered.

This is similar to GioUnix::MountPoint::get_options(), but it takes a GioUnix::MountEntry as an argument.

Returns

a string containing the options, or NULL if not available.

pub fn root_path(&self) -> Option<GString>

Gets the root of the mount within the filesystem. This is useful e.g. for mounts created by bind operation, or btrfs subvolumes.

For example, the root path is equal to / for a mount created by mount /dev/sda1 /mnt/foo and /bar for mount --bind /mnt/foo/bar /mnt/bar.

Returns

a string containing the root, or NULL if not supported

pub fn guess_can_eject(&self) -> bool

Guesses whether a Unix mount entry can be ejected.

Returns

true if @self is deemed to be ejectable; false otherwise

pub fn guess_icon(&self) -> Icon

Guesses the icon of a Unix mount entry.

Returns

a Icon

pub fn guess_name(&self) -> GString

Guesses the name of a Unix mount entry.

The result is a translated string.

Returns

a newly allocated translated string

pub fn guess_should_display(&self) -> bool

Guesses whether a Unix mount entry should be displayed in the UI.

Returns

true if @self is deemed to be displayable; false otherwise

pub fn guess_symbolic_icon(&self) -> Icon

Guesses the symbolic icon of a Unix mount entry.

Returns

a Icon

pub fn is_readonly(&self) -> bool

Checks if a Unix mount is mounted read only.

Returns

true if @self is read only; false otherwise

pub fn is_system_internal(&self) -> bool

Checks if a Unix mount is a system mount.

This is the Boolean OR of is_system_fs_type(), is_system_device_path() and is_mount_path_system_internal() on @selfโs properties.

The definition of what a โsystemโ mount entry is may change over time as new file system types and device paths are ignored.

Returns

true if the Unix mount is for a system path; false otherwise

pub fn for_mount_path( mount_path: impl AsRef<Path>, ) -> (Option<UnixMountEntry>, u64)

Gets a GioUnix::MountEntry for a given mount path.

If @time_read is set, it will be filled with a Unix timestamp for checking if the mounts have changed since with mount_entries_changed_since().

If more mounts have the same mount path, the last matching mount is returned.

This will return NULL if there is no mount point at @mount_path.

mount_path

path for a possible Unix mount

Returns

a GioUnix::MountEntry

time_read

return location for a timestamp

pub fn for_file_path( file_path: impl AsRef<Path>, ) -> (Option<UnixMountEntry>, u64)

Gets a GioUnix::MountEntry for a given file path.

If @time_read is set, it will be filled with a Unix timestamp for checking if the mounts have changed since with mount_entries_changed_since().

If more mounts have the same mount path, the last matching mount is returned.

This will return NULL if looking up the mount entry fails, if @file_path doesnโt exist or there is an I/O error.

file_path

file path on some Unix mount

Returns

a GioUnix::MountEntry

time_read

return location for a timestamp