Struct gio::UnixMountPoint

pub struct UnixMountPoint { /* private fields */ }

Available on Unix only.

Defines a Unix mount point (e.g. /dev). This corresponds roughly to a fstab entry.

GLib type: Boxed type with copy-on-clone semantics.

Implementations

impl UnixMountPoint

pub fn mount_points() -> (Vec<UnixMountPoint>, u64)

pub fn is_changed_since(time: u64) -> bool

impl UnixMountPoint

pub fn as_ptr(&self) -> *mut GUnixMountPoint

Return the inner pointer to the underlying C value.

pub unsafe fn from_glib_ptr_borrow(ptr: &*mut GUnixMountPoint) -> &Self

Borrows the underlying C value.

pub unsafe fn from_glib_ptr_borrow_mut( ptr: &mut *mut GUnixMountPoint, ) -> &mut Self

Borrows the underlying C value mutably.

impl UnixMountPoint

pub fn device_path(&self) -> PathBuf

Gets the device path for a unix mount point.

Returns

a string containing the device path.

pub fn fs_type(&self) -> GString

Gets the file system type for the mount point.

Returns

a string containing the file system type.

pub fn mount_path(&self) -> PathBuf

Gets the mount path for a unix mount point.

Returns

a string containing the mount path.

pub fn options(&self) -> Option<GString>

Gets the options for the mount point.

Returns

a string containing the options.

pub fn guess_can_eject(&self) -> bool

Guesses whether a Unix mount point can be ejected.

Returns

true if @self is deemed to be ejectable.

pub fn guess_icon(&self) -> Icon

Guesses the icon of a Unix mount point.

Returns

a #GIcon

pub fn guess_name(&self) -> GString

Guesses the name of a Unix mount point. The result is a translated string.

Returns

A newly allocated string that must be freed with g_free()

pub fn guess_symbolic_icon(&self) -> Icon

Guesses the symbolic icon of a Unix mount point.

Returns

a #GIcon

pub fn is_loopback(&self) -> bool

Checks if a unix mount point is a loopback device.

Returns

true if the mount point is a loopback. false otherwise.

pub fn is_readonly(&self) -> bool

Checks if a unix mount point is read only.

Returns

true if a mount point is read only.

pub fn is_user_mountable(&self) -> bool

Checks if a unix mount point is mountable by the user.

Returns

true if the mount point is user mountable.

pub fn at(mount_path: impl AsRef<Path>) -> (Option<UnixMountPoint>, u64)

Available on crate feature v2_66 only.

Gets a #GUnixMountPoint for a given mount path. If @time_read is set, it will be filled with a unix timestamp for checking if the mount points have changed since with g_unix_mount_points_changed_since().

If more mount points have the same mount path, the last matching mount point is returned.

mount_path

path for a possible unix mount point.

Returns

a #GUnixMountPoint, or None if no match is found.

time_read

guint64 to contain a timestamp.