gio
Struct UnixSocketAddress

pub struct UnixSocketAddress { /* private fields */ }

Available on Unix only.

Support for UNIX-domain (also known as local) sockets, corresponding to struct sockaddr_un.

UNIX domain sockets are generally visible in the filesystem. However, some systems support abstract socket names which are not visible in the filesystem and not affected by the filesystem permissions, visibility, etc. Currently this is only supported under Linux. If you attempt to use abstract sockets on other systems, function calls may return G_IO_ERROR_NOT_SUPPORTED errors. You can use abstract_names_supported() to see if abstract names are supported.

Since GLib 2.72, GUnixSocketAddress is available on all platforms. It requires underlying system support (such as Windows 10 with AF_UNIX) at run time.

Before GLib 2.72, <gio/gunixsocketaddress.h> belonged to the UNIX-specific GIO interfaces, thus you had to use the gio-unix-2.0.pc pkg-config file when using it. This is no longer necessary since GLib 2.72.

Properties

abstract

Whether or not this is an abstract address

Readable | Writeable | Construct Only

address-type

The type of Unix socket address.

Readable | Writeable | Construct Only

path

Unix socket path.

Readable | Writeable | Construct Only

path-as-array

Unix socket path, as a byte array.

Readable | Writeable | Construct Only

SocketAddress

family

The family of the socket address.

Readable

Implements

UnixSocketAddressExt, SocketAddressExt, [trait@glib::ObjectExt], SocketConnectableExt, UnixSocketAddressExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl UnixSocketAddress

pub fn new(path: &Path) -> UnixSocketAddress

Creates a new #GUnixSocketAddress for @path.

To create abstract socket addresses, on systems that support that, use g_unix_socket_address_new_abstract().

path

the socket path

Returns

a new #GUnixSocketAddress

pub fn with_type(address_type: UnixSocketAddressPath<'_>) -> Self

Creates a new #GUnixSocketAddress of type @type_ with name @path.

If @type_ is UnixSocketAddressType::Path, this is equivalent to calling g_unix_socket_address_new().

If @type_ is UnixSocketAddressType::Anonymous, @path and @path_len will be ignored.

If @path_type is UnixSocketAddressType::Abstract, then @path_len bytes of @path will be copied to the socket’s path, and only those bytes will be considered part of the name. (If @path_len is -1, then @path is assumed to be NUL-terminated.) For example, if @path was “test”, then calling g_socket_address_get_native_size() on the returned socket would return 7 (2 bytes of overhead, 1 byte for the abstract-socket indicator byte, and 4 bytes for the name “test”).

If @path_type is UnixSocketAddressType::AbstractPadded, then @path_len bytes of @path will be copied to the socket’s path, the rest of the path will be padded with 0 bytes, and the entire zero-padded buffer will be considered the name. (As above, if @path_len is -1, then @path is assumed to be NUL-terminated.) In this case, g_socket_address_get_native_size() will always return the full size of a struct sockaddr_un, although g_unix_socket_address_get_path_len() will still return just the length of @path.

UnixSocketAddressType::Abstract is preferred over UnixSocketAddressType::AbstractPadded for new programs. Of course, when connecting to a server created by another process, you must use the appropriate type corresponding to how that process created its listening socket.

path

the name

type_

a #GUnixSocketAddressType

Returns

a new #GUnixSocketAddress

impl UnixSocketAddress

pub const NONE: Option<&'static UnixSocketAddress> = None

pub fn abstract_names_supported() -> bool

Checks if abstract UNIX domain socket names are supported.

Returns

true if supported, false otherwise