Trait gio::ActionExt

pub trait ActionExt: 'static {

    fn activate(&self, parameter: Option<&Variant>);

    fn change_state(&self, value: &Variant);

    fn is_enabled(&self) -> bool;

    fn name(&self) -> GString;

    fn parameter_type(&self) -> Option<VariantType>;

    fn state(&self) -> Option<Variant>;

    fn state_hint(&self) -> Option<Variant>;

    fn state_type(&self) -> Option<VariantType>;

    fn connect_property_enabled_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_parameter_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_state_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_state_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

}

Trait containing all Action methods.

Implementors

Action, PropertyAction, SimpleAction

Required methods

fn activate(&self, parameter: Option<&Variant>)

Activates the action.

parameter must be the correct type of parameter for the action (ie: the parameter type given at construction time). If the parameter type was None then parameter must also be None.

If the parameter GVariant is floating, it is consumed.

parameter

the parameter to the activation

fn change_state(&self, value: &Variant)

Request for the state of self to be changed to value.

The action must be stateful and value must be of the correct type. See Action::get_state_type.

This call merely requests a change. The action may refuse to change its state or may change its state to something other than value. See Action::get_state_hint.

If the value GVariant is floating, it is consumed.

value

the new state

fn is_enabled(&self) -> bool

fn name(&self) -> GString

fn parameter_type(&self) -> Option<VariantType>

fn state(&self) -> Option<Variant>

fn state_hint(&self) -> Option<Variant>

fn state_type(&self) -> Option<VariantType>

fn connect_property_enabled_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_parameter_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_state_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_state_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...