Trait gio::FileExt

pub trait FileExt {
    fn delete<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        cancellable: P
    ) -> Result<(), Error>;

    fn dup(&self) -> Option<File>;

    fn equal<P: IsA<File>>(&self, file2: &P) -> bool;

    fn get_basename(&self) -> Option<PathBuf>;

    fn get_child<P: AsRef<Path>>(&self, name: P) -> Option<File>;

    fn get_child_for_display_name(
        &self, 
        display_name: &str
    ) -> Result<File, Error>;

    fn get_parent(&self) -> Option<File>;

    fn get_parse_name(&self) -> Option<String>;

    fn get_path(&self) -> Option<PathBuf>;

    fn get_relative_path<P: IsA<File>>(&self, descendant: &P) -> Option<PathBuf>;

    fn get_uri(&self) -> Option<String>;

    fn get_uri_scheme(&self) -> Option<String>;

    fn has_parent<'a, P: IsA<File> + 'a, Q: Into<Option<&'a P>>>(
        &self, 
        parent: Q
    ) -> bool;

    fn has_prefix<P: IsA<File>>(&self, prefix: &P) -> bool;

    fn has_uri_scheme(&self, uri_scheme: &str) -> bool;

    fn is_native(&self) -> bool;

    fn load_contents<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        cancellable: P
    ) -> Result<(Vec<u8>, String), Error>;

    fn make_directory<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        cancellable: P
    ) -> Result<(), Error>;

    fn make_directory_with_parents<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        cancellable: P
    ) -> Result<(), Error>;

    fn make_symbolic_link<'a, P: AsRef<Path>, Q: Into<Option<&'a Cancellable>>>(
        &self, 
        symlink_value: P, 
        cancellable: Q
    ) -> Result<(), Error>;

    fn query_default_handler<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        cancellable: P
    ) -> Result<AppInfo, Error>;

    fn query_exists<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        cancellable: P
    ) -> bool;

    fn query_file_type<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> FileType;

    fn query_filesystem_info<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        attributes: &str, 
        cancellable: P
    ) -> Result<FileInfo, Error>;

    fn query_info<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        attributes: &str, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> Result<FileInfo, Error>;

    fn replace_contents<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b Cancellable>>>(
        &self, 
        contents: &[u8], 
        etag: P, 
        make_backup: bool, 
        flags: FileCreateFlags, 
        cancellable: Q
    ) -> Result<String, Error>;

    fn resolve_relative_path<P: AsRef<Path>>(
        &self, 
        relative_path: P
    ) -> Option<File>;

    fn set_attribute_byte_string<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        attribute: &str, 
        value: &str, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> Result<(), Error>;

    fn set_attribute_int32<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        attribute: &str, 
        value: i32, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> Result<(), Error>;

    fn set_attribute_int64<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        attribute: &str, 
        value: i64, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> Result<(), Error>;

    fn set_attribute_string<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        attribute: &str, 
        value: &str, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> Result<(), Error>;

    fn set_attribute_uint32<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        attribute: &str, 
        value: u32, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> Result<(), Error>;

    fn set_attribute_uint64<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        attribute: &str, 
        value: u64, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> Result<(), Error>;

    fn set_attributes_from_info<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        info: &FileInfo, 
        flags: FileQueryInfoFlags, 
        cancellable: P
    ) -> Result<(), Error>;

    fn set_display_name<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        display_name: &str, 
        cancellable: P
    ) -> Result<File, Error>;

    fn supports_thread_contexts(&self) -> bool;

    fn trash<'a, P: Into<Option<&'a Cancellable>>>(
        &self, 
        cancellable: P
    ) -> Result<(), Error>;
}

Trait containing all File methods.

Implementors

File

Required Methods

fn delete<'a, P: Into<Option<&'a Cancellable>>>(
&self,
cancellable: P
) -> Result<(), Error>

Deletes a file. If the self is a directory, it will only be deleted if it is empty. This has the same semantics as g_unlink.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

cancellable

optional Cancellable object, None to ignore

Returns

true if the file was deleted. false otherwise.

fn dup(&self) -> Option<File>

Duplicates a File handle. This operation does not duplicate the actual file or directory represented by the File; see File::copy if attempting to copy a file.

This call does no blocking I/O.

Returns

a new File that is a duplicate of the given File.

fn equal<P: IsA<File>>(&self, file2: &P) -> bool

Checks if the two given GFiles refer to the same file.

Note that two GFiles that differ can still refer to the same file on the filesystem due to various forms of filename aliasing.

This call does no blocking I/O.

file2

the second File

Returns

true if self and file2 are equal.

fn get_basename(&self) -> Option<PathBuf>

Gets the base name (the last component of the path) for a given File.

If called for the top level of a system (such as the filesystem root or a uri like sftp://host/) it will return a single directory separator (and on Windows, possibly a drive letter).

The base name is a byte string (not UTF-8). It has no defined encoding or rules other than it may not contain zero bytes. If you want to use filenames in a user interface you should use the display name that you can get by requesting the G_FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME attribute with File::query_info.

This call does no blocking I/O.

Returns

string containing the File's base name, or None if given File is invalid. The returned string should be freed with g_free when no longer needed.

fn get_child<P: AsRef<Path>>(&self, name: P) -> Option<File>

Gets a child of self with basename equal to name.

Note that the file with that specific name might not exist, but you can still have a File that points to it. You can use this for instance to create that file.

This call does no blocking I/O.

name

string containing the child's basename

Returns

a File to a child specified by name. Free the returned object with gobject::ObjectExt::unref.

fn get_child_for_display_name(&self, display_name: &str) -> Result<File, Error>

Gets the child of self for a given display_name (i.e. a UTF-8 version of the name). If this function fails, it returns None and error will be set. This is very useful when constructing a File for a new file and the user entered the filename in the user interface, for instance when you select a directory and type a filename in the file selector.

This call does no blocking I/O.

display_name

string to a possible child

Returns

a File to the specified child, or None if the display name couldn't be converted. Free the returned object with gobject::ObjectExt::unref.

fn get_parent(&self) -> Option<File>

Gets the parent directory for the self. If the self represents the root directory of the file system, then None will be returned.

This call does no blocking I/O.

Returns

a File structure to the parent of the given File or None if there is no parent. Free the returned object with gobject::ObjectExt::unref.

fn get_parse_name(&self) -> Option<String>

Gets the parse name of the self. A parse name is a UTF-8 string that describes the file such that one can get the File back using File::parse_name.

This is generally used to show the File as a nice full-pathname kind of string in a user interface, like in a location entry.

For local files with names that can safely be converted to UTF-8 the pathname is used, otherwise the IRI is used (a form of URI that allows UTF-8 characters unescaped).

This call does no blocking I/O.

Returns

a string containing the File's parse name. The returned string should be freed with g_free when no longer needed.

fn get_path(&self) -> Option<PathBuf>

Gets the local pathname for File, if one exists. If non-None, this is guaranteed to be an absolute, canonical path. It might contain symlinks.

This call does no blocking I/O.

Returns

string containing the File's path, or None if no such path exists. The returned string should be freed with g_free when no longer needed.

fn get_relative_path<P: IsA<File>>(&self, descendant: &P) -> Option<PathBuf>

Gets the path for descendant relative to self.

This call does no blocking I/O.

descendant

input File

Returns

string with the relative path from descendant to self, or None if descendant doesn't have self as prefix. The returned string should be freed with g_free when no longer needed.

fn get_uri(&self) -> Option<String>

Gets the URI for the self.

This call does no blocking I/O.

Returns

a string containing the File's URI. The returned string should be freed with g_free when no longer needed.

fn get_uri_scheme(&self) -> Option<String>

Gets the URI scheme for a File. RFC 3986 decodes the scheme as:

URI = scheme ":" hier-part [ "?" query ] [ "#" fragment ]

Common schemes include "file", "http", "ftp", etc.

This call does no blocking I/O.

Returns

a string containing the URI scheme for the given File. The returned string should be freed with g_free when no longer needed.

fn has_parent<'a, P: IsA<File> + 'a, Q: Into<Option<&'a P>>>(
&self,
parent: Q
) -> bool

Checks if self has a parent, and optionally, if it is parent.

If parent is None then this function returns true if self has any parent at all. If parent is non-None then true is only returned if self is an immediate child of parent.

parent

the parent to check for, or None

Returns

true if self is an immediate child of parent (or any parent in the case that parent is None).

fn has_prefix<P: IsA<File>>(&self, prefix: &P) -> bool

Checks whether self has the prefix specified by prefix.

In other words, if the names of initial elements of self's pathname match prefix. Only full pathname elements are matched, so a path like /foo is not considered a prefix of /foobar, only of /foo/bar.

A File is not a prefix of itself. If you want to check for equality, use File::equal.

This call does no I/O, as it works purely on names. As such it can sometimes return false even if self is inside a prefix (from a filesystem point of view), because the prefix of self is an alias of prefix.

prefix

input File

Returns

true if the files's parent, grandparent, etc is prefix, false otherwise.

fn has_uri_scheme(&self, uri_scheme: &str) -> bool

Checks to see if a File has a given URI scheme.

This call does no blocking I/O.

uri_scheme

a string containing a URI scheme

Returns

true if File's backend supports the given URI scheme, false if URI scheme is None, not supported, or File is invalid.

fn is_native(&self) -> bool

Checks to see if a file is native to the platform.

A native file s one expressed in the platform-native filename format, e.g. "C:\Windows" or "/usr/bin/". This does not mean the file is local, as it might be on a locally mounted remote filesystem.

On some systems non-native files may be available using the native filesystem via a userspace filesystem (FUSE), in these cases this call will return false, but File::get_path will still return a native path.

This call does no blocking I/O.

Returns

true if self is native

fn load_contents<'a, P: Into<Option<&'a Cancellable>>>(
&self,
cancellable: P
) -> Result<(Vec<u8>, String), Error>

Loads the content of the file into memory. The data is always zero-terminated, but this is not included in the resultant length. The returned content should be freed with g_free when no longer needed.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

cancellable

optional Cancellable object, None to ignore

contents

a location to place the contents of the file

length

a location to place the length of the contents of the file, or None if the length is not needed

etag_out

a location to place the current entity tag for the file, or None if the entity tag is not needed

Returns

true if the self's contents were successfully loaded. false if there were errors.

fn make_directory<'a, P: Into<Option<&'a Cancellable>>>(
&self,
cancellable: P
) -> Result<(), Error>

Creates a directory. Note that this will only create a child directory of the immediate parent directory of the path or URI given by the File. To recursively create directories, see File::make_directory_with_parents. This function will fail if the parent directory does not exist, setting error to IOErrorEnum::NotFound. If the file system doesn't support creating directories, this function will fail, setting error to IOErrorEnum::NotSupported.

For a local File the newly created directory will have the default (current) ownership and permissions of the current process.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

cancellable

optional Cancellable object, None to ignore

Returns

true on successful creation, false otherwise.

fn make_directory_with_parents<'a, P: Into<Option<&'a Cancellable>>>(
&self,
cancellable: P
) -> Result<(), Error>

Creates a directory and any parent directories that may not exist similar to 'mkdir -p'. If the file system does not support creating directories, this function will fail, setting error to IOErrorEnum::NotSupported. If the directory itself already exists, this function will fail setting error to IOErrorEnum::Exists, unlike the similar g_mkdir_with_parents.

For a local File the newly created directories will have the default (current) ownership and permissions of the current process.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

cancellable

optional Cancellable object, None to ignore

Returns

true if all directories have been successfully created, false otherwise.

fn make_symbolic_link<'a, P: AsRef<Path>, Q: Into<Option<&'a Cancellable>>>(
&self,
symlink_value: P,
cancellable: Q
) -> Result<(), Error>

Creates a symbolic link named self which contains the string symlink_value.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

symlink_value

a string with the path for the target of the new symlink

cancellable

optional Cancellable object, None to ignore

Returns

true on the creation of a new symlink, false otherwise.

fn query_default_handler<'a, P: Into<Option<&'a Cancellable>>>(
&self,
cancellable: P
) -> Result<AppInfo, Error>

Returns the AppInfo that is registered as the default application to handle the file specified by self.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

cancellable

optional Cancellable object, None to ignore

Returns

a AppInfo if the handle was found, None if there were errors. When you are done with it, release it with gobject::ObjectExt::unref

fn query_exists<'a, P: Into<Option<&'a Cancellable>>>(
&self,
cancellable: P
) -> bool

Utility function to check if a particular file exists. This is implemented using File::query_info and as such does blocking I/O.

Note that in many cases it is racy to first check for file existence and then execute something based on the outcome of that, because the file might have been created or removed in between the operations. The general approach to handling that is to not check, but just do the operation and handle the errors as they come.

As an example of race-free checking, take the case of reading a file, and if it doesn't exist, creating it. There are two racy versions: read it, and on error create it; and: check if it exists, if not create it. These can both result in two processes creating the file (with perhaps a partially written file as the result). The correct approach is to always try to create the file with File::create which will either atomically create the file or fail with a IOErrorEnum::Exists error.

However, in many cases an existence check is useful in a user interface, for instance to make a menu item sensitive/insensitive, so that you don't have to fool users that something is possible and then just show an error dialog. If you do this, you should make sure to also handle the errors that can happen due to races when you execute the operation.

cancellable

optional Cancellable object, None to ignore

Returns

true if the file exists (and can be detected without error), false otherwise (or if cancelled).

fn query_file_type<'a, P: Into<Option<&'a Cancellable>>>(
&self,
flags: FileQueryInfoFlags,
cancellable: P
) -> FileType

Utility function to inspect the FileType of a file. This is implemented using File::query_info and as such does blocking I/O.

The primary use case of this method is to check if a file is a regular file, directory, or symlink.

flags

a set of FileQueryInfoFlags passed to File::query_info

cancellable

optional Cancellable object, None to ignore

Returns

The FileType of the file and FileType::Unknown if the file does not exist

fn query_filesystem_info<'a, P: Into<Option<&'a Cancellable>>>(
&self,
attributes: &str,
cancellable: P
) -> Result<FileInfo, Error>

Similar to File::query_info, but obtains information about the filesystem the self is on, rather than the file itself. For instance the amount of space available and the type of the filesystem.

The attributes value is a string that specifies the attributes that should be gathered. It is not an error if it's not possible to read a particular requested attribute from a file - it just won't be set. attributes should be a comma-separated list of attributes or attribute wildcards. The wildcard "*" means all attributes, and a wildcard like "filesystem::*" means all attributes in the filesystem namespace. The standard namespace for filesystem attributes is "filesystem". Common attributes of interest are G_FILE_ATTRIBUTE_FILESYSTEM_SIZE (the total size of the filesystem in bytes), G_FILE_ATTRIBUTE_FILESYSTEM_FREE (number of bytes available), and G_FILE_ATTRIBUTE_FILESYSTEM_TYPE (type of the filesystem).

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

If the file does not exist, the IOErrorEnum::NotFound error will be returned. Other errors are possible too, and depend on what kind of filesystem the file is on.

attributes

an attribute query string

cancellable

optional Cancellable object, None to ignore

Returns

a FileInfo or None if there was an error. Free the returned object with gobject::ObjectExt::unref.

fn query_info<'a, P: Into<Option<&'a Cancellable>>>(
&self,
attributes: &str,
flags: FileQueryInfoFlags,
cancellable: P
) -> Result<FileInfo, Error>

Gets the requested information about specified self. The result is a FileInfo object that contains key-value attributes (such as the type or size of the file).

The attributes value is a string that specifies the file attributes that should be gathered. It is not an error if it's not possible to read a particular requested attribute from a file - it just won't be set. attributes should be a comma-separated list of attributes or attribute wildcards. The wildcard "*" means all attributes, and a wildcard like "standard::*" means all attributes in the standard namespace. An example attribute query be "standard::*,owner::user". The standard attributes are available as defines, like G_FILE_ATTRIBUTE_STANDARD_NAME.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

For symlinks, normally the information about the target of the symlink is returned, rather than information about the symlink itself. However if you pass FileQueryInfoFlags::NofollowSymlinks in flags the information about the symlink itself will be returned. Also, for symlinks that point to non-existing files the information about the symlink itself will be returned.

If the file does not exist, the IOErrorEnum::NotFound error will be returned. Other errors are possible too, and depend on what kind of filesystem the file is on.

attributes

an attribute query string

flags

a set of FileQueryInfoFlags

cancellable

optional Cancellable object, None to ignore

Returns

a FileInfo for the given self, or None on error. Free the returned object with gobject::ObjectExt::unref.

fn replace_contents<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b Cancellable>>>(
&self,
contents: &[u8],
etag: P,
make_backup: bool,
flags: FileCreateFlags,
cancellable: Q
) -> Result<String, Error>

Replaces the contents of self with contents of length bytes.

If etag is specified (not None), any existing file must have that etag, or the error IOErrorEnum::WrongEtag will be returned.

If make_backup is true, this function will attempt to make a backup of self. Internally, it uses File::replace, so will try to replace the file contents in the safest way possible. For example, atomic renames are used when replacing local files’ contents.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

The returned new_etag can be used to verify that the file hasn't changed the next time it is saved over.

contents

a string containing the new contents for self

length

the length of contents in bytes

etag

the old [entity-tag][gfile-etag] for the document, or None

make_backup

true if a backup should be created

flags

a set of FileCreateFlags

new_etag

a location to a new [entity tag][gfile-etag] for the document. This should be freed with g_free when no longer needed, or None

cancellable

optional Cancellable object, None to ignore

Returns

true if successful. If an error has occurred, this function will return false and set error appropriately if present.

fn resolve_relative_path<P: AsRef<Path>>(
&self,
relative_path: P
) -> Option<File>

Resolves a relative path for self to an absolute path.

This call does no blocking I/O.

relative_path

a given relative path string

Returns

File to the resolved path. None if relative_path is None or if self is invalid. Free the returned object with gobject::ObjectExt::unref.

fn set_attribute_byte_string<'a, P: Into<Option<&'a Cancellable>>>(
&self,
attribute: &str,
value: &str,
flags: FileQueryInfoFlags,
cancellable: P
) -> Result<(), Error>

Sets attribute of type FileAttributeType::ByteString to value. If attribute is of a different type, this operation will fail, returning false.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

attribute

a string containing the attribute's name

value

a string containing the attribute's new value

flags

a FileQueryInfoFlags

cancellable

optional Cancellable object, None to ignore

Returns

true if the attribute was successfully set to value in the self, false otherwise.

fn set_attribute_int32<'a, P: Into<Option<&'a Cancellable>>>(
&self,
attribute: &str,
value: i32,
flags: FileQueryInfoFlags,
cancellable: P
) -> Result<(), Error>

Sets attribute of type FileAttributeType::Int32 to value. If attribute is of a different type, this operation will fail.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

attribute

a string containing the attribute's name

value

a gint32 containing the attribute's new value

flags

a FileQueryInfoFlags

cancellable

optional Cancellable object, None to ignore

Returns

true if the attribute was successfully set to value in the self, false otherwise.

fn set_attribute_int64<'a, P: Into<Option<&'a Cancellable>>>(
&self,
attribute: &str,
value: i64,
flags: FileQueryInfoFlags,
cancellable: P
) -> Result<(), Error>

Sets attribute of type FileAttributeType::Int64 to value. If attribute is of a different type, this operation will fail.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

attribute

a string containing the attribute's name

value

a guint64 containing the attribute's new value

flags

a FileQueryInfoFlags

cancellable

optional Cancellable object, None to ignore

Returns

true if the attribute was successfully set, false otherwise.

fn set_attribute_string<'a, P: Into<Option<&'a Cancellable>>>(
&self,
attribute: &str,
value: &str,
flags: FileQueryInfoFlags,
cancellable: P
) -> Result<(), Error>

Sets attribute of type FileAttributeType::String to value. If attribute is of a different type, this operation will fail.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

attribute

a string containing the attribute's name

value

a string containing the attribute's value

flags

FileQueryInfoFlags

cancellable

optional Cancellable object, None to ignore

Returns

true if the attribute was successfully set, false otherwise.

fn set_attribute_uint32<'a, P: Into<Option<&'a Cancellable>>>(
&self,
attribute: &str,
value: u32,
flags: FileQueryInfoFlags,
cancellable: P
) -> Result<(), Error>

Sets attribute of type FileAttributeType::Uint32 to value. If attribute is of a different type, this operation will fail.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

attribute

a string containing the attribute's name

value

a guint32 containing the attribute's new value

flags

a FileQueryInfoFlags

cancellable

optional Cancellable object, None to ignore

Returns

true if the attribute was successfully set to value in the self, false otherwise.

fn set_attribute_uint64<'a, P: Into<Option<&'a Cancellable>>>(
&self,
attribute: &str,
value: u64,
flags: FileQueryInfoFlags,
cancellable: P
) -> Result<(), Error>

Sets attribute of type FileAttributeType::Uint64 to value. If attribute is of a different type, this operation will fail.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

attribute

a string containing the attribute's name

value

a guint64 containing the attribute's new value

flags

a FileQueryInfoFlags

cancellable

optional Cancellable object, None to ignore

Returns

true if the attribute was successfully set to value in the self, false otherwise.

fn set_attributes_from_info<'a, P: Into<Option<&'a Cancellable>>>(
&self,
info: &FileInfo,
flags: FileQueryInfoFlags,
cancellable: P
) -> Result<(), Error>

Tries to set all attributes in the FileInfo on the target values, not stopping on the first error.

If there is any error during this operation then error will be set to the first error. Error on particular fields are flagged by setting the "status" field in the attribute value to FileAttributeStatus::ErrorSetting, which means you can also detect further errors.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

info

a FileInfo

flags

FileQueryInfoFlags

cancellable

optional Cancellable object, None to ignore

Returns

false if there was any error, true otherwise.

fn set_display_name<'a, P: Into<Option<&'a Cancellable>>>(
&self,
display_name: &str,
cancellable: P
) -> Result<File, Error>

Renames self to the specified display name.

The display name is converted from UTF-8 to the correct encoding for the target filesystem if possible and the self is renamed to this.

If you want to implement a rename operation in the user interface the edit name (G_FILE_ATTRIBUTE_STANDARD_EDIT_NAME) should be used as the initial value in the rename widget, and then the result after editing should be passed to File::set_display_name.

On success the resulting converted filename is returned.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

display_name

a string

cancellable

optional Cancellable object, None to ignore

Returns

a File specifying what self was renamed to, or None if there was an error. Free the returned object with gobject::ObjectExt::unref.

fn supports_thread_contexts(&self) -> bool

Checks if self supports [thread-default contexts][g-main-context-push-thread-default-context]. If this returns false, you cannot perform asynchronous operations on self in a thread that has a thread-default context.

Returns

Whether or not self supports thread-default contexts.

fn trash<'a, P: Into<Option<&'a Cancellable>>>(
&self,
cancellable: P
) -> Result<(), Error>

Sends self to the "Trashcan", if possible. This is similar to deleting it, but the user can recover it before emptying the trashcan. Not all file systems support trashing, so this call can return the IOErrorEnum::NotSupported error.

If cancellable is not None, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error IOErrorEnum::Cancelled will be returned.

cancellable

optional Cancellable object, None to ignore

Returns

true on successful trash, false otherwise.