Trait gio::FileInfoExt

pub trait FileInfoExt {
    fn clear_status(&self);

    fn copy_into(&self, dest_info: &FileInfo);

    fn dup(&self) -> Option<FileInfo>;

    fn get_attribute_as_string(&self, attribute: &str) -> Option<String>;

    fn get_attribute_boolean(&self, attribute: &str) -> bool;

    fn get_attribute_byte_string(&self, attribute: &str) -> Option<String>;

    fn get_attribute_int32(&self, attribute: &str) -> i32;

    fn get_attribute_int64(&self, attribute: &str) -> i64;

    fn get_attribute_object(&self, attribute: &str) -> Option<Object>;

    fn get_attribute_string(&self, attribute: &str) -> Option<String>;

    fn get_attribute_stringv(&self, attribute: &str) -> Vec<String>;

    fn get_attribute_uint32(&self, attribute: &str) -> u32;

    fn get_attribute_uint64(&self, attribute: &str) -> u64;

    fn get_content_type(&self) -> Option<String>;

    fn get_display_name(&self) -> Option<String>;

    fn get_edit_name(&self) -> Option<String>;

    fn get_etag(&self) -> Option<String>;

    fn get_file_type(&self) -> FileType;

    fn get_icon(&self) -> Option<Icon>;

    fn get_is_backup(&self) -> bool;

    fn get_is_hidden(&self) -> bool;

    fn get_is_symlink(&self) -> bool;

    fn get_name(&self) -> Option<PathBuf>;

    fn get_size(&self) -> i64;

    fn get_sort_order(&self) -> i32;

    fn get_symbolic_icon(&self) -> Option<Icon>;

    fn get_symlink_target(&self) -> Option<String>;

    fn has_attribute(&self, attribute: &str) -> bool;

    fn has_namespace(&self, name_space: &str) -> bool;

    fn list_attributes<'a, P: Into<Option<&'a str>>>(
        &self, 
        name_space: P
    ) -> Vec<String>;

    fn remove_attribute(&self, attribute: &str);

    fn set_attribute_boolean(&self, attribute: &str, attr_value: bool);

    fn set_attribute_byte_string(&self, attribute: &str, attr_value: &str);

    fn set_attribute_int32(&self, attribute: &str, attr_value: i32);

    fn set_attribute_int64(&self, attribute: &str, attr_value: i64);

    fn set_attribute_object<P: IsA<Object>>(
        &self, 
        attribute: &str, 
        attr_value: &P
    );

    fn set_attribute_string(&self, attribute: &str, attr_value: &str);

    fn set_attribute_stringv(&self, attribute: &str, attr_value: &[&str]);

    fn set_attribute_uint32(&self, attribute: &str, attr_value: u32);

    fn set_attribute_uint64(&self, attribute: &str, attr_value: u64);

    fn set_content_type(&self, content_type: &str);

    fn set_display_name(&self, display_name: &str);

    fn set_edit_name(&self, edit_name: &str);

    fn set_file_type(&self, type_: FileType);

    fn set_icon<P: IsA<Icon>>(&self, icon: &P);

    fn set_is_hidden(&self, is_hidden: bool);

    fn set_is_symlink(&self, is_symlink: bool);

    fn set_name<P: AsRef<Path>>(&self, name: P);

    fn set_size(&self, size: i64);

    fn set_sort_order(&self, sort_order: i32);

    fn set_symbolic_icon<P: IsA<Icon>>(&self, icon: &P);

    fn set_symlink_target(&self, symlink_target: &str);

    fn unset_attribute_mask(&self);
}

Trait containing all FileInfo methods.

Implementors

FileInfo

Required Methods

fn clear_status(&self)

Clears the status information from self.

fn copy_into(&self, dest_info: &FileInfo)

First clears all of the [GFileAttribute][gio-GFileAttribute] of dest_info, and then copies all of the file attributes from self to dest_info.

dest_info

destination to copy attributes to.

fn dup(&self) -> Option<FileInfo>

Duplicates a file info structure.

Returns

a duplicate FileInfo of self.

fn get_attribute_as_string(&self, attribute: &str) -> Option<String>

Gets the value of a attribute, formated as a string. This escapes things as needed to make the string valid utf8.

attribute

a file attribute key.

Returns

a UTF-8 string associated with the given attribute. When you're done with the string it must be freed with g_free.

fn get_attribute_boolean(&self, attribute: &str) -> bool

Gets the value of a boolean attribute. If the attribute does not contain a boolean value, false will be returned.

attribute

a file attribute key.

Returns

the boolean value contained within the attribute.

fn get_attribute_byte_string(&self, attribute: &str) -> Option<String>

Gets the value of a byte string attribute. If the attribute does not contain a byte string, None will be returned.

attribute

a file attribute key.

Returns

the contents of the attribute value as a byte string, or None otherwise.

fn get_attribute_int32(&self, attribute: &str) -> i32

Gets a signed 32-bit integer contained within the attribute. If the attribute does not contain a signed 32-bit integer, or is invalid, 0 will be returned.

attribute

a file attribute key.

Returns

a signed 32-bit integer from the attribute.

fn get_attribute_int64(&self, attribute: &str) -> i64

Gets a signed 64-bit integer contained within the attribute. If the attribute does not contain an signed 64-bit integer, or is invalid, 0 will be returned.

attribute

a file attribute key.

Returns

a signed 64-bit integer from the attribute.

fn get_attribute_object(&self, attribute: &str) -> Option<Object>

Gets the value of a gobject::Object attribute. If the attribute does not contain a gobject::Object, None will be returned.

attribute

a file attribute key.

Returns

a gobject::Object associated with the given attribute, or None otherwise.

fn get_attribute_string(&self, attribute: &str) -> Option<String>

Gets the value of a string attribute. If the attribute does not contain a string, None will be returned.

attribute

a file attribute key.

Returns

the contents of the attribute value as a UTF-8 string, or None otherwise.

fn get_attribute_stringv(&self, attribute: &str) -> Vec<String>

Gets the value of a stringv attribute. If the attribute does not contain a stringv, None will be returned.

attribute

a file attribute key.

Returns

the contents of the attribute value as a stringv, or None otherwise. Do not free. These returned strings are UTF-8.

fn get_attribute_uint32(&self, attribute: &str) -> u32

Gets an unsigned 32-bit integer contained within the attribute. If the attribute does not contain an unsigned 32-bit integer, or is invalid, 0 will be returned.

attribute

a file attribute key.

Returns

an unsigned 32-bit integer from the attribute.

fn get_attribute_uint64(&self, attribute: &str) -> u64

Gets a unsigned 64-bit integer contained within the attribute. If the attribute does not contain an unsigned 64-bit integer, or is invalid, 0 will be returned.

attribute

a file attribute key.

Returns

a unsigned 64-bit integer from the attribute.

fn get_content_type(&self) -> Option<String>

Gets the file's content type.

Returns

a string containing the file's content type.

fn get_display_name(&self) -> Option<String>

Gets a display name for a file.

Returns

a string containing the display name.

fn get_edit_name(&self) -> Option<String>

Gets the edit name for a file.

Returns

a string containing the edit name.

fn get_etag(&self) -> Option<String>

Gets the [entity tag][gfile-etag] for a given FileInfo. See G_FILE_ATTRIBUTE_ETAG_VALUE.

Returns

a string containing the value of the "etag:value" attribute.

fn get_file_type(&self) -> FileType

Gets a file's type (whether it is a regular file, symlink, etc). This is different from the file's content type, see FileInfoExt::get_content_type.

Returns

a FileType for the given file.

fn get_icon(&self) -> Option<Icon>

Gets the icon for a file.

Returns

Icon for the given self.

fn get_is_backup(&self) -> bool

Checks if a file is a backup file.

Returns

true if file is a backup file, false otherwise.

fn get_is_hidden(&self) -> bool

Checks if a file is hidden.

Returns

true if the file is a hidden file, false otherwise.

fn get_is_symlink(&self) -> bool

Checks if a file is a symlink.

Returns

true if the given self is a symlink.

fn get_name(&self) -> Option<PathBuf>

Gets the name for a file.

Returns

a string containing the file name.

fn get_size(&self) -> i64

Gets the file's size.

Returns

a goffset containing the file's size.

fn get_sort_order(&self) -> i32

Gets the value of the sort_order attribute from the FileInfo. See G_FILE_ATTRIBUTE_STANDARD_SORT_ORDER.

Returns

a gint32 containing the value of the "standard::sort_order" attribute.

fn get_symbolic_icon(&self) -> Option<Icon>

Gets the symbolic icon for a file.

Feature: v2_34

Returns

Icon for the given self.

fn get_symlink_target(&self) -> Option<String>

Gets the symlink target for a given FileInfo.

Returns

a string containing the symlink target.

fn has_attribute(&self, attribute: &str) -> bool

Checks if a file info structure has an attribute named attribute.

attribute

a file attribute key.

Returns

true if Ginfo has an attribute named attribute, false otherwise.

fn has_namespace(&self, name_space: &str) -> bool

Checks if a file info structure has an attribute in the specified name_space.

name_space

a file attribute namespace.

Returns

true if Ginfo has an attribute in name_space, false otherwise.

fn list_attributes<'a, P: Into<Option<&'a str>>>(
&self,
name_space: P
) -> Vec<String>

Lists the file info structure's attributes.

name_space

a file attribute key's namespace, or None to list all attributes.

Returns

a null-terminated array of strings of all of the possible attribute types for the given name_space, or None on error.

fn remove_attribute(&self, attribute: &str)

Removes all cases of attribute from self if it exists.

attribute

a file attribute key.

fn set_attribute_boolean(&self, attribute: &str, attr_value: bool)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a boolean value.

fn set_attribute_byte_string(&self, attribute: &str, attr_value: &str)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a byte string.

fn set_attribute_int32(&self, attribute: &str, attr_value: i32)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a signed 32-bit integer

fn set_attribute_int64(&self, attribute: &str, attr_value: i64)

Sets the attribute to contain the given attr_value, if possible.

attribute

attribute name to set.

attr_value

int64 value to set attribute to.

fn set_attribute_object<P: IsA<Object>>(&self, attribute: &str, attr_value: &P)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a gobject::Object.

fn set_attribute_string(&self, attribute: &str, attr_value: &str)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

a UTF-8 string.

fn set_attribute_stringv(&self, attribute: &str, attr_value: &[&str])

Sets the attribute to contain the given attr_value, if possible.

Sinze: 2.22

attribute

a file attribute key

attr_value

a None terminated array of UTF-8 strings.

fn set_attribute_uint32(&self, attribute: &str, attr_value: u32)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

an unsigned 32-bit integer.

fn set_attribute_uint64(&self, attribute: &str, attr_value: u64)

Sets the attribute to contain the given attr_value, if possible.

attribute

a file attribute key.

attr_value

an unsigned 64-bit integer.

fn set_content_type(&self, content_type: &str)

Sets the content type attribute for a given FileInfo. See G_FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE.

content_type

a content type. See [GContentType][gio-GContentType]

fn set_display_name(&self, display_name: &str)

Sets the display name for the current FileInfo. See G_FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME.

display_name

a string containing a display name.

fn set_edit_name(&self, edit_name: &str)

Sets the edit name for the current file. See G_FILE_ATTRIBUTE_STANDARD_EDIT_NAME.

edit_name

a string containing an edit name.

fn set_file_type(&self, type_: FileType)

Sets the file type in a FileInfo to type_. See G_FILE_ATTRIBUTE_STANDARD_TYPE.

type_

a FileType.

fn set_icon<P: IsA<Icon>>(&self, icon: &P)

Sets the icon for a given FileInfo. See G_FILE_ATTRIBUTE_STANDARD_ICON.

icon

a Icon.

fn set_is_hidden(&self, is_hidden: bool)

Sets the "is_hidden" attribute in a FileInfo according to is_hidden. See G_FILE_ATTRIBUTE_STANDARD_IS_HIDDEN.

is_hidden

a gboolean.

fn set_is_symlink(&self, is_symlink: bool)

Sets the "is_symlink" attribute in a FileInfo according to is_symlink. See G_FILE_ATTRIBUTE_STANDARD_IS_SYMLINK.

is_symlink

a gboolean.

fn set_name<P: AsRef<Path>>(&self, name: P)

Sets the name attribute for the current FileInfo. See G_FILE_ATTRIBUTE_STANDARD_NAME.

name

a string containing a name.

fn set_size(&self, size: i64)

Sets the G_FILE_ATTRIBUTE_STANDARD_SIZE attribute in the file info to the given size.

size

a goffset containing the file's size.

fn set_sort_order(&self, sort_order: i32)

Sets the sort order attribute in the file info structure. See G_FILE_ATTRIBUTE_STANDARD_SORT_ORDER.

sort_order

a sort order integer.

fn set_symbolic_icon<P: IsA<Icon>>(&self, icon: &P)

Sets the symbolic icon for a given FileInfo. See G_FILE_ATTRIBUTE_STANDARD_SYMBOLIC_ICON.

Feature: v2_34

icon

a Icon.

fn set_symlink_target(&self, symlink_target: &str)

Sets the G_FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET attribute in the file info to the given symlink target.

symlink_target

a static string containing a path to a symlink target.

fn unset_attribute_mask(&self)

Unsets a mask set by FileInfoExt::set_attribute_mask, if one is set.