Trait gio::FileMonitorExt

pub trait FileMonitorExt: 'static {
    fn cancel(&self) -> bool;

    fn emit_event<P: IsA<File>, Q: IsA<File>>(
        &self, 
        child: &P, 
        other_file: &Q, 
        event_type: FileMonitorEvent
    );

    fn is_cancelled(&self) -> bool;

    fn set_rate_limit(&self, limit_msecs: i32);

    fn get_property_cancelled(&self) -> bool;

    fn get_property_rate_limit(&self) -> i32;

    fn connect_changed<F: Fn(&Self, &File, Option<&File>, FileMonitorEvent) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_cancelled_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_rate_limit_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all FileMonitor methods.

Implementors

FileMonitor

Required methods

fn cancel(&self) -> bool

Cancels a file monitor.

Returns

always true

fn emit_event<P: IsA<File>, Q: IsA<File>>(
&self,
child: &P,
other_file: &Q,
event_type: FileMonitorEvent
)

Emits the FileMonitor::changed signal if a change has taken place. Should be called from file monitor implementations only.

Implementations are responsible to call this method from the [thread-default main context][g-main-context-push-thread-default] of the thread that the monitor was created in.

child

a File.

other_file

a File.

event_type

a set of FileMonitorEvent flags.

fn is_cancelled(&self) -> bool

Returns whether the monitor is canceled.

Returns

true if monitor is canceled. false otherwise.

fn set_rate_limit(&self, limit_msecs: i32)

Sets the rate limit to which the self will report consecutive change events to the same file.

limit_msecs

a non-negative integer with the limit in milliseconds to poll for changes

fn get_property_cancelled(&self) -> bool

fn get_property_rate_limit(&self) -> i32

fn connect_changed<F: Fn(&Self, &File, Option<&File>, FileMonitorEvent) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_cancelled_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_rate_limit_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...