Trait gio::FilterInputStreamExt

pub trait FilterInputStreamExt: 'static {
    fn get_base_stream(&self) -> Option<InputStream>;

    fn get_close_base_stream(&self) -> bool;

    fn set_close_base_stream(&self, close_base: bool);

    fn connect_property_close_base_stream_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all FilterInputStream methods.

Implementors

BufferedInputStream, ConverterInputStream, FilterInputStream

Required methods

fn get_base_stream(&self) -> Option<InputStream>

Gets the base stream for the filter stream.

Returns

a InputStream.

fn get_close_base_stream(&self) -> bool

Returns whether the base stream will be closed when self is closed.

Returns

true if the base stream will be closed.

fn set_close_base_stream(&self, close_base: bool)

Sets whether the base stream will be closed when self is closed.

close_base

true to close the base stream.

fn connect_property_close_base_stream_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...