Trait gio::IconExt

pub trait IconExt {
    fn equal<'a, P: IsA<Icon> + 'a, Q: Into<Option<&'a P>>>(
        &self, 
        icon2: Q
    ) -> bool;

    fn serialize(&self) -> Option<Variant>;

    fn to_string(&self) -> Option<String>;
}

Trait containing all Icon methods.

Implementors

Icon, ThemedIcon

Required Methods

fn equal<'a, P: IsA<Icon> + 'a, Q: Into<Option<&'a P>>>(&self, icon2: Q) -> bool

Checks if two icons are equal.

icon2

pointer to the second Icon.

Returns

true if self is equal to icon2. false otherwise.

fn serialize(&self) -> Option<Variant>

Serializes a Icon into a glib::Variant. An equivalent Icon can be retrieved back by calling Icon::deserialize on the returned value. As serialization will avoid using raw icon data when possible, it only makes sense to transfer the glib::Variant between processes on the same machine, (as opposed to over the network), and within the same file system namespace.

Feature: v2_38

Returns

a glib::Variant, or None when serialization fails.

fn to_string(&self) -> Option<String>

Generates a textual representation of self that can be used for serialization such as when passing self to a different process or saving it to persistent storage. Use Icon::new_for_string to get self back from the returned string.

The encoding of the returned string is proprietary to Icon except in the following two cases

If self is a FileIcon, the returned string is a native path (such as /path/to/my icon.png) without escaping if the File for self is a native file. If the file is not native, the returned string is the result of File::get_uri (such as sftp://path/to/my%20icon.png).

If self is a ThemedIcon with exactly one name, the encoding is simply the name (such as network-server).

Returns

An allocated NUL-terminated UTF8 string or None if self can't be serialized. Use g_free to free.