Trait gio::InetAddressMaskExt

pub trait InetAddressMaskExt: 'static {
    fn equal<P: IsA<InetAddressMask>>(&self, mask2: &P) -> bool;

    fn get_address(&self) -> InetAddress;

    fn get_family(&self) -> SocketFamily;

    fn get_length(&self) -> u32;

    fn matches<P: IsA<InetAddress>>(&self, address: &P) -> bool;

    fn to_string(&self) -> GString;

    fn set_property_address<P: IsA<InetAddress> + SetValueOptional>(
        &self, 
        address: Option<&P>
    );

    fn set_property_length(&self, length: u32);

    fn connect_property_address_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_family_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_length_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all InetAddressMask methods.

Implementors

InetAddressMask

Required methods

fn equal<P: IsA<InetAddressMask>>(&self, mask2: &P) -> bool

Tests if self and mask2 are the same mask.

mask2

another InetAddressMask

Returns

whether self and mask2 are the same mask

fn get_address(&self) -> InetAddress

Gets self's base address

Returns

self's base address

fn get_family(&self) -> SocketFamily

Gets the SocketFamily of self's address

Returns

the SocketFamily of self's address

fn get_length(&self) -> u32

Gets self's length

Returns

self's length

fn matches<P: IsA<InetAddress>>(&self, address: &P) -> bool

Tests if address falls within the range described by self.

address

a InetAddress

Returns

whether address falls within the range described by self.

fn to_string(&self) -> GString

Converts self back to its corresponding string form.

Returns

a string corresponding to self.

fn set_property_address<P: IsA<InetAddress> + SetValueOptional>(
&self,
address: Option<&P>
)

fn set_property_length(&self, length: u32)

fn connect_property_address_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_family_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_length_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...