Trait gio::InetSocketAddressExt

pub trait InetSocketAddressExt: 'static {
    fn get_address(&self) -> Option<InetAddress>;

    fn get_flowinfo(&self) -> u32;

    fn get_port(&self) -> u16;

    fn get_scope_id(&self) -> u32;
}

Trait containing all InetSocketAddress methods.

Implementors

InetSocketAddress, ProxyAddress

Required methods

fn get_address(&self) -> Option<InetAddress>

Gets self's InetAddress.

Returns

the InetAddress for self, which must be gobject::ObjectExt::ref'd if it will be stored

fn get_flowinfo(&self) -> u32

Gets the sin6_flowinfo field from self, which must be an IPv6 address.

Returns

the flowinfo field

fn get_port(&self) -> u16

Gets self's port.

Returns

the port for self

fn get_scope_id(&self) -> u32

Gets the sin6_scope_id field from self, which must be an IPv6 address.

Returns

the scope id field

Loading content...