Trait gio::MenuAttributeIterExt

pub trait MenuAttributeIterExt: 'static {
    fn get_name(&self) -> Option<GString>;

    fn get_next(&self) -> Option<(GString, Variant)>;

    fn get_value(&self) -> Option<Variant>;

    fn next(&self) -> bool;
}

Trait containing all MenuAttributeIter methods.

Implementors

MenuAttributeIter

Required methods

fn get_name(&self) -> Option<GString>

Gets the name of the attribute at the current iterator position, as a string.

The iterator is not advanced.

Returns

the name of the attribute

fn get_next(&self) -> Option<(GString, Variant)>

This function combines MenuAttributeIterExt::next with MenuAttributeIterExt::get_name and MenuAttributeIterExt::get_value.

First the iterator is advanced to the next (possibly first) attribute. If that fails, then false is returned and there are no other effects.

If successful, name and value are set to the name and value of the attribute that has just been advanced to. At this point, MenuAttributeIterExt::get_name and MenuAttributeIterExt::get_value will return the same values again.

The value returned in name remains valid for as long as the iterator remains at the current position. The value returned in value must be unreffed using glib::Variant::unref when it is no longer in use.

out_name

the type of the attribute

value

the attribute value

Returns

true on success, or false if there is no additional attribute

fn get_value(&self) -> Option<Variant>

Gets the value of the attribute at the current iterator position.

The iterator is not advanced.

Returns

the value of the current attribute

fn next(&self) -> bool

Attempts to advance the iterator to the next (possibly first) attribute.

true is returned on success, or false if there are no more attributes.

You must call this function when you first acquire the iterator to advance it to the first attribute (and determine if the first attribute exists at all).

Returns

true on success, or false when there are no more attributes

Loading content...