Trait gio::MenuExt

pub trait MenuExt {
    fn append<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
        &self, 
        label: P, 
        detailed_action: Q
    );

    fn append_item(&self, item: &MenuItem);

    fn append_section<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
        &self, 
        label: P, 
        section: &Q
    );

    fn append_submenu<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
        &self, 
        label: P, 
        submenu: &Q
    );

    fn freeze(&self);

    fn insert<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
        &self, 
        position: i32, 
        label: P, 
        detailed_action: Q
    );

    fn insert_item(&self, position: i32, item: &MenuItem);

    fn insert_section<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
        &self, 
        position: i32, 
        label: P, 
        section: &Q
    );

    fn insert_submenu<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
        &self, 
        position: i32, 
        label: P, 
        submenu: &Q
    );

    fn prepend<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
        &self, 
        label: P, 
        detailed_action: Q
    );

    fn prepend_item(&self, item: &MenuItem);

    fn prepend_section<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
        &self, 
        label: P, 
        section: &Q
    );

    fn prepend_submenu<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
        &self, 
        label: P, 
        submenu: &Q
    );

    fn remove(&self, position: i32);

    fn remove_all(&self);
}

Trait containing all Menu methods.

Implementors

Menu

Required Methods

fn append<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
&self,
label: P,
detailed_action: Q
)

Convenience function for appending a normal menu item to the end of self. Combine MenuItem::new and MenuExt::insert_item for a more flexible alternative.

label

the section label, or None

detailed_action

the detailed action string, or None

fn append_item(&self, item: &MenuItem)

Appends item to the end of self.

See MenuExt::insert_item for more information.

item

a MenuItem to append

fn append_section<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
&self,
label: P,
section: &Q
)

Convenience function for appending a section menu item to the end of self. Combine MenuItem::new_section and MenuExt::insert_item for a more flexible alternative.

label

the section label, or None

section

a MenuModel with the items of the section

fn append_submenu<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
&self,
label: P,
submenu: &Q
)

Convenience function for appending a submenu menu item to the end of self. Combine MenuItem::new_submenu and MenuExt::insert_item for a more flexible alternative.

label

the section label, or None

submenu

a MenuModel with the items of the submenu

fn freeze(&self)

Marks self as frozen.

After the menu is frozen, it is an error to attempt to make any changes to it. In effect this means that the Menu API must no longer be used.

This function causes MenuModelExt::is_mutable to begin returning false, which has some positive performance implications.

fn insert<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
&self,
position: i32,
label: P,
detailed_action: Q
)

Convenience function for inserting a normal menu item into self. Combine MenuItem::new and MenuExt::insert_item for a more flexible alternative.

position

the position at which to insert the item

label

the section label, or None

detailed_action

the detailed action string, or None

fn insert_item(&self, position: i32, item: &MenuItem)

Inserts item into self.

The "insertion" is actually done by copying all of the attribute and link values of item and using them to form a new item within self. As such, item itself is not really inserted, but rather, a menu item that is exactly the same as the one presently described by item.

This means that item is essentially useless after the insertion occurs. Any changes you make to it are ignored unless it is inserted again (at which point its updated values will be copied).

You should probably just free item once you're done.

There are many convenience functions to take care of common cases. See MenuExt::insert, MenuExt::insert_section and MenuExt::insert_submenu as well as "prepend" and "append" variants of each of these functions.

position

the position at which to insert the item

item

the MenuItem to insert

fn insert_section<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
&self,
position: i32,
label: P,
section: &Q
)

Convenience function for inserting a section menu item into self. Combine MenuItem::new_section and MenuExt::insert_item for a more flexible alternative.

position

the position at which to insert the item

label

the section label, or None

section

a MenuModel with the items of the section

fn insert_submenu<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
&self,
position: i32,
label: P,
submenu: &Q
)

Convenience function for inserting a submenu menu item into self. Combine MenuItem::new_submenu and MenuExt::insert_item for a more flexible alternative.

position

the position at which to insert the item

label

the section label, or None

submenu

a MenuModel with the items of the submenu

fn prepend<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
&self,
label: P,
detailed_action: Q
)

Convenience function for prepending a normal menu item to the start of self. Combine MenuItem::new and MenuExt::insert_item for a more flexible alternative.

label

the section label, or None

detailed_action

the detailed action string, or None

fn prepend_item(&self, item: &MenuItem)

Prepends item to the start of self.

See MenuExt::insert_item for more information.

item

a MenuItem to prepend

fn prepend_section<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
&self,
label: P,
section: &Q
)

Convenience function for prepending a section menu item to the start of self. Combine MenuItem::new_section and MenuExt::insert_item for a more flexible alternative.

label

the section label, or None

section

a MenuModel with the items of the section

fn prepend_submenu<'a, P: Into<Option<&'a str>>, Q: IsA<MenuModel>>(
&self,
label: P,
submenu: &Q
)

Convenience function for prepending a submenu menu item to the start of self. Combine MenuItem::new_submenu and MenuExt::insert_item for a more flexible alternative.

label

the section label, or None

submenu

a MenuModel with the items of the submenu

fn remove(&self, position: i32)

Removes an item from the menu.

position gives the index of the item to remove.

It is an error if position is not in range the range from 0 to one less than the number of items in the menu.

It is not possible to remove items by identity since items are added to the menu simply by copying their links and attributes (ie: identity of the item itself is not preserved).

position

the position of the item to remove

fn remove_all(&self)

Removes all items in the menu.

Feature: v2_38