Trait gio::MenuItemExt

pub trait MenuItemExt {
    fn get_attribute_value<'a, P: Into<Option<&'a VariantTy>>>(
        &self, 
        attribute: &str, 
        expected_type: P
    ) -> Option<Variant>;

    fn get_link(&self, link: &str) -> Option<MenuModel>;

    fn set_action_and_target_value<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b Variant>>>(
        &self, 
        action: P, 
        target_value: Q
    );

    fn set_attribute_value<'a, P: Into<Option<&'a Variant>>>(
        &self, 
        attribute: &str, 
        value: P
    );

    fn set_detailed_action(&self, detailed_action: &str);

    fn set_icon<P: IsA<Icon>>(&self, icon: &P);

    fn set_label<'a, P: Into<Option<&'a str>>>(&self, label: P);

    fn set_link<'a, P: IsA<MenuModel> + 'a, Q: Into<Option<&'a P>>>(
        &self, 
        link: &str, 
        model: Q
    );

    fn set_section<'a, P: IsA<MenuModel> + 'a, Q: Into<Option<&'a P>>>(
        &self, 
        section: Q
    );

    fn set_submenu<'a, P: IsA<MenuModel> + 'a, Q: Into<Option<&'a P>>>(
        &self, 
        submenu: Q
    );
}

Trait containing all MenuItem methods.

Implementors

MenuItem

Required Methods

fn get_attribute_value<'a, P: Into<Option<&'a VariantTy>>>(
&self,
attribute: &str,
expected_type: P
) -> Option<Variant>

Queries the named attribute on self.

If expected_type is specified and the attribute does not have this type, None is returned. None is also returned if the attribute simply does not exist.

Feature: v2_34

attribute

the attribute name to query

expected_type

the expected type of the attribute

Returns

the attribute value, or None

fn get_link(&self, link: &str) -> Option<MenuModel>

Queries the named link on self.

Feature: v2_34

link

the link name to query

Returns

the link, or None

fn set_action_and_target_value<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b Variant>>>(
&self,
action: P,
target_value: Q
)

Sets or unsets the "action" and "target" attributes of self.

If action is None then both the "action" and "target" attributes are unset (and target_value is ignored).

If action is non-None then the "action" attribute is set. The "target" attribute is then set to the value of target_value if it is non-None or unset otherwise.

Normal menu items (ie: not submenu, section or other custom item types) are expected to have the "action" attribute set to identify the action that they are associated with. The state type of the action help to determine the disposition of the menu item. See Action and ActionGroup for an overview of actions.

In general, clicking on the menu item will result in activation of the named action with the "target" attribute given as the parameter to the action invocation. If the "target" attribute is not set then the action is invoked with no parameter.

If the action has no state then the menu item is usually drawn as a plain menu item (ie: with no additional decoration).

If the action has a boolean state then the menu item is usually drawn as a toggle menu item (ie: with a checkmark or equivalent indication). The item should be marked as 'toggled' or 'checked' when the boolean state is true.

If the action has a string state then the menu item is usually drawn as a radio menu item (ie: with a radio bullet or equivalent indication). The item should be marked as 'selected' when the string state is equal to the value of the target property.

See MenuItemExt::set_action_and_target or MenuItemExt::set_detailed_action for two equivalent calls that are probably more convenient for most uses.

action

the name of the action for this item

target_value

a glib::Variant to use as the action target

fn set_attribute_value<'a, P: Into<Option<&'a Variant>>>(
&self,
attribute: &str,
value: P
)

Sets or unsets an attribute on self.

The attribute to set or unset is specified by attribute. This can be one of the standard attribute names G_MENU_ATTRIBUTE_LABEL, G_MENU_ATTRIBUTE_ACTION, G_MENU_ATTRIBUTE_TARGET, or a custom attribute name. Attribute names are restricted to lowercase characters, numbers and '-'. Furthermore, the names must begin with a lowercase character, must not end with a '-', and must not contain consecutive dashes.

must consist only of lowercase ASCII characters, digits and '-'.

If value is non-None then it is used as the new value for the attribute. If value is None then the attribute is unset. If the value glib::Variant is floating, it is consumed.

See also MenuItemExt::set_attribute for a more convenient way to do the same.

attribute

the attribute to set

value

a glib::Variant to use as the value, or None

fn set_detailed_action(&self, detailed_action: &str)

Sets the "action" and possibly the "target" attribute of self.

The format of detailed_action is the same format parsed by Action::parse_detailed_name.

See MenuItemExt::set_action_and_target or MenuItemExt::set_action_and_target_value for more flexible (but slightly less convenient) alternatives.

See also MenuItemExt::set_action_and_target_value for a description of the semantics of the action and target attributes.

detailed_action

the "detailed" action string

fn set_icon<P: IsA<Icon>>(&self, icon: &P)

Sets (or unsets) the icon on self.

This call is the same as calling Icon::serialize and using the result as the value to MenuItemExt::set_attribute_value for G_MENU_ATTRIBUTE_ICON.

This API is only intended for use with "noun" menu items; things like bookmarks or applications in an "Open With" menu. Don't use it on menu items corresponding to verbs (eg: stock icons for 'Save' or 'Quit').

If icon is None then the icon is unset.

Feature: v2_38

icon

a Icon, or None

fn set_label<'a, P: Into<Option<&'a str>>>(&self, label: P)

Sets or unsets the "label" attribute of self.

If label is non-None it is used as the label for the menu item. If it is None then the label attribute is unset.

label

the label to set, or None to unset

fn set_link<'a, P: IsA<MenuModel> + 'a, Q: Into<Option<&'a P>>>(
&self,
link: &str,
model: Q
)

Creates a link from self to model if non-None, or unsets it.

Links are used to establish a relationship between a particular menu item and another menu. For example, G_MENU_LINK_SUBMENU is used to associate a submenu with a particular menu item, and G_MENU_LINK_SECTION is used to create a section. Other types of link can be used, but there is no guarantee that clients will be able to make sense of them. Link types are restricted to lowercase characters, numbers and '-'. Furthermore, the names must begin with a lowercase character, must not end with a '-', and must not contain consecutive dashes.

link

type of link to establish or unset

model

the MenuModel to link to (or None to unset)

fn set_section<'a, P: IsA<MenuModel> + 'a, Q: Into<Option<&'a P>>>(
&self,
section: Q
)

Sets or unsets the "section" link of self to section.

The effect of having one menu appear as a section of another is exactly as it sounds: the items from section become a direct part of the menu that self is added to. See MenuItem::new_section for more information about what it means for a menu item to be a section.

section

a MenuModel, or None

fn set_submenu<'a, P: IsA<MenuModel> + 'a, Q: Into<Option<&'a P>>>(
&self,
submenu: Q
)

Sets or unsets the "submenu" link of self to submenu.

If submenu is non-None, it is linked to. If it is None then the link is unset.

The effect of having one menu appear as a submenu of another is exactly as it sounds.

submenu

a MenuModel, or None