Trait gio::SettingsExt

pub trait SettingsExt: 'static {

    fn apply(&self);

    fn bind_writable<P: IsA<Object>>(
        &self, 
        key: &str, 
        object: &P, 
        property: &str, 
        inverted: bool
    );

    fn create_action(&self, key: &str) -> Action;

    fn delay(&self);

    fn boolean(&self, key: &str) -> bool;

    fn child(&self, name: &str) -> Settings;

    fn default_value(&self, key: &str) -> Option<Variant>;

    fn double(&self, key: &str) -> f64;

    fn enum_(&self, key: &str) -> i32;

    fn flags(&self, key: &str) -> u32;

    fn has_unapplied(&self) -> bool;

    fn int(&self, key: &str) -> i32;

    fn int64(&self, key: &str) -> i64;

    fn string(&self, key: &str) -> GString;

    fn strv(&self, key: &str) -> Vec<GString>;

    fn uint(&self, key: &str) -> u32;

    fn uint64(&self, key: &str) -> u64;

    fn user_value(&self, key: &str) -> Option<Variant>;

    fn value(&self, key: &str) -> Variant;

    fn is_writable(&self, name: &str) -> bool;

    fn list_children(&self) -> Vec<GString>;

    fn list_keys(&self) -> Vec<GString>;

    fn reset(&self, key: &str);

    fn revert(&self);

    fn set_boolean(&self, key: &str, value: bool) -> Result<(), BoolError>;

    fn set_double(&self, key: &str, value: f64) -> Result<(), BoolError>;

    fn set_enum(&self, key: &str, value: i32) -> Result<(), BoolError>;

    fn set_flags(&self, key: &str, value: u32) -> Result<(), BoolError>;

    fn set_int(&self, key: &str, value: i32) -> Result<(), BoolError>;

    fn set_int64(&self, key: &str, value: i64) -> Result<(), BoolError>;

    fn set_string(&self, key: &str, value: &str) -> Result<(), BoolError>;

    fn set_strv(&self, key: &str, value: &[&str]) -> Result<(), BoolError>;

    fn set_uint(&self, key: &str, value: u32) -> Result<(), BoolError>;

    fn set_uint64(&self, key: &str, value: u64) -> Result<(), BoolError>;

    fn set_value(&self, key: &str, value: &Variant) -> Result<(), BoolError>;

    fn backend(&self) -> Option<SettingsBackend>;

    fn is_delay_apply(&self) -> bool;

    fn path(&self) -> Option<GString>;

    fn schema_id(&self) -> Option<GString>;

    fn settings_schema(&self) -> Option<SettingsSchema>;

    fn connect_changed<F: Fn(&Self, &str) + 'static>(
        &self, 
        detail: Option<&str>, 
        f: F
    ) -> SignalHandlerId;

    fn connect_writable_change_event<F: Fn(&Self, u32) -> Inhibit + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_writable_changed<F: Fn(&Self, &str) + 'static>(
        &self, 
        detail: Option<&str>, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_delay_apply_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_has_unapplied_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

}

Trait containing all Settings methods.

Implementors

Settings

Required methods

fn apply(&self)

Applies any changes that have been made to the settings. This function does nothing unless self is in ‘delay-apply’ mode; see SettingsExt::delay. In the normal case settings are always applied immediately.

fn bind_writable<P: IsA<Object>>(
&self,
key: &str,
object: &P,
property: &str,
inverted: bool
)

Create a binding between the writability of key in the self object and the property property of object. The property must be boolean; “sensitive” or “visible” properties of widgets are the most likely candidates.

Writable bindings are always uni-directional; changes of the writability of the setting will be propagated to the object property, not the other way.

When the inverted argument is true, the binding inverts the value as it passes from the setting to the object, i.e. property will be set to true if the key is not writable.

Note that the lifecycle of the binding is tied to object, and that you can have only one binding per object property. If you bind the same property twice on the same object, the second binding overrides the first one.

key

the key to bind

object

a gobject::Object

property

the name of a boolean property to bind

inverted

whether to ‘invert’ the value

fn create_action(&self, key: &str) -> Action

Creates a Action corresponding to a given Settings key.

The action has the same name as the key.

The value of the key becomes the state of the action and the action is enabled when the key is writable. Changing the state of the action results in the key being written to. Changes to the value or writability of the key cause appropriate change notifications to be emitted for the action.

For boolean-valued keys, action activations take no parameter and result in the toggling of the value. For all other types, activations take the new value for the key (which must have the correct type).

key

the name of a key in self

Returns

a new Action

fn delay(&self)

Changes the Settings object into ‘delay-apply’ mode. In this mode, changes to self are not immediately propagated to the backend, but kept locally until SettingsExt::apply is called.

fn boolean(&self, key: &str) -> bool

fn child(&self, name: &str) -> Settings

fn default_value(&self, key: &str) -> Option<Variant>

fn double(&self, key: &str) -> f64

fn enum_(&self, key: &str) -> i32

fn flags(&self, key: &str) -> u32

fn has_unapplied(&self) -> bool

fn int(&self, key: &str) -> i32

fn int64(&self, key: &str) -> i64

This is supported on crate feature v2_50 only.

fn string(&self, key: &str) -> GString

fn strv(&self, key: &str) -> Vec<GString>

fn uint(&self, key: &str) -> u32

fn uint64(&self, key: &str) -> u64

This is supported on crate feature v2_50 only.

fn user_value(&self, key: &str) -> Option<Variant>

fn value(&self, key: &str) -> Variant

fn is_writable(&self, name: &str) -> bool

Finds out if a key can be written or not

name

the name of a key

Returns

true if the key name is writable

fn list_children(&self) -> Vec<GString>

Gets the list of children on self.

The list is exactly the list of strings for which it is not an error to call SettingsExt::get_child.

There is little reason to call this function from “normal” code, since you should already know what children are in your schema. This function may still be useful there for introspection reasons, however.

You should free the return value with g_strfreev when you are done with it.

Returns

a list of the children on self, in no defined order

fn list_keys(&self) -> Vec<GString>

Introspects the list of keys on self.

You should probably not be calling this function from “normal” code (since you should already know what keys are in your schema). This function is intended for introspection reasons.

You should free the return value with g_strfreev when you are done with it.

Deprecated since 2.46

Use SettingsSchema::list_keys instead.

Returns

a list of the keys on self, in no defined order

fn reset(&self, key: &str)

Resets key to its default value.

This call resets the key, as much as possible, to its default value. That might be the value specified in the schema or the one set by the administrator.

key

the name of a key

fn revert(&self)

Reverts all non-applied changes to the settings. This function does nothing unless self is in ‘delay-apply’ mode; see SettingsExt::delay. In the normal case settings are always applied immediately.

Change notifications will be emitted for affected keys.

fn set_boolean(&self, key: &str, value: bool) -> Result<(), BoolError>

Sets key in self to value.

A convenience variant of SettingsExt::set for booleans.

It is a programmer error to give a key that isn’t specified as having a boolean type in the schema for self.

key

the name of the key to set

value

the value to set it to

Returns

true if setting the key succeeded, false if the key was not writable

fn set_double(&self, key: &str, value: f64) -> Result<(), BoolError>

Sets key in self to value.

A convenience variant of SettingsExt::set for doubles.

It is a programmer error to give a key that isn’t specified as having a ‘double’ type in the schema for self.

key

the name of the key to set

value

the value to set it to

Returns

true if setting the key succeeded, false if the key was not writable

fn set_enum(&self, key: &str, value: i32) -> Result<(), BoolError>

Looks up the enumerated type nick for value and writes it to key, within self.

It is a programmer error to give a key that isn’t contained in the schema for self or is not marked as an enumerated type, or for value not to be a valid value for the named type.

After performing the write, accessing key directly with SettingsExt::get_string will return the ‘nick’ associated with value.

key

a key, within self

value

an enumerated value

Returns

true, if the set succeeds

fn set_flags(&self, key: &str, value: u32) -> Result<(), BoolError>

Looks up the flags type nicks for the bits specified by value, puts them in an array of strings and writes the array to key, within self.

It is a programmer error to give a key that isn’t contained in the schema for self or is not marked as a flags type, or for value to contain any bits that are not value for the named type.

After performing the write, accessing key directly with SettingsExt::get_strv will return an array of ‘nicks’; one for each bit in value.

key

a key, within self

value

a flags value

Returns

true, if the set succeeds

fn set_int(&self, key: &str, value: i32) -> Result<(), BoolError>

Sets key in self to value.

A convenience variant of SettingsExt::set for 32-bit integers.

It is a programmer error to give a key that isn’t specified as having a int32 type in the schema for self.

key

the name of the key to set

value

the value to set it to

Returns

true if setting the key succeeded, false if the key was not writable

fn set_int64(&self, key: &str, value: i64) -> Result<(), BoolError>

This is supported on crate feature v2_50 only.

Sets key in self to value.

A convenience variant of SettingsExt::set for 64-bit integers.

It is a programmer error to give a key that isn’t specified as having a int64 type in the schema for self.

Feature: v2_50

key

the name of the key to set

value

the value to set it to

Returns

true if setting the key succeeded, false if the key was not writable

fn set_string(&self, key: &str, value: &str) -> Result<(), BoolError>

Sets key in self to value.

A convenience variant of SettingsExt::set for strings.

It is a programmer error to give a key that isn’t specified as having a string type in the schema for self.

key

the name of the key to set

value

the value to set it to

Returns

true if setting the key succeeded, false if the key was not writable

fn set_strv(&self, key: &str, value: &[&str]) -> Result<(), BoolError>

Sets key in self to value.

A convenience variant of SettingsExt::set for string arrays. If value is None, then key is set to be the empty array.

It is a programmer error to give a key that isn’t specified as having an array of strings type in the schema for self.

key

the name of the key to set

value

the value to set it to, or None

Returns

true if setting the key succeeded, false if the key was not writable

fn set_uint(&self, key: &str, value: u32) -> Result<(), BoolError>

Sets key in self to value.

A convenience variant of SettingsExt::set for 32-bit unsigned integers.

It is a programmer error to give a key that isn’t specified as having a uint32 type in the schema for self.

key

the name of the key to set

value

the value to set it to

Returns

true if setting the key succeeded, false if the key was not writable

fn set_uint64(&self, key: &str, value: u64) -> Result<(), BoolError>

This is supported on crate feature v2_50 only.

Sets key in self to value.

A convenience variant of SettingsExt::set for 64-bit unsigned integers.

It is a programmer error to give a key that isn’t specified as having a uint64 type in the schema for self.

Feature: v2_50

key

the name of the key to set

value

the value to set it to

Returns

true if setting the key succeeded, false if the key was not writable

fn set_value(&self, key: &str, value: &Variant) -> Result<(), BoolError>

Sets key in self to value.

It is a programmer error to give a key that isn’t contained in the schema for self or for value to have the incorrect type, per the schema.

If value is floating then this function consumes the reference.

key

the name of the key to set

value

a glib::Variant of the correct type

Returns

true if setting the key succeeded, false if the key was not writable

fn backend(&self) -> Option<SettingsBackend>

fn is_delay_apply(&self) -> bool

fn path(&self) -> Option<GString>

fn schema_id(&self) -> Option<GString>

fn settings_schema(&self) -> Option<SettingsSchema>

fn connect_changed<F: Fn(&Self, &str) + 'static>(
&self,
detail: Option<&str>,
f: F
) -> SignalHandlerId

The “changed” signal is emitted when a key has potentially changed. You should call one of the SettingsExt::get calls to check the new value.

This signal supports detailed connections. You can connect to the detailed signal “changed::x” in order to only receive callbacks when key “x” changes.

Note that settings only emits this signal if you have read key at least once while a signal handler was already connected for key.

key

the name of the key that changed

fn connect_writable_change_event<F: Fn(&Self, u32) -> Inhibit + 'static>(
&self,
f: F
) -> SignalHandlerId

The “writable-change-event” signal is emitted once per writability change event that affects this settings object. You should connect to this signal if you are interested in viewing groups of changes before they are split out into multiple emissions of the “writable-changed” signal. For most use cases it is more appropriate to use the “writable-changed” signal.

In the event that the writability change applies only to a single key, key will be set to the glib::Quark for that key. In the event that the writability change affects the entire settings object, key will be 0.

The default handler for this signal invokes the “writable-changed” and “changed” signals for each affected key. This is done because changes in writability might also imply changes in value (if for example, a new mandatory setting is introduced). If any other connected handler returns true then this default functionality will be suppressed.

key

the quark of the key, or 0

Returns

true to stop other handlers from being invoked for the event. FALSE to propagate the event further.

fn connect_writable_changed<F: Fn(&Self, &str) + 'static>(
&self,
detail: Option<&str>,
f: F
) -> SignalHandlerId

The “writable-changed” signal is emitted when the writability of a key has potentially changed. You should call SettingsExt::is_writable in order to determine the new status.

This signal supports detailed connections. You can connect to the detailed signal “writable-changed::x” in order to only receive callbacks when the writability of “x” changes.

key

the key

fn connect_property_delay_apply_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_has_unapplied_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...