Trait gio::ThemedIconExt

pub trait ThemedIconExt: 'static {
    fn append_name(&self, iconname: &str);

    fn get_names(&self) -> Vec<GString>;

    fn prepend_name(&self, iconname: &str);

    fn get_property_use_default_fallbacks(&self) -> bool;

    fn connect_property_names_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all ThemedIcon methods.

Implementors

ThemedIcon

Required methods

fn append_name(&self, iconname: &str)

Append a name to the list of icons from within self.

Note that doing so invalidates the hash computed by prior calls to Icon::hash.

iconname

name of icon to append to list of icons from within self.

fn get_names(&self) -> Vec<GString>

Gets the names of icons from within self.

Returns

a list of icon names.

fn prepend_name(&self, iconname: &str)

Prepend a name to the list of icons from within self.

Note that doing so invalidates the hash computed by prior calls to Icon::hash.

iconname

name of icon to prepend to list of icons from within self.

fn get_property_use_default_fallbacks(&self) -> bool

fn connect_property_names_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...