Trait gio::TlsBackendExt

pub trait TlsBackendExt: 'static {
    fn get_certificate_type(&self) -> Type;

    fn get_client_connection_type(&self) -> Type;

    fn get_default_database(&self) -> Option<TlsDatabase>;

    fn get_dtls_client_connection_type(&self) -> Type;

    fn get_dtls_server_connection_type(&self) -> Type;

    fn get_file_database_type(&self) -> Type;

    fn get_server_connection_type(&self) -> Type;

    fn set_default_database<P: IsA<TlsDatabase>>(&self, database: Option<&P>);

    fn supports_dtls(&self) -> bool;

    fn supports_tls(&self) -> bool;
}

Trait containing all TlsBackend methods.

Implementors

TlsBackend

Required methods

fn get_certificate_type(&self) -> Type

Gets the glib::Type of self's TlsCertificate implementation.

Returns

the glib::Type of self's TlsCertificate implementation.

fn get_client_connection_type(&self) -> Type

Gets the glib::Type of self's TlsClientConnection implementation.

Returns

the glib::Type of self's TlsClientConnection implementation.

fn get_default_database(&self) -> Option<TlsDatabase>

Gets the default TlsDatabase used to verify TLS connections.

Returns

the default database, which should be unreffed when done.

fn get_dtls_client_connection_type(&self) -> Type

This is supported on crate feature v2_48 only.

Gets the glib::Type of self’s DtlsClientConnection implementation.

Feature: v2_48

Returns

the glib::Type of self’s DtlsClientConnection implementation, or G_TYPE_INVALID if this backend doesn’t support DTLS.

fn get_dtls_server_connection_type(&self) -> Type

This is supported on crate feature v2_48 only.

Gets the glib::Type of self’s DtlsServerConnection implementation.

Feature: v2_48

Returns

the glib::Type of self’s DtlsServerConnection implementation, or G_TYPE_INVALID if this backend doesn’t support DTLS.

fn get_file_database_type(&self) -> Type

Gets the glib::Type of self's TlsFileDatabase implementation.

Returns

the glib::Type of backend's TlsFileDatabase implementation.

fn get_server_connection_type(&self) -> Type

Gets the glib::Type of self's TlsServerConnection implementation.

Returns

the glib::Type of self's TlsServerConnection implementation.

fn set_default_database<P: IsA<TlsDatabase>>(&self, database: Option<&P>)

This is supported on crate feature v2_60 only.

Set the default TlsDatabase used to verify TLS connections

Any subsequent call to TlsBackend::get_default_database will return the database set in this call. Existing databases and connections are not modified.

Setting a None default database will reset to using the system default database as if TlsBackend::set_default_database had never been called.

Feature: v2_60

database

the TlsDatabase

fn supports_dtls(&self) -> bool

This is supported on crate feature v2_48 only.

Checks if DTLS is supported. DTLS support may not be available even if TLS support is available, and vice-versa.

Feature: v2_48

Returns

whether DTLS is supported

fn supports_tls(&self) -> bool

Checks if TLS is supported; if this returns false for the default TlsBackend, it means no "real" TLS backend is available.

Returns

whether or not TLS is supported

Loading content...