Trait gio::TlsServerConnectionExt

pub trait TlsServerConnectionExt: 'static {
    fn get_property_authentication_mode(&self) -> TlsAuthenticationMode;

    fn set_property_authentication_mode(
        &self, 
        authentication_mode: TlsAuthenticationMode
    );

    fn connect_property_authentication_mode_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all TlsServerConnection methods.

Implementors

TlsServerConnection

Required methods

fn get_property_authentication_mode(&self) -> TlsAuthenticationMode

The TlsAuthenticationMode for the server. This can be changed before calling TlsConnectionExt::handshake if you want to rehandshake with a different mode from the initial handshake.

fn set_property_authentication_mode(
&self,
authentication_mode: TlsAuthenticationMode
)

The TlsAuthenticationMode for the server. This can be changed before calling TlsConnectionExt::handshake if you want to rehandshake with a different mode from the initial handshake.

fn connect_property_authentication_mode_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...