Trait gio::UnixInputStreamExt

pub trait UnixInputStreamExt: 'static {
    fn get_close_fd(&self) -> bool;
}

Trait containing all UnixInputStream methods.

Implementors

UnixInputStream

Required methods

fn get_close_fd(&self) -> bool

Returns whether the file descriptor of self will be closed when the stream is closed.

Returns

true if the file descriptor is closed when done

Loading content...