Trait gio::VfsExt

pub trait VfsExt: 'static {
    fn get_file_for_path(&self, path: &str) -> Option<File>;

    fn get_file_for_uri(&self, uri: &str) -> Option<File>;

    fn get_supported_uri_schemes(&self) -> Vec<GString>;

    fn is_active(&self) -> bool;

    fn parse_name(&self, parse_name: &str) -> Option<File>;
}

Trait containing all Vfs methods.

Implementors

Vfs

Required methods

fn get_file_for_path(&self, path: &str) -> Option<File>

Gets a File for path.

path

a string containing a VFS path.

Returns

a File. Free the returned object with gobject::ObjectExt::unref.

fn get_file_for_uri(&self, uri: &str) -> Option<File>

Gets a File for uri.

This operation never fails, but the returned object might not support any I/O operation if the URI is malformed or if the URI scheme is not supported.

uri

a string containing a URI

Returns

a File. Free the returned object with gobject::ObjectExt::unref.

fn get_supported_uri_schemes(&self) -> Vec<GString>

Gets a list of URI schemes supported by self.

Returns

a None-terminated array of strings. The returned array belongs to GIO and must not be freed or modified.

fn is_active(&self) -> bool

Checks if the VFS is active.

Returns

true if construction of the self was successful and it is now active.

fn parse_name(&self, parse_name: &str) -> Option<File>

This operation never fails, but the returned object might not support any I/O operations if the parse_name cannot be parsed by the Vfs module.

parse_name

a string to be parsed by the VFS module.

Returns

a File for the given parse_name. Free the returned object with gobject::ObjectExt::unref.

Loading content...