Trait gio::VolumeMonitorExt

pub trait VolumeMonitorExt: 'static {
    fn get_connected_drives(&self) -> Vec<Drive>;

    fn get_mount_for_uuid(&self, uuid: &str) -> Option<Mount>;

    fn get_mounts(&self) -> Vec<Mount>;

    fn get_volume_for_uuid(&self, uuid: &str) -> Option<Volume>;

    fn get_volumes(&self) -> Vec<Volume>;

    fn connect_drive_changed<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drive_connected<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drive_disconnected<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drive_eject_button<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drive_stop_button<F: Fn(&Self, &Drive) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_mount_added<F: Fn(&Self, &Mount) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_mount_changed<F: Fn(&Self, &Mount) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_mount_pre_unmount<F: Fn(&Self, &Mount) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_mount_removed<F: Fn(&Self, &Mount) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_volume_added<F: Fn(&Self, &Volume) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_volume_changed<F: Fn(&Self, &Volume) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_volume_removed<F: Fn(&Self, &Volume) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all VolumeMonitor methods.

Implementors

VolumeMonitor

Required methods

fn get_connected_drives(&self) -> Vec<Drive>

Gets a list of drives connected to the system.

The returned list should be freed with glib::List::free, after its elements have been unreffed with gobject::ObjectExt::unref.

Returns

a glib::List of connected Drive objects.

fn get_mount_for_uuid(&self, uuid: &str) -> Option<Mount>

Finds a Mount object by its UUID (see Mount::get_uuid)

uuid

the UUID to look for

Returns

a Mount or None if no such mount is available. Free the returned object with gobject::ObjectExt::unref.

fn get_mounts(&self) -> Vec<Mount>

Gets a list of the mounts on the system.

The returned list should be freed with glib::List::free, after its elements have been unreffed with gobject::ObjectExt::unref.

Returns

a glib::List of Mount objects.

fn get_volume_for_uuid(&self, uuid: &str) -> Option<Volume>

Finds a Volume object by its UUID (see Volume::get_uuid)

uuid

the UUID to look for

Returns

a Volume or None if no such volume is available. Free the returned object with gobject::ObjectExt::unref.

fn get_volumes(&self) -> Vec<Volume>

Gets a list of the volumes on the system.

The returned list should be freed with glib::List::free, after its elements have been unreffed with gobject::ObjectExt::unref.

Returns

a glib::List of Volume objects.

fn connect_drive_changed<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a drive changes.

drive

the drive that changed

fn connect_drive_connected<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a drive is connected to the system.

drive

a Drive that was connected.

fn connect_drive_disconnected<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a drive is disconnected from the system.

drive

a Drive that was disconnected.

fn connect_drive_eject_button<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when the eject button is pressed on drive.

drive

the drive where the eject button was pressed

fn connect_drive_stop_button<F: Fn(&Self, &Drive) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when the stop button is pressed on drive.

drive

the drive where the stop button was pressed

fn connect_mount_added<F: Fn(&Self, &Mount) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mount is added.

mount

a Mount that was added.

fn connect_mount_changed<F: Fn(&Self, &Mount) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mount changes.

mount

a Mount that changed.

fn connect_mount_pre_unmount<F: Fn(&Self, &Mount) + 'static>(
&self,
f: F
) -> SignalHandlerId

May be emitted when a mount is about to be removed.

This signal depends on the backend and is only emitted if GIO was used to unmount.

mount

a Mount that is being unmounted.

fn connect_mount_removed<F: Fn(&Self, &Mount) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mount is removed.

mount

a Mount that was removed.

fn connect_volume_added<F: Fn(&Self, &Volume) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mountable volume is added to the system.

volume

a Volume that was added.

fn connect_volume_changed<F: Fn(&Self, &Volume) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when mountable volume is changed.

volume

a Volume that changed.

fn connect_volume_removed<F: Fn(&Self, &Volume) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a mountable volume is removed from the system.

volume

a Volume that was removed.

Loading content...