Trait gio::ZlibDecompressorExt

pub trait ZlibDecompressorExt: 'static {
    fn get_file_info(&self) -> Option<FileInfo>;

    fn get_property_format(&self) -> ZlibCompressorFormat;

    fn connect_property_file_info_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all ZlibDecompressor methods.

Implementors

ZlibDecompressor

Required methods

fn get_file_info(&self) -> Option<FileInfo>

Retrieves the FileInfo constructed from the GZIP header data of compressed data processed by compressor, or None if self's ZlibDecompressor:format property is not ZlibCompressorFormat::Gzip, or the header data was not fully processed yet, or it not present in the data stream at all.

Returns

a FileInfo, or None

fn get_property_format(&self) -> ZlibCompressorFormat

fn connect_property_file_info_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...