Crate gio

Rust GIO bindings

Rust bindings and wrappers for GIO, part of gtk-rs-core.

GIO 2.56 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.80.0.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gio = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gio" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
gio = "0.13"
gio = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gio" }

See Also

glib

License

gio is available under the MIT License, please refer to it.

Re-exports

pub use crate::resource::resources_register_include_impl;

pub use gio_sys as ffi;

pub use glib;

pub use crate::functions::*;

Modules

builders

functions

prelude

Traits and essential types intended for blanket imports.

subclass

Macros

resources_register_include

Include gresources generated with glib_build_tools::compile_resources and register with glib. path is relative to OUTDIR.

Structs

Action

GAction represents a single named action.

ActionEntry

This struct defines a single action. It is for use with ActionMapExtManual::add_action_entries().

ActionEntryBuilder

ActionGroup

GActionGroup represents a group of actions.

ActionGroupExportId

ActionMap

GActionMap is an interface for action containers.

AppInfo

Information about an installed application and methods to launch it (with file arguments).

AppInfoCreateFlags

Flags used when creating a #GAppInfo.

AppInfoMonitor

GAppInfoMonitor monitors application information for changes.

AppLaunchContext

Integrating the launch with the launching application. This is used to handle for instance startup notification and launching the new application on the same screen as the launching window.

Application

GApplication is the core class for application support.

ApplicationBusyGuard

ApplicationCommandLine

GApplicationCommandLine represents a command-line invocation of an application.

ApplicationFlags

Flags used to define the behaviour of a #GApplication.

ApplicationHoldGuard

AskPasswordFlags

#GAskPasswordFlags are used to request specific information from the user, or to notify the user of their choices in an authentication situation.

AsyncInitable

GAsyncInitable is an interface for asynchronously initializable objects.

AsyncResult

GAsyncResult provides a base class for implementing asynchronous function results.

BindingBuilder

BufferedInputStream

Buffered input stream implements FilterInputStream and provides for buffered reads.

BufferedOutputStream

Buffered output stream implements FilterOutputStream and provides for buffered writes.

BusNameOwnerFlags

Flags used in g_bus_own_name().

BusNameWatcherFlags

Flags used in g_bus_watch_name().

BytesIcon

GBytesIcon specifies an image held in memory in a common format (usually PNG) to be used as icon.

Cancellable

GCancellable allows operations to be cancelled.

CancellableFuture

A future which can be cancelled via Cancellable.

Cancelled

Indicator that the CancellableFuture was cancelled.

CharsetConverter

GCharsetConverter is an implementation of Converter based on GLib::IConv.

Converter

GConverter is an interface for streaming conversions.

ConverterFlags

Flags used when calling a g_converter_convert().

ConverterInputStream

Converter input stream implements InputStream and allows conversion of data of various types during reading.

ConverterOutputStream

Converter output stream implements OutputStream and allows conversion of data of various types during reading.

Credentials

The GCredentials type is a reference-counted wrapper for native credentials.

DBusActionGroup

GDBusActionGroup is an implementation of the ActionGroup interface.

DBusArgInfo

Information about an argument for a method or a signal.

DBusAuthObserver

GDBusAuthObserver provides a mechanism for participating in how a DBusServer (or a DBusConnection) authenticates remote peers.

DBusCallFlags

Flags used in g_dbus_connection_call() and similar APIs.

DBusCapabilityFlags

Capabilities negotiated with the remote peer.

DBusConnection

The GDBusConnection type is used for D-Bus connections to remote peers such as a message buses.

DBusConnectionFlags

Flags used when creating a new #GDBusConnection.

DBusInterface

Base type for D-Bus interfaces.

DBusInterfaceInfo

Information about a D-Bus interface.

DBusInterfaceSkeleton

Abstract base class for D-Bus interfaces on the service side.

DBusInterfaceSkeletonFlags

Flags describing the behavior of a #GDBusInterfaceSkeleton instance.

DBusMenuModel

GDBusMenuModel is an implementation of MenuModel that can be used as a proxy for a menu model that is exported over D-Bus with DBusConnection::export_menu_model().

DBusMessage

A type for representing D-Bus messages that can be sent or received on a DBusConnection.

DBusMessageFlags

Message flags used in #GDBusMessage.

DBusMethodInfo

Information about a method on a D-Bus interface.

DBusMethodInvocation

Instances of the GDBusMethodInvocation class are used when handling D-Bus method calls. It provides a way to asynchronously return results and errors.

DBusNodeInfo

Information about nodes in a remote object hierarchy.

DBusObject

The GDBusObject type is the base type for D-Bus objects on both the service side (see DBusObjectSkeleton) and the client side (see DBusObjectProxy). It is essentially just a container of interfaces.

DBusObjectManager

The GDBusObjectManager type is the base type for service- and client-side implementations of the standardized org.freedesktop.DBus.ObjectManager interface.

DBusObjectManagerClientFlags

Flags used when constructing a #GDBusObjectManagerClient.

DBusObjectManagerServer

GDBusObjectManagerServer is used to export DBusObject instances using the standardized org.freedesktop.DBus.ObjectManager interface. For example, remote D-Bus clients can get all objects and properties in a single call. Additionally, any change in the object hierarchy is broadcast using signals. This means that D-Bus clients can keep caches up to date by only listening to D-Bus signals.

DBusObjectProxy

A GDBusObjectProxy is an object used to represent a remote object with one or more D-Bus interfaces. Normally, you don’t instantiate a GDBusObjectProxy yourself — typically Gio::DBusObjectManagerClient is used to obtain it.

DBusObjectSkeleton

A GDBusObjectSkeleton instance is essentially a group of D-Bus interfaces. The set of exported interfaces on the object may be dynamic and change at runtime.

DBusPropertyInfo

Information about a D-Bus property on a D-Bus interface.

DBusPropertyInfoFlags

Flags describing the access control of a D-Bus property.

DBusProxy

GDBusProxy is a base class used for proxies to access a D-Bus interface on a remote object. A GDBusProxy can be constructed for both well-known and unique names.

DBusProxyFlags

Flags used when constructing an instance of a #GDBusProxy derived class.

DBusSendMessageFlags

Flags used when sending #GDBusMessages on a #GDBusConnection.

DBusServer

GDBusServer is a helper for listening to and accepting D-Bus connections. This can be used to create a new D-Bus server, allowing two peers to use the D-Bus protocol for their own specialized communication. A server instance provided in this way will not perform message routing or implement the org.freedesktop.DBus interface.

DBusServerFlags

Flags used when creating a #GDBusServer.

DBusSignalFlags

Flags used when subscribing to signals via g_dbus_connection_signal_subscribe().

DBusSignalInfo

Information about a signal on a D-Bus interface.

DBusSubtreeFlags

Flags passed to g_dbus_connection_register_subtree().

DataInputStream

Data input stream implements InputStream and includes functions for reading structured data directly from a binary input stream.

DataOutputStream

Data output stream implements OutputStream and includes functions for writing data directly to an output stream.

DatagramBased

Interface for socket-like objects with datagram semantics.

DebugControllerv2_72

GDebugController is an interface to expose control of debugging features and debug output.

DebugControllerDBusv2_72

GDebugControllerDBus is an implementation of DebugController which exposes debug settings as a D-Bus object.

DesktopAppInfoNon-Windows and non-macOS

GDesktopAppInfo is an implementation of AppInfo based on desktop files.

Drive

GDrive represents a piece of hardware connected to the machine. It’s generally only created for removable hardware or hardware with removable media. For example, an optical disc drive, or a USB flash drive.

DriveStartFlags

Flags used when starting a drive.

DtlsClientConnection

GDtlsClientConnection is the client-side subclass of DtlsConnection, representing a client-side DTLS connection.

DtlsConnection

GDtlsConnection is the base DTLS connection class type, which wraps a DatagramBased and provides DTLS encryption on top of it. Its subclasses, DtlsClientConnection and DtlsServerConnection, implement client-side and server-side DTLS, respectively.

DtlsServerConnection

GDtlsServerConnection is the server-side subclass of DtlsConnection, representing a server-side DTLS connection.

Emblem

GEmblem is an implementation of Icon that supports having an emblem, which is an icon with additional properties. It can than be added to a EmblemedIcon.

EmblemedIcon

GEmblemedIcon is an implementation of Icon that supports adding an emblem to an icon. Adding multiple emblems to an icon is ensured via EmblemedIconExt::add_emblem().

File

GFile is a high level abstraction for manipulating files on a virtual file system. GFiles are lightweight, immutable objects that do no I/O upon creation. It is necessary to understand that GFile objects do not represent files, merely an identifier for a file. All file content I/O is implemented as streaming operations (see InputStream and OutputStream).

FileAttributeInfo

Information about a specific attribute.

FileAttributeInfoFlags

Flags specifying the behaviour of an attribute.

FileAttributeInfoList

Acts as a lightweight registry for possible valid file attributes. The registry stores Key-Value pair formats as #GFileAttributeInfos.

FileAttributeMatcher

Determines if a string matches a file attribute.

FileAttributematcherIter

FileCopyFlags

Flags used when copying or moving files.

FileCreateFlags

Flags used when an operation may create a file.

FileDescriptorBased

GFileDescriptorBased is an interface for file descriptor based IO.

FileEnumerator

GFileEnumerator allows you to operate on a set of File objects, returning a FileInfo structure for each file enumerated (e.g. FileExt::enumerate_children() will return a GFileEnumerator for each of the children within a directory).

FileEnumeratorStream

A Stream used to enumerate files in directories.

FileIOStream

GFileIOStream provides I/O streams that both read and write to the same file handle.

FileIcon

GFileIcon specifies an icon by pointing to an image file to be used as icon.

FileInfo

Stores information about a file system object referenced by a File.

FileInputStream

GFileInputStream provides input streams that take their content from a file.

FileMeasureFlags

Flags that can be used with g_file_measure_disk_usage().

FileMonitor

Monitors a file or directory for changes.

FileMonitorFlags

Flags used to set what a #GFileMonitor will watch for.

FileOutputStream

GFileOutputStream provides output streams that write their content to a file.

FileQueryInfoFlags

Flags used when querying a #GFileInfo.

FilenameCompleter

Completes partial file and directory names given a partial string by looking in the file system for clues. Can return a list of possible completion strings for widget implementations.

FilterId

FilterInputStream

Base class for input stream implementations that perform some kind of filtering operation on a base stream. Typical examples of filtering operations are character set conversion, compression and byte order flipping.

FilterOutputStream

Base class for output stream implementations that perform some kind of filtering operation on a base stream. Typical examples of filtering operations are character set conversion, compression and byte order flipping.

GioFuture

GioFutureResult

IOExtension

The implementation of an IOExtensionPoint. #GIOExtension is an opaque data structure and can only be accessed using the following functions.

IOExtensionPoint

An extension point provides a mechanism to extend the functionality of a library or application. Each extension point is identified by a name, and it may optionally require that any implementation must be of a certain type. GIOExtensionPoint provides a mechanism for modules to extend the functionality of the library or application that loaded it in an organized fashion.

IOExtensionPointBuilder

Builder for extension points.

IOStream

GIOStream represents an object that has both read and write streams. Generally the two streams act as separate input and output streams, but they share some common resources and state. For instance, for seekable streams, both streams may use the same position.

IOStreamAsyncReadWrite

IOStreamSpliceFlags

GIOStreamSpliceFlags determine how streams should be spliced.

Icon

GIcon is a very minimal interface for icons. It provides functions for checking the equality of two icons, hashing of icons and serializing an icon to and from strings.

InetAddress

GInetAddress represents an IPv4 or IPv6 internet address. Use ResolverExt::lookup_by_name() or ResolverExt::lookup_by_name_async() to look up the GInetAddress for a hostname. Use ResolverExt::lookup_by_address() or ResolverExt::lookup_by_address_async() to look up the hostname for a GInetAddress.

InetAddressMask

GInetAddressMask represents a range of IPv4 or IPv6 addresses described by a base address and a length indicating how many bits of the base address are relevant for matching purposes. These are often given in string form. For example, 10.0.0.0/8, or fe80::/10.

InetSocketAddress

An IPv4 or IPv6 socket address. That is, the combination of a InetAddress and a port number.

Initable

GInitable is implemented by objects that can fail during initialization. If an object implements this interface then it must be initialized as the first thing after construction, either via InitableExt::init() or AsyncInitableExt::init_async() (the latter is only available if it also implements AsyncInitable).

InputMessage

Structure used for scatter/gather data input when receiving multiple messages or packets in one go. You generally pass in an array of empty #GInputVectors and the operation will use all the buffers as if they were one buffer, and will set @bytes_received to the total number of bytes received across all #GInputVectors.

InputStream

GInputStream is a base class for implementing streaming input.

InputStreamAsyncBufRead

InputStreamAsyncRead

InputStreamRead

InputVector

JoinHandle

A handle to a task running on the I/O thread pool.

ListModel

GListModel is an interface that represents a mutable list of glib::Object. Its main intention is as a model for various widgets in user interfaces, such as list views, but it can also be used as a convenient method of returning lists of data, with support for updates.

ListStore

GListStore is a simple implementation of ListModel that stores all items in memory.

LoadableIcon

GLoadableIcon extends the Icon interface and adds the ability to load icons from streams.

LocalTask

LocalTask provides idiomatic access to gio’s GTask API, for instance by being generic over their value type, while not completely departing from the underlying C API. LocalTask does not require its value to be Send and Sync and thus is useful to to implement gio style asynchronous tasks that run in the glib main loop. If you need to run tasks in threads see the Task type.

MemoryInputStream

GMemoryInputStream is a class for using arbitrary memory chunks as input for GIO streaming input operations.

MemoryMonitorv2_64

GMemoryMonitor will monitor system memory and suggest to the application when to free memory so as to leave more room for other applications. It is implemented on Linux using the Low Memory Monitor (API documentation).

MemoryOutputStream

GMemoryOutputStream is a class for using arbitrary memory chunks as output for GIO streaming output operations.

Menu

GMenu is a simple implementation of MenuModel. You populate a GMenu by adding MenuItem instances to it.

MenuAttributeIter

#GMenuAttributeIter is an opaque structure type. You must access it using the functions below.

MenuItem

#GMenuItem is an opaque structure type. You must access it using the functions below.

MenuLinkIter

#GMenuLinkIter is an opaque structure type. You must access it using the functions below.

MenuModel

GMenuModel represents the contents of a menu — an ordered list of menu items. The items are associated with actions, which can be activated through them. Items can be grouped in sections, and may have submenus associated with them. Both items and sections usually have some representation data, such as labels or icons. The type of the associated action (ie whether it is stateful, and what kind of state it has) can influence the representation of the item.

MenuModelExportId

Mount

The GMount interface represents a user-visible mount, such as a mounted file system.

MountMountFlags

Flags used when mounting a mount.

MountOperation

GMountOperation provides a mechanism for interacting with the user. It can be used for authenticating mountable operations, such as loop mounting files, hard drive partitions or server locations. It can also be used to ask the user questions or show a list of applications preventing unmount or eject operations from completing.

MountUnmountFlags

Flags used when an unmounting a mount.

NativeSocketAddress

A socket address of some unknown native type.

NativeVolumeMonitor

This is an Abstract Base Class, you cannot instantiate it.

NetworkAddress

GNetworkAddress provides an easy way to resolve a hostname and then attempt to connect to that host, handling the possibility of multiple IP addresses and multiple address families.

NetworkMonitor

GNetworkMonitor provides an easy-to-use cross-platform API for monitoring network connectivity. On Linux, the available implementations are based on the kernel’s netlink interface and on NetworkManager.

NetworkService

Like NetworkAddress does with hostnames, GNetworkService provides an easy way to resolve a SRV record, and then attempt to connect to one of the hosts that implements that service, handling service priority/weighting, multiple IP addresses, and multiple address families.

Notification

GNotification is a mechanism for creating a notification to be shown to the user — typically as a pop-up notification presented by the desktop environment shell.

OutputMessage

Structure used for scatter/gather data output when sending multiple messages or packets in one go. You generally pass in an array of #GOutputVectors and the operation will use all the buffers as if they were one buffer.

OutputStream

GOutputStream is a base class for implementing streaming output.

OutputStreamAsyncWrite

OutputStreamSpliceFlags

GOutputStreamSpliceFlags determine how streams should be spliced.

OutputStreamWrite

OutputVector

OwnerId

Permission

A GPermission represents the status of the caller’s permission to perform a certain action.

PollableInputStream

GPollableInputStream is implemented by InputStreams that can be polled for readiness to read. This can be used when interfacing with a non-GIO API that expects UNIX-file-descriptor-style asynchronous I/O rather than GIO-style.

PollableOutputStream

GPollableOutputStream is implemented by OutputStreams that can be polled for readiness to write. This can be used when interfacing with a non-GIO API that expects UNIX-file-descriptor-style asynchronous I/O rather than GIO-style.

PowerProfileMonitorv2_70

GPowerProfileMonitor makes it possible for applications as well as OS components to monitor system power profiles and act upon them. It currently only exports whether the system is in “Power Saver” mode (known as “Low Power” mode on some systems).

PropertyAction

A GPropertyAction is a way to get a Action with a state value reflecting and controlling the value of a glib::Object property.

Proxy

A GProxy handles connecting to a remote host via a given type of proxy server. It is implemented by the gio-proxy extension point. The extensions are named after their proxy protocol name. As an example, a SOCKS5 proxy implementation can be retrieved with the name socks5 using the function IOExtensionPoint::extension_by_name().

ProxyAddress

A InetSocketAddress representing a connection via a proxy server.

ProxyAddressEnumerator

GProxyAddressEnumerator is a wrapper around SocketAddressEnumerator which takes the SocketAddress instances returned by the SocketAddressEnumerator and wraps them in ProxyAddress instances, using the given proxy-resolver.

ProxyResolver

GProxyResolver provides synchronous and asynchronous network proxy resolution. GProxyResolver is used within SocketClient through the method SocketConnectableExt::proxy_enumerate().

ReadInputStream

GLib type: GObject with reference counted clone semantics.

RegistrationBuilder

Build a registered DBus object, by handling different parts of DBus.

RegistrationId

RemoteActionGroup

The GRemoteActionGroup interface is implemented by ActionGroup instances that either transmit action invocations to other processes or receive action invocations in the local process from other processes.

Resolver

The object that handles DNS resolution. Use default() to get the default resolver.

ResolverNameLookupFlagsv2_60

Flags to modify lookup behavior.

Resource

Applications and libraries often contain binary or textual data that is really part of the application, rather than user data. For instance GtkBuilder .ui files, splashscreen images, Menu markup XML, CSS files, icons, etc. These are often shipped as files in $datadir/appname, or manually included as literal strings in the code.

ResourceFlags

GResourceFlags give information about a particular file inside a resource bundle.

ResourceLookupFlags

GResourceLookupFlags determine how resource path lookups are handled.

Seekable

GSeekable is implemented by streams (implementations of InputStream or OutputStream) that support seeking.

Settings

The GSettings class provides a convenient API for storing and retrieving application settings.

SettingsBackend

The GSettingsBackend interface defines a generic interface for non-strictly-typed data that is stored in a hierarchy. To implement an alternative storage backend for Settings, you need to implement the GSettingsBackend interface and then make it implement the extension point G_SETTINGS_BACKEND_EXTENSION_POINT_NAME.

SettingsBindFlags

Flags used when creating a binding. These flags determine in which direction the binding works. The default is to synchronize in both directions.

SettingsSchema

The SettingsSchemaSource and GSettingsSchema APIs provide a mechanism for advanced control over the loading of schemas and a mechanism for introspecting their content.

SettingsSchemaKey

#GSettingsSchemaKey is an opaque data structure and can only be accessed using the following functions.

SettingsSchemaSource

This is an opaque structure type. You may not access it directly.

SignalSubscriptionId

SimpleAction

A GSimpleAction is the obvious simple implementation of the Action interface. This is the easiest way to create an action for purposes of adding it to a SimpleActionGroup.

SimpleActionGroup

GSimpleActionGroup is a hash table filled with Action objects, implementing the ActionGroup and ActionMap interfaces.

SimpleIOStream

GSimpleIOStream creates a IOStream from an arbitrary InputStream and OutputStream. This allows any pair of input and output streams to be used with IOStream methods.

SimplePermission

GSimplePermission is a trivial implementation of Permission that represents a permission that is either always or never allowed. The value is given at construction and doesn’t change.

SimpleProxyResolver

GSimpleProxyResolver is a simple ProxyResolver implementation that handles a single default proxy, multiple URI-scheme-specific proxies, and a list of hosts that proxies should not be used for.

Socket

A GSocket is a low-level networking primitive. It is a more or less direct mapping of the BSD socket API in a portable GObject based API. It supports both the UNIX socket implementations and winsock2 on Windows.

SocketAddress

GSocketAddress is the equivalent of struct sockaddr and its subtypes in the BSD sockets API. This is an abstract class; use InetSocketAddress for internet sockets, or UnixSocketAddress for UNIX domain sockets.

SocketAddressEnumerator

GSocketAddressEnumerator is an enumerator type for SocketAddress instances. It is returned by enumeration functions such as SocketConnectableExt::enumerate(), which returns a GSocketAddressEnumerator to list each SocketAddress which could be used to connect to that SocketConnectable.

SocketClient

GSocketClient is a lightweight high-level utility class for connecting to a network host using a connection oriented socket type.

SocketConnectable

Objects that describe one or more potential socket endpoints implement GSocketConnectable. Callers can then use SocketConnectableExt::enumerate() to get a SocketAddressEnumerator to try out each socket address in turn until one succeeds, as shown in the sample code below.

SocketConnection

GSocketConnection is a IOStream for a connected socket. They can be created either by SocketClient when connecting to a host, or by SocketListener when accepting a new client.

SocketControlMessage

A GSocketControlMessage is a special-purpose utility message that can be sent to or received from a Socket. These types of messages are often called ‘ancillary data’.

SocketControlMessages

SocketListener

A GSocketListener is an object that keeps track of a set of server sockets and helps you accept sockets from any of the socket, either sync or async.

SocketMsgFlags

Flags used in g_socket_receive_message() and g_socket_send_message(). The flags listed in the enum are some commonly available flags, but the values used for them are the same as on the platform, and any other flags are passed in/out as is. So to use a platform specific flag, just include the right system header and pass in the flag.

SocketService

A GSocketService is an object that represents a service that is provided to the network or over local sockets. When a new connection is made to the service the incoming signal is emitted.

SrvTarget

A single target host/port that a network service is running on.

Subprocess

GSubprocess allows the creation of and interaction with child processes.

SubprocessFlags

Flags to define the behaviour of a #GSubprocess.

SubprocessLauncher

This class contains a set of options for launching child processes, such as where its standard input and output will be directed, the argument list, the environment, and more.

Task

Task provides idiomatic access to gio’s GTask API, for instance by being generic over their value type, while not completely departing from the underlying C API. Task is Send and Sync and requires its value to also be Send and Sync, thus is useful to to implement gio style asynchronous tasks that run in threads. If you need to only run tasks in glib main loop see the LocalTask type.

TcpConnection

This is the subclass of SocketConnection that is created for TCP/IP sockets.

TcpWrapperConnection

A GTcpWrapperConnection can be used to wrap a IOStream that is based on a Socket, but which is not actually a SocketConnection. This is used by SocketClient so that it can always return a SocketConnection, even when the connection it has actually created is not directly a SocketConnection.

TestDBus

A helper class for testing code which uses D-Bus without touching the user’s session bus.

TestDBusFlags

Flags to define future #GTestDBus behaviour.

ThemedIcon

GThemedIcon is an implementation of Icon that supports icon themes.

ThreadedSocketService

A GThreadedSocketService is a simple subclass of SocketService that handles incoming connections by creating a worker thread and dispatching the connection to it by emitting the run signal] in the new thread.

TlsBackend

TLS (Transport Layer Security, aka SSL) and DTLS backend. This is an internal type used to coordinate the different classes implemented by a TLS backend.

TlsCertificate

A certificate used for TLS authentication and encryption. This can represent either a certificate only (eg, the certificate received by a client from a server), or the combination of a certificate and a private key (which is needed when acting as a TlsServerConnection).

TlsCertificateFlags

A set of flags describing TLS certification validation. This can be used to describe why a particular certificate was rejected (for example, in #GTlsConnection::accept-certificate).

TlsClientConnection

GTlsClientConnection is the client-side subclass of TlsConnection, representing a client-side TLS connection.

TlsConnection

GTlsConnection is the base TLS connection class type, which wraps a IOStream and provides TLS encryption on top of it. Its subclasses, TlsClientConnection and TlsServerConnection, implement client-side and server-side TLS, respectively.

TlsDatabase

GTlsDatabase is used to look up certificates and other information from a certificate or key store. It is an abstract base class which TLS library specific subtypes override.

TlsDatabaseVerifyFlags

Flags for g_tls_database_verify_chain().

TlsFileDatabase

GTlsFileDatabase is implemented by TlsDatabase objects which load their certificate information from a file. It is an interface which TLS library specific subtypes implement.

TlsInteraction

GTlsInteraction provides a mechanism for the TLS connection and database code to interact with the user. It can be used to ask the user for passwords.

TlsPassword

An abstract interface representing a password used in TLS. Often used in user interaction such as unlocking a key storage token.

TlsPasswordFlags

Various flags for the password.

TlsServerConnection

GTlsServerConnection is the server-side subclass of TlsConnection, representing a server-side TLS connection.

UnixConnectionUnix

This is the subclass of SocketConnection that is created for UNIX domain sockets.

UnixCredentialsMessageUnix

This SocketControlMessage contains a Credentials instance. It may be sent using SocketExtManual::send_message() and received using SocketExtManual::receive_message() over UNIX sockets (ie: sockets in the G_SOCKET_FAMILY_UNIX family).

UnixFDListUnix

A GUnixFDList contains a list of file descriptors. It owns the file descriptors that it contains, closing them when finalized.

UnixFDMessageUnix

This SocketControlMessage contains a UnixFDList. It may be sent using SocketExtManual::send_message() and received using SocketExtManual::receive_message() over UNIX sockets (ie: sockets in the G_SOCKET_FAMILY_UNIX family). The file descriptors are copied between processes by the kernel.

UnixInputStreamUnix

GUnixInputStream implements InputStream for reading from a UNIX file descriptor, including asynchronous operations. (If the file descriptor refers to a socket or pipe, this will use poll() to do asynchronous I/O. If it refers to a regular file, it will fall back to doing asynchronous I/O in another thread.)

UnixMountEntryUnix

Defines a Unix mount entry (e.g. /media/cdrom). This corresponds roughly to a mtab entry.

UnixMountMonitorUnix

Watches #GUnixMounts for changes.

UnixMountPointUnix

Defines a Unix mount point (e.g. /dev). This corresponds roughly to a fstab entry.

UnixOutputStreamUnix

GUnixOutputStream implements OutputStream for writing to a UNIX file descriptor, including asynchronous operations. (If the file descriptor refers to a socket or pipe, this will use poll() to do asynchronous I/O. If it refers to a regular file, it will fall back to doing asynchronous I/O in another thread.)

UnixSocketAddressUnix

Support for UNIX-domain (also known as local) sockets, corresponding to struct sockaddr_un.

Vfs

Entry point for using GIO functionality.

Volume

The GVolume interface represents user-visible objects that can be mounted. For example, a file system partition on a USB flash drive, or an optical disc inserted into a disc drive.

VolumeMonitor

GVolumeMonitor is for listing the user interesting devices and volumes on the computer. In other words, what a file selector or file manager would show in a sidebar.

WatcherId

WriteOutputStream

GLib type: GObject with reference counted clone semantics.

ZlibCompressor

GZlibCompressor is an implementation of Converter that compresses data using zlib.

ZlibDecompressor

GZlibDecompressor is an implementation of Converter that decompresses data compressed with zlib.

Enums

BusType

An enumeration for well-known message buses.

ConverterResult

Results returned from g_converter_convert().

CredentialsType

Enumeration describing different kinds of native credential types.

DBusError

Error codes for the G_DBUS_ERROR error domain.

DBusMessageByteOrder

Enumeration used to describe the byte order of a D-Bus message.

DBusMessageHeaderField

Header fields used in #GDBusMessage.

DBusMessageType

Message types used in #GDBusMessage.

DataStreamByteOrder

#GDataStreamByteOrder is used to ensure proper endianness of streaming data sources across various machine architectures.

DataStreamNewlineType

#GDataStreamNewlineType is used when checking for or setting the line endings for a given file.

DriveStartStopType

Enumeration describing how a drive can be started/stopped.

EmblemOrigin

GEmblemOrigin is used to add information about the origin of the emblem to #GEmblem.

FileAttributeStatus

Used by g_file_set_attributes_from_info() when setting file attributes.

FileAttributeType

The data types for file attributes.

FileMonitorEvent

Specifies what type of event a monitor event is.

FileType

Indicates the file’s on-disk type.

FilesystemPreviewType

Indicates a hint from the file system whether files should be previewed in a file manager. Returned as the value of the key FILE_ATTRIBUTE_FILESYSTEM_USE_PREVIEW.

IOErrorEnum

Error codes returned by GIO functions.

IOModuleScopeFlags

Flags for use with g_io_module_scope_new().

InetAddressBytes

MemoryMonitorWarningLevelv2_64

Memory availability warning levels.

MountOperationResult

#GMountOperationResult is returned as a result when a request for information is send by the mounting operation.

NetworkConnectivity

The host’s network connectivity state, as reported by #GNetworkMonitor.

NotificationPriority

Priority levels for #GNotifications.

PasswordSave

#GPasswordSave is used to indicate the lifespan of a saved password.

PollableReturnv2_60

Return value for various IO operations that signal errors via the return value and not necessarily via a #GError.

ResolverError

An error code used with G_RESOLVER_ERROR in a #GError returned from a #GResolver routine.

ResolverRecordType

The type of record that g_resolver_lookup_records() or g_resolver_lookup_records_async() should retrieve. The records are returned as lists of #GVariant tuples. Each record type has different values in the variant tuples returned.

ResourceError

An error code used with G_RESOURCE_ERROR in a #GError returned from a #GResource routine.

SocketClientEvent

Describes an event occurring on a #GSocketClient. See the #GSocketClient::event signal for more details.

SocketFamily

The protocol family of a #GSocketAddress. (These values are identical to the system defines AF_INET, AF_INET6 and AF_UNIX, if available.)

SocketListenerEvent

Describes an event occurring on a #GSocketListener. See the #GSocketListener::event signal for more details.

SocketProtocol

A protocol identifier is specified when creating a #GSocket, which is a family/type specific identifier, where 0 means the default protocol for the particular family/type.

SocketType

Flags used when creating a #GSocket. Some protocols may not implement all the socket types.

TlsAuthenticationMode

The client authentication mode for a #GTlsServerConnection.

TlsCertificateRequestFlags

Flags for g_tls_interaction_request_certificate(), g_tls_interaction_request_certificate_async(), and g_tls_interaction_invoke_request_certificate().

TlsChannelBindingErrorv2_66

An error code used with G_TLS_CHANNEL_BINDING_ERROR in a #GError to indicate a TLS channel binding retrieval error.

TlsChannelBindingTypev2_66

The type of TLS channel binding data to retrieve from #GTlsConnection or #GDtlsConnection, as documented by RFC 5929 or RFC 9266. The tls-unique-for-telnet binding type is not currently implemented.

TlsDatabaseLookupFlags

Flags for g_tls_database_lookup_certificate_for_handle(), g_tls_database_lookup_certificate_issuer(), and g_tls_database_lookup_certificates_issued_by().

TlsError

An error code used with G_TLS_ERROR in a #GError returned from a TLS-related routine.

TlsInteractionResult

#GTlsInteractionResult is returned by various functions in #GTlsInteraction when finishing an interaction request.

TlsProtocolVersionv2_70

The TLS or DTLS protocol version used by a #GTlsConnection or #GDtlsConnection. The integer values of these versions are sequential to ensure newer known protocol versions compare greater than older known versions. Any known DTLS protocol version will compare greater than any SSL or TLS protocol version. The protocol version may be Unknown if the TLS backend supports a newer protocol version that GLib does not yet know about. This means that it’s possible for an unknown DTLS protocol version to compare less than the TLS protocol versions.

TlsRehandshakeModeDeprecated

When to allow rehandshaking. See g_tls_connection_set_rehandshake_mode().

UnixSocketAddressTypeUnix

The type of name used by a #GUnixSocketAddress. Path indicates a traditional unix domain socket bound to a filesystem path. Anonymous indicates a socket not bound to any name (eg, a client-side socket, or a socket created with socketpair()).

ZlibCompressorFormat

Used to select the type of data format to use for #GZlibDecompressor and #GZlibCompressor.

Statics

DEBUG_CONTROLLER_EXTENSION_POINT_NAMEv2_72

Extension point for debug control functionality. See [Extending GIO][extending-gio].

DRIVE_IDENTIFIER_KIND_UNIX_DEVICEv2_58

The string used to obtain a Unix device path with g_drive_get_identifier().

FILE_ATTRIBUTE_ACCESS_CAN_DELETE

A key in the “access” namespace for checking deletion privileges.

FILE_ATTRIBUTE_ACCESS_CAN_EXECUTE

A key in the “access” namespace for getting execution privileges.

FILE_ATTRIBUTE_ACCESS_CAN_READ

A key in the “access” namespace for getting read privileges.

FILE_ATTRIBUTE_ACCESS_CAN_RENAME

A key in the “access” namespace for checking renaming privileges.

FILE_ATTRIBUTE_ACCESS_CAN_TRASH

A key in the “access” namespace for checking trashing privileges.

FILE_ATTRIBUTE_ACCESS_CAN_WRITE

A key in the “access” namespace for getting write privileges.

FILE_ATTRIBUTE_DOS_IS_ARCHIVE

A key in the “dos” namespace for checking if the file’s archive flag is set.

FILE_ATTRIBUTE_DOS_IS_MOUNTPOINTv2_60

A key in the “dos” namespace for checking if the file is a NTFS mount point (a volume mount or a junction point).

FILE_ATTRIBUTE_DOS_IS_SYSTEM

A key in the “dos” namespace for checking if the file’s backup flag is set.

FILE_ATTRIBUTE_DOS_REPARSE_POINT_TAGv2_60

A key in the “dos” namespace for getting the file NTFS reparse tag.

FILE_ATTRIBUTE_ETAG_VALUE

A key in the “etag” namespace for getting the value of the file’s entity tag.

FILE_ATTRIBUTE_FILESYSTEM_FREE

A key in the “filesystem” namespace for getting the number of bytes of free space left on the file system.

FILE_ATTRIBUTE_FILESYSTEM_READONLY

A key in the “filesystem” namespace for checking if the file system is read only.

FILE_ATTRIBUTE_FILESYSTEM_REMOTE

A key in the “filesystem” namespace for checking if the file system is remote.

FILE_ATTRIBUTE_FILESYSTEM_SIZE

A key in the “filesystem” namespace for getting the total size (in bytes) of the file system, used in g_file_query_filesystem_info().

FILE_ATTRIBUTE_FILESYSTEM_TYPE

A key in the “filesystem” namespace for getting the file system’s type.

FILE_ATTRIBUTE_FILESYSTEM_USED

A key in the “filesystem” namespace for getting the number of bytes used by data on the file system.

FILE_ATTRIBUTE_FILESYSTEM_USE_PREVIEW

A key in the “filesystem” namespace for hinting a file manager application whether it should preview (e.g. thumbnail) files on the file system.

FILE_ATTRIBUTE_GVFS_BACKEND

A key in the “gvfs” namespace that gets the name of the current GVFS backend in use.

FILE_ATTRIBUTE_ID_FILE

A key in the “id” namespace for getting a file identifier.

FILE_ATTRIBUTE_ID_FILESYSTEM

A key in the “id” namespace for getting the file system identifier.

FILE_ATTRIBUTE_MOUNTABLE_CAN_EJECT

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be ejected.

FILE_ATTRIBUTE_MOUNTABLE_CAN_MOUNT

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) is mountable.

FILE_ATTRIBUTE_MOUNTABLE_CAN_POLL

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be polled.

FILE_ATTRIBUTE_MOUNTABLE_CAN_START

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be started.

FILE_ATTRIBUTE_MOUNTABLE_CAN_START_DEGRADED

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be started degraded.

FILE_ATTRIBUTE_MOUNTABLE_CAN_STOP

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be stopped.

FILE_ATTRIBUTE_MOUNTABLE_CAN_UNMOUNT

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) is unmountable.

FILE_ATTRIBUTE_MOUNTABLE_HAL_UDI

A key in the “mountable” namespace for getting the HAL UDI for the mountable file.

FILE_ATTRIBUTE_MOUNTABLE_IS_MEDIA_CHECK_AUTOMATIC

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) is automatically polled for media.

FILE_ATTRIBUTE_MOUNTABLE_START_STOP_TYPE

A key in the “mountable” namespace for getting the #GDriveStartStopType.

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE

A key in the “mountable” namespace for getting the unix device.

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE_FILE

A key in the “mountable” namespace for getting the unix device file.

FILE_ATTRIBUTE_OWNER_GROUP

A key in the “owner” namespace for getting the file owner’s group.

FILE_ATTRIBUTE_OWNER_USER

A key in the “owner” namespace for getting the user name of the file’s owner.

FILE_ATTRIBUTE_OWNER_USER_REAL

A key in the “owner” namespace for getting the real name of the user that owns the file.

FILE_ATTRIBUTE_PREVIEW_ICON

A key in the “preview” namespace for getting a #GIcon that can be used to get preview of the file.

FILE_ATTRIBUTE_RECENT_MODIFIED

A key in the “recent” namespace for getting time, when the metadata for the file in recent:/// was last changed.

FILE_ATTRIBUTE_SELINUX_CONTEXT

A key in the “selinux” namespace for getting the file’s SELinux context.

FILE_ATTRIBUTE_STANDARD_ALLOCATED_SIZE

A key in the “standard” namespace for getting the amount of disk space that is consumed by the file (in bytes).

FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE

A key in the “standard” namespace for getting the content type of the file.

FILE_ATTRIBUTE_STANDARD_COPY_NAME

A key in the “standard” namespace for getting the copy name of the file.

FILE_ATTRIBUTE_STANDARD_DESCRIPTION

A key in the “standard” namespace for getting the description of the file.

FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME

A key in the “standard” namespace for getting the display name of the file.

FILE_ATTRIBUTE_STANDARD_EDIT_NAME

A key in the “standard” namespace for edit name of the file.

FILE_ATTRIBUTE_STANDARD_FAST_CONTENT_TYPE

A key in the “standard” namespace for getting the fast content type.

FILE_ATTRIBUTE_STANDARD_ICON

A key in the “standard” namespace for getting the icon for the file.

FILE_ATTRIBUTE_STANDARD_IS_BACKUP

A key in the “standard” namespace for checking if a file is a backup file.

FILE_ATTRIBUTE_STANDARD_IS_HIDDEN

A key in the “standard” namespace for checking if a file is hidden.

FILE_ATTRIBUTE_STANDARD_IS_SYMLINK

A key in the “standard” namespace for checking if the file is a symlink. Typically the actual type is something else, if we followed the symlink to get the type.

FILE_ATTRIBUTE_STANDARD_IS_VIRTUAL

A key in the “standard” namespace for checking if a file is virtual.

FILE_ATTRIBUTE_STANDARD_IS_VOLATILE

A key in the “standard” namespace for checking if a file is volatile. This is meant for opaque, non-POSIX-like backends to indicate that the URI is not persistent. Applications should look at FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET for the persistent URI.

FILE_ATTRIBUTE_STANDARD_NAME

A key in the “standard” namespace for getting the name of the file.

FILE_ATTRIBUTE_STANDARD_SIZE

A key in the “standard” namespace for getting the file’s size (in bytes).

FILE_ATTRIBUTE_STANDARD_SORT_ORDER

A key in the “standard” namespace for setting the sort order of a file.

FILE_ATTRIBUTE_STANDARD_SYMBOLIC_ICON

A key in the “standard” namespace for getting the symbolic icon for the file.

FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET

A key in the “standard” namespace for getting the symlink target, if the file is a symlink.

FILE_ATTRIBUTE_STANDARD_TARGET_URI

A key in the “standard” namespace for getting the target URI for the file, in the case of FileType::Shortcut or FileType::Mountable files.

FILE_ATTRIBUTE_STANDARD_TYPE

A key in the “standard” namespace for storing file types.

FILE_ATTRIBUTE_THUMBNAILING_FAILED

A key in the “thumbnail” namespace for checking if thumbnailing failed.

FILE_ATTRIBUTE_THUMBNAILING_FAILED_LARGEv2_76

A key in the “thumbnail” namespace for checking if thumbnailing failed for the large image.

FILE_ATTRIBUTE_THUMBNAILING_FAILED_NORMALv2_76

A key in the “thumbnail” namespace for checking if thumbnailing failed for the normal image.

FILE_ATTRIBUTE_THUMBNAILING_FAILED_XLARGEv2_76

A key in the “thumbnail” namespace for checking if thumbnailing failed for the x-large image.

FILE_ATTRIBUTE_THUMBNAILING_FAILED_XXLARGEv2_76

A key in the “thumbnail” namespace for checking if thumbnailing failed for the xx-large image.

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID

A key in the “thumbnail” namespace for checking whether the thumbnail is outdated.

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID_LARGEv2_76

A key in the “thumbnail” namespace for checking whether the large thumbnail is outdated.

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID_NORMALv2_76

A key in the “thumbnail” namespace for checking whether the normal thumbnail is outdated.

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID_XLARGEv2_76

A key in the “thumbnail” namespace for checking whether the x-large thumbnail is outdated.

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID_XXLARGEv2_76

A key in the “thumbnail” namespace for checking whether the xx-large thumbnail is outdated.

FILE_ATTRIBUTE_THUMBNAIL_PATH

A key in the “thumbnail” namespace for getting the path to the thumbnail image with the biggest size available.

FILE_ATTRIBUTE_THUMBNAIL_PATH_LARGEv2_76

A key in the “thumbnail” namespace for getting the path to the large thumbnail image.

FILE_ATTRIBUTE_THUMBNAIL_PATH_NORMALv2_76

A key in the “thumbnail” namespace for getting the path to the normal thumbnail image.

FILE_ATTRIBUTE_THUMBNAIL_PATH_XLARGEv2_76

A key in the “thumbnail” namespace for getting the path to the x-large thumbnail image.

FILE_ATTRIBUTE_THUMBNAIL_PATH_XXLARGEv2_76

A key in the “thumbnail” namespace for getting the path to the xx-large thumbnail image.

FILE_ATTRIBUTE_TIME_ACCESS

A key in the “time” namespace for getting the time the file was last accessed.

FILE_ATTRIBUTE_TIME_ACCESS_NSECv2_74

A key in the “time” namespace for getting the nanoseconds of the time the file was last accessed. This should be used in conjunction with #G_FILE_ATTRIBUTE_TIME_ACCESS. Corresponding #GFileAttributeType is FileAttributeType::Uint32.

FILE_ATTRIBUTE_TIME_ACCESS_USEC

A key in the “time” namespace for getting the microseconds of the time the file was last accessed.

FILE_ATTRIBUTE_TIME_CHANGED

A key in the “time” namespace for getting the time the file was last changed.

FILE_ATTRIBUTE_TIME_CHANGED_NSECv2_74

A key in the “time” namespace for getting the nanoseconds of the time the file was last changed. This should be used in conjunction with #G_FILE_ATTRIBUTE_TIME_CHANGED. Corresponding #GFileAttributeType is FileAttributeType::Uint32.

FILE_ATTRIBUTE_TIME_CHANGED_USEC

A key in the “time” namespace for getting the microseconds of the time the file was last changed.

FILE_ATTRIBUTE_TIME_CREATED

A key in the “time” namespace for getting the time the file was created.

FILE_ATTRIBUTE_TIME_CREATED_NSECv2_74

A key in the “time” namespace for getting the nanoseconds of the time the file was created. This should be used in conjunction with #G_FILE_ATTRIBUTE_TIME_CREATED. Corresponding #GFileAttributeType is FileAttributeType::Uint32.

FILE_ATTRIBUTE_TIME_CREATED_USEC

A key in the “time” namespace for getting the microseconds of the time the file was created.

FILE_ATTRIBUTE_TIME_MODIFIED

A key in the “time” namespace for getting the time the file was last modified.

FILE_ATTRIBUTE_TIME_MODIFIED_NSECv2_74

A key in the “time” namespace for getting the nanoseconds of the time the file was last modified. This should be used in conjunction with #G_FILE_ATTRIBUTE_TIME_MODIFIED. Corresponding #GFileAttributeType is FileAttributeType::Uint32.

FILE_ATTRIBUTE_TIME_MODIFIED_USEC

A key in the “time” namespace for getting the microseconds of the time the file was last modified.

FILE_ATTRIBUTE_TRASH_DELETION_DATE

A key in the “trash” namespace for getting the deletion date and time of a file inside the trash:/// folder.

FILE_ATTRIBUTE_TRASH_ITEM_COUNT

A key in the “trash” namespace for getting the number of (toplevel) items that are present in the trash:/// folder.

FILE_ATTRIBUTE_TRASH_ORIG_PATH

A key in the “trash” namespace for getting the original path of a file inside the trash:/// folder before it was trashed.

FILE_ATTRIBUTE_UNIX_BLOCKS

A key in the “unix” namespace for getting the number of blocks allocated for the file.

FILE_ATTRIBUTE_UNIX_BLOCK_SIZE

A key in the “unix” namespace for getting the block size for the file system.

FILE_ATTRIBUTE_UNIX_DEVICE

A key in the “unix” namespace for getting the device id of the device the file is located on (see stat() documentation).

FILE_ATTRIBUTE_UNIX_GID

A key in the “unix” namespace for getting the group ID for the file.

FILE_ATTRIBUTE_UNIX_INODE

A key in the “unix” namespace for getting the inode of the file.

FILE_ATTRIBUTE_UNIX_IS_MOUNTPOINT

A key in the “unix” namespace for checking if the file represents a UNIX mount point.

FILE_ATTRIBUTE_UNIX_MODE

A key in the “unix” namespace for getting the mode of the file (e.g. whether the file is a regular file, symlink, etc).

FILE_ATTRIBUTE_UNIX_NLINK

A key in the “unix” namespace for getting the number of hard links for a file.

FILE_ATTRIBUTE_UNIX_RDEV

A key in the “unix” namespace for getting the device ID for the file (if it is a special file).

FILE_ATTRIBUTE_UNIX_UID

A key in the “unix” namespace for getting the user ID for the file.

MEMORY_MONITOR_EXTENSION_POINT_NAMEv2_64

Extension point for memory usage monitoring functionality. See [Extending GIO][extending-gio].

MENU_ATTRIBUTE_ACTION

The menu item attribute which holds the action name of the item. Action names are namespaced with an identifier for the action group in which the action resides. For example, “win.” for window-specific actions and “app.” for application-wide actions.

MENU_ATTRIBUTE_ACTION_NAMESPACE

The menu item attribute that holds the namespace for all action names in menus that are linked from this item.

MENU_ATTRIBUTE_ICON

The menu item attribute which holds the icon of the item.

MENU_ATTRIBUTE_LABEL

The menu item attribute which holds the label of the item.

MENU_ATTRIBUTE_TARGET

The menu item attribute which holds the target with which the item’s action will be activated.

MENU_LINK_SECTION

The name of the link that associates a menu item with a section. The linked menu will usually be shown in place of the menu item, using the item’s label as a header.

MENU_LINK_SUBMENU

The name of the link that associates a menu item with a submenu.

NATIVE_VOLUME_MONITOR_EXTENSION_POINT_NAME

NETWORK_MONITOR_EXTENSION_POINT_NAME

Extension point for network status monitoring functionality. See [Extending GIO][extending-gio].

POWER_PROFILE_MONITOR_EXTENSION_POINT_NAMEv2_70

Extension point for power profile usage monitoring functionality. See [Extending GIO][extending-gio].

PROXY_EXTENSION_POINT_NAME

Extension point for proxy functionality. See [Extending GIO][extending-gio].

PROXY_RESOLVER_EXTENSION_POINT_NAME

Extension point for proxy resolving functionality. See [Extending GIO][extending-gio].

SETTINGS_BACKEND_EXTENSION_POINT_NAME

Extension point for #GSettingsBackend functionality.

TLS_BACKEND_EXTENSION_POINT_NAME

Extension point for TLS functionality via #GTlsBackend. See [Extending GIO][extending-gio].

TLS_DATABASE_PURPOSE_AUTHENTICATE_CLIENT

The purpose used to verify the client certificate in a TLS connection. Used by TLS servers.

TLS_DATABASE_PURPOSE_AUTHENTICATE_SERVER

The purpose used to verify the server certificate in a TLS connection. This is the most common purpose in use. Used by TLS clients.

VFS_EXTENSION_POINT_NAME

Extension point for #GVfs functionality. See [Extending GIO][extending-gio].

VOLUME_IDENTIFIER_KIND_CLASS

The string used to obtain the volume class with g_volume_get_identifier().

VOLUME_IDENTIFIER_KIND_HAL_UDIDeprecated

The string used to obtain a Hal UDI with g_volume_get_identifier().

VOLUME_IDENTIFIER_KIND_LABEL

The string used to obtain a filesystem label with g_volume_get_identifier().

VOLUME_IDENTIFIER_KIND_NFS_MOUNT

The string used to obtain a NFS mount with g_volume_get_identifier().

VOLUME_IDENTIFIER_KIND_UNIX_DEVICE

The string used to obtain a Unix device path with g_volume_get_identifier().

VOLUME_IDENTIFIER_KIND_UUID

The string used to obtain a UUID with g_volume_get_identifier().

VOLUME_MONITOR_EXTENSION_POINT_NAME

Extension point for volume monitor functionality. See [Extending GIO][extending-gio].

Functions

bus_own_name

Starts acquiring @name on the bus specified by @bus_type and calls @name_acquired_handler and @name_lost_handler when the name is acquired respectively lost. Callbacks will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this function from.

bus_own_name_on_connection

Like g_bus_own_name() but takes a #GDBusConnection instead of a #GBusType.

bus_unown_name

bus_unwatch_name

bus_watch_name

Starts watching @name on the bus specified by @bus_type and calls @name_appeared_handler and @name_vanished_handler when the name is known to have an owner respectively known to lose its owner. Callbacks will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this function from.

bus_watch_name_on_connection

Like g_bus_watch_name() but takes a #GDBusConnection instead of a #GBusType.

spawn_blocking

Runs a blocking I/O task on the I/O thread pool.