Crate gio

Rust GIO bindings

Rust bindings and wrappers for GIO, part of gtk-rs-core.

GIO 2.48 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.56.0.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gio = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gio" }

Avoid mixing versioned and git crates like this:

[dependencies]
gio = "0.13"
gio = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gio" }

See Also

glib

License

gio is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use glib;

Modules

functions

prelude

Traits and essential types intended for blanket imports.

subclass

Macros

resources_register_include

Include gresources generated with compile_resources and register with glib. path is relative to OUTDIR.

Structs

Action

Action represents a single named action.

ActionGroup

ActionGroup represents a group of actions. Actions can be used to expose functionality in a structured way, either from one part of a program to another, or to the outside world. Action groups are often used together with a MenuModel that provides additional representation data for displaying the actions to the user, e.g. in a menu.

ActionGroupExportId

ActionMap

The GActionMap interface is implemented by ActionGroup implementations that operate by containing a number of named Action instances, such as SimpleActionGroup.

AppInfo

AppInfo and AppLaunchContext are used for describing and launching applications installed on the system.

AppInfoCreateFlags

Flags used when creating a AppInfo.

AppInfoMonitor

AppInfoMonitor is a very simple object used for monitoring the app info database for changes (ie: newly installed or removed applications).

AppLaunchContext

Integrating the launch with the launching application. This is used to handle for instance startup notification and launching the new application on the same screen as the launching window.

Application

A Application is the foundation of an application. It wraps some low-level platform-specific services and is intended to act as the foundation for higher-level application classes such as GtkApplication or MxApplication. In general, you should not use this class outside of a higher level framework.

ApplicationCommandLine

ApplicationCommandLine represents a command-line invocation of an application. It is created by Application and emitted in the signal::Application::command-line signal and virtual function.

ApplicationFlags

Flags used to define the behaviour of a Application.

AskPasswordFlags

AskPasswordFlags are used to request specific information from the user, or to notify the user of their choices in an authentication situation.

AsyncInitable

This is the asynchronous version of Initable; it behaves the same in all ways except that initialization is asynchronous. For more details see the descriptions on Initable.

AsyncResult

Provides a base class for implementing asynchronous function results.

BindingBuilder

BufferedInputStream

Buffered input stream implements FilterInputStream and provides for buffered reads.

BufferedOutputStream

Buffered output stream implements FilterOutputStream and provides for buffered writes.

BusNameOwnerFlags

Flags used in g_bus_own_name().

BusNameWatcherFlags

Flags used in g_bus_watch_name().

BytesIcon

BytesIcon specifies an image held in memory in a common format (usually png) to be used as icon.

Cancellable

GCancellable is a thread-safe operation cancellation stack used throughout GIO to allow for cancellation of synchronous and asynchronous operations.

CharsetConverter

CharsetConverter is an implementation of Converter based on GIConv.

Converter

Converter is implemented by objects that convert binary data in various ways. The conversion can be stateful and may fail at any place.

ConverterFlags

Flags used when calling a ConverterExtManual::convert().

ConverterInputStream

Converter input stream implements InputStream and allows conversion of data of various types during reading.

ConverterOutputStream

Converter output stream implements OutputStream and allows conversion of data of various types during reading.

Credentials

The Credentials type is a reference-counted wrapper for native credentials. This information is typically used for identifying, authenticating and authorizing other processes.

DBusActionGroup

DBusActionGroup is an implementation of the ActionGroup interface that can be used as a proxy for an action group that is exported over D-Bus with DBusConnection::export_action_group().

DBusArgInfo

Information about an argument for a method or a signal.

DBusAuthObserver

The DBusAuthObserver type provides a mechanism for participating in how a DBusServer (or a DBusConnection) authenticates remote peers. Simply instantiate a DBusAuthObserver and connect to the signals you are interested in. Note that new signals may be added in the future

DBusCallFlags

Flags used in DBusConnection::call() and similar APIs.

DBusCapabilityFlags

Capabilities negotiated with the remote peer.

DBusConnection

The DBusConnection type is used for D-Bus connections to remote peers such as a message buses. It is a low-level API that offers a lot of flexibility. For instance, it lets you establish a connection over any transport that can by represented as a IOStream.

DBusConnectionFlags

Flags used when creating a new DBusConnection.

DBusInterface

The DBusInterface type is the base type for D-Bus interfaces both on the service side (see DBusInterfaceSkeleton) and client side (see DBusProxy).

DBusInterfaceInfo

Information about a D-Bus interface.

DBusInterfaceSkeleton

Abstract base class for D-Bus interfaces on the service side.

DBusInterfaceSkeletonFlags

Flags describing the behavior of a DBusInterfaceSkeleton instance.

DBusMenuModel

DBusMenuModel is an implementation of MenuModel that can be used as a proxy for a menu model that is exported over D-Bus with DBusConnection::export_menu_model().

DBusMessage

A type for representing D-Bus messages that can be sent or received on a DBusConnection.

DBusMessageFlags

Message flags used in DBusMessage.

DBusMethodInfo

Information about a method on an D-Bus interface.

DBusMethodInvocation

Instances of the DBusMethodInvocation class are used when handling D-Bus method calls. It provides a way to asynchronously return results and errors.

DBusNodeInfo

Information about nodes in a remote object hierarchy.

DBusObject

The DBusObject type is the base type for D-Bus objects on both the service side (see GDBusObjectSkeleton) and the client side (see GDBusObjectProxy). It is essentially just a container of interfaces.

DBusPropertyInfo

Information about a D-Bus property on a D-Bus interface.

DBusProxy

DBusProxy is a base class used for proxies to access a D-Bus interface on a remote object. A DBusProxy can be constructed for both well-known and unique names.

DBusProxyFlags

Flags used when constructing an instance of a DBusProxy derived class.

DBusSendMessageFlags

Flags used when sending GDBusMessages on a DBusConnection.

DBusServer

DBusServer is a helper for listening to and accepting D-Bus connections. This can be used to create a new D-Bus server, allowing two peers to use the D-Bus protocol for their own specialized communication. A server instance provided in this way will not perform message routing or implement the org.freedesktop.DBus interface.

DBusServerFlags

Flags used when creating a DBusServer.

DBusSignalFlags

Flags used when subscribing to signals via DBusConnection::signal_subscribe().

DBusSignalInfo

Information about a signal on a D-Bus interface.

DataInputStream

Data input stream implements InputStream and includes functions for reading structured data directly from a binary input stream.

DataOutputStream

Data output stream implements OutputStream and includes functions for writing data directly to an output stream.

DebugController

DebugController is an interface to expose control of debugging features and debug output.

DebugControllerDBus

DebugControllerDBus is an implementation of DebugController which exposes debug settings as a D-Bus object.

DesktopAppInfo

DesktopAppInfo is an implementation of AppInfo based on desktop files.

Drive

Drive - this represent a piece of hardware connected to the machine. It’s generally only created for removable hardware or hardware with removable media.

DriveStartFlags

Flags used when starting a drive.

Emblem

Emblem is an implementation of Icon that supports having an emblem, which is an icon with additional properties. It can than be added to a EmblemedIcon.

EmblemedIcon

EmblemedIcon is an implementation of Icon that supports adding an emblem to an icon. Adding multiple emblems to an icon is ensured via EmblemedIconExt::add_emblem().

File

File is a high level abstraction for manipulating files on a virtual file system. GFiles are lightweight, immutable objects that do no I/O upon creation. It is necessary to understand that File objects do not represent files, merely an identifier for a file. All file content I/O is implemented as streaming operations (see InputStream and OutputStream).

FileAttributeInfo

Information about a specific attribute.

FileAttributeInfoFlags

Flags specifying the behaviour of an attribute.

FileAttributeInfoList

Acts as a lightweight registry for possible valid file attributes. The registry stores Key-Value pair formats as GFileAttributeInfos.

FileAttributeMatcher

Determines if a string matches a file attribute.

FileAttributematcherIter

FileCopyFlags

Flags used when copying or moving files.

FileCreateFlags

Flags used when an operation may create a file.

FileEnumerator

FileEnumerator allows you to operate on a set of GFiles, returning a FileInfo structure for each file enumerated (e.g. FileExt::enumerate_children() will return a FileEnumerator for each of the children within a directory).

FileIOStream

GFileIOStream provides io streams that both read and write to the same file handle.

FileIcon

FileIcon specifies an icon by pointing to an image file to be used as icon.

FileInfo

Functionality for manipulating basic metadata for files. FileInfo implements methods for getting information that all files should contain, and allows for manipulation of extended attributes.

FileInputStream

GFileInputStream provides input streams that take their content from a file.

FileMeasureFlags

Flags that can be used with FileExtManual::measure_disk_usage().

FileMonitor

Monitors a file or directory for changes.

FileMonitorFlags

Flags used to set what a FileMonitor will watch for.

FileOutputStream

GFileOutputStream provides output streams that write their content to a file.

FileQueryInfoFlags

Flags used when querying a FileInfo.

FilenameCompleter

Completes partial file and directory names given a partial string by looking in the file system for clues. Can return a list of possible completion strings for widget implementations.

FilterId

FilterInputStream

Base class for input stream implementations that perform some kind of filtering operation on a base stream. Typical examples of filtering operations are character set conversion, compression and byte order flipping.

FilterOutputStream

Base class for output stream implementations that perform some kind of filtering operation on a base stream. Typical examples of filtering operations are character set conversion, compression and byte order flipping.

GioFuture

GioFutureResult

IOExtension

The implementation of an IOExtensionPoint. IOExtension is an opaque data structure and can only be accessed using the following functions.

IOExtensionPoint

An extension point provides a mechanism to extend the functionality of a library or application. Each extension point is identified by a name, and it may optionally require that any implementation must be of a certain type. IOExtensionPoint is an opaque data structure and can only be accessed using the following functions.

IOExtensionPointBuilder

Builder for extension points.

IOStream

GIOStream represents an object that has both read and write streams. Generally the two streams act as separate input and output streams, but they share some common resources and state. For instance, for seekable streams, both streams may use the same position.

IOStreamAsyncReadWrite

IOStreamSpliceFlags

GIOStreamSpliceFlags determine how streams should be spliced.

Icon

Icon is a very minimal interface for icons. It provides functions for checking the equality of two icons, hashing of icons and serializing an icon to and from strings.

InetAddress

InetAddress represents an IPv4 or IPv6 internet address. Use ResolverExt::lookup_by_name() or ResolverExt::lookup_by_name_async() to look up the InetAddress for a hostname. Use ResolverExt::lookup_by_address() or ResolverExt::lookup_by_address_async() to look up the hostname for a InetAddress.

InetAddressMask

InetAddressMask represents a range of IPv4 or IPv6 addresses described by a base address and a length indicating how many bits of the base address are relevant for matching purposes. These are often given in string form. Eg, “10.0.0.0/8”, or “fe80::/10”.

InetSocketAddress

An IPv4 or IPv6 socket address; that is, the combination of a InetAddress and a port number.

Initable

Initable is implemented by objects that can fail during initialization. If an object implements this interface then it must be initialized as the first thing after construction, either via InitableExt::init() or AsyncInitableExt::init_async() (the latter is only available if it also implements AsyncInitable).

InputStream

InputStream has functions to read from a stream (InputStreamExtManual::read()), to close a stream (InputStreamExt::close()) and to skip some content (InputStreamExt::skip()).

InputStreamAsyncBufRead

InputStreamAsyncRead

InputStreamRead

ListModel

ListModel is an interface that represents a mutable list of GObjects. Its main intention is as a model for various widgets in user interfaces, such as list views, but it can also be used as a convenient method of returning lists of data, with support for updates.

ListStore

ListStore is a simple implementation of ListModel that stores all items in memory.

LoadableIcon

Extends the Icon interface and adds the ability to load icons from streams.

LocalTask

LocalTask provides idiomatic access to gio’s GTask API, for instance by being generic over their value type, while not completely departing from the underlying C API. LocalTask does not require its value to be Send and Sync and thus is useful to to implement gio style asynchronous tasks that run in the glib main loop. If you need to run tasks in threads see the Task type.

MemoryInputStream

MemoryInputStream is a class for using arbitrary memory chunks as input for GIO streaming input operations.

MemoryMonitor

MemoryMonitor will monitor system memory and suggest to the application when to free memory so as to leave more room for other applications. It is implemented on Linux using the Low Memory Monitor (API documentation).

MemoryOutputStream

MemoryOutputStream is a class for using arbitrary memory chunks as output for GIO streaming output operations.

Menu

Menu is a simple implementation of MenuModel. You populate a Menu by adding MenuItem instances to it.

MenuAttributeIter

MenuAttributeIter is an opaque structure type. You must access it using the functions below.

MenuItem

MenuItem is an opaque structure type. You must access it using the functions below.

MenuLinkIter

MenuLinkIter is an opaque structure type. You must access it using the functions below.

MenuModel

MenuModel represents the contents of a menu – an ordered list of menu items. The items are associated with actions, which can be activated through them. Items can be grouped in sections, and may have submenus associated with them. Both items and sections usually have some representation data, such as labels or icons. The type of the associated action (ie whether it is stateful, and what kind of state it has) can influence the representation of the item.

MenuModelExportId

Mount

The Mount interface represents user-visible mounts. Note, when porting from GnomeVFS, Mount is the moral equivalent of GnomeVFSVolume.

MountMountFlags

Flags used when mounting a mount.

MountOperation

MountOperation provides a mechanism for interacting with the user. It can be used for authenticating mountable operations, such as loop mounting files, hard drive partitions or server locations. It can also be used to ask the user questions or show a list of applications preventing unmount or eject operations from completing.

MountUnmountFlags

Flags used when an unmounting a mount.

NetworkAddress

NetworkAddress provides an easy way to resolve a hostname and then attempt to connect to that host, handling the possibility of multiple IP addresses and multiple address families.

NetworkMonitor

NetworkMonitor provides an easy-to-use cross-platform API for monitoring network connectivity. On Linux, the available implementations are based on the kernel’s netlink interface and on NetworkManager.

NetworkService

Like NetworkAddress does with hostnames, NetworkService provides an easy way to resolve a SRV record, and then attempt to connect to one of the hosts that implements that service, handling service priority/weighting, multiple IP addresses, and multiple address families.

Notification

Notification is a mechanism for creating a notification to be shown to the user – typically as a pop-up notification presented by the desktop environment shell.

OutputStream

OutputStream has functions to write to a stream (OutputStreamExt::write()), to close a stream (OutputStreamExt::close()) and to flush pending writes (OutputStreamExt::flush()).

OutputStreamAsyncWrite

OutputStreamSpliceFlags

GOutputStreamSpliceFlags determine how streams should be spliced.

OutputStreamWrite

OwnerId

Permission

A Permission represents the status of the caller’s permission to perform a certain action.

PollableInputStream

PollableInputStream is implemented by GInputStreams that can be polled for readiness to read. This can be used when interfacing with a non-GIO API that expects UNIX-file-descriptor-style asynchronous I/O rather than GIO-style.

PollableOutputStream

PollableOutputStream is implemented by GOutputStreams that can be polled for readiness to write. This can be used when interfacing with a non-GIO API that expects UNIX-file-descriptor-style asynchronous I/O rather than GIO-style.

PowerProfileMonitor

PowerProfileMonitor makes it possible for applications as well as OS components to monitor system power profiles and act upon them. It currently only exports whether the system is in “Power Saver” mode (known as “Low Power” mode on some systems).

PropertyAction

A PropertyAction is a way to get a Action with a state value reflecting and controlling the value of a glib::Object property.

Proxy

A Proxy handles connecting to a remote host via a given type of proxy server. It is implemented by the ‘gio-proxy’ extension point. The extensions are named after their proxy protocol name. As an example, a SOCKS5 proxy implementation can be retrieved with the name ‘socks5’ using the function IOExtensionPoint::extension_by_name().

ProxyAddress

Support for proxied InetSocketAddress.

ProxyResolver

ProxyResolver provides synchronous and asynchronous network proxy resolution. ProxyResolver is used within SocketClient through the method SocketConnectableExt::proxy_enumerate().

ReadInputStream

RegistrationId

RemoteActionGroup

The GRemoteActionGroup interface is implemented by ActionGroup instances that either transmit action invocations to other processes or receive action invocations in the local process from other processes.

Resolver

Resolver provides cancellable synchronous and asynchronous DNS resolution, for hostnames (ResolverExt::lookup_by_address(), ResolverExt::lookup_by_name() and their async variants) and SRV (service) records (ResolverExt::lookup_service()).

ResolverNameLookupFlagsv2_60

Flags to modify lookup behavior.

Resource

Applications and libraries often contain binary or textual data that is really part of the application, rather than user data. For instance GtkBuilder .ui files, splashscreen images, GMenu markup XML, CSS files, icons, etc. These are often shipped as files in $datadir/appname, or manually included as literal strings in the code.

ResourceLookupFlags

GResourceLookupFlags determine how resource path lookups are handled.

Seekable

Seekable is implemented by streams (implementations of InputStream or OutputStream) that support seeking.

Settings

The Settings class provides a convenient API for storing and retrieving application settings.

SettingsBackend

The SettingsBackend interface defines a generic interface for non-strictly-typed data that is stored in a hierarchy. To implement an alternative storage backend for Settings, you need to implement the SettingsBackend interface and then make it implement the extension point SETTINGS_BACKEND_EXTENSION_POINT_NAME.

SettingsBindFlags

Flags used when creating a binding. These flags determine in which direction the binding works. The default is to synchronize in both directions.

SettingsSchema

The SettingsSchemaSource and SettingsSchema APIs provide a mechanism for advanced control over the loading of schemas and a mechanism for introspecting their content.

SettingsSchemaKey

SettingsSchemaKey is an opaque data structure and can only be accessed using the following functions.

SettingsSchemaSource

This is an opaque structure type. You may not access it directly.

SignalSubscriptionId

SimpleAction

A SimpleAction is the obvious simple implementation of the Action interface. This is the easiest way to create an action for purposes of adding it to a SimpleActionGroup.

SimpleActionGroup

SimpleActionGroup is a hash table filled with Action objects, implementing the ActionGroup and ActionMap interfaces.

SimpleIOStream

GSimpleIOStream creates a IOStream from an arbitrary InputStream and OutputStream. This allows any pair of input and output streams to be used with IOStream methods.

SimplePermission

SimplePermission is a trivial implementation of Permission that represents a permission that is either always or never allowed. The value is given at construction and doesn’t change.

SimpleProxyResolver

SimpleProxyResolver is a simple ProxyResolver implementation that handles a single default proxy, multiple URI-scheme-specific proxies, and a list of hosts that proxies should not be used for.

Socket

A Socket is a low-level networking primitive. It is a more or less direct mapping of the BSD socket API in a portable GObject based API. It supports both the UNIX socket implementations and winsock2 on Windows.

SocketAddress

SocketAddress is the equivalent of struct sockaddr in the BSD sockets API. This is an abstract class; use InetSocketAddress for internet sockets, or UnixSocketAddress for UNIX domain sockets.

SocketAddressEnumerator

SocketAddressEnumerator is an enumerator type for SocketAddress instances. It is returned by enumeration functions such as SocketConnectableExt::enumerate(), which returns a SocketAddressEnumerator to list each SocketAddress which could be used to connect to that SocketConnectable.

SocketClient

SocketClient is a lightweight high-level utility class for connecting to a network host using a connection oriented socket type.

SocketConnectable

Objects that describe one or more potential socket endpoints implement SocketConnectable. Callers can then use SocketConnectableExt::enumerate() to get a SocketAddressEnumerator to try out each socket address in turn until one succeeds, as shown in the sample code below.

SocketConnection

SocketConnection is a IOStream for a connected socket. They can be created either by SocketClient when connecting to a host, or by SocketListener when accepting a new client.

SocketListener

A SocketListener is an object that keeps track of a set of server sockets and helps you accept sockets from any of the socket, either sync or async.

SocketService

A SocketService is an object that represents a service that is provided to the network or over local sockets. When a new connection is made to the service the signal::SocketService::incoming signal is emitted.

SrvTarget

SRV (service) records are used by some network protocols to provide service-specific aliasing and load-balancing. For example, XMPP (Jabber) uses SRV records to locate the XMPP server for a domain; rather than connecting directly to “example.com” or assuming a specific server hostname like “xmpp.example.com”, an XMPP client would look up the “xmpp-client” SRV record for “example.com”, and then connect to whatever host was pointed to by that record.

Subprocess

Subprocess allows the creation of and interaction with child processes.

SubprocessFlags

Flags to define the behaviour of a Subprocess.

SubprocessLauncher

This class contains a set of options for launching child processes, such as where its standard input and output will be directed, the argument list, the environment, and more.

Task

Task provides idiomatic access to gio’s GTask API, for instance by being generic over their value type, while not completely departing from the underlying C API. Task is Send and Sync and requires its value to also be Send and Sync, thus is useful to to implement gio style asynchronous tasks that run in threads. If you need to only run tasks in glib main loop see the LocalTask type.

TcpConnection

This is the subclass of SocketConnection that is created for TCP/IP sockets.

ThemedIcon

ThemedIcon is an implementation of Icon that supports icon themes. ThemedIcon contains a list of all of the icons present in an icon theme, so that icons can be looked up quickly. ThemedIcon does not provide actual pixmaps for icons, just the icon names. Ideally something like gtk_icon_theme_choose_icon() should be used to resolve the list of names so that fallback icons work nicely with themes that inherit other themes.

ThreadedSocketService

A ThreadedSocketService is a simple subclass of SocketService that handles incoming connections by creating a worker thread and dispatching the connection to it by emitting the signal::ThreadedSocketService::run signal in the new thread.

TlsBackend

TLS (Transport Layer Security, aka SSL) and DTLS backend.

TlsCertificate

A certificate used for TLS authentication and encryption. This can represent either a certificate only (eg, the certificate received by a client from a server), or the combination of a certificate and a private key (which is needed when acting as a TlsServerConnection).

TlsCertificateFlags

A set of flags describing TLS certification validation. This can be used to describe why a particular certificate was rejected (for example, in signal::TlsConnection::accept-certificate).

TlsClientConnection

TlsClientConnection is the client-side subclass of TlsConnection, representing a client-side TLS connection.

TlsConnection

TlsConnection is the base TLS connection class type, which wraps a IOStream and provides TLS encryption on top of it. Its subclasses, TlsClientConnection and TlsServerConnection, implement client-side and server-side TLS, respectively.

TlsDatabase

TlsDatabase is used to look up certificates and other information from a certificate or key store. It is an abstract base class which TLS library specific subtypes override.

TlsDatabaseVerifyFlags

Flags for TlsDatabaseExt::verify_chain().

TlsFileDatabase

TlsFileDatabase is implemented by TlsDatabase objects which load their certificate information from a file. It is an interface which TLS library specific subtypes implement.

TlsInteraction

TlsInteraction provides a mechanism for the TLS connection and database code to interact with the user. It can be used to ask the user for passwords.

TlsPassword

Holds a password used in TLS.

TlsPasswordFlags

Various flags for the password.

TlsServerConnection

TlsServerConnection is the server-side subclass of TlsConnection, representing a server-side TLS connection.

UnixFDList

A UnixFDList contains a list of file descriptors. It owns the file descriptors that it contains, closing them when finalized.

UnixInputStream

UnixInputStream implements InputStream for reading from a UNIX file descriptor, including asynchronous operations. (If the file descriptor refers to a socket or pipe, this will use poll() to do asynchronous I/O. If it refers to a regular file, it will fall back to doing asynchronous I/O in another thread.)

UnixMountEntry

Defines a Unix mount entry (e.g. <filename>/media/cdrom</filename>). This corresponds roughly to a mtab entry.

UnixMountPoint

Defines a Unix mount point (e.g. <filename>/dev</filename>). This corresponds roughly to a fstab entry.

UnixOutputStream

UnixOutputStream implements OutputStream for writing to a UNIX file descriptor, including asynchronous operations. (If the file descriptor refers to a socket or pipe, this will use poll() to do asynchronous I/O. If it refers to a regular file, it will fall back to doing asynchronous I/O in another thread.)

UnixSocketAddress

Support for UNIX-domain (also known as local) sockets.

Vfs

Entry point for using GIO functionality.

Volume

The Volume interface represents user-visible objects that can be mounted. Note, when porting from GnomeVFS, Volume is the moral equivalent of GnomeVFSDrive.

VolumeMonitor

VolumeMonitor is for listing the user interesting devices and volumes on the computer. In other words, what a file selector or file manager would show in a sidebar.

WatcherId

WriteOutputStream

ZlibCompressor

ZlibCompressor is an implementation of Converter that compresses data using zlib.

ZlibDecompressor

ZlibDecompressor is an implementation of Converter that decompresses data compressed with zlib.

Enums

BusType

An enumeration for well-known message buses.

ConverterResult

Results returned from ConverterExtManual::convert().

CredentialsType

Enumeration describing different kinds of native credential types.

DBusMessageByteOrder

Enumeration used to describe the byte order of a D-Bus message.

DBusMessageHeaderField

Header fields used in DBusMessage.

DBusMessageType

Message types used in DBusMessage.

DataStreamByteOrder

DataStreamByteOrder is used to ensure proper endianness of streaming data sources across various machine architectures.

DataStreamNewlineType

DataStreamNewlineType is used when checking for or setting the line endings for a given file.

DriveStartStopType

Enumeration describing how a drive can be started/stopped.

EmblemOrigin

GEmblemOrigin is used to add information about the origin of the emblem to Emblem.

FileAttributeStatus

Used by FileExt::set_attributes_from_info() when setting file attributes.

FileAttributeType

The data types for file attributes.

FileMonitorEvent

Specifies what type of event a monitor event is.

FileType

Indicates the file’s on-disk type.

IOErrorEnum

Error codes returned by GIO functions.

InetAddressBytes

MemoryMonitorWarningLevelv2_64

Memory availability warning levels.

MountOperationResult

MountOperationResult is returned as a result when a request for information is send by the mounting operation.

NetworkConnectivity

The host’s network connectivity state, as reported by NetworkMonitor.

NotificationPriority

Priority levels for GNotifications.

PasswordSave

PasswordSave is used to indicate the lifespan of a saved password.

ResolverError

An error code used with G_RESOLVER_ERROR in a glib::Error returned from a Resolver routine.

ResolverRecordType

The type of record that ResolverExt::lookup_records() or ResolverExt::lookup_records_async() should retrieve. The records are returned as lists of glib::Variant tuples. Each record type has different values in the variant tuples returned.

ResourceError

An error code used with G_RESOURCE_ERROR in a glib::Error returned from a Resource routine.

SocketClientEvent

Describes an event occurring on a SocketClient. See the signal::SocketClient::event signal for more details.

SocketFamily

The protocol family of a SocketAddress. (These values are identical to the system defines AF_INET, AF_INET6 and AF_UNIX, if available.)

SocketListenerEvent

Describes an event occurring on a SocketListener. See the signal::SocketListener::event signal for more details.

SocketProtocol

A protocol identifier is specified when creating a Socket, which is a family/type specific identifier, where 0 means the default protocol for the particular family/type.

SocketType

Flags used when creating a Socket. Some protocols may not implement all the socket types.

TlsAuthenticationMode

The client authentication mode for a TlsServerConnection.

TlsCertificateRequestFlags

Flags for TlsInteractionExt::request_certificate(), TlsInteractionExt::request_certificate_async(), and TlsInteractionExt::invoke_request_certificate().

TlsChannelBindingTypev2_66

The type of TLS channel binding data to retrieve from TlsConnection or GDtlsConnection, as documented by RFC 5929. The tls-unique-for-telnet binding type is not currently implemented.

TlsDatabaseLookupFlags

Flags for TlsDatabaseExt::lookup_certificate_for_handle(), TlsDatabaseExt::lookup_certificate_issuer(), and TlsDatabaseExt::lookup_certificates_issued_by().

TlsError

An error code used with G_TLS_ERROR in a glib::Error returned from a TLS-related routine.

TlsInteractionResult

TlsInteractionResult is returned by various functions in TlsInteraction when finishing an interaction request.

TlsProtocolVersionv2_70

The TLS or DTLS protocol version used by a TlsConnection or GDtlsConnection. The integer values of these versions are sequential to ensure newer known protocol versions compare greater than older known versions. Any known DTLS protocol version will compare greater than any SSL or TLS protocol version. The protocol version may be Unknown if the TLS backend supports a newer protocol version that GLib does not yet know about. This means that it’s possible for an unknown DTLS protocol version to compare less than the TLS protocol versions.

TlsRehandshakeMode

When to allow rehandshaking. See TlsConnectionExt::set_rehandshake_mode().

UnixSocketAddressTypeUnix

The type of name used by a UnixSocketAddress. Path indicates a traditional unix domain socket bound to a filesystem path. Anonymous indicates a socket not bound to any name (eg, a client-side socket, or a socket created with socketpair()).

ZlibCompressorFormat

Used to select the type of data format to use for ZlibDecompressor and ZlibCompressor.

Statics

DEBUG_CONTROLLER_EXTENSION_POINT_NAMEv2_72

Extension point for debug control functionality. See [Extending GIO][extending-gio].

DRIVE_IDENTIFIER_KIND_UNIX_DEVICEv2_58

The string used to obtain a Unix device path with DriveExt::identifier().

FILE_ATTRIBUTE_ACCESS_CAN_DELETE

A key in the “access” namespace for checking deletion privileges.

FILE_ATTRIBUTE_ACCESS_CAN_EXECUTE

A key in the “access” namespace for getting execution privileges.

FILE_ATTRIBUTE_ACCESS_CAN_READ

A key in the “access” namespace for getting read privileges.

FILE_ATTRIBUTE_ACCESS_CAN_RENAME

A key in the “access” namespace for checking renaming privileges.

FILE_ATTRIBUTE_ACCESS_CAN_TRASH

A key in the “access” namespace for checking trashing privileges.

FILE_ATTRIBUTE_ACCESS_CAN_WRITE

A key in the “access” namespace for getting write privileges.

FILE_ATTRIBUTE_DOS_IS_ARCHIVE

A key in the “dos” namespace for checking if the file’s archive flag is set.

FILE_ATTRIBUTE_DOS_IS_MOUNTPOINTv2_60

A key in the “dos” namespace for checking if the file is a NTFS mount point (a volume mount or a junction point).

FILE_ATTRIBUTE_DOS_IS_SYSTEM

A key in the “dos” namespace for checking if the file’s backup flag is set.

FILE_ATTRIBUTE_DOS_REPARSE_POINT_TAGv2_60

A key in the “dos” namespace for getting the file NTFS reparse tag.

FILE_ATTRIBUTE_ETAG_VALUE

A key in the “etag” namespace for getting the value of the file’s entity tag.

FILE_ATTRIBUTE_FILESYSTEM_FREE

A key in the “filesystem” namespace for getting the number of bytes of free space left on the file system.

FILE_ATTRIBUTE_FILESYSTEM_READONLY

A key in the “filesystem” namespace for checking if the file system is read only.

FILE_ATTRIBUTE_FILESYSTEM_REMOTE

A key in the “filesystem” namespace for checking if the file system is remote.

FILE_ATTRIBUTE_FILESYSTEM_SIZE

A key in the “filesystem” namespace for getting the total size (in bytes) of the file system, used in FileExt::query_filesystem_info().

FILE_ATTRIBUTE_FILESYSTEM_TYPE

A key in the “filesystem” namespace for getting the file system’s type.

FILE_ATTRIBUTE_FILESYSTEM_USED

A key in the “filesystem” namespace for getting the number of bytes used by data on the file system.

FILE_ATTRIBUTE_FILESYSTEM_USE_PREVIEW

A key in the “filesystem” namespace for hinting a file manager application whether it should preview (e.g. thumbnail) files on the file system.

FILE_ATTRIBUTE_GVFS_BACKEND

A key in the “gvfs” namespace that gets the name of the current GVFS backend in use.

FILE_ATTRIBUTE_ID_FILE

A key in the “id” namespace for getting a file identifier.

FILE_ATTRIBUTE_ID_FILESYSTEM

A key in the “id” namespace for getting the file system identifier.

FILE_ATTRIBUTE_MOUNTABLE_CAN_EJECT

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be ejected.

FILE_ATTRIBUTE_MOUNTABLE_CAN_MOUNT

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) is mountable.

FILE_ATTRIBUTE_MOUNTABLE_CAN_POLL

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be polled.

FILE_ATTRIBUTE_MOUNTABLE_CAN_START

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be started.

FILE_ATTRIBUTE_MOUNTABLE_CAN_START_DEGRADED

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be started degraded.

FILE_ATTRIBUTE_MOUNTABLE_CAN_STOP

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) can be stopped.

FILE_ATTRIBUTE_MOUNTABLE_CAN_UNMOUNT

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) is unmountable.

FILE_ATTRIBUTE_MOUNTABLE_HAL_UDI

A key in the “mountable” namespace for getting the HAL UDI for the mountable file.

FILE_ATTRIBUTE_MOUNTABLE_IS_MEDIA_CHECK_AUTOMATIC

A key in the “mountable” namespace for checking if a file (of type G_FILE_TYPE_MOUNTABLE) is automatically polled for media.

FILE_ATTRIBUTE_MOUNTABLE_START_STOP_TYPE

A key in the “mountable” namespace for getting the DriveStartStopType.

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE

A key in the “mountable” namespace for getting the unix device.

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE_FILE

A key in the “mountable” namespace for getting the unix device file.

FILE_ATTRIBUTE_OWNER_GROUP

A key in the “owner” namespace for getting the file owner’s group.

FILE_ATTRIBUTE_OWNER_USER

A key in the “owner” namespace for getting the user name of the file’s owner.

FILE_ATTRIBUTE_OWNER_USER_REAL

A key in the “owner” namespace for getting the real name of the user that owns the file.

FILE_ATTRIBUTE_PREVIEW_ICON

A key in the “preview” namespace for getting a Icon that can be used to get preview of the file.

FILE_ATTRIBUTE_RECENT_MODIFIEDv2_52

A key in the “recent” namespace for getting time, when the metadata for the file in recent:/// was last changed.

FILE_ATTRIBUTE_SELINUX_CONTEXT

A key in the “selinux” namespace for getting the file’s SELinux context.

FILE_ATTRIBUTE_STANDARD_ALLOCATED_SIZE

A key in the “standard” namespace for getting the amount of disk space that is consumed by the file (in bytes).

FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE

A key in the “standard” namespace for getting the content type of the file.

FILE_ATTRIBUTE_STANDARD_COPY_NAME

A key in the “standard” namespace for getting the copy name of the file.

FILE_ATTRIBUTE_STANDARD_DESCRIPTION

A key in the “standard” namespace for getting the description of the file.

FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME

A key in the “standard” namespace for getting the display name of the file.

FILE_ATTRIBUTE_STANDARD_EDIT_NAME

A key in the “standard” namespace for edit name of the file.

FILE_ATTRIBUTE_STANDARD_FAST_CONTENT_TYPE

A key in the “standard” namespace for getting the fast content type.

FILE_ATTRIBUTE_STANDARD_ICON

A key in the “standard” namespace for getting the icon for the file.

FILE_ATTRIBUTE_STANDARD_IS_BACKUP

A key in the “standard” namespace for checking if a file is a backup file.

FILE_ATTRIBUTE_STANDARD_IS_HIDDEN

A key in the “standard” namespace for checking if a file is hidden.

FILE_ATTRIBUTE_STANDARD_IS_SYMLINK

A key in the “standard” namespace for checking if the file is a symlink. Typically the actual type is something else, if we followed the symlink to get the type.

FILE_ATTRIBUTE_STANDARD_IS_VIRTUAL

A key in the “standard” namespace for checking if a file is virtual.

FILE_ATTRIBUTE_STANDARD_IS_VOLATILE

A key in the “standard” namespace for checking if a file is volatile. This is meant for opaque, non-POSIX-like backends to indicate that the URI is not persistent. Applications should look at FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET for the persistent URI.

FILE_ATTRIBUTE_STANDARD_NAME

A key in the “standard” namespace for getting the name of the file.

FILE_ATTRIBUTE_STANDARD_SIZE

A key in the “standard” namespace for getting the file’s size (in bytes).

FILE_ATTRIBUTE_STANDARD_SORT_ORDER

A key in the “standard” namespace for setting the sort order of a file.

FILE_ATTRIBUTE_STANDARD_SYMBOLIC_ICON

A key in the “standard” namespace for getting the symbolic icon for the file.

FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET

A key in the “standard” namespace for getting the symlink target, if the file is a symlink.

FILE_ATTRIBUTE_STANDARD_TARGET_URI

A key in the “standard” namespace for getting the target URI for the file, in the case of FileType::Shortcut or FileType::Mountable files.

FILE_ATTRIBUTE_STANDARD_TYPE

A key in the “standard” namespace for storing file types.

FILE_ATTRIBUTE_THUMBNAILING_FAILED

A key in the “thumbnail” namespace for checking if thumbnailing failed.

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID

A key in the “thumbnail” namespace for checking whether the thumbnail is outdated.

FILE_ATTRIBUTE_THUMBNAIL_PATH

A key in the “thumbnail” namespace for getting the path to the thumbnail image.

FILE_ATTRIBUTE_TIME_ACCESS

A key in the “time” namespace for getting the time the file was last accessed.

FILE_ATTRIBUTE_TIME_ACCESS_USEC

A key in the “time” namespace for getting the microseconds of the time the file was last accessed.

FILE_ATTRIBUTE_TIME_CHANGED

A key in the “time” namespace for getting the time the file was last changed.

FILE_ATTRIBUTE_TIME_CHANGED_USEC

A key in the “time” namespace for getting the microseconds of the time the file was last changed.

FILE_ATTRIBUTE_TIME_CREATED

A key in the “time” namespace for getting the time the file was created.

FILE_ATTRIBUTE_TIME_CREATED_USEC

A key in the “time” namespace for getting the microseconds of the time the file was created.

FILE_ATTRIBUTE_TIME_MODIFIED

A key in the “time” namespace for getting the time the file was last modified.

FILE_ATTRIBUTE_TIME_MODIFIED_USEC

A key in the “time” namespace for getting the microseconds of the time the file was last modified.

FILE_ATTRIBUTE_TRASH_DELETION_DATE

A key in the “trash” namespace for getting the deletion date and time of a file inside the trash:/// folder.

FILE_ATTRIBUTE_TRASH_ITEM_COUNT

A key in the “trash” namespace for getting the number of (toplevel) items that are present in the trash:/// folder.

FILE_ATTRIBUTE_TRASH_ORIG_PATH

A key in the “trash” namespace for getting the original path of a file inside the trash:/// folder before it was trashed.

FILE_ATTRIBUTE_UNIX_BLOCKS

A key in the “unix” namespace for getting the number of blocks allocated for the file.

FILE_ATTRIBUTE_UNIX_BLOCK_SIZE

A key in the “unix” namespace for getting the block size for the file system.

FILE_ATTRIBUTE_UNIX_DEVICE

A key in the “unix” namespace for getting the device id of the device the file is located on (see stat() documentation).

FILE_ATTRIBUTE_UNIX_GID

A key in the “unix” namespace for getting the group ID for the file.

FILE_ATTRIBUTE_UNIX_INODE

A key in the “unix” namespace for getting the inode of the file.

FILE_ATTRIBUTE_UNIX_IS_MOUNTPOINT

A key in the “unix” namespace for checking if the file represents a UNIX mount point.

FILE_ATTRIBUTE_UNIX_MODE

A key in the “unix” namespace for getting the mode of the file (e.g. whether the file is a regular file, symlink, etc).

FILE_ATTRIBUTE_UNIX_NLINK

A key in the “unix” namespace for getting the number of hard links for a file.

FILE_ATTRIBUTE_UNIX_RDEV

A key in the “unix” namespace for getting the device ID for the file (if it is a special file).

FILE_ATTRIBUTE_UNIX_UID

A key in the “unix” namespace for getting the user ID for the file.

MEMORY_MONITOR_EXTENSION_POINT_NAMEv2_64

Extension point for memory usage monitoring functionality. See [Extending GIO][extending-gio].

MENU_ATTRIBUTE_ACTION

The menu item attribute which holds the action name of the item. Action names are namespaced with an identifier for the action group in which the action resides. For example, “win.” for window-specific actions and “app.” for application-wide actions.

MENU_ATTRIBUTE_ACTION_NAMESPACE

The menu item attribute that holds the namespace for all action names in menus that are linked from this item.

MENU_ATTRIBUTE_ICON

The menu item attribute which holds the icon of the item.

MENU_ATTRIBUTE_LABEL

The menu item attribute which holds the label of the item.

MENU_ATTRIBUTE_TARGET

The menu item attribute which holds the target with which the item’s action will be activated.

MENU_LINK_SECTION

The name of the link that associates a menu item with a section. The linked menu will usually be shown in place of the menu item, using the item’s label as a header.

MENU_LINK_SUBMENU

The name of the link that associates a menu item with a submenu.

NATIVE_VOLUME_MONITOR_EXTENSION_POINT_NAME

NETWORK_MONITOR_EXTENSION_POINT_NAME

Extension point for network status monitoring functionality. See [Extending GIO][extending-gio].

POWER_PROFILE_MONITOR_EXTENSION_POINT_NAMEv2_70

Extension point for power profile usage monitoring functionality. See [Extending GIO][extending-gio].

PROXY_EXTENSION_POINT_NAME

Extension point for proxy functionality. See [Extending GIO][extending-gio].

PROXY_RESOLVER_EXTENSION_POINT_NAME

Extension point for proxy resolving functionality. See [Extending GIO][extending-gio].

SETTINGS_BACKEND_EXTENSION_POINT_NAME

Extension point for SettingsBackend functionality.

TLS_BACKEND_EXTENSION_POINT_NAME

Extension point for TLS functionality via TlsBackend. See [Extending GIO][extending-gio].

TLS_DATABASE_PURPOSE_AUTHENTICATE_CLIENT

The purpose used to verify the client certificate in a TLS connection. Used by TLS servers.

TLS_DATABASE_PURPOSE_AUTHENTICATE_SERVER

The purpose used to verify the server certificate in a TLS connection. This is the most common purpose in use. Used by TLS clients.

VFS_EXTENSION_POINT_NAME

Extension point for Vfs functionality. See [Extending GIO][extending-gio].

VOLUME_IDENTIFIER_KIND_CLASS

The string used to obtain the volume class with VolumeExt::identifier().

VOLUME_IDENTIFIER_KIND_HAL_UDI

The string used to obtain a Hal UDI with VolumeExt::identifier().

VOLUME_IDENTIFIER_KIND_LABEL

The string used to obtain a filesystem label with VolumeExt::identifier().

VOLUME_IDENTIFIER_KIND_NFS_MOUNT

The string used to obtain a NFS mount with VolumeExt::identifier().

VOLUME_IDENTIFIER_KIND_UNIX_DEVICE

The string used to obtain a Unix device path with VolumeExt::identifier().

VOLUME_IDENTIFIER_KIND_UUID

The string used to obtain a UUID with VolumeExt::identifier().

VOLUME_MONITOR_EXTENSION_POINT_NAME

Extension point for volume monitor functionality. See [Extending GIO][extending-gio].

Functions

bus_get

Asynchronously connects to the message bus specified by bus_type.

bus_get_future

bus_get_sync

Synchronously connects to the message bus specified by bus_type. Note that the returned object may shared with other callers, e.g. if two separate parts of a process calls this function with the same bus_type, they will share the same object.

bus_own_name

Starts acquiring name on the bus specified by bus_type and calls name_acquired_handler and name_lost_handler when the name is acquired respectively lost. Callbacks will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this function from.

bus_own_name_on_connection

Like g_bus_own_name() but takes a DBusConnection instead of a BusType.

bus_unown_name

bus_unwatch_name

bus_watch_name

Starts watching name on the bus specified by bus_type and calls name_appeared_handler and name_vanished_handler when the name is known to have an owner respectively known to lose its owner. Callbacks will be invoked in the [thread-default main context][g-main-context-push-thread-default] of the thread you are calling this function from.

bus_watch_name_on_connection

Like g_bus_watch_name() but takes a DBusConnection instead of a BusType.

compile_resources

Call from build script to run glib-compile-resources to generate compiled gresources to embed in binary with resources_register_include. target is relative to OUT_DIR.

content_type_can_be_executable

Checks if a content type can be executable. Note that for instance things like text files can be executables (i.e. scripts and batch files).

content_type_equals

Compares two content types for equality.

content_type_from_mime_type

Tries to find a content type based on the mime type name.

content_type_get_description

Gets the human readable description of the content type.

content_type_get_generic_icon_name

Gets the generic icon name for a content type.

content_type_get_icon

Gets the icon for a content type.

content_type_get_mime_dirsv2_60

Get the list of directories which MIME data is loaded from. See content_type_set_mime_dirs() for details.

content_type_get_mime_type

Gets the mime type for the content type, if one is registered.

content_type_get_symbolic_icon

Gets the symbolic icon for a content type.

content_type_guess

Guesses the content type based on example data. If the function is uncertain, result_uncertain will be set to true. Either filename or data may be None, in which case the guess will be based solely on the other argument.

content_type_guess_for_tree

Tries to guess the type of the tree with root root, by looking at the files it contains. The result is an array of content types, with the best guess coming first.

content_type_is_a

Determines if type_ is a subset of supertype.

content_type_is_mime_typev2_52

Determines if type_ is a subset of mime_type. Convenience wrapper around content_type_is_a().

content_type_is_unknown

Checks if the content type is the generic “unknown” type. On UNIX this is the “application/octet-stream” mimetype, while on win32 it is “*” and on OSX it is a dynamic type or octet-stream.

content_type_set_mime_dirsv2_60

Set the list of directories used by GIO to load the MIME database. If dirs is None, the directories used are the default:

content_types_get_registered

Gets a list of strings containing all the registered content types known to the system. The list and its data should be freed using g_list_free_full (list, g_free).

dbus_address_escape_value

Escape string so it can appear in a D-Bus address as the value part of a key-value pair.

dbus_address_get_for_bus_sync

Synchronously looks up the D-Bus address for the well-known message bus instance specified by bus_type. This may involve using various platform specific mechanisms.

dbus_address_get_stream

Asynchronously connects to an endpoint specified by address and sets up the connection so it is in a state to run the client-side of the D-Bus authentication conversation. address must be in the D-Bus address format.

dbus_address_get_stream_future

dbus_address_get_stream_sync

Synchronously connects to an endpoint specified by address and sets up the connection so it is in a state to run the client-side of the D-Bus authentication conversation. address must be in the D-Bus address format.

dbus_escape_object_pathv2_68

This is a language binding friendly version of g_dbus_escape_object_path_bytestring().

dbus_generate_guid

Generate a D-Bus GUID that can be used with e.g. DBusConnection::new().

dbus_gvalue_to_gvariant

Converts a glib::Value to a glib::Variant of the type indicated by the type_ parameter.

dbus_gvariant_to_gvalue

Converts a glib::Variant to a glib::Value. If value is floating, it is consumed.

dbus_is_address

Checks if string is a D-Bus address.

dbus_is_error_namev2_70

Check whether string is a valid D-Bus error name.

dbus_is_guid

Checks if string is a D-Bus GUID.

dbus_is_interface_name

Checks if string is a valid D-Bus interface name.

dbus_is_member_name

Checks if string is a valid D-Bus member (e.g. signal or method) name.

dbus_is_name

Checks if string is a valid D-Bus bus name (either unique or well-known).

dbus_is_supported_address

Like dbus_is_address() but also checks if the library supports the transports in string and that key/value pairs for each transport are valid. See the specification of the D-Bus address format.

dbus_is_unique_name

Checks if string is a valid D-Bus unique bus name.

io_error_from_errno

Converts errno.h error codes into GIO error codes. The fallback value IOErrorEnum::Failed is returned for error codes not currently handled (but note that future GLib releases may return a more specific value instead).

io_modules_scan_all_in_directory

Scans all the modules in the specified directory, ensuring that any extension point implemented by a module is registered.

keyfile_settings_backend_new

Creates a keyfile-backed SettingsBackend.

memory_settings_backend_new

Creates a memory-backed SettingsBackend.

null_settings_backend_new

Creates a readonly SettingsBackend.

resources_enumerate_children

Returns all the names of children at the specified path in the set of globally registered resources. The return result is a None terminated list of strings which should be released with g_strfreev().

resources_get_info

Looks for a file at the specified path in the set of globally registered resources and if found returns information about it.

resources_lookup_data

Looks for a file at the specified path in the set of globally registered resources and returns a glib::Bytes that lets you directly access the data in memory.

resources_open_stream

Looks for a file at the specified path in the set of globally registered resources and returns a InputStream that lets you read the data.

resources_register

Registers the resource with the process-global set of resources. Once a resource is registered the files in it can be accessed with the global resource lookup functions like resources_lookup_data().

resources_unregister

Unregisters the resource from the process-global set of resources.

unix_is_mount_path_system_internalUnix

Determines if mount_path is considered an implementation of the OS. This is primarily used for hiding mountable and mounted volumes that only are used in the OS and has little to no relevance to the casual user.

unix_is_system_device_pathUnix and v2_56

Determines if device_path is considered a block device path which is only used in implementation of the OS. This is primarily used for hiding mounted volumes that are intended as APIs for programs to read, and system administrators at a shell; rather than something that should, for example, appear in a GUI. For example, the Linux /proc filesystem.

unix_is_system_fs_typeUnix and v2_56

Determines if fs_type is considered a type of file system which is only used in implementation of the OS. This is primarily used for hiding mounted volumes that are intended as APIs for programs to read, and system administrators at a shell; rather than something that should, for example, appear in a GUI. For example, the Linux /proc filesystem.