Crate gio

Re-exports

pub use ffi;

pub use glib;

Modules

functions

prelude
Traits and essential types intended for blanket imports.

subclass

task

Macros

resources_register_include
Include gresources generated with compile_resources and register with glib. path is relative to OUTDIR.

Structs

Action
crate::Action represents a single named action.

ActionGroup
crate::ActionGroup represents a group of actions. Actions can be used to expose functionality in a structured way, either from one part of a program to another, or to the outside world. Action groups are often used together with a crate::MenuModel that provides additional representation data for displaying the actions to the user, e.g. in a menu.

ActionGroupExportId

ActionMap
The GActionMap interface is implemented by crate::ActionGroup implementations that operate by containing a number of named crate::Action instances, such as crate::SimpleActionGroup.

AppInfo
crate::AppInfo and crate::AppLaunchContext are used for describing and launching applications installed on the system.

AppInfoCreateFlags
Flags used when creating a crate::AppInfo.

AppInfoMonitor
crate::AppInfoMonitor is a very simple object used for monitoring the app info database for changes (ie: newly installed or removed applications).

AppLaunchContext
Integrating the launch with the launching application. This is used to handle for instance startup notification and launching the new application on the same screen as the launching window.

Application
A crate::Application is the foundation of an application. It wraps some low-level platform-specific services and is intended to act as the foundation for higher-level application classes such as GtkApplication or MxApplication. In general, you should not use this class outside of a higher level framework.

ApplicationBuilder

ApplicationCommandLine
crate::ApplicationCommandLine represents a command-line invocation of an application. It is created by crate::Application and emitted in the [crate::Application::command-line] signal and virtual function.

ApplicationFlags
Flags used to define the behaviour of a crate::Application.

AskPasswordFlags
crate::AskPasswordFlags are used to request specific information from the user, or to notify the user of their choices in an authentication situation.

AsyncResult
Provides a base class for implementing asynchronous function results.

BindingBuilder

BufferedInputStream
Buffered input stream implements crate::FilterInputStream and provides for buffered reads.

BufferedInputStreamBuilder

BufferedOutputStream
Buffered output stream implements crate::FilterOutputStream and provides for buffered writes.

BufferedOutputStreamBuilder

BusNameOwnerFlags
Flags used in g_bus_own_name().

BusNameWatcherFlags
Flags used in g_bus_watch_name().

BytesIcon
crate::BytesIcon specifies an image held in memory in a common format (usually png) to be used as icon.

Cancellable
GCancellable is a thread-safe operation cancellation stack used throughout GIO to allow for cancellation of synchronous and asynchronous operations.

CharsetConverter
crate::CharsetConverter is an implementation of crate::Converter based on GIConv.

CharsetConverterBuilder

Converter
crate::Converter is implemented by objects that convert binary data in various ways. The conversion can be stateful and may fail at any place.

ConverterFlags
Flags used when calling a [crate::Converter::convert()].

ConverterInputStream
Converter input stream implements crate::InputStream and allows conversion of data of various types during reading.

ConverterInputStreamBuilder

ConverterOutputStream
Converter output stream implements crate::OutputStream and allows conversion of data of various types during reading.

ConverterOutputStreamBuilder

Credentials
The crate::Credentials type is a reference-counted wrapper for native credentials. This information is typically used for identifying, authenticating and authorizing other processes.

DBusActionGroup
crate::DBusActionGroup is an implementation of the crate::ActionGroup interface that can be used as a proxy for an action group that is exported over D-Bus with crate::DBusConnection::export_action_group().

DBusArgInfo
Information about an argument for a method or a signal.

DBusAuthObserver
The crate::DBusAuthObserver type provides a mechanism for participating in how a crate::DBusServer (or a crate::DBusConnection) authenticates remote peers. Simply instantiate a crate::DBusAuthObserver and connect to the signals you are interested in. Note that new signals may be added in the future

DBusCallFlags
Flags used in crate::DBusConnection::call() and similar APIs.

DBusCapabilityFlags
Capabilities negotiated with the remote peer.

DBusConnection
The crate::DBusConnection type is used for D-Bus connections to remote peers such as a message buses. It is a low-level API that offers a lot of flexibility. For instance, it lets you establish a connection over any transport that can by represented as a crate::IOStream.

DBusConnectionFlags
Flags used when creating a new crate::DBusConnection.

DBusInterface
The crate::DBusInterface type is the base type for D-Bus interfaces both on the service side (see crate::DBusInterfaceSkeleton) and client side (see crate::DBusProxy).

DBusInterfaceInfo
Information about a D-Bus interface.

DBusInterfaceSkeleton
Abstract base class for D-Bus interfaces on the service side.

DBusInterfaceSkeletonFlags
Flags describing the behavior of a crate::DBusInterfaceSkeleton instance.

DBusMenuModel
crate::DBusMenuModel is an implementation of crate::MenuModel that can be used as a proxy for a menu model that is exported over D-Bus with crate::DBusConnection::export_menu_model().

DBusMessage
A type for representing D-Bus messages that can be sent or received on a crate::DBusConnection.

DBusMessageFlags
Message flags used in crate::DBusMessage.

DBusMethodInfo
Information about a method on an D-Bus interface.

DBusMethodInvocation
Instances of the crate::DBusMethodInvocation class are used when handling D-Bus method calls. It provides a way to asynchronously return results and errors.

DBusNodeInfo
Information about nodes in a remote object hierarchy.

DBusObject
The crate::DBusObject type is the base type for D-Bus objects on both the service side (see [crate::DBusObjectSkeleton]) and the client side (see [crate::DBusObjectProxy]). It is essentially just a container of interfaces.

DBusPropertyInfo
Information about a D-Bus property on a D-Bus interface.

DBusProxy
crate::DBusProxy is a base class used for proxies to access a D-Bus interface on a remote object. A crate::DBusProxy can be constructed for both well-known and unique names.

DBusProxyFlags
Flags used when constructing an instance of a crate::DBusProxy derived class.

DBusSendMessageFlags
Flags used when sending GDBusMessages on a crate::DBusConnection.

DBusServer
crate::DBusServer is a helper for listening to and accepting D-Bus connections. This can be used to create a new D-Bus server, allowing two peers to use the D-Bus protocol for their own specialized communication. A server instance provided in this way will not perform message routing or implement the org.freedesktop.DBus interface.

DBusServerFlags
Flags used when creating a crate::DBusServer.

DBusSignalFlags
Flags used when subscribing to signals via crate::DBusConnection::signal_subscribe().

DBusSignalInfo
Information about a signal on a D-Bus interface.

DataInputStream
Data input stream implements crate::InputStream and includes functions for reading structured data directly from a binary input stream.

DataInputStreamBuilder

DataOutputStream
Data output stream implements crate::OutputStream and includes functions for writing data directly to an output stream.

DataOutputStreamBuilder

DesktopAppInfo
crate::DesktopAppInfo is an implementation of crate::AppInfo based on desktop files.

Drive
crate::Drive - this represent a piece of hardware connected to the machine. It’s generally only created for removable hardware or hardware with removable media.

DriveStartFlags
Flags used when starting a drive.

Emblem
crate::Emblem is an implementation of crate::Icon that supports having an emblem, which is an icon with additional properties. It can than be added to a crate::EmblemedIcon.

EmblemedIcon
crate::EmblemedIcon is an implementation of crate::Icon that supports adding an emblem to an icon. Adding multiple emblems to an icon is ensured via crate::prelude::EmblemedIconExt::add_emblem().

File
crate::File is a high level abstraction for manipulating files on a virtual file system. GFiles are lightweight, immutable objects that do no I/O upon creation. It is necessary to understand that crate::File objects do not represent files, merely an identifier for a file. All file content I/O is implemented as streaming operations (see crate::InputStream and crate::OutputStream).

FileAttributeMatcher
Determines if a string matches a file attribute.

FileAttributematcherIter

FileCopyFlags
Flags used when copying or moving files.

FileCreateFlags
Flags used when an operation may create a file.

FileEnumerator
crate::FileEnumerator allows you to operate on a set of GFiles, returning a crate::FileInfo structure for each file enumerated (e.g. [crate::File::enumerate_children()] will return a crate::FileEnumerator for each of the children within a directory).

FileIOStream
GFileIOStream provides io streams that both read and write to the same file handle.

FileIcon
crate::FileIcon specifies an icon by pointing to an image file to be used as icon.

FileInfo
Functionality for manipulating basic metadata for files. crate::FileInfo implements methods for getting information that all files should contain, and allows for manipulation of extended attributes.

FileInputStream
GFileInputStream provides input streams that take their content from a file.

FileMeasureFlags
Flags that can be used with [crate::File::measure_disk_usage()].

FileMonitor
Monitors a file or directory for changes.

FileMonitorFlags
Flags used to set what a crate::FileMonitor will watch for.

FileOutputStream
GFileOutputStream provides output streams that write their content to a file.

FileQueryInfoFlags
Flags used when querying a crate::FileInfo.

FilenameCompleter
Completes partial file and directory names given a partial string by looking in the file system for clues. Can return a list of possible completion strings for widget implementations.

FilterId

FilterInputStream
Base class for input stream implementations that perform some kind of filtering operation on a base stream. Typical examples of filtering operations are character set conversion, compression and byte order flipping.

FilterOutputStream
Base class for output stream implementations that perform some kind of filtering operation on a base stream. Typical examples of filtering operations are character set conversion, compression and byte order flipping.

GioFuture

GioFutureResult

IOStream
GIOStream represents an object that has both read and write streams. Generally the two streams act as separate input and output streams, but they share some common resources and state. For instance, for seekable streams, both streams may use the same position.

IOStreamAsyncReadWrite

IOStreamSpliceFlags
GIOStreamSpliceFlags determine how streams should be spliced.

Icon
crate::Icon is a very minimal interface for icons. It provides functions for checking the equality of two icons, hashing of icons and serializing an icon to and from strings.

InetAddress
crate::InetAddress represents an IPv4 or IPv6 internet address. Use crate::prelude::ResolverExt::lookup_by_name() or crate::prelude::ResolverExt::lookup_by_name_async() to look up the crate::InetAddress for a hostname. Use crate::prelude::ResolverExt::lookup_by_address() or crate::prelude::ResolverExt::lookup_by_address_async() to look up the hostname for a crate::InetAddress.

InetAddressMask
crate::InetAddressMask represents a range of IPv4 or IPv6 addresses described by a base address and a length indicating how many bits of the base address are relevant for matching purposes. These are often given in string form. Eg, “10.0.0.0/8”, or “fe80::/10”.

InetSocketAddress
An IPv4 or IPv6 socket address; that is, the combination of a crate::InetAddress and a port number.

InputStream
crate::InputStream has functions to read from a stream ([Self::read()]), to close a stream (crate::prelude::InputStreamExt::close()) and to skip some content (crate::prelude::InputStreamExt::skip()).

InputStreamAsyncBufRead

InputStreamAsyncRead

InputStreamRead

ListModel
crate::ListModel is an interface that represents a mutable list of GObjects. Its main intention is as a model for various widgets in user interfaces, such as list views, but it can also be used as a convenient method of returning lists of data, with support for updates.

ListStore
crate::ListStore is a simple implementation of crate::ListModel that stores all items in memory.

ListStoreBuilder

LoadableIcon
Extends the crate::Icon interface and adds the ability to load icons from streams.

MemoryInputStream
crate::MemoryInputStream is a class for using arbitrary memory chunks as input for GIO streaming input operations.

MemoryMonitor
crate::MemoryMonitor will monitor system memory and suggest to the application when to free memory so as to leave more room for other applications. It is implemented on Linux using the Low Memory Monitor (API documentation).

MemoryOutputStream
crate::MemoryOutputStream is a class for using arbitrary memory chunks as output for GIO streaming output operations.

Menu
crate::Menu is a simple implementation of crate::MenuModel. You populate a crate::Menu by adding crate::MenuItem instances to it.

MenuAttributeIter
crate::MenuAttributeIter is an opaque structure type. You must access it using the functions below.

MenuItem
crate::MenuItem is an opaque structure type. You must access it using the functions below.

MenuLinkIter
crate::MenuLinkIter is an opaque structure type. You must access it using the functions below.

MenuModel
crate::MenuModel represents the contents of a menu – an ordered list of menu items. The items are associated with actions, which can be activated through them. Items can be grouped in sections, and may have submenus associated with them. Both items and sections usually have some representation data, such as labels or icons. The type of the associated action (ie whether it is stateful, and what kind of state it has) can influence the representation of the item.

MenuModelExportId

Mount
The crate::Mount interface represents user-visible mounts. Note, when porting from GnomeVFS, crate::Mount is the moral equivalent of GnomeVFSVolume.

MountMountFlags
Flags used when mounting a mount.

MountOperation
crate::MountOperation provides a mechanism for interacting with the user. It can be used for authenticating mountable operations, such as loop mounting files, hard drive partitions or server locations. It can also be used to ask the user questions or show a list of applications preventing unmount or eject operations from completing.

MountUnmountFlags
Flags used when an unmounting a mount.

NetworkAddress
crate::NetworkAddress provides an easy way to resolve a hostname and then attempt to connect to that host, handling the possibility of multiple IP addresses and multiple address families.

NetworkMonitor
crate::NetworkMonitor provides an easy-to-use cross-platform API for monitoring network connectivity. On Linux, the available implementations are based on the kernel’s netlink interface and on NetworkManager.

NetworkService
Like crate::NetworkAddress does with hostnames, crate::NetworkService provides an easy way to resolve a SRV record, and then attempt to connect to one of the hosts that implements that service, handling service priority/weighting, multiple IP addresses, and multiple address families.

Notification
crate::Notification is a mechanism for creating a notification to be shown to the user – typically as a pop-up notification presented by the desktop environment shell.

OutputStream
crate::OutputStream has functions to write to a stream (crate::prelude::OutputStreamExt::write()), to close a stream (crate::prelude::OutputStreamExt::close()) and to flush pending writes (crate::prelude::OutputStreamExt::flush()).

OutputStreamAsyncWrite

OutputStreamSpliceFlags
GOutputStreamSpliceFlags determine how streams should be spliced.

OutputStreamWrite

OwnerId

Permission
A crate::Permission represents the status of the caller’s permission to perform a certain action.

PollableInputStream
crate::PollableInputStream is implemented by GInputStreams that can be polled for readiness to read. This can be used when interfacing with a non-GIO API that expects UNIX-file-descriptor-style asynchronous I/O rather than GIO-style.

PollableOutputStream
crate::PollableOutputStream is implemented by GOutputStreams that can be polled for readiness to write. This can be used when interfacing with a non-GIO API that expects UNIX-file-descriptor-style asynchronous I/O rather than GIO-style.

PropertyAction
A crate::PropertyAction is a way to get a crate::Action with a state value reflecting and controlling the value of a crate::glib::object::Object property.

Proxy
A crate::Proxy handles connecting to a remote host via a given type of proxy server. It is implemented by the ‘gio-proxy’ extension point. The extensions are named after their proxy protocol name. As an example, a SOCKS5 proxy implementation can be retrieved with the name ‘socks5’ using the function [crate::IOExtensionPoint::get_extension_by_name()].

ProxyAddress
Support for proxied crate::InetSocketAddress.

ProxyResolver
crate::ProxyResolver provides synchronous and asynchronous network proxy resolution. crate::ProxyResolver is used within crate::SocketClient through the method [crate::SocketConnectable::proxy_enumerate()].

ReadInputStream

RegistrationId

RemoteActionGroup
The GRemoteActionGroup interface is implemented by crate::ActionGroup instances that either transmit action invocations to other processes or receive action invocations in the local process from other processes.

Resolver
crate::Resolver provides cancellable synchronous and asynchronous DNS resolution, for hostnames (crate::prelude::ResolverExt::lookup_by_address(), crate::prelude::ResolverExt::lookup_by_name() and their async variants) and SRV (service) records (crate::prelude::ResolverExt::lookup_service()).

ResolverNameLookupFlags
Flags to modify lookup behavior.

Resource
Applications and libraries often contain binary or textual data that is really part of the application, rather than user data. For instance GtkBuilder .ui files, splashscreen images, GMenu markup XML, CSS files, icons, etc. These are often shipped as files in $datadir/appname, or manually included as literal strings in the code.

ResourceLookupFlags
GResourceLookupFlags determine how resource path lookups are handled.

Seekable
crate::Seekable is implemented by streams (implementations of crate::InputStream or crate::OutputStream) that support seeking.

Settings
The crate::Settings class provides a convenient API for storing and retrieving application settings.

SettingsBackend
The crate::SettingsBackend interface defines a generic interface for non-strictly-typed data that is stored in a hierarchy. To implement an alternative storage backend for crate::Settings, you need to implement the crate::SettingsBackend interface and then make it implement the extension point G_SETTINGS_BACKEND_EXTENSION_POINT_NAME.

SettingsBindFlags
Flags used when creating a binding. These flags determine in which direction the binding works. The default is to synchronize in both directions.

SettingsSchema
The crate::SettingsSchemaSource and crate::SettingsSchema APIs provide a mechanism for advanced control over the loading of schemas and a mechanism for introspecting their content.

SettingsSchemaKey
crate::SettingsSchemaKey is an opaque data structure and can only be accessed using the following functions.

SettingsSchemaSource
This is an opaque structure type. You may not access it directly.

SignalSubscriptionId

SimpleAction
A crate::SimpleAction is the obvious simple implementation of the crate::Action interface. This is the easiest way to create an action for purposes of adding it to a crate::SimpleActionGroup.

SimpleActionGroup
crate::SimpleActionGroup is a hash table filled with crate::Action objects, implementing the crate::ActionGroup and crate::ActionMap interfaces.

SimpleIOStream
GSimpleIOStream creates a crate::IOStream from an arbitrary crate::InputStream and crate::OutputStream. This allows any pair of input and output streams to be used with crate::IOStream methods.

SimplePermission
crate::SimplePermission is a trivial implementation of crate::Permission that represents a permission that is either always or never allowed. The value is given at construction and doesn’t change.

Socket
A crate::Socket is a low-level networking primitive. It is a more or less direct mapping of the BSD socket API in a portable GObject based API. It supports both the UNIX socket implementations and winsock2 on Windows.

SocketAddress
crate::SocketAddress is the equivalent of struct sockaddr in the BSD sockets API. This is an abstract class; use crate::InetSocketAddress for internet sockets, or crate::UnixSocketAddress for UNIX domain sockets.

SocketAddressEnumerator
crate::SocketAddressEnumerator is an enumerator type for crate::SocketAddress instances. It is returned by enumeration functions such as [crate::SocketConnectable::enumerate()], which returns a crate::SocketAddressEnumerator to list each crate::SocketAddress which could be used to connect to that crate::SocketConnectable.

SocketClient
crate::SocketClient is a lightweight high-level utility class for connecting to a network host using a connection oriented socket type.

SocketConnectable
Objects that describe one or more potential socket endpoints implement crate::SocketConnectable. Callers can then use [Self::enumerate()] to get a crate::SocketAddressEnumerator to try out each socket address in turn until one succeeds, as shown in the sample code below.

SocketConnection
crate::SocketConnection is a crate::IOStream for a connected socket. They can be created either by crate::SocketClient when connecting to a host, or by crate::SocketListener when accepting a new client.

SocketListener
A crate::SocketListener is an object that keeps track of a set of server sockets and helps you accept sockets from any of the socket, either sync or async.

SocketService
A crate::SocketService is an object that represents a service that is provided to the network or over local sockets. When a new connection is made to the service the [crate::SocketService::incoming] signal is emitted.

SrvTarget
SRV (service) records are used by some network protocols to provide service-specific aliasing and load-balancing. For example, XMPP (Jabber) uses SRV records to locate the XMPP server for a domain; rather than connecting directly to “example.com” or assuming a specific server hostname like “xmpp.example.com”, an XMPP client would look up the “xmpp-client” SRV record for “example.com”, and then connect to whatever host was pointed to by that record.

Subprocess
crate::Subprocess allows the creation of and interaction with child processes.

SubprocessFlags
Flags to define the behaviour of a crate::Subprocess.

SubprocessLauncher
This class contains a set of options for launching child processes, such as where its standard input and output will be directed, the argument list, the environment, and more.

Task
A crate::Task represents and manages a cancellable “task”.

TcpConnection
This is the subclass of crate::SocketConnection that is created for TCP/IP sockets.

ThemedIcon
crate::ThemedIcon is an implementation of crate::Icon that supports icon themes. crate::ThemedIcon contains a list of all of the icons present in an icon theme, so that icons can be looked up quickly. crate::ThemedIcon does not provide actual pixmaps for icons, just the icon names. Ideally something like gtk_icon_theme_choose_icon() should be used to resolve the list of names so that fallback icons work nicely with themes that inherit other themes.

ThreadedSocketService
A crate::ThreadedSocketService is a simple subclass of crate::SocketService that handles incoming connections by creating a worker thread and dispatching the connection to it by emitting the [crate::ThreadedSocketService::run] signal in the new thread.

TlsBackend
TLS (Transport Layer Security, aka SSL) and DTLS backend.

TlsCertificate
A certificate used for TLS authentication and encryption. This can represent either a certificate only (eg, the certificate received by a client from a server), or the combination of a certificate and a private key (which is needed when acting as a crate::TlsServerConnection).

TlsCertificateFlags
A set of flags describing TLS certification validation. This can be used to set which validation steps to perform (eg, with [crate::TlsClientConnection::set_validation_flags()]), or to describe why a particular certificate was rejected (eg, in [crate::TlsConnection::accept-certificate]).

TlsClientConnection
crate::TlsClientConnection is the client-side subclass of crate::TlsConnection, representing a client-side TLS connection.

TlsConnection
crate::TlsConnection is the base TLS connection class type, which wraps a crate::IOStream and provides TLS encryption on top of it. Its subclasses, crate::TlsClientConnection and crate::TlsServerConnection, implement client-side and server-side TLS, respectively.

TlsDatabase
crate::TlsDatabase is used to look up certificates and other information from a certificate or key store. It is an abstract base class which TLS library specific subtypes override.

TlsDatabaseVerifyFlags
Flags for crate::prelude::TlsDatabaseExt::verify_chain().

TlsFileDatabase
crate::TlsFileDatabase is implemented by crate::TlsDatabase objects which load their certificate information from a file. It is an interface which TLS library specific subtypes implement.

TlsInteraction
crate::TlsInteraction provides a mechanism for the TLS connection and database code to interact with the user. It can be used to ask the user for passwords.

TlsPassword
Holds a password used in TLS.

TlsPasswordFlags
Various flags for the password.

TlsServerConnection
crate::TlsServerConnection is the server-side subclass of crate::TlsConnection, representing a server-side TLS connection.

UnixFDList
A crate::UnixFDList contains a list of file descriptors. It owns the file descriptors that it contains, closing them when finalized.

UnixInputStream
crate::UnixInputStream implements crate::InputStream for reading from a UNIX file descriptor, including asynchronous operations. (If the file descriptor refers to a socket or pipe, this will use poll() to do asynchronous I/O. If it refers to a regular file, it will fall back to doing asynchronous I/O in another thread.)

UnixMountEntry
Defines a Unix mount entry (e.g. <filename>/media/cdrom</filename>). This corresponds roughly to a mtab entry.

UnixMountPoint
Defines a Unix mount point (e.g. <filename>/dev</filename>). This corresponds roughly to a fstab entry.

UnixOutputStream
crate::UnixOutputStream implements crate::OutputStream for writing to a UNIX file descriptor, including asynchronous operations. (If the file descriptor refers to a socket or pipe, this will use poll() to do asynchronous I/O. If it refers to a regular file, it will fall back to doing asynchronous I/O in another thread.)

UnixSocketAddress
Support for UNIX-domain (also known as local) sockets.

Vfs
Entry point for using GIO functionality.

Volume
The crate::Volume interface represents user-visible objects that can be mounted. Note, when porting from GnomeVFS, crate::Volume is the moral equivalent of GnomeVFSDrive.

VolumeMonitor
crate::VolumeMonitor is for listing the user interesting devices and volumes on the computer. In other words, what a file selector or file manager would show in a sidebar.

WatcherId

WriteOutputStream

ZlibCompressor
Zlib decompression

ZlibDecompressor
Zlib decompression

Enums

BusType
An enumeration for well-known message buses.

ConverterResult
Results returned from [crate::Converter::convert()].

CredentialsType
Enumeration describing different kinds of native credential types.

DBusMessageByteOrder
Enumeration used to describe the byte order of a D-Bus message.

DBusMessageHeaderField
Header fields used in crate::DBusMessage.

DBusMessageType
Message types used in crate::DBusMessage.

DataStreamByteOrder
crate::DataStreamByteOrder is used to ensure proper endianness of streaming data sources across various machine architectures.

DataStreamNewlineType
crate::DataStreamNewlineType is used when checking for or setting the line endings for a given file.

DriveStartStopType
Enumeration describing how a drive can be started/stopped.

EmblemOrigin
GEmblemOrigin is used to add information about the origin of the emblem to crate::Emblem.

FileAttributeStatus
Used by [crate::File::set_attributes_from_info()] when setting file attributes.

FileAttributeType
The data types for file attributes.

FileMonitorEvent
Specifies what type of event a monitor event is.

FileType
Indicates the file’s on-disk type.

IOErrorEnum
Error codes returned by GIO functions.

InetAddressBytes

MemoryMonitorWarningLevel	v2_64
Memory availability warning levels.

MountOperationResult
crate::MountOperationResult is returned as a result when a request for information is send by the mounting operation.

NetworkConnectivity
The host’s network connectivity state, as reported by crate::NetworkMonitor.

NotificationPriority
Priority levels for GNotifications.

PasswordSave
crate::PasswordSave is used to indicate the lifespan of a saved password.

ResolverRecordType
The type of record that crate::prelude::ResolverExt::lookup_records() or crate::prelude::ResolverExt::lookup_records_async() should retrieve. The records are returned as lists of crate::glib::Variant tuples. Each record type has different values in the variant tuples returned.

ResourceError
An error code used with G_RESOURCE_ERROR in a crate::glib::Error returned from a crate::Resource routine.

SocketClientEvent
Describes an event occurring on a crate::SocketClient. See the [crate::SocketClient::event] signal for more details.

SocketFamily
The protocol family of a crate::SocketAddress. (These values are identical to the system defines AF_INET, AF_INET6 and AF_UNIX, if available.)

SocketListenerEvent
Describes an event occurring on a crate::SocketListener. See the [crate::SocketListener::event] signal for more details.

SocketProtocol
A protocol identifier is specified when creating a crate::Socket, which is a family/type specific identifier, where 0 means the default protocol for the particular family/type.

SocketType
Flags used when creating a crate::Socket. Some protocols may not implement all the socket types.

TlsAuthenticationMode
The client authentication mode for a crate::TlsServerConnection.

TlsCertificateRequestFlags
Flags for crate::prelude::TlsInteractionExt::request_certificate(), crate::prelude::TlsInteractionExt::request_certificate_async(), and crate::prelude::TlsInteractionExt::invoke_request_certificate().

TlsChannelBindingType	v2_66
The type of TLS channel binding data to retrieve from crate::TlsConnection or [crate::DtlsConnection], as documented by RFC 5929. The tls-unique-for-telnet binding type is not currently implemented.

TlsDatabaseLookupFlags
Flags for crate::prelude::TlsDatabaseExt::lookup_certificate_for_handle(), crate::prelude::TlsDatabaseExt::lookup_certificate_issuer(), and crate::prelude::TlsDatabaseExt::lookup_certificates_issued_by().

TlsInteractionResult
crate::TlsInteractionResult is returned by various functions in crate::TlsInteraction when finishing an interaction request.

TlsRehandshakeMode
When to allow rehandshaking. See crate::prelude::TlsConnectionExt::set_rehandshake_mode().

UnixSocketAddressType
The type of name used by a crate::UnixSocketAddress. Self::Path indicates a traditional unix domain socket bound to a filesystem path. Self::Anonymous indicates a socket not bound to any name (eg, a client-side socket, or a socket created with socketpair()).

ZlibCompressorFormat
Used to select the type of data format to use for crate::ZlibDecompressor and crate::ZlibCompressor.

Constants

NONE_ACTION

NONE_ACTION_GROUP

NONE_ACTION_MAP

NONE_APPLICATION

NONE_APPLICATION_COMMAND_LINE

NONE_APP_INFO

NONE_APP_LAUNCH_CONTEXT

NONE_ASYNC_RESULT

NONE_BUFFERED_INPUT_STREAM

NONE_BUFFERED_OUTPUT_STREAM

NONE_CANCELLABLE

NONE_CONVERTER

NONE_CONVERTER_INPUT_STREAM

NONE_CONVERTER_OUTPUT_STREAM

NONE_DATA_INPUT_STREAM

NONE_DATA_OUTPUT_STREAM

NONE_DBUS_INTERFACE

NONE_DBUS_INTERFACE_SKELETON

NONE_DBUS_OBJECT

NONE_DBUS_PROXY

NONE_DRIVE

NONE_EMBLEMED_ICON

NONE_FILE

NONE_FILE_ENUMERATOR

NONE_FILE_INPUT_STREAM

NONE_FILE_IO_STREAM

NONE_FILE_MONITOR

NONE_FILE_OUTPUT_STREAM

NONE_FILTER_INPUT_STREAM

NONE_FILTER_OUTPUT_STREAM

NONE_ICON

NONE_INET_ADDRESS

NONE_INET_ADDRESS_MASK

NONE_INET_SOCKET_ADDRESS

NONE_INPUT_STREAM

NONE_IO_STREAM

NONE_LIST_MODEL

NONE_LOADABLE_ICON

NONE_MEMORY_INPUT_STREAM

NONE_MEMORY_MONITOR

NONE_MEMORY_OUTPUT_STREAM

NONE_MENU_ATTRIBUTE_ITER

NONE_MENU_LINK_ITER

NONE_MENU_MODEL

NONE_MOUNT

NONE_MOUNT_OPERATION

NONE_NETWORK_ADDRESS

NONE_NETWORK_MONITOR

NONE_NETWORK_SERVICE

NONE_OUTPUT_STREAM

NONE_PERMISSION

NONE_POLLABLE_INPUT_STREAM

NONE_POLLABLE_OUTPUT_STREAM

NONE_PROXY

NONE_PROXY_ADDRESS

NONE_PROXY_RESOLVER

NONE_REMOTE_ACTION_GROUP

NONE_RESOLVER

NONE_SEEKABLE

NONE_SETTINGS

NONE_SETTINGS_BACKEND

NONE_SIMPLE_ACTION_GROUP

NONE_SOCKET

NONE_SOCKET_ADDRESS

NONE_SOCKET_ADDRESS_ENUMERATOR

NONE_SOCKET_CLIENT

NONE_SOCKET_CONNECTABLE

NONE_SOCKET_CONNECTION

NONE_SOCKET_LISTENER

NONE_SOCKET_SERVICE

NONE_TCP_CONNECTION

NONE_THREADED_SOCKET_SERVICE

NONE_TLS_BACKEND

NONE_TLS_CERTIFICATE

NONE_TLS_CLIENT_CONNECTION

NONE_TLS_CONNECTION

NONE_TLS_DATABASE

NONE_TLS_FILE_DATABASE

NONE_TLS_INTERACTION

NONE_TLS_PASSWORD

NONE_TLS_SERVER_CONNECTION

NONE_UNIX_FD_LIST

NONE_UNIX_INPUT_STREAM

NONE_UNIX_OUTPUT_STREAM

NONE_UNIX_SOCKET_ADDRESS

NONE_VFS

NONE_VOLUME

NONE_VOLUME_MONITOR

Statics

DRIVE_IDENTIFIER_KIND_UNIX_DEVICE	v2_58

FILE_ATTRIBUTE_ACCESS_CAN_DELETE

FILE_ATTRIBUTE_ACCESS_CAN_EXECUTE

FILE_ATTRIBUTE_ACCESS_CAN_READ

FILE_ATTRIBUTE_ACCESS_CAN_RENAME

FILE_ATTRIBUTE_ACCESS_CAN_TRASH

FILE_ATTRIBUTE_ACCESS_CAN_WRITE

FILE_ATTRIBUTE_DOS_IS_ARCHIVE

FILE_ATTRIBUTE_DOS_IS_MOUNTPOINT	v2_60

FILE_ATTRIBUTE_DOS_IS_SYSTEM

FILE_ATTRIBUTE_DOS_REPARSE_POINT_TAG	v2_60

FILE_ATTRIBUTE_ETAG_VALUE

FILE_ATTRIBUTE_FILESYSTEM_FREE

FILE_ATTRIBUTE_FILESYSTEM_READONLY

FILE_ATTRIBUTE_FILESYSTEM_REMOTE

FILE_ATTRIBUTE_FILESYSTEM_SIZE

FILE_ATTRIBUTE_FILESYSTEM_TYPE

FILE_ATTRIBUTE_FILESYSTEM_USED

FILE_ATTRIBUTE_FILESYSTEM_USE_PREVIEW

FILE_ATTRIBUTE_GVFS_BACKEND

FILE_ATTRIBUTE_ID_FILE

FILE_ATTRIBUTE_ID_FILESYSTEM

FILE_ATTRIBUTE_MOUNTABLE_CAN_EJECT

FILE_ATTRIBUTE_MOUNTABLE_CAN_MOUNT

FILE_ATTRIBUTE_MOUNTABLE_CAN_POLL

FILE_ATTRIBUTE_MOUNTABLE_CAN_START

FILE_ATTRIBUTE_MOUNTABLE_CAN_START_DEGRADED

FILE_ATTRIBUTE_MOUNTABLE_CAN_STOP

FILE_ATTRIBUTE_MOUNTABLE_CAN_UNMOUNT

FILE_ATTRIBUTE_MOUNTABLE_HAL_UDI

FILE_ATTRIBUTE_MOUNTABLE_IS_MEDIA_CHECK_AUTOMATIC

FILE_ATTRIBUTE_MOUNTABLE_START_STOP_TYPE

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE_FILE

FILE_ATTRIBUTE_OWNER_GROUP

FILE_ATTRIBUTE_OWNER_USER

FILE_ATTRIBUTE_OWNER_USER_REAL

FILE_ATTRIBUTE_PREVIEW_ICON

FILE_ATTRIBUTE_RECENT_MODIFIED	v2_52

FILE_ATTRIBUTE_SELINUX_CONTEXT

FILE_ATTRIBUTE_STANDARD_ALLOCATED_SIZE

FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE

FILE_ATTRIBUTE_STANDARD_COPY_NAME

FILE_ATTRIBUTE_STANDARD_DESCRIPTION

FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME

FILE_ATTRIBUTE_STANDARD_EDIT_NAME

FILE_ATTRIBUTE_STANDARD_FAST_CONTENT_TYPE

FILE_ATTRIBUTE_STANDARD_ICON

FILE_ATTRIBUTE_STANDARD_IS_BACKUP

FILE_ATTRIBUTE_STANDARD_IS_HIDDEN

FILE_ATTRIBUTE_STANDARD_IS_SYMLINK

FILE_ATTRIBUTE_STANDARD_IS_VIRTUAL

FILE_ATTRIBUTE_STANDARD_IS_VOLATILE

FILE_ATTRIBUTE_STANDARD_NAME

FILE_ATTRIBUTE_STANDARD_SIZE

FILE_ATTRIBUTE_STANDARD_SORT_ORDER

FILE_ATTRIBUTE_STANDARD_SYMBOLIC_ICON

FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET

FILE_ATTRIBUTE_STANDARD_TARGET_URI

FILE_ATTRIBUTE_STANDARD_TYPE

FILE_ATTRIBUTE_THUMBNAILING_FAILED

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID

FILE_ATTRIBUTE_THUMBNAIL_PATH

FILE_ATTRIBUTE_TIME_ACCESS

FILE_ATTRIBUTE_TIME_ACCESS_USEC

FILE_ATTRIBUTE_TIME_CHANGED

FILE_ATTRIBUTE_TIME_CHANGED_USEC

FILE_ATTRIBUTE_TIME_CREATED

FILE_ATTRIBUTE_TIME_CREATED_USEC

FILE_ATTRIBUTE_TIME_MODIFIED

FILE_ATTRIBUTE_TIME_MODIFIED_USEC

FILE_ATTRIBUTE_TRASH_DELETION_DATE

FILE_ATTRIBUTE_TRASH_ITEM_COUNT

FILE_ATTRIBUTE_TRASH_ORIG_PATH

FILE_ATTRIBUTE_UNIX_BLOCKS

FILE_ATTRIBUTE_UNIX_BLOCK_SIZE

FILE_ATTRIBUTE_UNIX_DEVICE

FILE_ATTRIBUTE_UNIX_GID

FILE_ATTRIBUTE_UNIX_INODE

FILE_ATTRIBUTE_UNIX_IS_MOUNTPOINT

FILE_ATTRIBUTE_UNIX_MODE

FILE_ATTRIBUTE_UNIX_NLINK

FILE_ATTRIBUTE_UNIX_RDEV

FILE_ATTRIBUTE_UNIX_UID

MEMORY_MONITOR_EXTENSION_POINT_NAME	v2_64

MENU_ATTRIBUTE_ACTION

MENU_ATTRIBUTE_ACTION_NAMESPACE

MENU_ATTRIBUTE_ICON

MENU_ATTRIBUTE_LABEL

MENU_ATTRIBUTE_TARGET

MENU_LINK_SECTION

MENU_LINK_SUBMENU

NATIVE_VOLUME_MONITOR_EXTENSION_POINT_NAME

NETWORK_MONITOR_EXTENSION_POINT_NAME

PROXY_EXTENSION_POINT_NAME

PROXY_RESOLVER_EXTENSION_POINT_NAME

SETTINGS_BACKEND_EXTENSION_POINT_NAME

TLS_BACKEND_EXTENSION_POINT_NAME

TLS_DATABASE_PURPOSE_AUTHENTICATE_CLIENT

TLS_DATABASE_PURPOSE_AUTHENTICATE_SERVER

VFS_EXTENSION_POINT_NAME

VOLUME_IDENTIFIER_KIND_CLASS

VOLUME_IDENTIFIER_KIND_HAL_UDI

VOLUME_IDENTIFIER_KIND_LABEL

VOLUME_IDENTIFIER_KIND_NFS_MOUNT

VOLUME_IDENTIFIER_KIND_UNIX_DEVICE

VOLUME_IDENTIFIER_KIND_UUID

VOLUME_MONITOR_EXTENSION_POINT_NAME

Functions

bus_get

bus_get_future

bus_get_sync

bus_own_name

bus_own_name_on_connection

bus_unown_name

bus_unwatch_name

bus_watch_name

bus_watch_name_on_connection

compile_resources
Call from build script to run glib-compile-resources to generate compiled gresources to embed in binary with resources_register_include. target is relative to OUT_DIR.

content_type_can_be_executable

content_type_equals

content_type_from_mime_type

content_type_get_description

content_type_get_generic_icon_name

content_type_get_icon

content_type_get_mime_dirs	v2_60

content_type_get_mime_type

content_type_get_symbolic_icon

content_type_guess

content_type_guess_for_tree

content_type_is_a

content_type_is_mime_type	v2_52

content_type_is_unknown

content_type_set_mime_dirs	v2_60

content_types_get_registered

dbus_address_escape_value

dbus_address_get_for_bus_sync

dbus_address_get_stream

dbus_address_get_stream_future

dbus_address_get_stream_sync

dbus_generate_guid

dbus_gvalue_to_gvariant

dbus_gvariant_to_gvalue

dbus_is_address

dbus_is_guid

dbus_is_interface_name

dbus_is_member_name

dbus_is_name

dbus_is_supported_address

dbus_is_unique_name

io_error_from_errno

io_modules_scan_all_in_directory

io_scheduler_cancel_all_jobs

keyfile_settings_backend_new

memory_settings_backend_new

networking_init

null_settings_backend_new

resources_enumerate_children

resources_get_info

resources_lookup_data

resources_open_stream

resources_register

resources_unregister

unix_is_mount_path_system_internal	Unix

unix_is_system_device_path	Unix and v2_56

unix_is_system_fs_type	Unix and v2_56