Crate gio

Re-exports

pub use ffi;

pub use glib;

Modules

functions

prelude
Traits and essential types intended for blanket imports.

subclass

task

Macros

resources_register_include
Include gresources generated with compile_resources and register with glib. path is relative to OUTDIR.

Structs

Action

ActionGroup

ActionGroupExportId

ActionMap

AppInfo

AppInfoCreateFlags

AppInfoMonitor

AppLaunchContext

Application

ApplicationBuilder

ApplicationCommandLine

ApplicationFlags

AskPasswordFlags

AsyncResult

BufferedInputStream

BufferedInputStreamBuilder

BufferedOutputStream

BufferedOutputStreamBuilder

BusNameOwnerFlags

BusNameWatcherFlags

BytesIcon

Cancellable

CharsetConverter

CharsetConverterBuilder

Converter

ConverterFlags

ConverterInputStream

ConverterInputStreamBuilder

ConverterOutputStream

ConverterOutputStreamBuilder

Credentials

DBusArgInfo

DBusAuthObserver

DBusCallFlags

DBusCapabilityFlags

DBusConnection

DBusConnectionFlags

DBusInterface

DBusInterfaceInfo

DBusInterfaceSkeleton

DBusInterfaceSkeletonFlags

DBusMenuModel

DBusMessage

DBusMessageFlags

DBusMethodInfo

DBusMethodInvocation

DBusNodeInfo

DBusObject

DBusPropertyInfo

DBusProxy

DBusProxyFlags

DBusSendMessageFlags

DBusServer

DBusServerFlags

DBusSignalFlags

DBusSignalInfo

DataInputStream

DataInputStreamBuilder

DataOutputStream

DataOutputStreamBuilder

DesktopAppInfo

Drive

DriveStartFlags

Emblem

EmblemedIcon

File

FileAttributeMatcher

FileAttributematcherIter

FileCopyFlags

FileCreateFlags

FileEnumerator

FileIOStream

FileIcon

FileInfo

FileInputStream

FileMeasureFlags

FileMonitor

FileMonitorFlags

FileOutputStream

FileQueryInfoFlags

FilenameCompleter

FilterId

FilterInputStream

FilterOutputStream

GioFuture

GioFutureResult

IOStream

IOStreamAsyncReadWrite

IOStreamSpliceFlags

Icon

InetAddress

InetAddressMask

InetSocketAddress

InputStream

InputStreamAsyncBufRead

InputStreamAsyncRead

InputStreamRead

ListModel

ListStore

ListStoreBuilder

LoadableIcon

MemoryInputStream

MemoryMonitor

MemoryOutputStream

Menu

MenuAttributeIter

MenuItem

MenuLinkIter

MenuModel

MenuModelExportId

Mount

MountMountFlags

MountOperation

MountUnmountFlags

NetworkAddress

NetworkMonitor

NetworkService

Notification

OutputStream

OutputStreamAsyncWrite

OutputStreamSpliceFlags

OutputStreamWrite

OwnerId

Permission

PollableInputStream

PollableOutputStream

PropertyAction

Proxy

ProxyAddress

ProxyResolver

ReadInputStream

RegistrationId

RemoteActionGroup

Resolver

ResolverNameLookupFlags

Resource

ResourceLookupFlags

Seekable

Settings

SettingsBackend

SettingsBindFlags

SettingsSchema

SettingsSchemaKey

SettingsSchemaSource

SignalSubscriptionId

SimpleAction

SimpleActionGroup

SimpleIOStream

SimplePermission

Socket

SocketAddress

SocketAddressEnumerator

SocketClient

SocketConnectable

SocketConnection

SocketListener

SocketService

SrvTarget

Subprocess

SubprocessFlags

SubprocessLauncher

Task

TcpConnection

ThemedIcon

ThreadedSocketService

TlsBackend

TlsCertificate

TlsCertificateFlags

TlsClientConnection

TlsConnection

TlsDatabase

TlsDatabaseVerifyFlags

TlsFileDatabase

TlsInteraction

TlsPassword

TlsPasswordFlags

TlsServerConnection

UnixFDList

UnixInputStream

UnixMountEntry

UnixMountPoint

UnixOutputStream

UnixSocketAddress

Vfs

Volume

VolumeMonitor

WatcherId

WriteOutputStream

ZlibCompressor

ZlibDecompressor

Enums

BusType
An enumeration for well-known message buses.

ConverterResult
Results returned from Converter::convert.

CredentialsType
Enumeration describing different kinds of native credential types.

DBusMessageByteOrder
Enumeration used to describe the byte order of a D-Bus message.

DBusMessageHeaderField
Header fields used in DBusMessage.

DBusMessageType
Message types used in DBusMessage.

DataStreamByteOrder
DataStreamByteOrder is used to ensure proper endianness of streaming data sources across various machine architectures.

DataStreamNewlineType
DataStreamNewlineType is used when checking for or setting the line endings for a given file.

DriveStartStopType
Enumeration describing how a drive can be started/stopped.

EmblemOrigin
GEmblemOrigin is used to add information about the origin of the emblem to Emblem.

FileAttributeStatus
Used by File::set_attributes_from_info when setting file attributes.

FileAttributeType
The data types for file attributes.

FileMonitorEvent
Specifies what type of event a monitor event is.

FileType
Indicates the file's on-disk type.

IOErrorEnum
Error codes returned by GIO functions.

InetAddressBytes

MemoryMonitorWarningLevel	v2_64
Memory availability warning levels.

MountOperationResult
MountOperationResult is returned as a result when a request for information is send by the mounting operation.

NetworkConnectivity	v2_44
The host's network connectivity state, as reported by NetworkMonitor.

NotificationPriority
Priority levels for GNotifications.

PasswordSave
PasswordSave is used to indicate the lifespan of a saved password.

ResolverRecordType
The type of record that ResolverExt::lookup_records or ResolverExt::lookup_records_async should retrieve. The records are returned as lists of glib::Variant tuples. Each record type has different values in the variant tuples returned.

ResourceError
An error code used with G_RESOURCE_ERROR in a glib::Error returned from a Resource routine.

SocketClientEvent
Describes an event occurring on a SocketClient. See the SocketClient::event signal for more details.

SocketFamily
The protocol family of a SocketAddress. (These values are identical to the system defines AF_INET, AF_INET6 and AF_UNIX, if available.)

SocketListenerEvent	v2_46
Describes an event occurring on a SocketListener. See the SocketListener::event signal for more details.

SocketProtocol
A protocol identifier is specified when creating a Socket, which is a family/type specific identifier, where 0 means the default protocol for the particular family/type.

SocketType
Flags used when creating a Socket. Some protocols may not implement all the socket types.

TlsAuthenticationMode
The client authentication mode for a TlsServerConnection.

TlsCertificateRequestFlags
Flags for TlsInteractionExt::request_certificate, TlsInteractionExt::request_certificate_async, and TlsInteractionExt::invoke_request_certificate.

TlsChannelBindingType	v2_66
The type of TLS channel binding data to retrieve from TlsConnection or DtlsConnection, as documented by RFC 5929. The tls-unique-for-telnet binding type is not currently implemented.

TlsDatabaseLookupFlags
Flags for TlsDatabaseExt::lookup_certificate_for_handle, TlsDatabaseExt::lookup_certificate_issuer, and TlsDatabaseExt::lookup_certificates_issued_by.

TlsInteractionResult
TlsInteractionResult is returned by various functions in TlsInteraction when finishing an interaction request.

TlsRehandshakeMode	Deprecated
When to allow rehandshaking. See TlsConnectionExt::set_rehandshake_mode.

UnixSocketAddressType
The type of name used by a UnixSocketAddress. UnixSocketAddressType::Path indicates a traditional unix domain socket bound to a filesystem path. UnixSocketAddressType::Anonymous indicates a socket not bound to any name (eg, a client-side socket, or a socket created with socketpair).

ZlibCompressorFormat
Used to select the type of data format to use for ZlibDecompressor and ZlibCompressor.

Constants

NONE_ACTION

NONE_ACTION_GROUP

NONE_ACTION_MAP

NONE_APPLICATION

NONE_APPLICATION_COMMAND_LINE

NONE_APP_INFO

NONE_APP_LAUNCH_CONTEXT

NONE_ASYNC_RESULT

NONE_BUFFERED_INPUT_STREAM

NONE_BUFFERED_OUTPUT_STREAM

NONE_CANCELLABLE

NONE_CONVERTER

NONE_CONVERTER_INPUT_STREAM

NONE_CONVERTER_OUTPUT_STREAM

NONE_DATA_INPUT_STREAM

NONE_DATA_OUTPUT_STREAM

NONE_DBUS_INTERFACE

NONE_DBUS_INTERFACE_SKELETON

NONE_DBUS_OBJECT

NONE_DBUS_PROXY

NONE_DRIVE

NONE_EMBLEMED_ICON

NONE_FILE

NONE_FILE_ENUMERATOR

NONE_FILE_INPUT_STREAM

NONE_FILE_IO_STREAM

NONE_FILE_MONITOR

NONE_FILE_OUTPUT_STREAM

NONE_FILTER_INPUT_STREAM

NONE_FILTER_OUTPUT_STREAM

NONE_ICON

NONE_INET_ADDRESS

NONE_INET_ADDRESS_MASK

NONE_INET_SOCKET_ADDRESS

NONE_INPUT_STREAM

NONE_IO_STREAM

NONE_LIST_MODEL

NONE_LOADABLE_ICON

NONE_MEMORY_INPUT_STREAM

NONE_MEMORY_MONITOR

NONE_MEMORY_OUTPUT_STREAM

NONE_MENU_ATTRIBUTE_ITER

NONE_MENU_LINK_ITER

NONE_MENU_MODEL

NONE_MOUNT

NONE_MOUNT_OPERATION

NONE_NETWORK_ADDRESS

NONE_NETWORK_MONITOR

NONE_NETWORK_SERVICE

NONE_OUTPUT_STREAM

NONE_PERMISSION

NONE_POLLABLE_INPUT_STREAM

NONE_POLLABLE_OUTPUT_STREAM

NONE_PROXY

NONE_PROXY_ADDRESS

NONE_PROXY_RESOLVER

NONE_REMOTE_ACTION_GROUP

NONE_RESOLVER

NONE_SEEKABLE

NONE_SETTINGS

NONE_SETTINGS_BACKEND

NONE_SIMPLE_ACTION_GROUP

NONE_SOCKET

NONE_SOCKET_ADDRESS

NONE_SOCKET_ADDRESS_ENUMERATOR

NONE_SOCKET_CLIENT

NONE_SOCKET_CONNECTABLE

NONE_SOCKET_CONNECTION

NONE_SOCKET_LISTENER

NONE_SOCKET_SERVICE

NONE_TCP_CONNECTION

NONE_THREADED_SOCKET_SERVICE

NONE_TLS_BACKEND

NONE_TLS_CERTIFICATE

NONE_TLS_CLIENT_CONNECTION

NONE_TLS_CONNECTION

NONE_TLS_DATABASE

NONE_TLS_FILE_DATABASE

NONE_TLS_INTERACTION

NONE_TLS_PASSWORD

NONE_TLS_SERVER_CONNECTION

NONE_UNIX_FD_LIST

NONE_UNIX_INPUT_STREAM

NONE_UNIX_OUTPUT_STREAM

NONE_UNIX_SOCKET_ADDRESS

NONE_VFS

NONE_VOLUME

NONE_VOLUME_MONITOR

Statics

DRIVE_IDENTIFIER_KIND_UNIX_DEVICE	v2_58

FILE_ATTRIBUTE_ACCESS_CAN_DELETE

FILE_ATTRIBUTE_ACCESS_CAN_EXECUTE

FILE_ATTRIBUTE_ACCESS_CAN_READ

FILE_ATTRIBUTE_ACCESS_CAN_RENAME

FILE_ATTRIBUTE_ACCESS_CAN_TRASH

FILE_ATTRIBUTE_ACCESS_CAN_WRITE

FILE_ATTRIBUTE_DOS_IS_ARCHIVE

FILE_ATTRIBUTE_DOS_IS_MOUNTPOINT	v2_60

FILE_ATTRIBUTE_DOS_IS_SYSTEM

FILE_ATTRIBUTE_DOS_REPARSE_POINT_TAG	v2_60

FILE_ATTRIBUTE_ETAG_VALUE

FILE_ATTRIBUTE_FILESYSTEM_FREE

FILE_ATTRIBUTE_FILESYSTEM_READONLY

FILE_ATTRIBUTE_FILESYSTEM_REMOTE

FILE_ATTRIBUTE_FILESYSTEM_SIZE

FILE_ATTRIBUTE_FILESYSTEM_TYPE

FILE_ATTRIBUTE_FILESYSTEM_USED

FILE_ATTRIBUTE_FILESYSTEM_USE_PREVIEW

FILE_ATTRIBUTE_GVFS_BACKEND

FILE_ATTRIBUTE_ID_FILE

FILE_ATTRIBUTE_ID_FILESYSTEM

FILE_ATTRIBUTE_MOUNTABLE_CAN_EJECT

FILE_ATTRIBUTE_MOUNTABLE_CAN_MOUNT

FILE_ATTRIBUTE_MOUNTABLE_CAN_POLL

FILE_ATTRIBUTE_MOUNTABLE_CAN_START

FILE_ATTRIBUTE_MOUNTABLE_CAN_START_DEGRADED

FILE_ATTRIBUTE_MOUNTABLE_CAN_STOP

FILE_ATTRIBUTE_MOUNTABLE_CAN_UNMOUNT

FILE_ATTRIBUTE_MOUNTABLE_HAL_UDI

FILE_ATTRIBUTE_MOUNTABLE_IS_MEDIA_CHECK_AUTOMATIC

FILE_ATTRIBUTE_MOUNTABLE_START_STOP_TYPE

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE_FILE

FILE_ATTRIBUTE_OWNER_GROUP

FILE_ATTRIBUTE_OWNER_USER

FILE_ATTRIBUTE_OWNER_USER_REAL

FILE_ATTRIBUTE_PREVIEW_ICON

FILE_ATTRIBUTE_RECENT_MODIFIED	v2_52

FILE_ATTRIBUTE_SELINUX_CONTEXT

FILE_ATTRIBUTE_STANDARD_ALLOCATED_SIZE

FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE

FILE_ATTRIBUTE_STANDARD_COPY_NAME

FILE_ATTRIBUTE_STANDARD_DESCRIPTION

FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME

FILE_ATTRIBUTE_STANDARD_EDIT_NAME

FILE_ATTRIBUTE_STANDARD_FAST_CONTENT_TYPE

FILE_ATTRIBUTE_STANDARD_ICON

FILE_ATTRIBUTE_STANDARD_IS_BACKUP

FILE_ATTRIBUTE_STANDARD_IS_HIDDEN

FILE_ATTRIBUTE_STANDARD_IS_SYMLINK

FILE_ATTRIBUTE_STANDARD_IS_VIRTUAL

FILE_ATTRIBUTE_STANDARD_IS_VOLATILE	v2_46

FILE_ATTRIBUTE_STANDARD_NAME

FILE_ATTRIBUTE_STANDARD_SIZE

FILE_ATTRIBUTE_STANDARD_SORT_ORDER

FILE_ATTRIBUTE_STANDARD_SYMBOLIC_ICON

FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET

FILE_ATTRIBUTE_STANDARD_TARGET_URI

FILE_ATTRIBUTE_STANDARD_TYPE

FILE_ATTRIBUTE_THUMBNAILING_FAILED

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID

FILE_ATTRIBUTE_THUMBNAIL_PATH

FILE_ATTRIBUTE_TIME_ACCESS

FILE_ATTRIBUTE_TIME_ACCESS_USEC

FILE_ATTRIBUTE_TIME_CHANGED

FILE_ATTRIBUTE_TIME_CHANGED_USEC

FILE_ATTRIBUTE_TIME_CREATED

FILE_ATTRIBUTE_TIME_CREATED_USEC

FILE_ATTRIBUTE_TIME_MODIFIED

FILE_ATTRIBUTE_TIME_MODIFIED_USEC

FILE_ATTRIBUTE_TRASH_DELETION_DATE

FILE_ATTRIBUTE_TRASH_ITEM_COUNT

FILE_ATTRIBUTE_TRASH_ORIG_PATH

FILE_ATTRIBUTE_UNIX_BLOCKS

FILE_ATTRIBUTE_UNIX_BLOCK_SIZE

FILE_ATTRIBUTE_UNIX_DEVICE

FILE_ATTRIBUTE_UNIX_GID

FILE_ATTRIBUTE_UNIX_INODE

FILE_ATTRIBUTE_UNIX_IS_MOUNTPOINT

FILE_ATTRIBUTE_UNIX_MODE

FILE_ATTRIBUTE_UNIX_NLINK

FILE_ATTRIBUTE_UNIX_RDEV

FILE_ATTRIBUTE_UNIX_UID

MEMORY_MONITOR_EXTENSION_POINT_NAME	v2_64

MENU_ATTRIBUTE_ACTION

MENU_ATTRIBUTE_ACTION_NAMESPACE

MENU_ATTRIBUTE_ICON

MENU_ATTRIBUTE_LABEL

MENU_ATTRIBUTE_TARGET

MENU_LINK_SECTION

MENU_LINK_SUBMENU

NATIVE_VOLUME_MONITOR_EXTENSION_POINT_NAME

NETWORK_MONITOR_EXTENSION_POINT_NAME

PROXY_EXTENSION_POINT_NAME

PROXY_RESOLVER_EXTENSION_POINT_NAME

SETTINGS_BACKEND_EXTENSION_POINT_NAME

TLS_BACKEND_EXTENSION_POINT_NAME

TLS_DATABASE_PURPOSE_AUTHENTICATE_CLIENT

TLS_DATABASE_PURPOSE_AUTHENTICATE_SERVER

VFS_EXTENSION_POINT_NAME

VOLUME_IDENTIFIER_KIND_CLASS

VOLUME_IDENTIFIER_KIND_HAL_UDI	Deprecated

VOLUME_IDENTIFIER_KIND_LABEL

VOLUME_IDENTIFIER_KIND_NFS_MOUNT

VOLUME_IDENTIFIER_KIND_UNIX_DEVICE

VOLUME_IDENTIFIER_KIND_UUID

VOLUME_MONITOR_EXTENSION_POINT_NAME

Traits

ActionExt
Trait containing all Action methods.

ActionGroupExt
Trait containing all ActionGroup methods.

ActionMapExt
Trait containing all ActionMap methods.

AppInfoExt
Trait containing all AppInfo methods.

AppLaunchContextExt
Trait containing all AppLaunchContext methods.

ApplicationCommandLineExt
Trait containing all ApplicationCommandLine methods.

ApplicationExt
Trait containing all Application methods.

AsyncResultExt
Trait containing all AsyncResult methods.

BufferedInputStreamExt
Trait containing all BufferedInputStream methods.

BufferedOutputStreamExt
Trait containing all BufferedOutputStream methods.

CancellableExt
Trait containing all Cancellable methods.

ConverterExt
Trait containing all Converter methods.

ConverterInputStreamExt
Trait containing all ConverterInputStream methods.

ConverterOutputStreamExt
Trait containing all ConverterOutputStream methods.

DBusInterfaceExt
Trait containing all DBusInterface methods.

DBusInterfaceSkeletonExt
Trait containing all DBusInterfaceSkeleton methods.

DBusObjectExt
Trait containing all DBusObject methods.

DBusProxyExt
Trait containing all DBusProxy methods.

DataInputStreamExt
Trait containing all DataInputStream methods.

DataOutputStreamExt
Trait containing all DataOutputStream methods.

DriveExt
Trait containing all Drive methods.

EmblemedIconExt
Trait containing all EmblemedIcon methods.

FileEnumeratorExt
Trait containing all FileEnumerator methods.

FileExt
Trait containing all File methods.

FileIOStreamExt
Trait containing all FileIOStream methods.

FileInputStreamExt
Trait containing all FileInputStream methods.

FileMonitorExt
Trait containing all FileMonitor methods.

FileOutputStreamExt
Trait containing all FileOutputStream methods.

FilterInputStreamExt
Trait containing all FilterInputStream methods.

FilterOutputStreamExt
Trait containing all FilterOutputStream methods.

IOStreamExt
Trait containing all IOStream methods.

IconExt
Trait containing all Icon methods.

InetAddressExt
Trait containing all InetAddress methods.

InetAddressMaskExt
Trait containing all InetAddressMask methods.

InetSocketAddressExt
Trait containing all InetSocketAddress methods.

InputStreamExt
Trait containing all InputStream methods.

ListModelExt
Trait containing all ListModel methods.

LoadableIconExt
Trait containing all LoadableIcon methods.

MemoryInputStreamExt
Trait containing all MemoryInputStream methods.

MemoryMonitorExt
Trait containing all MemoryMonitor methods.

MemoryOutputStreamExt
Trait containing all MemoryOutputStream methods.

MenuAttributeIterExt
Trait containing all MenuAttributeIter methods.

MenuLinkIterExt
Trait containing all MenuLinkIter methods.

MenuModelExt
Trait containing all MenuModel methods.

MountExt
Trait containing all Mount methods.

MountOperationExt
Trait containing all MountOperation methods.

NetworkAddressExt
Trait containing all NetworkAddress methods.

NetworkMonitorExt
Trait containing all NetworkMonitor methods.

NetworkServiceExt
Trait containing all NetworkService methods.

OutputStreamExt
Trait containing all OutputStream methods.

PermissionExt
Trait containing all Permission methods.

PollableInputStreamExt
Trait containing all PollableInputStream methods.

PollableOutputStreamExt
Trait containing all PollableOutputStream methods.

ProxyAddressExt
Trait containing all ProxyAddress methods.

ProxyExt
Trait containing all Proxy methods.

ProxyResolverExt
Trait containing all ProxyResolver methods.

RemoteActionGroupExt
Trait containing all RemoteActionGroup methods.

ResolverExt
Trait containing all Resolver methods.

SeekableExt
Trait containing all Seekable methods.

SettingsBackendExt
Trait containing all SettingsBackend methods.

SettingsExt
Trait containing all Settings methods.

SocketAddressEnumeratorExt
Trait containing all SocketAddressEnumerator methods.

SocketAddressExt
Trait containing all SocketAddress methods.

SocketClientExt
Trait containing all SocketClient methods.

SocketConnectableExt
Trait containing all SocketConnectable methods.

SocketConnectionExt
Trait containing all SocketConnection methods.

SocketExt
Trait containing all Socket methods.

SocketListenerExt
Trait containing all SocketListener methods.

SocketServiceExt
Trait containing all SocketService methods.

TcpConnectionExt
Trait containing all TcpConnection methods.

ThreadedSocketServiceExt
Trait containing all ThreadedSocketService methods.

TlsBackendExt
Trait containing all TlsBackend methods.

TlsCertificateExt
Trait containing all TlsCertificate methods.

TlsClientConnectionExt
Trait containing all TlsClientConnection methods.

TlsConnectionExt
Trait containing all TlsConnection methods.

TlsConnectionManualExt

TlsDatabaseExt
Trait containing all TlsDatabase methods.

TlsFileDatabaseExt
Trait containing all TlsFileDatabase methods.

TlsInteractionExt
Trait containing all TlsInteraction methods.

TlsPasswordExt
Trait containing all TlsPassword methods.

TlsServerConnectionExt
Trait containing all TlsServerConnection methods.

UnixFDListExt
Trait containing all UnixFDList methods.

UnixInputStreamExt
Trait containing all UnixInputStream methods.

UnixOutputStreamExt
Trait containing all UnixOutputStream methods.

UnixSocketAddressExt
Trait containing all UnixSocketAddress methods.

VfsExt
Trait containing all Vfs methods.

VolumeExt
Trait containing all Volume methods.

VolumeMonitorExt
Trait containing all VolumeMonitor methods.

Functions

bus_get

bus_get_future

bus_get_sync

bus_own_name

bus_own_name_on_connection

bus_unown_name

bus_unwatch_name

bus_watch_name

bus_watch_name_on_connection

compile_resources
Call from build script to run glib-compile-resources to generate compiled gresources to embed in binary with resources_register_include. target is relative to OUT_DIR.

content_type_can_be_executable

content_type_equals

content_type_from_mime_type

content_type_get_description

content_type_get_generic_icon_name

content_type_get_icon

content_type_get_mime_dirs	v2_60

content_type_get_mime_type

content_type_get_symbolic_icon

content_type_guess

content_type_guess_for_tree

content_type_is_a

content_type_is_mime_type	v2_52

content_type_is_unknown

content_type_set_mime_dirs	v2_60

content_types_get_registered

dbus_address_escape_value

dbus_address_get_for_bus_sync

dbus_address_get_stream

dbus_address_get_stream_future

dbus_address_get_stream_sync

dbus_generate_guid

dbus_gvalue_to_gvariant

dbus_gvariant_to_gvalue

dbus_is_address

dbus_is_guid

dbus_is_interface_name

dbus_is_member_name

dbus_is_name

dbus_is_supported_address

dbus_is_unique_name

io_error_from_errno

io_modules_scan_all_in_directory

io_scheduler_cancel_all_jobs

keyfile_settings_backend_new

memory_settings_backend_new

networking_init

null_settings_backend_new

resources_enumerate_children

resources_get_info

resources_lookup_data

resources_open_stream

resources_register

resources_unregister

unix_is_mount_path_system_internal	Unix

unix_is_system_device_path	Unix and v2_56

unix_is_system_fs_type	Unix and v2_56