Crate gio

Rust GIO bindings

Rust bindings and wrappers for GIO, part of gtk-rs-core.

GIO 2.48 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.56.0.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gio = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gio" }

Avoid mixing versioned and git crates like this:

[dependencies]
gio = "0.13"
gio = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gio" }

See Also

glib

License

gio is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use glib;

Modules

functions

prelude

Traits and essential types intended for blanket imports.

subclass

Macros

resources_register_include

Include gresources generated with compile_resources and register with glib. path is relative to OUTDIR.

Structs

Action

ActionGroup

ActionGroupExportId

ActionMap

AppInfo

AppInfoCreateFlags

AppInfoMonitor

AppLaunchContext

Application

ApplicationCommandLine

ApplicationFlags

AskPasswordFlags

AsyncInitable

AsyncResult

BindingBuilder

BufferedInputStream

BufferedOutputStream

BusNameOwnerFlags

BusNameWatcherFlags

BytesIcon

Cancellable

CharsetConverter

Converter

ConverterFlags

ConverterInputStream

ConverterOutputStream

Credentials

DBusActionGroup

DBusArgInfo

DBusAuthObserver

DBusCallFlags

DBusCapabilityFlags

DBusConnection

DBusConnectionFlags

DBusInterface

DBusInterfaceInfo

DBusInterfaceSkeleton

DBusInterfaceSkeletonFlags

DBusMenuModel

DBusMessage

DBusMessageFlags

DBusMethodInfo

DBusMethodInvocation

DBusNodeInfo

DBusObject

DBusPropertyInfo

DBusProxy

DBusProxyFlags

DBusSendMessageFlags

DBusServer

DBusServerFlags

DBusSignalFlags

DBusSignalInfo

DataInputStream

DataOutputStream

DesktopAppInfo

Drive

DriveStartFlags

Emblem

EmblemedIcon

File

FileAttributeInfo

FileAttributeInfoFlags

FileAttributeInfoList

FileAttributeMatcher

FileAttributematcherIter

FileCopyFlags

FileCreateFlags

FileEnumerator

FileIOStream

FileIcon

FileInfo

FileInputStream

FileMeasureFlags

FileMonitor

FileMonitorFlags

FileOutputStream

FileQueryInfoFlags

FilenameCompleter

FilterId

FilterInputStream

FilterOutputStream

GioFuture

GioFutureResult

IOExtension

The implementation of an IOExtensionPoint.

IOExtensionPoint

An extension point provides a mechanism to extend the functionality of a library or application. Each extension point is identified by a name, and it may optionally require that any implementation must be of a certain type.

IOExtensionPointBuilder

Builder for extension points.

IOStream

IOStreamAsyncReadWrite

IOStreamSpliceFlags

Icon

InetAddress

InetAddressMask

InetSocketAddress

Initable

InputStream

InputStreamAsyncBufRead

InputStreamAsyncRead

InputStreamRead

ListModel

ListStore

LoadableIcon

LocalTask

LocalTask provides idiomatic access to gio’s GTask API, for instance by being generic over their value type, while not completely departing from the underlying C API. LocalTask does not require its value to be Send and Sync and thus is useful to to implement gio style asynchronous tasks that run in the glib main loop. If you need to run tasks in threads see the Task type.

MemoryInputStream

MemoryOutputStream

Menu

MenuAttributeIter

MenuItem

MenuLinkIter

MenuModel

MenuModelExportId

Mount

MountMountFlags

MountOperation

MountUnmountFlags

NetworkAddress

NetworkMonitor

NetworkService

Notification

OutputStream

OutputStreamAsyncWrite

OutputStreamSpliceFlags

OutputStreamWrite

OwnerId

Permission

PollableInputStream

PollableOutputStream

PropertyAction

Proxy

ProxyAddress

ProxyResolver

ReadInputStream

RegistrationId

RemoteActionGroup

Resolver

Resource

ResourceLookupFlags

Seekable

Settings

SettingsBackend

SettingsBindFlags

SettingsSchema

SettingsSchemaKey

SettingsSchemaSource

SignalSubscriptionId

SimpleAction

SimpleActionGroup

SimpleIOStream

SimplePermission

SimpleProxyResolver

Socket

SocketAddress

SocketAddressEnumerator

SocketClient

SocketConnectable

SocketConnection

SocketListener

SocketService

SrvTarget

Subprocess

SubprocessFlags

SubprocessLauncher

Task

Task provides idiomatic access to gio’s GTask API, for instance by being generic over their value type, while not completely departing from the underlying C API. Task is Send and Sync and requires its value to also be Send and Sync, thus is useful to to implement gio style asynchronous tasks that run in threads. If you need to only run tasks in glib main loop see the LocalTask type.

TcpConnection

ThemedIcon

ThreadedSocketService

TlsBackend

TlsCertificate

TlsCertificateFlags

TlsClientConnection

TlsConnection

TlsDatabase

TlsDatabaseVerifyFlags

TlsFileDatabase

TlsInteraction

TlsPassword

TlsPasswordFlags

TlsServerConnection

UnixFDList

UnixInputStream

UnixOutputStream

UnixSocketAddress

Vfs

Volume

VolumeMonitor

WatcherId

WriteOutputStream

ZlibCompressor

ZlibDecompressor

Enums

BusType

ConverterResult

CredentialsType

DBusMessageByteOrder

DBusMessageHeaderField

DBusMessageType

DataStreamByteOrder

DataStreamNewlineType

DriveStartStopType

EmblemOrigin

FileAttributeStatus

FileAttributeType

FileMonitorEvent

FileType

IOErrorEnum

InetAddressBytes

MountOperationResult

NetworkConnectivity

NotificationPriority

PasswordSave

ResolverError

ResolverRecordType

ResourceError

SocketClientEvent

SocketFamily

SocketListenerEvent

SocketProtocol

SocketType

TlsAuthenticationMode

TlsCertificateRequestFlags

TlsDatabaseLookupFlags

TlsError

TlsInteractionResult

TlsRehandshakeMode

UnixSocketAddressType

ZlibCompressorFormat

Statics

FILE_ATTRIBUTE_ACCESS_CAN_DELETE

FILE_ATTRIBUTE_ACCESS_CAN_EXECUTE

FILE_ATTRIBUTE_ACCESS_CAN_READ

FILE_ATTRIBUTE_ACCESS_CAN_RENAME

FILE_ATTRIBUTE_ACCESS_CAN_TRASH

FILE_ATTRIBUTE_ACCESS_CAN_WRITE

FILE_ATTRIBUTE_DOS_IS_ARCHIVE

FILE_ATTRIBUTE_DOS_IS_SYSTEM

FILE_ATTRIBUTE_ETAG_VALUE

FILE_ATTRIBUTE_FILESYSTEM_FREE

FILE_ATTRIBUTE_FILESYSTEM_READONLY

FILE_ATTRIBUTE_FILESYSTEM_REMOTE

FILE_ATTRIBUTE_FILESYSTEM_SIZE

FILE_ATTRIBUTE_FILESYSTEM_TYPE

FILE_ATTRIBUTE_FILESYSTEM_USED

FILE_ATTRIBUTE_FILESYSTEM_USE_PREVIEW

FILE_ATTRIBUTE_GVFS_BACKEND

FILE_ATTRIBUTE_ID_FILE

FILE_ATTRIBUTE_ID_FILESYSTEM

FILE_ATTRIBUTE_MOUNTABLE_CAN_EJECT

FILE_ATTRIBUTE_MOUNTABLE_CAN_MOUNT

FILE_ATTRIBUTE_MOUNTABLE_CAN_POLL

FILE_ATTRIBUTE_MOUNTABLE_CAN_START

FILE_ATTRIBUTE_MOUNTABLE_CAN_START_DEGRADED

FILE_ATTRIBUTE_MOUNTABLE_CAN_STOP

FILE_ATTRIBUTE_MOUNTABLE_CAN_UNMOUNT

FILE_ATTRIBUTE_MOUNTABLE_HAL_UDI

FILE_ATTRIBUTE_MOUNTABLE_IS_MEDIA_CHECK_AUTOMATIC

FILE_ATTRIBUTE_MOUNTABLE_START_STOP_TYPE

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE

FILE_ATTRIBUTE_MOUNTABLE_UNIX_DEVICE_FILE

FILE_ATTRIBUTE_OWNER_GROUP

FILE_ATTRIBUTE_OWNER_USER

FILE_ATTRIBUTE_OWNER_USER_REAL

FILE_ATTRIBUTE_PREVIEW_ICON

FILE_ATTRIBUTE_SELINUX_CONTEXT

FILE_ATTRIBUTE_STANDARD_ALLOCATED_SIZE

FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE

FILE_ATTRIBUTE_STANDARD_COPY_NAME

FILE_ATTRIBUTE_STANDARD_DESCRIPTION

FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME

FILE_ATTRIBUTE_STANDARD_EDIT_NAME

FILE_ATTRIBUTE_STANDARD_FAST_CONTENT_TYPE

FILE_ATTRIBUTE_STANDARD_ICON

FILE_ATTRIBUTE_STANDARD_IS_BACKUP

FILE_ATTRIBUTE_STANDARD_IS_HIDDEN

FILE_ATTRIBUTE_STANDARD_IS_SYMLINK

FILE_ATTRIBUTE_STANDARD_IS_VIRTUAL

FILE_ATTRIBUTE_STANDARD_IS_VOLATILE

FILE_ATTRIBUTE_STANDARD_NAME

FILE_ATTRIBUTE_STANDARD_SIZE

FILE_ATTRIBUTE_STANDARD_SORT_ORDER

FILE_ATTRIBUTE_STANDARD_SYMBOLIC_ICON

FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET

FILE_ATTRIBUTE_STANDARD_TARGET_URI

FILE_ATTRIBUTE_STANDARD_TYPE

FILE_ATTRIBUTE_THUMBNAILING_FAILED

FILE_ATTRIBUTE_THUMBNAIL_IS_VALID

FILE_ATTRIBUTE_THUMBNAIL_PATH

FILE_ATTRIBUTE_TIME_ACCESS

FILE_ATTRIBUTE_TIME_ACCESS_USEC

FILE_ATTRIBUTE_TIME_CHANGED

FILE_ATTRIBUTE_TIME_CHANGED_USEC

FILE_ATTRIBUTE_TIME_CREATED

FILE_ATTRIBUTE_TIME_CREATED_USEC

FILE_ATTRIBUTE_TIME_MODIFIED

FILE_ATTRIBUTE_TIME_MODIFIED_USEC

FILE_ATTRIBUTE_TRASH_DELETION_DATE

FILE_ATTRIBUTE_TRASH_ITEM_COUNT

FILE_ATTRIBUTE_TRASH_ORIG_PATH

FILE_ATTRIBUTE_UNIX_BLOCKS

FILE_ATTRIBUTE_UNIX_BLOCK_SIZE

FILE_ATTRIBUTE_UNIX_DEVICE

FILE_ATTRIBUTE_UNIX_GID

FILE_ATTRIBUTE_UNIX_INODE

FILE_ATTRIBUTE_UNIX_IS_MOUNTPOINT

FILE_ATTRIBUTE_UNIX_MODE

FILE_ATTRIBUTE_UNIX_NLINK

FILE_ATTRIBUTE_UNIX_RDEV

FILE_ATTRIBUTE_UNIX_UID

MENU_ATTRIBUTE_ACTION

MENU_ATTRIBUTE_ACTION_NAMESPACE

MENU_ATTRIBUTE_ICON

MENU_ATTRIBUTE_LABEL

MENU_ATTRIBUTE_TARGET

MENU_LINK_SECTION

MENU_LINK_SUBMENU

NATIVE_VOLUME_MONITOR_EXTENSION_POINT_NAME

NETWORK_MONITOR_EXTENSION_POINT_NAME

PROXY_EXTENSION_POINT_NAME

PROXY_RESOLVER_EXTENSION_POINT_NAME

SETTINGS_BACKEND_EXTENSION_POINT_NAME

TLS_BACKEND_EXTENSION_POINT_NAME

TLS_DATABASE_PURPOSE_AUTHENTICATE_CLIENT

TLS_DATABASE_PURPOSE_AUTHENTICATE_SERVER

VFS_EXTENSION_POINT_NAME

VOLUME_IDENTIFIER_KIND_CLASS

VOLUME_IDENTIFIER_KIND_HAL_UDI

VOLUME_IDENTIFIER_KIND_LABEL

VOLUME_IDENTIFIER_KIND_NFS_MOUNT

VOLUME_IDENTIFIER_KIND_UNIX_DEVICE

VOLUME_IDENTIFIER_KIND_UUID

VOLUME_MONITOR_EXTENSION_POINT_NAME

Functions

bus_get

bus_get_future

bus_get_sync

bus_own_name

bus_own_name_on_connection

bus_unown_name

bus_unwatch_name

bus_watch_name

bus_watch_name_on_connection

compile_resources

Call from build script to run glib-compile-resources to generate compiled gresources to embed in binary with resources_register_include. target is relative to OUT_DIR.

content_type_can_be_executable

content_type_equals

content_type_from_mime_type

content_type_get_description

content_type_get_generic_icon_name

content_type_get_icon

content_type_get_mime_type

content_type_get_symbolic_icon

content_type_guess

content_type_guess_for_tree

content_type_is_a

content_type_is_unknown

content_types_get_registered

dbus_address_escape_value

dbus_address_get_for_bus_sync

dbus_address_get_stream

dbus_address_get_stream_future

dbus_address_get_stream_sync

dbus_generate_guid

dbus_gvalue_to_gvariant

dbus_gvariant_to_gvalue

dbus_is_address

dbus_is_guid

dbus_is_interface_name

dbus_is_member_name

dbus_is_name

dbus_is_supported_address

dbus_is_unique_name

io_error_from_errno

io_modules_scan_all_in_directory

keyfile_settings_backend_new

memory_settings_backend_new

null_settings_backend_new

resources_enumerate_children

resources_get_info

resources_lookup_data

resources_open_stream

resources_register

resources_unregister

unix_is_mount_path_system_internal