Function gio::functions::dbus_gvalue_to_gvariant

pub fn dbus_gvalue_to_gvariant(gvalue: &Value, type_: &VariantTy) -> Variant

Converts a glib::Value to a glib::Variant of the type indicated by the type_ parameter.

The conversion is using the following rules:

G_TYPE_STRING: ‘s’, ‘o’, ‘g’ or ‘ay’

G_TYPE_STRV: ‘as’, ‘ao’ or ‘aay’

G_TYPE_BOOLEAN: ‘b’

G_TYPE_UCHAR: ‘y’

G_TYPE_INT: ‘i’, ‘n’

G_TYPE_UINT: ‘u’, ‘q’

G_TYPE_INT64: ‘x’

G_TYPE_UINT64: ‘t’

G_TYPE_DOUBLE: ‘d’

G_TYPE_VARIANT: Any glib::VariantType

This can fail if e.g. gvalue is of type G_TYPE_STRING and type_ is ‘i’, i.e. G_VARIANT_TYPE_INT32. It will also fail for any GType (including e.g. G_TYPE_OBJECT and G_TYPE_BOXED derived-types) not in the table above.

Note that if gvalue is of type G_TYPE_VARIANT and its value is None, the empty glib::Variant instance (never None) for type_ is returned (e.g. 0 for scalar types, the empty string for string types, ‘/’ for object path types, the empty array for any array type and so on).

See the dbus_gvariant_to_gvalue() function for how to convert a glib::Variant to a glib::Value.

gvalue

A glib::Value to convert to a glib::Variant

type_

A glib::VariantType

Returns

A glib::Variant (never floating) of glib::VariantType type_ holding the data from gvalue or an empty glib::Variant in case of failure. Free with g_variant_unref().