Function gio::functions::io_error_from_errno

pub fn io_error_from_errno(err_no: i32) -> IOErrorEnum

Converts errno.h error codes into GIO error codes. The fallback value IOErrorEnum::Failed is returned for error codes not currently handled (but note that future GLib releases may return a more specific value instead).

As errno is global and may be modified by intermediate function calls, you should save its value as soon as the call which sets it

err_no

Error number as defined in errno.h.

Returns

IOErrorEnum value for the given errno.h error number.