gio::functions

Function io_modules_load_all_in_directory_with_scope

pub fn io_modules_load_all_in_directory_with_scope(
    dirname: impl AsRef<Path>,
    scope: &mut IOModuleScope,
) -> Vec<IOModule>

Loads all the modules in the specified directory.

If don’t require all modules to be initialized (and thus registering all gtypes) then you can use g_io_modules_scan_all_in_directory() which allows delayed/lazy loading of modules.

dirname

pathname for a directory containing modules to load.

scope

a scope to use when scanning the modules.

Returns

a list of #GIOModules loaded from the directory, All the modules are loaded into memory, if you want to unload them (enabling on-demand loading) you must call g_type_module_unuse() on all the modules. Free the list with g_list_free().