Function gio::functions::unix_is_system_fs_type

pub fn unix_is_system_fs_type(fs_type: &str) -> bool

Available on Unix and crate feature v2_56 only.

Determines if fs_type is considered a type of file system which is only used in implementation of the OS. This is primarily used for hiding mounted volumes that are intended as APIs for programs to read, and system administrators at a shell; rather than something that should, for example, appear in a GUI. For example, the Linux /proc filesystem.

The list of file system types considered ‘system’ ones may change over time.

fs_type

a file system type, e.g. procfs or tmpfs

Returns

true if fs_type is considered an implementation detail of the OS.