Trait gio::prelude::AppInfoExt

pub trait AppInfoExt: 'static {

    // Required methods
    fn add_supports_type(&self, content_type: &str) -> Result<(), Error>;
    fn can_delete(&self) -> bool;
    fn can_remove_supports_type(&self) -> bool;
    fn delete(&self) -> bool;
    fn dup(&self) -> AppInfo;
    fn equal(&self, appinfo2: &impl IsA<AppInfo>) -> bool;
    fn commandline(&self) -> Option<PathBuf>;
    fn description(&self) -> Option<GString>;
    fn display_name(&self) -> GString;
    fn executable(&self) -> PathBuf;
    fn icon(&self) -> Option<Icon>;
    fn id(&self) -> Option<GString>;
    fn name(&self) -> GString;
    fn supported_types(&self) -> Vec<GString>;
    fn launch(
        &self,
        files: &[File],
        context: Option<&impl IsA<AppLaunchContext>>
    ) -> Result<(), Error>;
    fn launch_uris(
        &self,
        uris: &[&str],
        context: Option<&impl IsA<AppLaunchContext>>
    ) -> Result<(), Error>;
    fn remove_supports_type(&self, content_type: &str) -> Result<(), Error>;
    fn set_as_default_for_extension(
        &self,
        extension: impl AsRef<Path>
    ) -> Result<(), Error>;
    fn set_as_default_for_type(&self, content_type: &str) -> Result<(), Error>;
    fn set_as_last_used_for_type(&self, content_type: &str) -> Result<(), Error>;
    fn should_show(&self) -> bool;
    fn supports_files(&self) -> bool;
    fn supports_uris(&self) -> bool;

}

Trait containing all AppInfo methods.

Implementors

AppInfo, DesktopAppInfo

Required Methods

fn add_supports_type(&self, content_type: &str) -> Result<(), Error>

Adds a content type to the application information to indicate the application is capable of opening files with the given content type.

content_type

a string.

Returns

true on success, false on error.

fn can_delete(&self) -> bool

Obtains the information whether the AppInfo can be deleted. See delete().

Returns

true if self can be deleted

fn can_remove_supports_type(&self) -> bool

Checks if a supported content type can be removed from an application.

Returns

true if it is possible to remove supported content types from a given self, false if not.

fn delete(&self) -> bool

Tries to delete a AppInfo.

On some platforms, there may be a difference between user-defined GAppInfos which can be deleted, and system-wide ones which cannot. See can_delete().

Returns

true if self has been deleted

fn dup(&self) -> AppInfo

Creates a duplicate of a AppInfo.

Returns

a duplicate of self.

fn equal(&self, appinfo2: &impl IsA<AppInfo>) -> bool

fn commandline(&self) -> Option<PathBuf>

Gets the commandline with which the application will be started.

Returns

a string containing the self’s commandline, or None if this information is not available

fn description(&self) -> Option<GString>

Gets a human-readable description of an installed application.

Returns

a string containing a description of the application self, or None if none.

fn display_name(&self) -> GString

Gets the display name of the application. The display name is often more descriptive to the user than the name itself.

Returns

the display name of the application for self, or the name if no display name is available.

fn executable(&self) -> PathBuf

Gets the executable’s name for the installed application.

Returns

a string containing the self’s application binaries name

fn icon(&self) -> Option<Icon>

Gets the icon for the application.

Returns

the default Icon for self or None if there is no default icon.

fn id(&self) -> Option<GString>

Gets the ID of an application. An id is a string that identifies the application. The exact format of the id is platform dependent. For instance, on Unix this is the desktop file id from the xdg menu specification.

Note that the returned ID may be None, depending on how the self has been constructed.

Returns

a string containing the application’s ID.

fn name(&self) -> GString

Gets the installed name of the application.

Returns

the name of the application for self.

fn supported_types(&self) -> Vec<GString>

Retrieves the list of content types that app_info claims to support. If this information is not provided by the environment, this function will return None. This function does not take in consideration associations added with add_supports_type(), but only those exported directly by the application.

Returns

a list of content types.

fn launch( &self, files: &[File], context: Option<&impl IsA<AppLaunchContext>> ) -> Result<(), Error>

Launches the application. Passes files to the launched application as arguments, using the optional context to get information about the details of the launcher (like what screen it is on). On error, error will be set accordingly.

To launch the application without arguments pass a None files list.

Note that even if the launch is successful the application launched can fail to start if it runs into problems during startup. There is no way to detect this.

Some URIs can be changed when passed through a GFile (for instance unsupported URIs with strange formats like mailto:), so if you have a textual URI you want to pass in as argument, consider using launch_uris() instead.

The launched application inherits the environment of the launching process, but it can be modified with AppLaunchContextExt::setenv() and AppLaunchContextExt::unsetenv().

On UNIX, this function sets the GIO_LAUNCHED_DESKTOP_FILE environment variable with the path of the launched desktop file and GIO_LAUNCHED_DESKTOP_FILE_PID to the process id of the launched process. This can be used to ignore GIO_LAUNCHED_DESKTOP_FILE, should it be inherited by further processes. The DISPLAY and DESKTOP_STARTUP_ID environment variables are also set, based on information provided in context.

files

a GList of File objects

context

a AppLaunchContext or None

Returns

true on successful launch, false otherwise.

fn launch_uris( &self, uris: &[&str], context: Option<&impl IsA<AppLaunchContext>> ) -> Result<(), Error>

Launches the application. This passes the uris to the launched application as arguments, using the optional context to get information about the details of the launcher (like what screen it is on). On error, error will be set accordingly.

To launch the application without arguments pass a None uris list.

Note that even if the launch is successful the application launched can fail to start if it runs into problems during startup. There is no way to detect this.

uris

a GList containing URIs to launch.

context

a AppLaunchContext or None

Returns

true on successful launch, false otherwise.

fn remove_supports_type(&self, content_type: &str) -> Result<(), Error>

Removes a supported type from an application, if possible.

content_type

a string.

Returns

true on success, false on error.

fn set_as_default_for_extension( &self, extension: impl AsRef<Path> ) -> Result<(), Error>

Sets the application as the default handler for the given file extension.

extension

a string containing the file extension (without the dot).

Returns

true on success, false on error.

fn set_as_default_for_type(&self, content_type: &str) -> Result<(), Error>

fn set_as_last_used_for_type(&self, content_type: &str) -> Result<(), Error>

fn should_show(&self) -> bool

fn supports_files(&self) -> bool

fn supports_uris(&self) -> bool