Trait gio::prelude::AppLaunchContextExt

pub trait AppLaunchContextExt: 'static {
    fn display(
        &self, 
        info: &impl IsA<AppInfo>, 
        files: &[File]
    ) -> Option<GString>;
    fn environment(&self) -> Vec<OsString>;
    fn startup_notify_id(
        &self, 
        info: &impl IsA<AppInfo>, 
        files: &[File]
    ) -> Option<GString>;
    fn launch_failed(&self, startup_notify_id: &str);
    fn setenv(&self, variable: impl AsRef<OsStr>, value: impl AsRef<OsStr>);
    fn unsetenv(&self, variable: impl AsRef<OsStr>);
    fn connect_launch_failed<F: Fn(&Self, &str) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_launch_started<F: Fn(&Self, &AppInfo, Option<&Variant>) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_launched<F: Fn(&Self, &AppInfo, &Variant) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all AppLaunchContext methods.

Implementors

AppLaunchContext

Required Methods

fn display(&self, info: &impl IsA<AppInfo>, files: &[File]) -> Option<GString>

Gets the display string for the self. This is used to ensure new applications are started on the same display as the launching application, by setting the DISPLAY environment variable.

info

a AppInfo

files

a GList of File objects

Returns

a display string for the display.

fn environment(&self) -> Vec<OsString>

Gets the complete environment variable list to be passed to the child process when self is used to launch an application. This is a None-terminated array of strings, where each string has the form KEY=VALUE.

Returns

the child’s environment

fn startup_notify_id(
&self,
info: &impl IsA<AppInfo>,
files: &[File]
) -> Option<GString>

Initiates startup notification for the application and returns the DESKTOP_STARTUP_ID for the launched operation, if supported.

Startup notification IDs are defined in the FreeDesktop.Org Startup Notifications standard.

info

a AppInfo

files

a GList of of File objects

Returns

a startup notification ID for the application, or None if not supported.

fn launch_failed(&self, startup_notify_id: &str)

Called when an application has failed to launch, so that it can cancel the application startup notification started in startup_notify_id().

startup_notify_id

the startup notification id that was returned by startup_notify_id().

fn setenv(&self, variable: impl AsRef<OsStr>, value: impl AsRef<OsStr>)

Arranges for variable to be set to value in the child’s environment when self is used to launch an application.

variable

the environment variable to set

value

the value for to set the variable to.

fn unsetenv(&self, variable: impl AsRef<OsStr>)

fn connect_launch_failed<F: Fn(&Self, &str) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_launch_started<F: Fn(&Self, &AppInfo, Option<&Variant>) + 'static>(
&self,
f: F
) -> SignalHandlerId

Available on crate feature v2_72 only.

fn connect_launched<F: Fn(&Self, &AppInfo, &Variant) + 'static>(
&self,
f: F
) -> SignalHandlerId