Trait gio::prelude::ApplicationCommandLineExt

pub trait ApplicationCommandLineExt: 'static {
    fn create_file_for_arg(&self, arg: impl AsRef<OsStr>) -> File;
    fn arguments(&self) -> Vec<OsString>;
    fn cwd(&self) -> Option<PathBuf>;
    fn environ(&self) -> Vec<OsString>;
    fn exit_status(&self) -> i32;
    fn is_remote(&self) -> bool;
    fn options_dict(&self) -> VariantDict;
    fn platform_data(&self) -> Option<Variant>;
    fn stdin(&self) -> Option<InputStream>;
    fn getenv(&self, name: impl AsRef<OsStr>) -> Option<GString>;
    fn set_exit_status(&self, exit_status: i32);
    fn connect_is_remote_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all ApplicationCommandLine methods.

Implementors

ApplicationCommandLine

Required Methods

fn create_file_for_arg(&self, arg: impl AsRef<OsStr>) -> File

Creates a File corresponding to a filename that was given as part of the invocation of self.

This differs from File::for_commandline_arg() in that it resolves relative pathnames using the current working directory of the invoking process rather than the local process.

arg

an argument from self

Returns

a new File

fn arguments(&self) -> Vec<OsString>

fn cwd(&self) -> Option<PathBuf>

fn environ(&self) -> Vec<OsString>

fn exit_status(&self) -> i32

fn is_remote(&self) -> bool

fn options_dict(&self) -> VariantDict

fn platform_data(&self) -> Option<Variant>

fn stdin(&self) -> Option<InputStream>

fn getenv(&self, name: impl AsRef<OsStr>) -> Option<GString>

fn set_exit_status(&self, exit_status: i32)

fn connect_is_remote_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId