Trait gio::prelude::CancellableExt

pub trait CancellableExt: IsA<Cancellable> + Sealed + 'static {
    // Provided methods
    fn cancel(&self) { ... }
    fn fd(&self) -> i32 { ... }
    fn is_cancelled(&self) -> bool { ... }
    fn pop_current(&self) { ... }
    fn push_current(&self) { ... }
    fn release_fd(&self) { ... }
}

Trait containing all Cancellable methods.

Implementors

Cancellable

Provided Methods

fn cancel(&self)

Will set self to cancelled, and will emit the cancelled signal. (However, see the warning about race conditions in the documentation for that signal if you are planning to connect to it.)

This function is thread-safe. In other words, you can safely call it from a thread other than the one running the operation that was passed the self.

If self is None, this function returns immediately for convenience.

The convention within GIO is that cancelling an asynchronous operation causes it to complete asynchronously. That is, if you cancel the operation from the same thread in which it is running, then the operation’s GAsyncReadyCallback will not be invoked until the application returns to the main loop.

fn fd(&self) -> i32

Gets the file descriptor for a cancellable job. This can be used to implement cancellable operations on Unix systems. The returned fd will turn readable when self is cancelled.

You are not supposed to read from the fd yourself, just check for readable status. Reading to unset the readable status is done with g_cancellable_reset().

After a successful return from this function, you should use release_fd() to free up resources allocated for the returned file descriptor.

See also g_cancellable_make_pollfd().

Returns

A valid file descriptor. -1 if the file descriptor is not supported, or on errors.

fn is_cancelled(&self) -> bool

Checks if a cancellable job has been cancelled.

Returns

true if self is cancelled, FALSE if called with None or if item is not cancelled.

fn pop_current(&self)

Pops self off the cancellable stack (verifying that self is on the top of the stack).

fn push_current(&self)

Pushes self onto the cancellable stack. The current cancellable can then be received using Cancellable::current().

This is useful when implementing cancellable operations in code that does not allow you to pass down the cancellable object.

This is typically called automatically by e.g. File operations, so you rarely have to call this yourself.

fn release_fd(&self)

Releases a resources previously allocated by fd() or g_cancellable_make_pollfd().

For compatibility reasons with older releases, calling this function is not strictly required, the resources will be automatically freed when the self is finalized. However, the self will block scarce file descriptors until it is finalized if this function is not called. This can cause the application to run out of file descriptors when many GCancellables are used at the same time.

Object Safety

This trait is not object safe.