Trait gio::prelude::DataOutputStreamExt

pub trait DataOutputStreamExt: 'static {
    // Required methods
    fn byte_order(&self) -> DataStreamByteOrder;
    fn put_byte(
        &self,
        data: u8,
        cancellable: Option<&impl IsA<Cancellable>>
    ) -> Result<(), Error>;
    fn put_int16(
        &self,
        data: i16,
        cancellable: Option<&impl IsA<Cancellable>>
    ) -> Result<(), Error>;
    fn put_int32(
        &self,
        data: i32,
        cancellable: Option<&impl IsA<Cancellable>>
    ) -> Result<(), Error>;
    fn put_int64(
        &self,
        data: i64,
        cancellable: Option<&impl IsA<Cancellable>>
    ) -> Result<(), Error>;
    fn put_string(
        &self,
        str: &str,
        cancellable: Option<&impl IsA<Cancellable>>
    ) -> Result<(), Error>;
    fn put_uint16(
        &self,
        data: u16,
        cancellable: Option<&impl IsA<Cancellable>>
    ) -> Result<(), Error>;
    fn put_uint32(
        &self,
        data: u32,
        cancellable: Option<&impl IsA<Cancellable>>
    ) -> Result<(), Error>;
    fn put_uint64(
        &self,
        data: u64,
        cancellable: Option<&impl IsA<Cancellable>>
    ) -> Result<(), Error>;
    fn set_byte_order(&self, order: DataStreamByteOrder);
    fn connect_byte_order_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F
    ) -> SignalHandlerId;
}

Trait containing all DataOutputStream methods.

Implementors

DataOutputStream

Required Methods

fn byte_order(&self) -> DataStreamByteOrder

Gets the byte order for the stream.

Returns

the DataStreamByteOrder for the self.

fn put_byte( &self, data: u8, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<(), Error>

Puts a byte into the output stream.

data

a guchar.

cancellable

optional Cancellable object, None to ignore.

Returns

true if data was successfully added to the self.

fn put_int16( &self, data: i16, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<(), Error>

Puts a signed 16-bit integer into the output stream.

data

a gint16.

cancellable

optional Cancellable object, None to ignore.

Returns

true if data was successfully added to the self.

fn put_int32( &self, data: i32, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<(), Error>

Puts a signed 32-bit integer into the output stream.

data

a gint32.

cancellable

optional Cancellable object, None to ignore.

Returns

true if data was successfully added to the self.

fn put_int64( &self, data: i64, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<(), Error>

Puts a signed 64-bit integer into the stream.

data

a gint64.

cancellable

optional Cancellable object, None to ignore.

Returns

true if data was successfully added to the self.

fn put_string( &self, str: &str, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<(), Error>

Puts a string into the output stream.

str

a string.

cancellable

optional Cancellable object, None to ignore.

Returns

true if string was successfully added to the self.

fn put_uint16( &self, data: u16, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<(), Error>

Puts an unsigned 16-bit integer into the output stream.

data

a guint16.

cancellable

optional Cancellable object, None to ignore.

Returns

true if data was successfully added to the self.

fn put_uint32( &self, data: u32, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<(), Error>

Puts an unsigned 32-bit integer into the stream.

data

a guint32.

cancellable

optional Cancellable object, None to ignore.

Returns

true if data was successfully added to the self.

fn put_uint64( &self, data: u64, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<(), Error>

Puts an unsigned 64-bit integer into the stream.

data

a guint64.

cancellable

optional Cancellable object, None to ignore.

Returns

true if data was successfully added to the self.

fn set_byte_order(&self, order: DataStreamByteOrder)

Sets the byte order of the data output stream to order.

order

a GDataStreamByteOrder.

fn connect_byte_order_notify<F: Fn(&Self) + 'static>( &self, f: F ) -> SignalHandlerId