Trait gio::prelude::InetAddressExt

pub trait InetAddressExt: 'static {

    fn equal(&self, other_address: &impl IsA<InetAddress>) -> bool;
    fn family(&self) -> SocketFamily;
    fn is_any(&self) -> bool;
    fn is_link_local(&self) -> bool;
    fn is_loopback(&self) -> bool;
    fn is_mc_global(&self) -> bool;
    fn is_mc_link_local(&self) -> bool;
    fn is_mc_node_local(&self) -> bool;
    fn is_mc_org_local(&self) -> bool;
    fn is_mc_site_local(&self) -> bool;
    fn is_multicast(&self) -> bool;
    fn is_site_local(&self) -> bool;
    fn native_size(&self) -> usize;
    fn to_str(&self) -> GString;
    fn connect_is_any_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_link_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_loopback_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_mc_global_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_mc_link_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_mc_node_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_mc_org_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_mc_site_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_multicast_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_is_site_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

}

Trait containing all InetAddress methods.

Implementors

InetAddress

Required Methods

fn equal(&self, other_address: &impl IsA<InetAddress>) -> bool

fn family(&self) -> SocketFamily

Gets self’s family

Returns

self’s family

fn is_any(&self) -> bool

Tests whether self is the “any” address for its family.

Returns

true if self is the “any” address for its family.

fn is_link_local(&self) -> bool

Tests whether self is a link-local address (that is, if it identifies a host on a local network that is not connected to the Internet).

Returns

true if self is a link-local address.

fn is_loopback(&self) -> bool

Tests whether self is the loopback address for its family.

Returns

true if self is the loopback address for its family.

fn is_mc_global(&self) -> bool

Tests whether self is a global multicast address.

Returns

true if self is a global multicast address.

fn is_mc_link_local(&self) -> bool

Tests whether self is a link-local multicast address.

Returns

true if self is a link-local multicast address.

fn is_mc_node_local(&self) -> bool

Tests whether self is a node-local multicast address.

Returns

true if self is a node-local multicast address.

fn is_mc_org_local(&self) -> bool

Tests whether self is an organization-local multicast address.

Returns

true if self is an organization-local multicast address.

fn is_mc_site_local(&self) -> bool

Tests whether self is a site-local multicast address.

Returns

true if self is a site-local multicast address.

fn is_multicast(&self) -> bool

Tests whether self is a multicast address.

Returns

true if self is a multicast address.

fn is_site_local(&self) -> bool

Tests whether self is a site-local address such as 10.0.0.1 (that is, the address identifies a host on a local network that can not be reached directly from the Internet, but which may have outgoing Internet connectivity via a NAT or firewall).

Returns

true if self is a site-local address.

fn native_size(&self) -> usize

Gets the size of the native raw binary address for self. This is the size of the data that you get from InetAddressExtManual::to_bytes().

Returns

the number of bytes used for the native version of self.

fn to_str(&self) -> GString

Converts self to string form.

Returns

a representation of self as a string, which should be freed after use.

fn connect_is_any_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_link_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_loopback_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_mc_global_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_mc_link_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_mc_node_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_mc_org_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_mc_site_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_multicast_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_is_site_local_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId