Trait gio::prelude::InetAddressMaskExt

pub trait InetAddressMaskExt: 'static {
    fn equal(&self, mask2: &impl IsA<InetAddressMask>) -> bool;
    fn address(&self) -> InetAddress;
    fn family(&self) -> SocketFamily;
    fn length(&self) -> u32;
    fn matches(&self, address: &impl IsA<InetAddress>) -> bool;
    fn to_str(&self) -> GString;
    fn set_address<P: IsA<InetAddress>>(&self, address: Option<&P>);
    fn set_length(&self, length: u32);
    fn connect_address_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_family_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_length_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all InetAddressMask methods.

Implementors

InetAddressMask

Required Methods

fn equal(&self, mask2: &impl IsA<InetAddressMask>) -> bool

fn address(&self) -> InetAddress

Gets self’s base address

Returns

self’s base address

fn family(&self) -> SocketFamily

Gets the SocketFamily of self’s address

Returns

the SocketFamily of self’s address

fn length(&self) -> u32

Gets self’s length

Returns

self’s length

fn matches(&self, address: &impl IsA<InetAddress>) -> bool

Tests if address falls within the range described by self.

address

a InetAddress

Returns

whether address falls within the range described by self.

fn to_str(&self) -> GString

Converts self back to its corresponding string form.

Returns

a string corresponding to self.

fn set_address<P: IsA<InetAddress>>(&self, address: Option<&P>)

fn set_length(&self, length: u32)

fn connect_address_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_family_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_length_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId