gio::prelude

Trait InetSocketAddressExt

pub trait InetSocketAddressExt: IsA<InetSocketAddress> + 'static {
    // Provided methods
    fn address(&self) -> InetAddress { ... }
    fn flowinfo(&self) -> u32 { ... }
    fn port(&self) -> u16 { ... }
    fn scope_id(&self) -> u32 { ... }
}

Trait containing all InetSocketAddress methods.

Implementors

InetSocketAddress, ProxyAddress

Provided Methods

fn address(&self) -> InetAddress

Gets @self’s #GInetAddress.

Returns

the #GInetAddress for @self, which must be g_object_ref()’d if it will be stored

fn flowinfo(&self) -> u32

Gets the sin6_flowinfo field from @self, which must be an IPv6 address.

If not overridden this value will be inherited from address.

Returns

the flowinfo field

fn port(&self) -> u16

Gets @self’s port.

Returns

the port for @self

fn scope_id(&self) -> u32

Gets the sin6_scope_id field from @self, which must be an IPv6 address.

If not overridden this value will be inherited from address.

Returns

the scope id field

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.