Trait gio::prelude::InetSocketAddressExt

pub trait InetSocketAddressExt: 'static {
    fn address(&self) -> InetAddress;
    fn flowinfo(&self) -> u32;
    fn port(&self) -> u16;
    fn scope_id(&self) -> u32;
}

Trait containing all InetSocketAddress methods.

Implementors

InetSocketAddress, ProxyAddress

Required Methods

fn address(&self) -> InetAddress

Gets self’s InetAddress.

Returns

the InetAddress for self, which must be g_object_ref()’d if it will be stored

fn flowinfo(&self) -> u32

Gets the sin6_flowinfo field from self, which must be an IPv6 address.

Returns

the flowinfo field

fn port(&self) -> u16

Gets self’s port.

Returns

the port for self

fn scope_id(&self) -> u32

Gets the sin6_scope_id field from self, which must be an IPv6 address.

Returns

the scope id field