gio::prelude
Trait LoadableIconExt

pub trait LoadableIconExt: IsA<LoadableIcon> + 'static {
    // Provided methods
    fn load(
        &self,
        size: i32,
        cancellable: Option<&impl IsA<Cancellable>>,
    ) -> Result<(InputStream, GString), Error> { ... }
    fn load_async<P: FnOnce(Result<(InputStream, GString), Error>) + 'static>(
        &self,
        size: i32,
        cancellable: Option<&impl IsA<Cancellable>>,
        callback: P,
    ) { ... }
    fn load_future(
        &self,
        size: i32,
    ) -> Pin<Box_<dyn Future<Output = Result<(InputStream, GString), Error>> + 'static>> { ... }
}

Trait containing all LoadableIcon methods.

Implementors

BytesIcon, FileIcon, LoadableIcon

Provided Methods

fn load( &self, size: i32, cancellable: Option<&impl IsA<Cancellable>>, ) -> Result<(InputStream, GString), Error>

Loads a loadable icon. For the asynchronous version of this function, see g_loadable_icon_load_async().

size

an integer.

cancellable

optional #GCancellable object, None to ignore.

Returns

a #GInputStream to read the icon from.

type_

a location to store the type of the loaded icon, None to ignore.

fn load_async<P: FnOnce(Result<(InputStream, GString), Error>) + 'static>( &self, size: i32, cancellable: Option<&impl IsA<Cancellable>>, callback: P, )

Loads an icon asynchronously. To finish this function, see g_loadable_icon_load_finish(). For the synchronous, blocking version of this function, see g_loadable_icon_load().

size

an integer.

cancellable

optional #GCancellable object, None to ignore.

callback

a #GAsyncReadyCallback to call when the request is satisfied

fn load_future( &self, size: i32, ) -> Pin<Box_<dyn Future<Output = Result<(InputStream, GString), Error>> + 'static>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.