gio::prelude

Trait MemoryMonitorExt

pub trait MemoryMonitorExt: IsA<MemoryMonitor> + 'static {
    // Provided method
    fn connect_low_memory_warning<F: Fn(&Self, MemoryMonitorWarningLevel) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Available on crate feature v2_64 only.

Trait containing all MemoryMonitor methods.

Implementors

MemoryMonitor

Provided Methods

fn connect_low_memory_warning<F: Fn(&Self, MemoryMonitorWarningLevel) + 'static>( &self, f: F, ) -> SignalHandlerId

Emitted when the system is running low on free memory. The signal handler should then take the appropriate action depending on the warning level. See the #GMemoryMonitorWarningLevel documentation for details.

level

the #GMemoryMonitorWarningLevel warning level

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.