gio::prelude
Trait NetworkServiceExt

pub trait NetworkServiceExt: IsA<NetworkService> + 'static {
    // Provided methods
    fn domain(&self) -> GString { ... }
    fn protocol(&self) -> GString { ... }
    fn scheme(&self) -> GString { ... }
    fn service(&self) -> GString { ... }
    fn set_scheme(&self, scheme: &str) { ... }
    fn connect_scheme_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all NetworkService methods.

Implementors

NetworkService

Provided Methods

fn domain(&self) -> GString

Gets the domain that @self serves. This might be either UTF-8 or ASCII-encoded, depending on what @self was created with.

Returns

@self’s domain name

fn protocol(&self) -> GString

Gets @self’s protocol name (eg, “tcp”).

Returns

@self’s protocol name

fn scheme(&self) -> GString

Gets the URI scheme used to resolve proxies. By default, the service name is used as scheme.

Returns

@self’s scheme name

fn service(&self) -> GString

Gets @self’s service name (eg, “ldap”).

Returns

@self’s service name

fn set_scheme(&self, scheme: &str)

Set’s the URI scheme used to resolve proxies. By default, the service name is used as scheme.

scheme

a URI scheme

fn connect_scheme_notify<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.