Trait gio::prelude::NotificationExt

pub trait NotificationExt {
    fn add_button(&self, label: &str, detailed_action: &str);

    fn add_button_with_target_value<'a, P: Into<Option<&'a Variant>>>(
        &self, 
        label: &str, 
        action: &str, 
        target: P
    );

    fn set_body<'a, P: Into<Option<&'a str>>>(&self, body: P);

    fn set_default_action(&self, detailed_action: &str);

    fn set_default_action_and_target_value<'a, P: Into<Option<&'a Variant>>>(
        &self, 
        action: &str, 
        target: P
    );

    fn set_icon<P: IsA<Icon>>(&self, icon: &P);

    fn set_priority(&self, priority: NotificationPriority);

    fn set_title(&self, title: &str);

    fn set_urgent(&self, urgent: bool);
}

Trait containing all Notification methods.

Feature: v2_40

Implementors

Notification

Required Methods

fn add_button(&self, label: &str, detailed_action: &str)

Adds a button to self that activates the action in detailed_action when clicked. That action must be an application-wide action (starting with "app."). If detailed_action contains a target, the action will be activated with that target as its parameter.

See Action::parse_detailed_name for a description of the format for detailed_action.

Feature: v2_40

label

label of the button

detailed_action

a detailed action name

fn add_button_with_target_value<'a, P: Into<Option<&'a Variant>>>(
&self,
label: &str,
action: &str,
target: P
)

Adds a button to self that activates action when clicked. action must be an application-wide action (it must start with "app.").

If target is non-None, action will be activated with target as its parameter.

Feature: v2_40

label

label of the button

action

an action name

target

a glib::Variant to use as action's parameter, or None

fn set_body<'a, P: Into<Option<&'a str>>>(&self, body: P)

Sets the body of self to body.

Feature: v2_40

body

the new body for self, or None

fn set_default_action(&self, detailed_action: &str)

Sets the default action of self to detailed_action. This action is activated when the notification is clicked on.

The action in detailed_action must be an application-wide action (it must start with "app."). If detailed_action contains a target, the given action will be activated with that target as its parameter. See Action::parse_detailed_name for a description of the format for detailed_action.

When no default action is set, the application that the notification was sent on is activated.

Feature: v2_40

detailed_action

a detailed action name

fn set_default_action_and_target_value<'a, P: Into<Option<&'a Variant>>>(
&self,
action: &str,
target: P
)

Sets the default action of self to action. This action is activated when the notification is clicked on. It must be an application-wide action (start with "app.").

If target is non-None, action will be activated with target as its parameter.

When no default action is set, the application that the notification was sent on is activated.

Feature: v2_40

action

an action name

target

a glib::Variant to use as action's parameter, or None

fn set_icon<P: IsA<Icon>>(&self, icon: &P)

Sets the icon of self to icon.

Feature: v2_40

icon

the icon to be shown in self, as a Icon

fn set_priority(&self, priority: NotificationPriority)

Sets the priority of self to priority. See NotificationPriority for possible values.

Feature: v2_42

priority

a NotificationPriority

fn set_title(&self, title: &str)

Sets the title of self to title.

Feature: v2_40

title

the new title for self

fn set_urgent(&self, urgent: bool)

Deprecated in favor of NotificationExt::set_priority.

Feature: v2_40

urgent

true if self is urgent