Trait gio::prelude::ProxyAddressExt

pub trait ProxyAddressExt: 'static {
    fn get_destination_hostname(&self) -> GString;

    fn get_destination_port(&self) -> u16;

    fn get_destination_protocol(&self) -> Option<GString>;

    fn get_password(&self) -> Option<GString>;

    fn get_protocol(&self) -> GString;

    fn get_uri(&self) -> Option<GString>;

    fn get_username(&self) -> Option<GString>;
}

Trait containing all ProxyAddress methods.

Implementors

ProxyAddress

Required methods

fn get_destination_hostname(&self) -> GString

Gets self's destination hostname; that is, the name of the host that will be connected to via the proxy, not the name of the proxy itself.

Returns

the self's destination hostname

fn get_destination_port(&self) -> u16

Gets self's destination port; that is, the port on the destination host that will be connected to via the proxy, not the port number of the proxy itself.

Returns

the self's destination port

fn get_destination_protocol(&self) -> Option<GString>

Gets the protocol that is being spoken to the destination server; eg, "http" or "ftp".

Returns

the self's destination protocol

fn get_password(&self) -> Option<GString>

Gets self's password.

Returns

the self's password

fn get_protocol(&self) -> GString

Gets self's protocol. eg, "socks" or "http"

Returns

the self's protocol

fn get_uri(&self) -> Option<GString>

Gets the proxy URI that self was constructed from.

Returns

the self's URI, or None if unknown

fn get_username(&self) -> Option<GString>

Gets self's username.

Returns

the self's username

Loading content...