Trait gio::prelude::ProxyAddressExt

pub trait ProxyAddressExt:
    IsA<ProxyAddress>
    + Sealed
    + 'static {
    // Provided methods
    fn destination_hostname(&self) -> GString { ... }
    fn destination_port(&self) -> u16 { ... }
    fn destination_protocol(&self) -> GString { ... }
    fn password(&self) -> Option<GString> { ... }
    fn protocol(&self) -> GString { ... }
    fn uri(&self) -> Option<GString> { ... }
    fn username(&self) -> Option<GString> { ... }
}

Trait containing all ProxyAddress methods.

Implementors

ProxyAddress

Provided Methods

fn destination_hostname(&self) -> GString

Gets @self’s destination hostname; that is, the name of the host that will be connected to via the proxy, not the name of the proxy itself.

Returns

the @self’s destination hostname

fn destination_port(&self) -> u16

Gets @self’s destination port; that is, the port on the destination host that will be connected to via the proxy, not the port number of the proxy itself.

Returns

the @self’s destination port

fn destination_protocol(&self) -> GString

Gets the protocol that is being spoken to the destination server; eg, “http” or “ftp”.

Returns

the @self’s destination protocol

fn password(&self) -> Option<GString>

Gets @self’s password.

Returns

the @self’s password

fn protocol(&self) -> GString

Gets @self’s protocol. eg, “socks” or “http”

Returns

the @self’s protocol

fn uri(&self) -> Option<GString>

Gets the proxy URI that @self was constructed from.

Returns

the @self’s URI, or None if unknown

fn username(&self) -> Option<GString>

Gets @self’s username.

Returns

the @self’s username

Object Safety

This trait is not object safe.