gio::prelude

Trait ProxyResolverExt

pub trait ProxyResolverExt: IsA<ProxyResolver> + 'static {
    // Provided methods
    fn is_supported(&self) -> bool { ... }
    fn lookup(
        &self,
        uri: &str,
        cancellable: Option<&impl IsA<Cancellable>>,
    ) -> Result<Vec<GString>, Error> { ... }
    fn lookup_async<P: FnOnce(Result<Vec<GString>, Error>) + 'static>(
        &self,
        uri: &str,
        cancellable: Option<&impl IsA<Cancellable>>,
        callback: P,
    ) { ... }
    fn lookup_future(
        &self,
        uri: &str,
    ) -> Pin<Box_<dyn Future<Output = Result<Vec<GString>, Error>> + 'static>> { ... }
}

Trait containing all ProxyResolver methods.

Implementors

ProxyResolver, SimpleProxyResolver

Provided Methods

fn is_supported(&self) -> bool

Checks if @self can be used on this system. (This is used internally; g_proxy_resolver_get_default() will only return a proxy resolver that returns true for this method.)

Returns

true if @self is supported.

fn lookup( &self, uri: &str, cancellable: Option<&impl IsA<Cancellable>>, ) -> Result<Vec<GString>, Error>

Looks into the system proxy configuration to determine what proxy, if any, to use to connect to @uri. The returned proxy URIs are of the form <protocol>://[user[:password]@]host[:port] or direct://, where <protocol> could be http, rtsp, socks or other proxying protocol.

If you don’t know what network protocol is being used on the socket, you should use none as the URI protocol. In this case, the resolver might still return a generic proxy type (such as SOCKS), but would not return protocol-specific proxy types (such as http).

direct:// is used when no proxy is needed. Direct connection should not be attempted unless it is part of the returned array of proxies.

uri

a URI representing the destination to connect to

cancellable

a #GCancellable, or None

Returns

A NULL-terminated array of proxy URIs. Must be freed with g_strfreev().

fn lookup_async<P: FnOnce(Result<Vec<GString>, Error>) + 'static>( &self, uri: &str, cancellable: Option<&impl IsA<Cancellable>>, callback: P, )

Asynchronous lookup of proxy. See g_proxy_resolver_lookup() for more details.

uri

a URI representing the destination to connect to

cancellable

a #GCancellable, or None

callback

callback to call after resolution completes

fn lookup_future( &self, uri: &str, ) -> Pin<Box_<dyn Future<Output = Result<Vec<GString>, Error>> + 'static>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.