gio::prelude

Trait RemoteActionGroupExt

pub trait RemoteActionGroupExt: IsA<RemoteActionGroup> + 'static {
    // Provided methods
    fn activate_action_full(
        &self,
        action_name: &str,
        parameter: Option<&Variant>,
        platform_data: &Variant,
    ) { ... }
    fn change_action_state_full(
        &self,
        action_name: &str,
        value: &Variant,
        platform_data: &Variant,
    ) { ... }
}

Trait containing all RemoteActionGroup methods.

Implementors

DBusActionGroup, RemoteActionGroup

Provided Methods

fn activate_action_full( &self, action_name: &str, parameter: Option<&Variant>, platform_data: &Variant, )

Activates the remote action.

This is the same as g_action_group_activate_action() except that it allows for provision of “platform data” to be sent along with the activation request. This typically contains details such as the user interaction timestamp or startup notification information.

@platform_data must be non-None and must have the type G_VARIANT_TYPE_VARDICT. If it is floating, it will be consumed.

action_name

the name of the action to activate

parameter

the optional parameter to the activation

platform_data

the platform data to send

fn change_action_state_full( &self, action_name: &str, value: &Variant, platform_data: &Variant, )

Changes the state of a remote action.

This is the same as g_action_group_change_action_state() except that it allows for provision of “platform data” to be sent along with the state change request. This typically contains details such as the user interaction timestamp or startup notification information.

@platform_data must be non-None and must have the type G_VARIANT_TYPE_VARDICT. If it is floating, it will be consumed.

action_name

the name of the action to change the state of

value

the new requested value for the state

platform_data

the platform data to send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.