Trait gio::prelude::SettingsBackendExt

pub trait SettingsBackendExt: 'static {
    fn path_writable_changed(&self, path: &str);
    fn writable_changed(&self, key: &str);
}

Trait containing all SettingsBackend methods.

Implementors

SettingsBackend

Required Methods

fn path_writable_changed(&self, path: &str)

Signals that the writability of all keys below a given path may have changed.

Since GSettings performs no locking operations for itself, this call will always be made in response to external events.

path

the name of the path

fn writable_changed(&self, key: &str)

Signals that the writability of a single key has possibly changed.

Since GSettings performs no locking operations for itself, this call will always be made in response to external events.

key

the name of the key