Trait gio::prelude::SimpleProxyResolverExt

pub trait SimpleProxyResolverExt: 'static {
    fn set_default_proxy(&self, default_proxy: &str);
    fn set_uri_proxy(&self, uri_scheme: &str, proxy: &str);
    fn default_proxy(&self) -> Option<GString>;
    fn ignore_hosts(&self) -> Vec<GString>;
    fn connect_default_proxy_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
    fn connect_ignore_hosts_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all SimpleProxyResolver methods.

Implementors

SimpleProxyResolver

Required Methods

fn set_default_proxy(&self, default_proxy: &str)

Sets the default proxy on self, to be used for any URIs that don’t match property::SimpleProxyResolver::ignore-hosts or a proxy set via set_uri_proxy().

If default_proxy starts with “socks://”, SimpleProxyResolver will treat it as referring to all three of the socks5, socks4a, and socks4 proxy types.

default_proxy

the default proxy to use

fn set_uri_proxy(&self, uri_scheme: &str, proxy: &str)

Adds a URI-scheme-specific proxy to self; URIs whose scheme matches uri_scheme (and which don’t match property::SimpleProxyResolver::ignore-hosts) will be proxied via proxy.

As with property::SimpleProxyResolver::default-proxy, if proxy starts with “socks://”, SimpleProxyResolver will treat it as referring to all three of the socks5, socks4a, and socks4 proxy types.

uri_scheme

the URI scheme to add a proxy for

proxy

the proxy to use for uri_scheme

fn default_proxy(&self) -> Option<GString>

The default proxy URI that will be used for any URI that doesn’t match property::SimpleProxyResolver::ignore-hosts, and doesn’t match any of the schemes set with set_uri_proxy().

Note that as a special case, if this URI starts with “socks://”, SimpleProxyResolver will treat it as referring to all three of the socks5, socks4a, and socks4 proxy types.

fn ignore_hosts(&self) -> Vec<GString>

A list of hostnames and IP addresses that the resolver should allow direct connections to.

Entries can be in one of 4 formats:

A hostname, such as “example.com”, “.example.com”, or “*.example.com”, any of which match “example.com” or any subdomain of it.

An IPv4 or IPv6 address, such as “192.168.1.1”, which matches only that address.

A hostname or IP address followed by a port, such as “example.com:80”, which matches whatever the hostname or IP address would match, but only for URLs with the (explicitly) indicated port. In the case of an IPv6 address, the address part must appear in brackets: “[::1]:443”

An IP address range, given by a base address and prefix length, such as “fe80::/10”, which matches any address in that range.

Note that when dealing with Unicode hostnames, the matching is done against the ASCII form of the name.

Also note that hostname exclusions apply only to connections made to hosts identified by name, and IP address exclusions apply only to connections made to hosts identified by address. That is, if example.com has an address of 192.168.1.1, and the :ignore-hosts list contains only “192.168.1.1”, then a connection to “example.com” (eg, via a NetworkAddress) will use the proxy, and a connection to “192.168.1.1” (eg, via a InetSocketAddress) will not.

These rules match the “ignore-hosts”/“noproxy” rules most commonly used by other applications.

fn connect_default_proxy_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_ignore_hosts_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId