Trait gio::prelude::SocketConnectableExt

pub trait SocketConnectableExt: 'static {
    fn enumerate(&self) -> SocketAddressEnumerator;

    fn proxy_enumerate(&self) -> SocketAddressEnumerator;

    fn to_string(&self) -> GString;
}

Trait containing all SocketConnectable methods.

Implementors

InetSocketAddress, NetworkAddress, NetworkService, ProxyAddress, SocketAddress, SocketConnectable, UnixSocketAddress

Required methods

fn enumerate(&self) -> SocketAddressEnumerator

Creates a SocketAddressEnumerator for self.

Returns

a new SocketAddressEnumerator.

fn proxy_enumerate(&self) -> SocketAddressEnumerator

Creates a SocketAddressEnumerator for self that will return a ProxyAddress for each of its addresses that you must connect to via a proxy.

If self does not implement proxy_enumerate(), this will fall back to calling enumerate().

Returns

a new SocketAddressEnumerator.

fn to_string(&self) -> GString

Format a SocketConnectable as a string. This is a human-readable format for use in debugging output, and is not a stable serialization format. It is not suitable for use in user interfaces as it exposes too much information for a user.

If the SocketConnectable implementation does not support string formatting, the implementation’s type name will be returned as a fallback.

Returns

the formatted string