Trait gio::prelude::SocketConnectableExt

pub trait SocketConnectableExt: 'static {
    fn enumerate(&self) -> Option<SocketAddressEnumerator>;

    fn proxy_enumerate(&self) -> Option<SocketAddressEnumerator>;

    fn to_string(&self) -> Option<GString>;
}

Trait containing all SocketConnectable methods.

Implementors

InetSocketAddress, NetworkAddress, NetworkService, ProxyAddress, SocketAddress, SocketConnectable, UnixSocketAddress

Required methods

fn enumerate(&self) -> Option<SocketAddressEnumerator>

Creates a SocketAddressEnumerator for self.

Returns

a new SocketAddressEnumerator.

fn proxy_enumerate(&self) -> Option<SocketAddressEnumerator>

Creates a SocketAddressEnumerator for self that will return a ProxyAddress for each of its addresses that you must connect to via a proxy.

If self does not implement SocketConnectable::proxy_enumerate, this will fall back to calling SocketConnectable::enumerate.

Returns

a new SocketAddressEnumerator.

fn to_string(&self) -> Option<GString>

This is supported on crate feature v2_48 only.

Format a SocketConnectable as a string. This is a human-readable format for use in debugging output, and is not a stable serialization format. It is not suitable for use in user interfaces as it exposes too much information for a user.

If the SocketConnectable implementation does not support string formatting, the implementation’s type name will be returned as a fallback.

Feature: v2_48

Returns

the formatted string

Loading content...