Trait gio::prelude::SocketListenerExtManual

pub trait SocketListenerExtManual: SocketListenerExt {
    // Required method
    fn incoming(
        &self,
    ) -> Pin<Box<dyn Stream<Item = Result<(SocketConnection, Option<Object>), Error>>>>;
}

Required Methods

fn incoming( &self, ) -> Pin<Box<dyn Stream<Item = Result<(SocketConnection, Option<Object>), Error>>>>

Returns a stream of incoming connections

Iterating over this stream is equivalent to calling SocketListenerExt::accept_future in a loop. The stream of connections is infinite, i.e awaiting the next connection will never result in None.

Object Safety

This trait is not object safe.