Trait gio::prelude::ThreadedSocketServiceExt

pub trait ThreadedSocketServiceExt: 'static {
    fn max_threads(&self) -> i32;

    fn connect_run<F: Fn(&Self, &SocketConnection, Option<&Object>) -> bool + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all ThreadedSocketService methods.

Implementors

ThreadedSocketService

Required methods

fn max_threads(&self) -> i32

fn connect_run<F: Fn(&Self, &SocketConnection, Option<&Object>) -> bool + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::run signal is emitted in a worker thread in response to an incoming connection. This thread is dedicated to handling connection and may perform blocking IO. The signal handler need not return until the connection is closed.

connection

a new SocketConnection object.

source_object

the source_object passed to SocketListenerExt::add_address().

Returns

true to stop further signal handlers from being called