Trait gio::prelude::TlsBackendExt

pub trait TlsBackendExt:
    IsA<TlsBackend>
    + Sealed
    + 'static {
    // Provided methods
    fn certificate_type(&self) -> Type { ... }
    fn client_connection_type(&self) -> Type { ... }
    fn default_database(&self) -> TlsDatabase { ... }
    fn dtls_client_connection_type(&self) -> Type { ... }
    fn dtls_server_connection_type(&self) -> Type { ... }
    fn file_database_type(&self) -> Type { ... }
    fn server_connection_type(&self) -> Type { ... }
    fn set_default_database(&self, database: Option<&impl IsA<TlsDatabase>>) { ... }
    fn supports_dtls(&self) -> bool { ... }
    fn supports_tls(&self) -> bool { ... }
}

Trait containing all TlsBackend methods.

Implementors

TlsBackend

Provided Methods

fn certificate_type(&self) -> Type

Gets the #GType of @self’s #GTlsCertificate implementation.

Returns

the #GType of @self’s #GTlsCertificate implementation.

fn client_connection_type(&self) -> Type

Gets the #GType of @self’s #GTlsClientConnection implementation.

Returns

the #GType of @self’s #GTlsClientConnection implementation.

fn default_database(&self) -> TlsDatabase

Gets the default #GTlsDatabase used to verify TLS connections.

Returns

the default database, which should be unreffed when done.

fn dtls_client_connection_type(&self) -> Type

Gets the #GType of @self’s #GDtlsClientConnection implementation.

Returns

the #GType of @self’s #GDtlsClientConnection implementation, or G_TYPE_INVALID if this backend doesn’t support DTLS.

fn dtls_server_connection_type(&self) -> Type

Gets the #GType of @self’s #GDtlsServerConnection implementation.

Returns

the #GType of @self’s #GDtlsServerConnection implementation, or G_TYPE_INVALID if this backend doesn’t support DTLS.

fn file_database_type(&self) -> Type

Gets the #GType of @self’s #GTlsFileDatabase implementation.

Returns

the #GType of backend’s #GTlsFileDatabase implementation.

fn server_connection_type(&self) -> Type

Gets the #GType of @self’s #GTlsServerConnection implementation.

Returns

the #GType of @self’s #GTlsServerConnection implementation.

fn set_default_database(&self, database: Option<&impl IsA<TlsDatabase>>)

Available on crate feature v2_60 only.

Set the default #GTlsDatabase used to verify TLS connections

Any subsequent call to g_tls_backend_get_default_database() will return the database set in this call. Existing databases and connections are not modified.

Setting a None default database will reset to using the system default database as if g_tls_backend_set_default_database() had never been called.

database

the #GTlsDatabase

fn supports_dtls(&self) -> bool

Checks if DTLS is supported. DTLS support may not be available even if TLS support is available, and vice-versa.

Returns

whether DTLS is supported

fn supports_tls(&self) -> bool

Checks if TLS is supported; if this returns false for the default #GTlsBackend, it means no “real” TLS backend is available.

Returns

whether or not TLS is supported

Object Safety

This trait is not object safe.