gio::prelude

Trait TlsClientConnectionExt

pub trait TlsClientConnectionExt: IsA<TlsClientConnection> + 'static {
    // Provided methods
    fn copy_session_state(&self, source: &impl IsA<TlsClientConnection>) { ... }
    fn accepted_cas(&self) -> Vec<ByteArray> { ... }
    fn server_identity(&self) -> Option<SocketConnectable> { ... }
    fn validation_flags(&self) -> TlsCertificateFlags { ... }
    fn set_server_identity(&self, identity: &impl IsA<SocketConnectable>) { ... }
    fn set_validation_flags(&self, flags: TlsCertificateFlags) { ... }
    fn connect_accepted_cas_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
    fn connect_server_identity_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
    fn connect_validation_flags_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all TlsClientConnection methods.

Implementors

TlsClientConnection

Provided Methods

fn copy_session_state(&self, source: &impl IsA<TlsClientConnection>)

Possibly copies session state from one connection to another, for use in TLS session resumption. This is not normally needed, but may be used when the same session needs to be used between different endpoints, as is required by some protocols, such as FTP over TLS. @source should have already completed a handshake and, since TLS 1.3, it should have been used to read data at least once. @self should not have completed a handshake.

It is not possible to know whether a call to this function will actually do anything. Because session resumption is normally used only for performance benefit, the TLS backend might not implement this function. Even if implemented, it may not actually succeed in allowing @self to resume @source’s TLS session, because the server may not have sent a session resumption token to @source, or it may refuse to accept the token from @self. There is no way to know whether a call to this function is actually successful.

Using this function is not required to benefit from session resumption. If the TLS backend supports session resumption, the session will be resumed automatically if it is possible to do so without weakening the privacy guarantees normally provided by TLS, without need to call this function. For example, with TLS 1.3, a session ticket will be automatically copied from any #GTlsClientConnection that has previously received session tickets from the server, provided a ticket is available that has not previously been used for session resumption, since session ticket reuse would be a privacy weakness. Using this function causes the ticket to be copied without regard for privacy considerations.

source

a #GTlsClientConnection

fn accepted_cas(&self) -> Vec<ByteArray>

Gets the list of distinguished names of the Certificate Authorities that the server will accept certificates from. This will be set during the TLS handshake if the server requests a certificate. Otherwise, it will be None.

Each item in the list is a #GByteArray which contains the complete subject DN of the certificate authority.

Returns

the list of CA DNs. You should unref each element with g_byte_array_unref() and then the free the list with g_list_free().

fn server_identity(&self) -> Option<SocketConnectable>

Gets @self’s expected server identity

Returns

a #GSocketConnectable describing the expected server identity, or None if the expected identity is not known.

fn validation_flags(&self) -> TlsCertificateFlags

👎Deprecated: Since 2.72

Gets @self’s validation flags

This function does not work as originally designed and is impossible to use correctly. See #GTlsClientConnection:validation-flags for more information.

Deprecated since 2.72

Do not attempt to ignore validation errors.

Returns

the validation flags

fn set_server_identity(&self, identity: &impl IsA<SocketConnectable>)

Sets @self’s expected server identity, which is used both to tell servers on virtual hosts which certificate to present, and also to let @self know what name to look for in the certificate when performing TlsCertificateFlags::BAD_IDENTITY validation, if enabled.

identity

a #GSocketConnectable describing the expected server identity

fn set_validation_flags(&self, flags: TlsCertificateFlags)

👎Deprecated: Since 2.72

Sets @self’s validation flags, to override the default set of checks performed when validating a server certificate. By default, TlsCertificateFlags::VALIDATE_ALL is used.

This function does not work as originally designed and is impossible to use correctly. See #GTlsClientConnection:validation-flags for more information.

Deprecated since 2.72

Do not attempt to ignore validation errors.

flags

the #GTlsCertificateFlags to use

fn connect_accepted_cas_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

fn connect_server_identity_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

fn connect_validation_flags_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

👎Deprecated: Since 2.72

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.