Trait gio::prelude::TlsConnectionExt

pub trait TlsConnectionExt: 'static {
    fn emit_accept_certificate<P: IsA<TlsCertificate>>(
        &self, 
        peer_cert: &P, 
        errors: TlsCertificateFlags
    ) -> bool;

    fn get_certificate(&self) -> Option<TlsCertificate>;

    fn get_database(&self) -> Option<TlsDatabase>;

    fn get_interaction(&self) -> Option<TlsInteraction>;

    fn get_negotiated_protocol(&self) -> Option<GString>;

    fn get_peer_certificate(&self) -> Option<TlsCertificate>;

    fn get_peer_certificate_errors(&self) -> TlsCertificateFlags;

    fn get_rehandshake_mode(&self) -> TlsRehandshakeMode;

    fn get_require_close_notify(&self) -> bool;

    fn handshake<P: IsA<Cancellable>>(
        &self, 
        cancellable: Option<&P>
    ) -> Result<(), Error>;

    fn handshake_async<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
        &self, 
        io_priority: Priority, 
        cancellable: Option<&P>, 
        callback: Q
    );

    fn handshake_async_future(
        &self, 
        io_priority: Priority
    ) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>;

    fn set_advertised_protocols(&self, protocols: &[&str]);

    fn set_certificate<P: IsA<TlsCertificate>>(&self, certificate: &P);

    fn set_database<P: IsA<TlsDatabase>>(&self, database: Option<&P>);

    fn set_interaction<P: IsA<TlsInteraction>>(&self, interaction: Option<&P>);

    fn set_rehandshake_mode(&self, mode: TlsRehandshakeMode);

    fn set_require_close_notify(&self, require_close_notify: bool);

    fn get_property_advertised_protocols(&self) -> Vec<GString>;

    fn get_property_base_io_stream(&self) -> Option<IOStream>;

    fn connect_accept_certificate<F: Fn(&Self, &TlsCertificate, TlsCertificateFlags) -> bool + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_advertised_protocols_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_certificate_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_database_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_interaction_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_negotiated_protocol_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_peer_certificate_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_peer_certificate_errors_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_rehandshake_mode_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_require_close_notify_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all TlsConnection methods.

Implementors

TlsClientConnection, TlsConnection, TlsServerConnection

Required methods

fn emit_accept_certificate<P: IsA<TlsCertificate>>(
&self,
peer_cert: &P,
errors: TlsCertificateFlags
) -> bool

Used by TlsConnection implementations to emit the TlsConnection::accept-certificate signal.

peer_cert

the peer's TlsCertificate

errors

the problems with peer_cert

Returns

true if one of the signal handlers has returned true to accept peer_cert

fn get_certificate(&self) -> Option<TlsCertificate>

Gets self's certificate, as set by TlsConnectionExt::set_certificate.

Returns

self's certificate, or None

fn get_database(&self) -> Option<TlsDatabase>

Gets the certificate database that self uses to verify peer certificates. See TlsConnectionExt::set_database.

Returns

the certificate database that self uses or None

fn get_interaction(&self) -> Option<TlsInteraction>

Get the object that will be used to interact with the user. It will be used for things like prompting the user for passwords. If None is returned, then no user interaction will occur for this connection.

Returns

The interaction object.

fn get_negotiated_protocol(&self) -> Option<GString>

This is supported on crate feature v2_60 only.

Gets the name of the application-layer protocol negotiated during the handshake.

If the peer did not use the ALPN extension, or did not advertise a protocol that matched one of self's protocols, or the TLS backend does not support ALPN, then this will be None. See TlsConnectionExt::set_advertised_protocols.

Feature: v2_60

Returns

the negotiated protocol, or None

fn get_peer_certificate(&self) -> Option<TlsCertificate>

Gets self's peer's certificate after the handshake has completed or failed. (It is not set during the emission of TlsConnection::accept-certificate.)

Returns

self's peer's certificate, or None

fn get_peer_certificate_errors(&self) -> TlsCertificateFlags

Gets the errors associated with validating self's peer's certificate, after the handshake has completed or failed. (It is not set during the emission of TlsConnection::accept-certificate.)

Returns

self's peer's certificate errors

fn get_rehandshake_mode(&self) -> TlsRehandshakeMode

👎 Deprecated

Gets self rehandshaking mode. See TlsConnectionExt::set_rehandshake_mode for details.

Deprecated since 2.60

Changing the rehandshake mode is no longer required for compatibility. Also, rehandshaking has been removed from the TLS protocol in TLS 1.3.

Returns

TlsRehandshakeMode::Safely

fn get_require_close_notify(&self) -> bool

Tests whether or not self expects a proper TLS close notification when the connection is closed. See TlsConnectionExt::set_require_close_notify for details.

Returns

true if self requires a proper TLS close notification.

fn handshake<P: IsA<Cancellable>>(
&self,
cancellable: Option<&P>
) -> Result<(), Error>

Attempts a TLS handshake on self.

On the client side, it is never necessary to call this method; although the connection needs to perform a handshake after connecting (or after sending a "STARTTLS"-type command), TlsConnection will handle this for you automatically when you try to send or receive data on the connection. You can call TlsConnectionExt::handshake manually if you want to know whether the initial handshake succeeded or failed (as opposed to just immediately trying to use self to read or write, in which case, if it fails, it may not be possible to tell if it failed before or after completing the handshake), but beware that servers may reject client authentication after the handshake has completed, so a successful handshake does not indicate the connection will be usable.

Likewise, on the server side, although a handshake is necessary at the beginning of the communication, you do not need to call this function explicitly unless you want clearer error reporting.

Previously, calling TlsConnectionExt::handshake after the initial handshake would trigger a rehandshake; however, this usage was deprecated in GLib 2.60 because rehandshaking was removed from the TLS protocol in TLS 1.3. Since GLib 2.64, calling this function after the initial handshake will no longer do anything.

When using a TlsConnection created by SocketClient, the SocketClient performs the initial handshake, so calling this function manually is not recommended.

TlsConnection::accept_certificate may be emitted during the handshake.

cancellable

a Cancellable, or None

Returns

success or failure

fn handshake_async<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
io_priority: Priority,
cancellable: Option<&P>,
callback: Q
)

Asynchronously performs a TLS handshake on self. See TlsConnectionExt::handshake for more information.

io_priority

the [I/O priority][io-priority] of the request

cancellable

a Cancellable, or None

callback

callback to call when the handshake is complete

user_data

the data to pass to the callback function

fn handshake_async_future(
&self,
io_priority: Priority
) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

fn set_advertised_protocols(&self, protocols: &[&str])

This is supported on crate feature v2_60 only.

Sets the list of application-layer protocols to advertise that the caller is willing to speak on this connection. The Application-Layer Protocol Negotiation (ALPN) extension will be used to negotiate a compatible protocol with the peer; use TlsConnectionExt::get_negotiated_protocol to find the negotiated protocol after the handshake. Specifying None for the the value of protocols will disable ALPN negotiation.

See IANA TLS ALPN Protocol IDs for a list of registered protocol IDs.

Feature: v2_60

protocols

a None-terminated array of ALPN protocol names (eg, "http/1.1", "h2"), or None

fn set_certificate<P: IsA<TlsCertificate>>(&self, certificate: &P)

This sets the certificate that self will present to its peer during the TLS handshake. For a TlsServerConnection, it is mandatory to set this, and that will normally be done at construct time.

For a TlsClientConnection, this is optional. If a handshake fails with TlsError::CertificateRequired, that means that the server requires a certificate, and if you try connecting again, you should call this method first. You can call TlsClientConnection::get_accepted_cas on the failed connection to get a list of Certificate Authorities that the server will accept certificates from.

(It is also possible that a server will allow the connection with or without a certificate; in that case, if you don't provide a certificate, you can tell that the server requested one by the fact that TlsClientConnection::get_accepted_cas will return non-None.)

certificate

the certificate to use for self

fn set_database<P: IsA<TlsDatabase>>(&self, database: Option<&P>)

Sets the certificate database that is used to verify peer certificates. This is set to the default database by default. See TlsBackend::get_default_database. If set to None, then peer certificate validation will always set the TlsCertificateFlags::UnknownCa error (meaning TlsConnection::accept-certificate will always be emitted on client-side connections, unless that bit is not set in TlsClientConnection:validation-flags).

database

a TlsDatabase

fn set_interaction<P: IsA<TlsInteraction>>(&self, interaction: Option<&P>)

Set the object that will be used to interact with the user. It will be used for things like prompting the user for passwords.

The interaction argument will normally be a derived subclass of TlsInteraction. None can also be provided if no user interaction should occur for this connection.

interaction

an interaction object, or None

fn set_rehandshake_mode(&self, mode: TlsRehandshakeMode)

👎 Deprecated

Since GLib 2.64, changing the rehandshake mode is no longer supported and will have no effect. With TLS 1.3, rehandshaking has been removed from the TLS protocol, replaced by separate post-handshake authentication and rekey operations.

Deprecated since 2.60

Changing the rehandshake mode is no longer required for compatibility. Also, rehandshaking has been removed from the TLS protocol in TLS 1.3.

mode

the rehandshaking mode

fn set_require_close_notify(&self, require_close_notify: bool)

Sets whether or not self expects a proper TLS close notification before the connection is closed. If this is true (the default), then self will expect to receive a TLS close notification from its peer before the connection is closed, and will return a TlsError::Eof error if the connection is closed without proper notification (since this may indicate a network error, or man-in-the-middle attack).

In some protocols, the application will know whether or not the connection was closed cleanly based on application-level data (because the application-level data includes a length field, or is somehow self-delimiting); in this case, the close notify is redundant and sometimes omitted. (TLS 1.1 explicitly allows this; in TLS 1.0 it is technically an error, but often done anyway.) You can use TlsConnectionExt::set_require_close_notify to tell self to allow an "unannounced" connection close, in which case the close will show up as a 0-length read, as in a non-TLS SocketConnection, and it is up to the application to check that the data has been fully received.

Note that this only affects the behavior when the peer closes the connection; when the application calls IOStreamExt::close itself on self, this will send a close notification regardless of the setting of this property. If you explicitly want to do an unclean close, you can close self's TlsConnection:base-io-stream rather than closing self itself, but note that this may only be done when no other operations are pending on self or the base I/O stream.

require_close_notify

whether or not to require close notification

fn get_property_advertised_protocols(&self) -> Vec<GString>

This is supported on crate feature v2_60 only.

The list of application-layer protocols that the connection advertises that it is willing to speak. See TlsConnectionExt::set_advertised_protocols.

Feature: v2_60

fn get_property_base_io_stream(&self) -> Option<IOStream>

The IOStream that the connection wraps. The connection holds a reference to this stream, and may run operations on the stream from other threads throughout its lifetime. Consequently, after the IOStream has been constructed, application code may only run its own operations on this stream when no IOStream operations are running.

fn connect_accept_certificate<F: Fn(&Self, &TlsCertificate, TlsCertificateFlags) -> bool + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted during the TLS handshake after the peer certificate has been received. You can examine peer_cert's certification path by calling TlsCertificateExt::get_issuer on it.

For a client-side connection, peer_cert is the server's certificate, and the signal will only be emitted if the certificate was not acceptable according to conn's TlsClientConnection:validation_flags. If you would like the certificate to be accepted despite errors, return true from the signal handler. Otherwise, if no handler accepts the certificate, the handshake will fail with TlsError::BadCertificate.

For a server-side connection, peer_cert is the certificate presented by the client, if this was requested via the server's TlsServerConnection:authentication_mode. On the server side, the signal is always emitted when the client presents a certificate, and the certificate will only be accepted if a handler returns true.

Note that if this signal is emitted as part of asynchronous I/O in the main thread, then you should not attempt to interact with the user before returning from the signal handler. If you want to let the user decide whether or not to accept the certificate, you would have to return false from the signal handler on the first attempt, and then after the connection attempt returns a TlsError::BadCertificate, you can interact with the user, and if the user decides to accept the certificate, remember that fact, create a new connection, and return true from the signal handler the next time.

If you are doing I/O in another thread, you do not need to worry about this, and can simply block in the signal handler until the UI thread returns an answer.

peer_cert

the peer's TlsCertificate

errors

the problems with peer_cert.

Returns

true to accept peer_cert (which will also immediately end the signal emission). false to allow the signal emission to continue, which will cause the handshake to fail if no one else overrides it.

fn connect_property_advertised_protocols_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

This is supported on crate feature v2_60 only.

fn connect_property_certificate_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_database_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_interaction_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_negotiated_protocol_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

This is supported on crate feature v2_60 only.

fn connect_property_peer_certificate_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_peer_certificate_errors_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_rehandshake_mode_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

👎 Deprecated

fn connect_property_require_close_notify_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...