Trait gio::prelude::TlsConnectionExtManual

pub trait TlsConnectionExtManual {
    fn channel_binding_data(
        &self, 
        type_: TlsChannelBindingType
    ) -> Result<ByteArray, Error>;
}

Required Methods

fn channel_binding_data(
&self,
type_: TlsChannelBindingType
) -> Result<ByteArray, Error>

Available on crate feature v2_66 only.

Query the TLS backend for TLS channel binding data of type_ for self.

This call retrieves TLS channel binding data as specified in RFC 5056, RFC 5929, and related RFCs. The binding data is returned in data. The data is resized by the callee using glib::ByteArray buffer management and will be freed when the data is destroyed by g_byte_array_unref(). If data is None, it will only check whether TLS backend is able to fetch the data (e.g. whether type_ is supported by the TLS backend). It does not guarantee that the data will be available though. That could happen if TLS connection does not support type_ or the binding data is not available yet due to additional negotiation or input required.

type_

TlsChannelBindingType type of data to fetch

Returns

true on success, false otherwise

data

glib::ByteArray is filled with the binding data, or None