Trait gio::prelude::TlsPasswordExt

pub trait TlsPasswordExt:
    IsA<TlsPassword>
    + Sealed
    + 'static {
    // Provided methods
    fn description(&self) -> GString { ... }
    fn flags(&self) -> TlsPasswordFlags { ... }
    fn warning(&self) -> GString { ... }
    fn set_description(&self, description: &str) { ... }
    fn set_flags(&self, flags: TlsPasswordFlags) { ... }
    fn set_warning(&self, warning: &str) { ... }
    fn connect_description_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
    fn connect_flags_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
    fn connect_warning_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all TlsPassword methods.

Implementors

TlsPassword

Provided Methods

fn description(&self) -> GString

Get a description string about what the password will be used for.

Returns

The description of the password.

fn flags(&self) -> TlsPasswordFlags

Get flags about the password.

Returns

The flags about the password.

fn warning(&self) -> GString

Get a user readable translated warning. Usually this warning is a representation of the password flags returned from g_tls_password_get_flags().

Returns

The warning.

fn set_description(&self, description: &str)

Set a description string about what the password will be used for.

description

The description of the password

fn set_flags(&self, flags: TlsPasswordFlags)

Set flags about the password.

flags

The flags about the password

fn set_warning(&self, warning: &str)

Set a user readable translated warning. Usually this warning is a representation of the password flags returned from g_tls_password_get_flags().

warning

The user readable warning

fn connect_description_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

fn connect_flags_notify<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

fn connect_warning_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

Object Safety

This trait is not object safe.