Trait gio::prelude::UnixOutputStreamExtManual

pub trait UnixOutputStreamExtManual: Sized {
    fn fd<T: FromRawFd>(&self) -> T;
    unsafe fn set_close_fd(&self, close_fd: bool);
}

Required Methods

fn fd<T: FromRawFd>(&self) -> T

Return the UNIX file descriptor that the stream writes to.

Returns

The file descriptor of self

unsafe fn set_close_fd(&self, close_fd: bool)

Sets whether the fd of this stream will be closed when the stream is closed.

Safety

If you pass in false as the parameter, you may only close the fd if the all references to the stream have been dropped. If you pass in true, you must never call close.