gio
Struct AppInfo

pub struct AppInfo { /* private fields */ }

Information about an installed application and methods to launch it (with file arguments).

GAppInfo and GAppLaunchContext are used for describing and launching applications installed on the system.

As of GLib 2.20, URIs will always be converted to POSIX paths (using FileExt::path()) when using AppInfoExt::launch() even if the application requested an URI and not a POSIX path. For example for a desktop-file based application with the following Exec key:

Exec=totem %U

and a single URI, sftp://foo/file.avi, then /home/user/.gvfs/sftp on foo/file.avi will be passed. This will only work if a set of suitable GIO extensions (such as GVfs 2.26 compiled with FUSE support), is available and operational; if this is not the case, the URI will be passed unmodified to the application. Some URIs, such as mailto:, of course cannot be mapped to a POSIX path (in GVfs there’s no FUSE mount for it); such URIs will be passed unmodified to the application.

Specifically for GVfs 2.26 and later, the POSIX URI will be mapped back to the GIO URI in the File constructors (since GVfs implements the GVfs extension point). As such, if the application needs to examine the URI, it needs to use FileExt::uri() or similar on File. In other words, an application cannot assume that the URI passed to e.g. File::for_commandline_arg() is equal to the result of FileExt::uri(). The following snippet illustrates this:

⚠️ The following code is in c ⚠️

GFile *f;
char *uri;

file = g_file_new_for_commandline_arg (uri_from_commandline);

uri = g_file_get_uri (file);
strcmp (uri, uri_from_commandline) == 0;
g_free (uri);

if (g_file_has_uri_scheme (file, "cdda"))
  {
    // do something special with uri
  }
g_object_unref (file);

This code will work when both cdda://sr0/Track 1.wav and /home/user/.gvfs/cdda on sr0/Track 1.wav is passed to the application. It should be noted that it’s generally not safe for applications to rely on the format of a particular URIs. Different launcher applications (e.g. file managers) may have different ideas of what a given URI means.

Implements

AppInfoExt, AppInfoExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl AppInfo

pub const NONE: Option<&'static AppInfo> = None

pub fn create_from_commandline( commandline: impl AsRef<OsStr>, application_name: Option<&str>, flags: AppInfoCreateFlags, ) -> Result<AppInfo, Error>

Creates a new AppInfo from the given information.

Note that for @commandline, the quoting rules of the Exec key of the freedesktop.org Desktop Entry Specification are applied. For example, if the @commandline contains percent-encoded URIs, the percent-character must be doubled in order to prevent it from being swallowed by Exec key unquoting. See the specification for exact quoting rules.

commandline

the command line to use

application_name

the application name, or NULL to use @commandline

flags

flags that can specify details of the created AppInfo

Returns

new AppInfo for given command.

pub fn all() -> Vec<AppInfo>

Gets a list of all of the applications currently registered on this system.

For desktop files, this includes applications that have NoDisplay=true set or are excluded from display by means of OnlyShowIn or NotShowIn. See AppInfoExt::should_show().

The returned list does not include applications which have the Hidden key set.

Returns

a newly allocated list of references to AppInfos.

pub fn all_for_type(content_type: &str) -> Vec<AppInfo>

Gets a list of all AppInfos for a given content type, including the recommended and fallback AppInfos. See recommended_for_type() and fallback_for_type().

content_type

the content type to find a AppInfo for

Returns

list of AppInfos for given @content_type.

pub fn default_for_type( content_type: &str, must_support_uris: bool, ) -> Option<AppInfo>

Gets the default AppInfo for a given content type.

content_type

the content type to find a AppInfo for

must_support_uris

if TRUE, the AppInfo is expected to support URIs

Returns

AppInfo for given @content_type or NULL on error.

pub fn default_for_type_async<P: FnOnce(Result<AppInfo, Error>) + 'static>( content_type: &str, must_support_uris: bool, cancellable: Option<&impl IsA<Cancellable>>, callback: P, )

Available on crate feature v2_74 only.

Asynchronously gets the default AppInfo for a given content type.

content_type

the content type to find a AppInfo for

must_support_uris

if TRUE, the AppInfo is expected to support URIs

cancellable

a Cancellable

callback

a [type@Gio.AsyncReadyCallback] to call when the request is done

pub fn default_for_type_future( content_type: &str, must_support_uris: bool, ) -> Pin<Box_<dyn Future<Output = Result<AppInfo, Error>> + 'static>>

Available on crate feature v2_74 only.

pub fn default_for_uri_scheme(uri_scheme: &str) -> Option<AppInfo>

Gets the default application for handling URIs with the given URI scheme.

A URI scheme is the initial part of the URI, up to but not including the :. For example, http, ftp or sip.

uri_scheme

a string containing a URI scheme.

Returns

AppInfo for given @uri_scheme or NULL on error.

pub fn default_for_uri_scheme_async<P: FnOnce(Result<AppInfo, Error>) + 'static>( uri_scheme: &str, cancellable: Option<&impl IsA<Cancellable>>, callback: P, )

Available on crate feature v2_74 only.

Asynchronously gets the default application for handling URIs with the given URI scheme. A URI scheme is the initial part of the URI, up to but not including the :, e.g. http, ftp or sip.

uri_scheme

a string containing a URI scheme.

cancellable

a Cancellable

callback

a [type@Gio.AsyncReadyCallback] to call when the request is done

pub fn default_for_uri_scheme_future( uri_scheme: &str, ) -> Pin<Box_<dyn Future<Output = Result<AppInfo, Error>> + 'static>>

Available on crate feature v2_74 only.

pub fn fallback_for_type(content_type: &str) -> Vec<AppInfo>

Gets a list of fallback AppInfos for a given content type, i.e. those applications which claim to support the given content type by MIME type subclassing and not directly.

content_type

the content type to find a AppInfo for

Returns

list of AppInfos for given @content_type or NULL on error.

pub fn recommended_for_type(content_type: &str) -> Vec<AppInfo>

Gets a list of recommended AppInfos for a given content type, i.e. those applications which claim to support the given content type exactly, and not by MIME type subclassing.

Note that the first application of the list is the last used one, i.e. the last one for which AppInfoExt::set_as_last_used_for_type() has been called.

content_type

the content type to find a AppInfo for

Returns

list of AppInfos for given @content_type or NULL on error.

pub fn launch_default_for_uri( uri: &str, context: Option<&impl IsA<AppLaunchContext>>, ) -> Result<(), Error>

Utility function that launches the default application registered to handle the specified uri. Synchronous I/O is done on the uri to detect the type of the file if required.

The D-Bus–activated applications don’t have to be started if your application terminates too soon after this function. To prevent this, use launch_default_for_uri_async() instead.

uri

the uri to show

context

optional launch context

Returns

TRUE on success, FALSE on error.

pub fn launch_default_for_uri_async<P: FnOnce(Result<(), Error>) + 'static>( uri: &str, context: Option<&impl IsA<AppLaunchContext>>, cancellable: Option<&impl IsA<Cancellable>>, callback: P, )

Async version of launch_default_for_uri().

This version is useful if you are interested in receiving error information in the case where the application is sandboxed and the portal may present an application chooser dialog to the user.

This is also useful if you want to be sure that the D-Bus–activated applications are really started before termination and if you are interested in receiving error information from their activation.

uri

the uri to show

context

optional launch context

cancellable

a Cancellable

callback

a [type@Gio.AsyncReadyCallback] to call when the request is done

pub fn launch_default_for_uri_future( uri: &str, context: Option<&(impl IsA<AppLaunchContext> + Clone + 'static)>, ) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

pub fn reset_type_associations(content_type: &str)

Removes all changes to the type associations done by AppInfoExt::set_as_default_for_type(), AppInfoExt::set_as_default_for_extension(), AppInfoExt::add_supports_type() or AppInfoExt::remove_supports_type().

content_type

a content type