gio
Struct Application

pub struct Application { /* private fields */ }

GApplication is the core class for application support.

A GApplication is the foundation of an application. It wraps some low-level platform-specific services and is intended to act as the foundation for higher-level application classes such as GtkApplication or MxApplication. In general, you should not use this class outside of a higher level framework.

GApplication provides convenient life-cycle management by maintaining a “use count” for the primary application instance. The use count can be changed using ApplicationExtManual::hold() and [ApplicationExtManual::release()][crate::prelude::ApplicationExtManual::release()]. If it drops to zero, the application exits. Higher-level classes such as GtkApplication employ the use count to ensure that the application stays alive as long as it has any opened windows.

Another feature that GApplication (optionally) provides is process uniqueness. Applications can make use of this functionality by providing a unique application ID. If given, only one application with this ID can be running at a time per session. The session concept is platform-dependent, but corresponds roughly to a graphical desktop login. When your application is launched again, its arguments are passed through platform communication to the already running program. The already running instance of the program is called the “primary instance”; for non-unique applications this is always the current instance. On Linux, the D-Bus session bus is used for communication.

The use of GApplication differs from some other commonly-used uniqueness libraries (such as libunique) in important ways. The application is not expected to manually register itself and check if it is the primary instance. Instead, the main() function of a GApplication should do very little more than instantiating the application instance, possibly connecting signal handlers, then calling ApplicationExtManual::run(). All checks for uniqueness are done internally. If the application is the primary instance then the startup signal is emitted and the mainloop runs. If the application is not the primary instance then a signal is sent to the primary instance and ApplicationExtManual::run() promptly returns. See the code examples below.

If used, the expected form of an application identifier is the same as that of a D-Bus well-known bus name. Examples include: com.example.MyApp, org.example.internal_apps.Calculator, org._7_zip.Archiver. For details on valid application identifiers, see id_is_valid().

On Linux, the application identifier is claimed as a well-known bus name on the user’s session bus. This means that the uniqueness of your application is scoped to the current session. It also means that your application may provide additional services (through registration of other object paths) at that bus name. The registration of these object paths should be done with the shared GDBus session bus. Note that due to the internal architecture of GDBus, method calls can be dispatched at any time (even if a main loop is not running). For this reason, you must ensure that any object paths that you wish to register are registered before #GApplication attempts to acquire the bus name of your application (which happens in ApplicationExt::register()). Unfortunately, this means that you cannot use is-remote to decide if you want to register object paths.

GApplication also implements the ActionGroup and ActionMap interfaces and lets you easily export actions by adding them with ActionMapExt::add_action(). When invoking an action by calling ActionGroupExt::activate_action() on the application, it is always invoked in the primary instance. The actions are also exported on the session bus, and GIO provides the DBusActionGroup wrapper to conveniently access them remotely. GIO provides a DBusMenuModel wrapper for remote access to exported MenuModels.

Note: Due to the fact that actions are exported on the session bus, using maybe parameters is not supported, since D-Bus does not support maybe types.

There is a number of different entry points into a GApplication:

via ‘Activate’ (i.e. just starting the application)

via ‘Open’ (i.e. opening some files)

by handling a command-line

via activating an action

The startup signal lets you handle the application initialization for all of these in a single place.

Regardless of which of these entry points is used to start the application, GApplication passes some ‘platform data’ from the launching instance to the primary instance, in the form of a glib::Variant dictionary mapping strings to variants. To use platform data, override the ApplicationImpl::before_emit() or ApplicationImpl::after_emit() virtual functions in your GApplication subclass. When dealing with ApplicationCommandLine objects, the platform data is directly available via ApplicationCommandLineExt::cwd(), ApplicationCommandLineExt::environ() and ApplicationCommandLineExt::platform_data().

As the name indicates, the platform data may vary depending on the operating system, but it always includes the current directory (key cwd), and optionally the environment (ie the set of environment variables and their values) of the calling process (key environ). The environment is only added to the platform data if the G_APPLICATION_SEND_ENVIRONMENT flag is set. GApplication subclasses can add their own platform data by overriding the vfunc::Gio::Application::add_platform_data virtual function. For instance, GtkApplication adds startup notification data in this way.

To parse commandline arguments you may handle the command-line signal or override the vfunc::Gio::Application::local_command_line virtual function, to parse them in either the primary instance or the local instance, respectively.

For an example of opening files with a GApplication, see gapplication-example-open.c.

For an example of using actions with GApplication, see gapplication-example-actions.c.

For an example of using extra D-Bus hooks with GApplication, see gapplication-example-dbushooks.c.

Properties

action-group

The group of actions that the application exports.

Writeable

application-id

The unique identifier for the application.

Readable | Writeable | Construct

flags

Flags specifying the behaviour of the application.

Readable | Writeable

inactivity-timeout

Time (in milliseconds) to stay alive after becoming idle.

Readable | Writeable

is-busy

Whether the application is currently marked as busy through g_application_mark_busy() or g_application_bind_busy_property().

Readable

is-registered

Whether ApplicationExt::register() has been called.

Readable

is-remote

Whether this application instance is remote.

Readable

resource-base-path

The base resource path for the application.

Readable | Writeable

version

The human-readable version number of the application.

Readable | Writeable

Signals

activate

The ::activate signal is emitted on the primary instance when an activation occurs. See g_application_activate().

command-line

The ::command-line signal is emitted on the primary instance when a commandline is not handled locally. See g_application_run() and the #GApplicationCommandLine documentation for more information.

handle-local-options

The ::handle-local-options signal is emitted on the local instance after the parsing of the commandline options has occurred.

You can add options to be recognised during commandline option parsing using g_application_add_main_option_entries() and g_application_add_option_group().

Signal handlers can inspect @options (along with values pointed to from the @arg_data of an installed #GOptionEntrys) in order to decide to perform certain actions, including direct local handling (which may be useful for options like –version).

In the event that the application is marked ApplicationFlags::HANDLES_COMMAND_LINE the “normal processing” will send the @options dictionary to the primary instance where it can be read with g_application_command_line_get_options_dict(). The signal handler can modify the dictionary before returning, and the modified dictionary will be sent.

In the event that ApplicationFlags::HANDLES_COMMAND_LINE is not set, “normal processing” will treat the remaining uncollected command line arguments as filenames or URIs. If there are no arguments, the application is activated by g_application_activate(). One or more arguments results in a call to g_application_open().

If you want to handle the local commandline arguments for yourself by converting them to calls to g_application_open() or g_action_group_activate_action() then you must be sure to register the application first. You should probably not call g_application_activate() for yourself, however: just return -1 and allow the default handler to do it for you. This will ensure that the --gapplication-service switch works properly (i.e. no activation in that case).

Note that this signal is emitted from the default implementation of local_command_line(). If you override that function and don’t chain up then this signal will never be emitted.

You can override local_command_line() if you need more powerful capabilities than what is provided here, but this should not normally be required.

name-lost

The ::name-lost signal is emitted only on the registered primary instance when a new instance has taken over. This can only happen if the application is using the ApplicationFlags::ALLOW_REPLACEMENT flag.

The default handler for this signal calls g_application_quit().

open

The ::open signal is emitted on the primary instance when there are files to open. See g_application_open() for more information.

shutdown

The ::shutdown signal is emitted only on the registered primary instance immediately after the main loop terminates.

startup

The ::startup signal is emitted on the primary instance immediately after registration. See g_application_register().

ActionGroup

action-added

Signals that a new action was just added to the group.

This signal is emitted after the action has been added and is now visible.

Detailed

action-enabled-changed

Signals that the enabled status of the named action has changed.

Detailed

action-removed

Signals that an action is just about to be removed from the group.

This signal is emitted before the action is removed, so the action is still visible and can be queried from the signal handler.

Detailed

action-state-changed

Signals that the state of the named action has changed.

Detailed

Implements

ApplicationExt, [trait@glib::ObjectExt], ActionGroupExt, ActionMapExt, ApplicationExtManual, ActionMapExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl Application

pub const NONE: Option<&'static Application> = None

pub fn new(application_id: Option<&str>, flags: ApplicationFlags) -> Application

Creates a new #GApplication instance.

If non-None, the application id must be valid. See g_application_id_is_valid().

If no application ID is given then some features of #GApplication (most notably application uniqueness) will be disabled.

application_id

the application id

flags

the application flags

Returns

a new #GApplication instance

pub fn builder() -> ApplicationBuilder

Creates a new builder-pattern struct instance to construct Application objects.

This method returns an instance of ApplicationBuilder which can be used to create Application objects.

pub fn default() -> Option<Application>

Returns the default #GApplication instance for this process.

Normally there is only one #GApplication per process and it becomes the default when it is created. You can exercise more control over this by using g_application_set_default().

If there is no default application then None is returned.

Returns

the default application for this process, or None

pub fn id_is_valid(application_id: &str) -> bool

Checks if @application_id is a valid application identifier.

A valid ID is required for calls to g_application_new() and g_application_set_application_id().

Application identifiers follow the same format as D-Bus well-known bus names. For convenience, the restrictions on application identifiers are reproduced here:

Application identifiers are composed of 1 or more elements separated by a period (.) character. All elements must contain at least one character.

Each element must only contain the ASCII characters [A-Z][a-z][0-9]_-, with - discouraged in new application identifiers. Each element must not begin with a digit.

Application identifiers must contain at least one . (period) character (and thus at least two elements).

Application identifiers must not begin with a . (period) character.

Application identifiers must not exceed 255 characters.

Note that the hyphen (-) character is allowed in application identifiers, but is problematic or not allowed in various specifications and APIs that refer to D-Bus, such as Flatpak application IDs, the DBusActivatable interface in the Desktop Entry Specification, and the convention that an application’s “main” interface and object path resemble its application identifier and bus name. To avoid situations that require special-case handling, it is recommended that new application identifiers consistently replace hyphens with underscores.

Like D-Bus interface names, application identifiers should start with the reversed DNS domain name of the author of the interface (in lower-case), and it is conventional for the rest of the application identifier to consist of words run together, with initial capital letters.

As with D-Bus interface names, if the author’s DNS domain name contains hyphen/minus characters they should be replaced by underscores, and if it contains leading digits they should be escaped by prepending an underscore. For example, if the owner of 7-zip.org used an application identifier for an archiving application, it might be named org._7_zip.Archiver.

application_id

a potential application identifier

Returns

true if @application_id is valid