Struct gio::ApplicationFlags

pub struct ApplicationFlags { /* fields omitted */ }

Implementations

impl ApplicationFlags

pub const FLAGS_NONE: ApplicationFlags

pub const IS_SERVICE: ApplicationFlags

pub const IS_LAUNCHER: ApplicationFlags

pub const HANDLES_OPEN: ApplicationFlags

pub const HANDLES_COMMAND_LINE: ApplicationFlags

pub const SEND_ENVIRONMENT: ApplicationFlags

pub const NON_UNIQUE: ApplicationFlags

pub const CAN_OVERRIDE_APP_ID: ApplicationFlags

pub const ALLOW_REPLACEMENT: ApplicationFlags

pub const REPLACE: ApplicationFlags

pub const fn empty() -> ApplicationFlags

Returns an empty set of flags

Uses found in examples/

examples/overlay/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85

//! # Overlay example
//!
//! This example demonstrates how to create an element "floating" above others.

use gtk::prelude::*;
use gtk::{gdk, gio, glib};

fn button_clicked(button: &gtk::Button, overlay_text: &gtk::Label) {
    overlay_text.set_text(&button.label().expect("Couldn't get button label"));
}

fn build_ui(application: &gtk::Application) {
    let window = gtk::ApplicationWindow::new(application);

    window.set_title("Overlay");
    window.set_position(gtk::WindowPosition::Center);

    // The overlay container.
    let overlay = gtk::Overlay::new();

    // The overlay label.
    let overlay_text = gtk::Label::new(Some("0"));
    // We need to name it in order to apply CSS on it.
    overlay_text.set_widget_name("overlay-label");
    // We put the overlay in the top-right corner of the window.
    overlay_text.set_halign(gtk::Align::End);
    overlay_text.set_valign(gtk::Align::Start);

    // We add into the overlay container as the overlay element.
    overlay.add_overlay(&overlay_text);

    let hbox = gtk::Box::new(gtk::Orientation::Horizontal, 0);

    let but1 = gtk::Button::with_label("Click me!");
    let but2 = gtk::Button::with_label("Or me!");
    let but3 = gtk::Button::with_label("Why not me?");

    // When a button is clicked on, we set its label to the overlay label.
    but1.connect_clicked(glib::clone!(@weak overlay_text => move |b| {
        button_clicked(b, &overlay_text);
    }));
    but2.connect_clicked(glib::clone!(@weak overlay_text => move |b| {
        button_clicked(b, &overlay_text);
    }));
    but3.connect_clicked(glib::clone!(@weak overlay_text => move |b| {
        button_clicked(b, &overlay_text);
    }));

    hbox.add(&but1);
    hbox.add(&but2);
    hbox.add(&but3);

    // We add the horizontal box into the overlay container "normally" (so this won't be an overlay
    // element).
    overlay.add(&hbox);
    // Then we add the overlay container inside our window.
    window.add(&overlay);

    window.show_all();
}

fn main() {
    let application =
        gtk::Application::new(Some("com.github.overlay"), gio::ApplicationFlags::empty());

    application.connect_startup(|_| {
        // We add a bit of CSS in order to make the overlay label easier to be seen.
        let provider = gtk::CssProvider::new();
        // Basic CSS: we change background color, we set font color to black and we set it as bold.
        let style = include_bytes!("style.css");
        provider.load_from_data(style).expect("Failed to load CSS");
        gtk::StyleContext::add_provider_for_screen(
            &gdk::Screen::default().expect("Error initializing gtk css provider."),
            &provider,
            gtk::STYLE_PROVIDER_PRIORITY_APPLICATION,
        );
    });

    application.connect_activate(|app| {
        // We build the application UI.
        build_ui(app);
    });

    application.run();
}

examples/css/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62

use gtk::prelude::*;
use gtk::{gdk, gio};

fn main() {
    let application = gtk::Application::new(Some("com.github.css"), gio::ApplicationFlags::empty());

    application.connect_startup(|app| {
        // The CSS "magic" happens here.
        let provider = gtk::CssProvider::new();
        // Load the CSS file
        let style = include_bytes!("style.css");
        provider.load_from_data(style).expect("Failed to load CSS");
        // We give the CssProvided to the default screen so the CSS rules we added
        // can be applied to our window.
        gtk::StyleContext::add_provider_for_screen(
            &gdk::Screen::default().expect("Error initializing gtk css provider."),
            &provider,
            gtk::STYLE_PROVIDER_PRIORITY_APPLICATION,
        );

        // We build the application UI.
        build_ui(app);
    });

    application.run();
}

fn build_ui(application: &gtk::Application) {
    let window = gtk::ApplicationWindow::new(application);

    window.set_title("CSS");
    window.set_position(gtk::WindowPosition::Center);

    // The container container.
    let vbox = gtk::Box::new(gtk::Orientation::Vertical, 0);

    let label = gtk::Button::with_label("hover me!");
    // We need to name it in order to be able to use its name as a CSS label to
    // apply CSS on it.
    label.set_widget_name("label1");

    let entry = gtk::Entry::new();
    // We need to name it in order to apply CSS on it.
    entry.set_widget_name("entry1");
    entry.set_text("Some text");

    let combo = gtk::ComboBoxText::new();
    combo.append_text("option 1");
    combo.append_text("option 2");
    combo.append_text("option 3");
    combo.set_active(Some(0));

    vbox.add(&label);
    vbox.add(&entry);
    vbox.add(&combo);
    // Then we add the container inside our window.
    window.add(&vbox);

    application.connect_activate(move |_| {
        window.show_all();
    });
}

examples/progress_tracker/main.rs
≺ ≻↕

  1
  2
  3
  4
  5
  6
  7
  8
  9
 10
 11
 12
 13
 14
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24
 25
 26
 27
 28
 29
 30
 31
 32
 33
 34
 35
 36
 37
 38
 39
 40
 41
 42
 43
 44
 45
 46
 47
 48
 49
 50
 51
 52
 53
 54
 55
 56
 57
 58
 59
 60
 61
 62
 63
 64
 65
 66
 67
 68
 69
 70
 71
 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216

use gtk::prelude::*;
use gtk::{gio, glib};

use std::cell::{Cell, RefCell};
use std::rc::Rc;
use std::thread;
use std::time::Duration;

pub fn main() {
    glib::set_program_name(Some("Progress Tracker"));

    let application = gtk::Application::new(
        Some("com.github.progress-tracker"),
        gio::ApplicationFlags::empty(),
    );

    application.connect_startup(|app| {
        let application = Application::new(app);

        let application_container = RefCell::new(Some(application));
        app.connect_shutdown(move |_| {
            let application = application_container
                .borrow_mut()
                .take()
                .expect("Shutdown called multiple times");
            // Here we could do whatever we need to do for shutdown now
            drop(application);
        });
    });

    application.connect_activate(|_| {});
    application.run();
}

pub struct Application {
    pub widgets: Rc<Widgets>,
}

impl Application {
    pub fn new(app: &gtk::Application) -> Self {
        let app = Application {
            widgets: Rc::new(Widgets::new(app)),
        };

        app.connect_progress();

        app
    }

    fn connect_progress(&self) {
        let active = Rc::new(Cell::new(false));
        self.widgets.main_view.button.connect_clicked(
            glib::clone!(@weak self.widgets as widgets => move |_| {
                if active.get() {
                    return;
                }

                active.set(true);

                let (tx, rx) = glib::MainContext::channel(glib::PRIORITY_DEFAULT);
                thread::spawn(move || {
                    for v in 1..=10 {
                        let _ = tx.send(Some(v));
                        thread::sleep(Duration::from_millis(500));
                    }
                    let _ = tx.send(None);
                });

                rx.attach(None, glib::clone!(@weak active, @weak widgets => @default-return glib::Continue(false), move |value| match value {
                    Some(value) => {
                        widgets
                            .main_view
                            .progress
                            .set_fraction(f64::from(value) / 10.0);

                        if value == 10 {
                            widgets
                                .view_stack
                                .set_visible_child(&widgets.complete_view.container);

                            glib::timeout_add_local(Duration::from_millis(1500), glib::clone!(@weak widgets => @default-return glib::Continue(false), move || {
                                widgets.main_view.progress.set_fraction(0.0);
                                widgets
                                    .view_stack
                                    .set_visible_child(&widgets.main_view.container);
                                glib::Continue(false)
                            }));
                        }

                        glib::Continue(true)
                    }
                    None => {
                        active.set(false);
                        glib::Continue(false)
                    }
                }));
            }),
        );
    }
}

pub struct Widgets {
    pub window: gtk::ApplicationWindow,
    pub header: Header,
    pub view_stack: gtk::Stack,
    pub main_view: MainView,
    pub complete_view: CompleteView,
}

impl Widgets {
    pub fn new(application: &gtk::Application) -> Self {
        let complete_view = CompleteView::new();
        let main_view = MainView::new();

        let view_stack = gtk::Stack::new();
        view_stack.set_border_width(6);
        view_stack.set_vexpand(true);
        view_stack.set_hexpand(true);
        view_stack.add(&main_view.container);
        view_stack.add(&complete_view.container);

        let header = Header::new();

        let window = gtk::ApplicationWindow::new(application);
        window.set_icon_name(Some("package-x-generic"));
        window.set_window_position(gtk::WindowPosition::Center);
        window.set_titlebar(Some(&header.container));
        window.add(&view_stack);
        window.show_all();
        window.set_default_size(500, 250);
        window.connect_delete_event(move |window, _| {
            window.close();
            Inhibit(false)
        });

        Self {
            window,
            header,
            view_stack,
            main_view,
            complete_view,
        }
    }
}

pub struct Header {
    container: gtk::HeaderBar,
}

impl Header {
    #[allow(clippy::new_without_default)]
    pub fn new() -> Self {
        let container = gtk::HeaderBar::new();
        container.set_title(Some("Progress Tracker"));
        container.set_show_close_button(true);

        Self { container }
    }
}

pub struct CompleteView {
    pub container: gtk::Grid,
}

impl CompleteView {
    #[allow(clippy::new_without_default)]
    pub fn new() -> Self {
        let label = gtk::Label::new(None);
        label.set_markup("Task complete");
        label.set_halign(gtk::Align::Center);
        label.set_valign(gtk::Align::Center);
        label.set_vexpand(true);
        label.set_hexpand(true);

        let container = gtk::Grid::new();
        container.set_vexpand(true);
        container.set_hexpand(true);
        container.add(&label);

        Self { container }
    }
}

pub struct MainView {
    pub container: gtk::Grid,
    pub progress: gtk::ProgressBar,
    pub button: gtk::Button,
}

impl MainView {
    #[allow(clippy::new_without_default)]
    pub fn new() -> Self {
        let progress = gtk::ProgressBar::new();
        progress.set_text(Some("Progress Bar"));
        progress.set_show_text(true);
        progress.set_hexpand(true);

        let button = gtk::Button::new();
        button.set_label("start");
        button.set_halign(gtk::Align::Center);

        let container = gtk::Grid::new();
        container.attach(&progress, 0, 0, 1, 1);
        container.attach(&button, 0, 1, 1, 1);
        container.set_row_spacing(12);
        container.set_border_width(6);
        container.set_vexpand(true);
        container.set_hexpand(true);

        Self {
            container,
            progress,
            button,
        }
    }
}

examples/basic_subclass/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12

pub mod simple_application;
pub mod simple_window;

use gtk::prelude::*;

fn main() {
    gtk::init().expect("Failed to initialize gtk");

    let app = simple_application::SimpleApplication::new();

    app.run();
}

examples/accessibility/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52

use gtk::prelude::*;
use gtk::{atk, gio};

fn build_ui(application: &gtk::Application) {
    let window = gtk::ApplicationWindow::new(application);

    window.set_title("Accessibility");
    window.set_position(gtk::WindowPosition::Center);

    let button = gtk::Button::with_label("Click me!");
    let label = gtk::Label::new(Some("0"));
    let vbox = gtk::Box::new(gtk::Orientation::Vertical, 0);

    if let (Some(button_obj), Some(label_obj)) = (button.accessible(), label.accessible()) {
        // We set the description
        button_obj.set_description("Button to increase label value");

        // Then we setup the relation saying that the label is linked to the button.
        let relation_set = label_obj
            .ref_relation_set()
            .expect("Failed to get relation for label");
        let relation = atk::Relation::new(&[button_obj], atk::RelationType::LabelFor);

        relation_set.add(&relation);
    }

    vbox.add(&button);
    vbox.add(&label);

    window.add(&vbox);

    button.connect_clicked(move |_| {
        let value = label.text().parse().unwrap_or(0) + 1;
        label.set_text(&value.to_string());
    });

    window.show_all();
}

fn main() {
    let application = gtk::Application::new(
        Some("com.github.accessibility"),
        gio::ApplicationFlags::empty(),
    );

    application.connect_activate(|app| {
        // We build the application UI.
        build_ui(app);
    });

    application.run();
}

examples/tree_view_model_sort/main.rs
≺ ≻↕

 1
 2
 3
 4
 5
 6
 7
 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59

use gtk::prelude::*;
use gtk::{gio, glib};

fn build_ui(application: &gtk::Application) {
    let window = gtk::ApplicationWindow::new(application);

    window.set_title("Tree Model Sort Window");
    window.set_border_width(10);
    window.set_position(gtk::WindowPosition::Center);
    window.set_default_size(350, 70);

    let store = gtk::TreeStore::new(&[glib::Type::STRING]);
    store.insert_with_values(None, None, &[(0, &"One")]);
    store.insert_with_values(None, None, &[(0, &"Two")]);
    store.insert_with_values(None, None, &[(0, &"Three")]);
    store.insert_with_values(None, None, &[(0, &"Four")]);

    // We create the `TreeModelSort` and we give it the `TreeStore` as
    // parameter.
    let sortable_store = gtk::TreeModelSort::new(&store);

    // Then we create the `TreeView` from the `TreeModelSort`.
    let treeview = gtk::TreeView::with_model(&sortable_store);

    let column = gtk::TreeViewColumn::new();
    column.set_title("Value");
    column.set_clickable(true);
    column.set_sort_indicator(true);
    column.set_sort_column_id(0);

    let renderer = gtk::CellRendererText::new();
    column.pack_end(&renderer, true);
    column.add_attribute(&renderer, "text", 0);

    treeview.append_column(&column);

    treeview.connect_row_activated(move |_, path, _column| {
        let real_path = sortable_store
            .convert_path_to_child_path(&path)
            .expect("Sorted path does not correspond to real path");
        println!(
            "Clicked on sorted: {:?}, real: {:?}",
            path.indices(),
            real_path.indices()
        );
    });

    // We finally add the `TreeView` to the window.
    window.add(&treeview);
    window.show_all();
}

fn main() {
    let application =
        gtk::Application::new(Some("com.github.basic"), gio::ApplicationFlags::empty());

    application.connect_activate(build_ui);
    application.run();
}

pub const fn all() -> ApplicationFlags

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u32) -> Option<ApplicationFlags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> ApplicationFlags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> ApplicationFlags

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: ApplicationFlags) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: ApplicationFlags) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: ApplicationFlags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: ApplicationFlags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: ApplicationFlags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: ApplicationFlags, value: bool)

Inserts or removes the specified flags depending on the passed value.