Struct gio::DBusProxy

#[repr(transparent)]
pub struct DBusProxy { /* private fields */ }

DBusProxy is a base class used for proxies to access a D-Bus interface on a remote object. A DBusProxy can be constructed for both well-known and unique names.

By default, DBusProxy will cache all properties (and listen to changes) of the remote object, and proxy all signals that get emitted. This behaviour can be changed by passing suitable DBusProxyFlags when the proxy is created. If the proxy is for a well-known name, the property cache is flushed when the name owner vanishes and reloaded when a name owner appears.

The unique name owner of the proxy’s name is tracked and can be read from property::DBusProxy::g-name-owner. Connect to the signal::glib::Object::notify signal to get notified of changes. Additionally, only signals and property changes emitted from the current name owner are considered and calls are always sent to the current name owner. This avoids a number of race conditions when the name is lost by one owner and claimed by another. However, if no name owner currently exists, then calls will be sent to the well-known name which may result in the message bus launching an owner (unless DBusProxyFlags::DO_NOT_AUTO_START is set).

If the proxy is for a stateless D-Bus service, where the name owner may be started and stopped between calls, the property::DBusProxy::g-name-owner tracking of DBusProxy will cause the proxy to drop signal and property changes from the service after it has restarted for the first time. When interacting with a stateless D-Bus service, do not use DBusProxy — use direct D-Bus method calls and signal connections.

The generic signal::DBusProxy::g-properties-changed and signal::DBusProxy::g-signal signals are not very convenient to work with. Therefore, the recommended way of working with proxies is to subclass DBusProxy, and have more natural properties and signals in your derived class. This [example][gdbus-example-gdbus-codegen] shows how this can easily be done using the [gdbus-codegen][gdbus-codegen] tool.

A DBusProxy instance can be used from multiple threads but note that all signals (e.g. signal::DBusProxy::g-signal, signal::DBusProxy::g-properties-changed and signal::glib::Object::notify) are emitted in the [thread-default main context][g-main-context-push-thread-default] of the thread where the instance was constructed.

An example using a proxy for a well-known name can be found in gdbus-example-watch-proxy.c

Implements

DBusProxyExt, glib::ObjectExt, AsyncInitableExt, DBusInterfaceExt, InitableExt, DBusProxyExtManual

Implementations

impl DBusProxy

pub const NONE: Option<&'static DBusProxy> = None

pub fn for_bus_sync( bus_type: BusType, flags: DBusProxyFlags, info: Option<&DBusInterfaceInfo>, name: &str, object_path: &str, interface_name: &str, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<DBusProxy, Error>

Like new_sync() but takes a BusType instead of a DBusConnection.

DBusProxy is used in this [example][gdbus-wellknown-proxy].

bus_type

A BusType.

flags

Flags used when constructing the proxy.

info

A DBusInterfaceInfo specifying the minimal interface that proxy conforms to or None.

name

A bus name (well-known or unique).

object_path

An object path.

interface_name

A D-Bus interface name.

cancellable

A Cancellable or None.

Returns

A DBusProxy or None if error is set. Free with g_object_unref().

pub fn new_sync( connection: &DBusConnection, flags: DBusProxyFlags, info: Option<&DBusInterfaceInfo>, name: Option<&str>, object_path: &str, interface_name: &str, cancellable: Option<&impl IsA<Cancellable>> ) -> Result<DBusProxy, Error>

Creates a proxy for accessing interface_name on the remote object at object_path owned by name at connection and synchronously loads D-Bus properties unless the DBusProxyFlags::DO_NOT_LOAD_PROPERTIES flag is used.

If the DBusProxyFlags::DO_NOT_CONNECT_SIGNALS flag is not set, also sets up match rules for signals. Connect to the signal::DBusProxy::g-signal signal to handle signals from the remote object.

If both DBusProxyFlags::DO_NOT_LOAD_PROPERTIES and DBusProxyFlags::DO_NOT_CONNECT_SIGNALS are set, this constructor is guaranteed to return immediately without blocking.

If name is a well-known name and the DBusProxyFlags::DO_NOT_AUTO_START and DBusProxyFlags::DO_NOT_AUTO_START_AT_CONSTRUCTION flags aren’t set and no name owner currently exists, the message bus will be requested to launch a name owner for the name.

This is a synchronous failable constructor. See new() and g_dbus_proxy_new_finish() for the asynchronous version.

DBusProxy is used in this [example][gdbus-wellknown-proxy].

connection

A DBusConnection.

flags

Flags used when constructing the proxy.

info

A DBusInterfaceInfo specifying the minimal interface that proxy conforms to or None.

name

A bus name (well-known or unique) or None if connection is not a message bus connection.

object_path

An object path.

interface_name

A D-Bus interface name.

cancellable

A Cancellable or None.

Returns

A DBusProxy or None if error is set. Free with g_object_unref().

pub fn new<P: FnOnce(Result<DBusProxy, Error>) + 'static>( connection: &DBusConnection, flags: DBusProxyFlags, info: Option<&DBusInterfaceInfo>, name: Option<&str>, object_path: &str, interface_name: &str, cancellable: Option<&impl IsA<Cancellable>>, callback: P )

Creates a proxy for accessing interface_name on the remote object at object_path owned by name at connection and asynchronously loads D-Bus properties unless the DBusProxyFlags::DO_NOT_LOAD_PROPERTIES flag is used. Connect to the signal::DBusProxy::g-properties-changed signal to get notified about property changes.

If the DBusProxyFlags::DO_NOT_CONNECT_SIGNALS flag is not set, also sets up match rules for signals. Connect to the signal::DBusProxy::g-signal signal to handle signals from the remote object.

If both DBusProxyFlags::DO_NOT_LOAD_PROPERTIES and DBusProxyFlags::DO_NOT_CONNECT_SIGNALS are set, this constructor is guaranteed to complete immediately without blocking.

If name is a well-known name and the DBusProxyFlags::DO_NOT_AUTO_START and DBusProxyFlags::DO_NOT_AUTO_START_AT_CONSTRUCTION flags aren’t set and no name owner currently exists, the message bus will be requested to launch a name owner for the name.

This is a failable asynchronous constructor - when the proxy is ready, callback will be invoked and you can use g_dbus_proxy_new_finish() to get the result.

See new_sync() and for a synchronous version of this constructor.

DBusProxy is used in this [example][gdbus-wellknown-proxy].

connection

A DBusConnection.

flags

Flags used when constructing the proxy.

info

A DBusInterfaceInfo specifying the minimal interface that proxy conforms to or None.

name

A bus name (well-known or unique) or None if connection is not a message bus connection.

object_path

An object path.

interface_name

A D-Bus interface name.

cancellable

A Cancellable or None.

callback

Callback function to invoke when the proxy is ready.

pub fn new_future( connection: &DBusConnection, flags: DBusProxyFlags, info: Option<&DBusInterfaceInfo>, name: Option<&str>, object_path: &str, interface_name: &str ) -> Pin<Box_<dyn Future<Output = Result<DBusProxy, Error>> + 'static>>

pub fn for_bus<P: FnOnce(Result<DBusProxy, Error>) + 'static>( bus_type: BusType, flags: DBusProxyFlags, info: Option<&DBusInterfaceInfo>, name: &str, object_path: &str, interface_name: &str, cancellable: Option<&impl IsA<Cancellable>>, callback: P )

Like new() but takes a BusType instead of a DBusConnection.

DBusProxy is used in this [example][gdbus-wellknown-proxy].

bus_type

A BusType.

flags

Flags used when constructing the proxy.

info

A DBusInterfaceInfo specifying the minimal interface that proxy conforms to or None.

name

A bus name (well-known or unique).

object_path

An object path.

interface_name

A D-Bus interface name.

cancellable

A Cancellable or None.

callback

Callback function to invoke when the proxy is ready.

pub fn for_bus_future( bus_type: BusType, flags: DBusProxyFlags, info: Option<&DBusInterfaceInfo>, name: &str, object_path: &str, interface_name: &str ) -> Pin<Box_<dyn Future<Output = Result<DBusProxy, Error>> + 'static>>