gio

Struct DesktopAppInfo

pub struct DesktopAppInfo { /* private fields */ }

Available on non-Windows and non-macOS only.

GDesktopAppInfo is an implementation of AppInfo based on desktop files.

Note that <gio/gdesktopappinfo.h> belongs to the UNIX-specific GIO interfaces, thus you have to use the gio-unix-2.0.pc pkg-config file or the GioUnix-2.0 GIR namespace when using it.

Properties

filename

The origin filename of this DesktopAppInfo

Readable | Writeable | Construct Only

Implements

[trait@glib::ObjectExt], AppInfoExt, DesktopAppInfoExtManual, AppInfoExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl DesktopAppInfo

pub fn search(search_string: &str) -> Vec<Vec<GString>>

Searches desktop files for ones that match @search_string.

The return value is an array of strvs. Each strv contains a list of applications that matched @search_string with an equal score. The outer list is sorted by score so that the first strv contains the best-matching applications, and so on. The algorithm for determining matches is undefined and may change at any time.

None of the search results are subjected to the normal validation checks performed by new() (for example, checking that the executable referenced by a result exists), and so it is possible for new() to return NULL when passed an app ID returned by this function. It is expected that calling code will do this when subsequently creating a DesktopAppInfo for each result.

search_string

the search string to use

Returns

a list of strvs. Free each item with strfreev() and free the outer list with free().

impl DesktopAppInfo

pub fn new(desktop_id: &str) -> Option<DesktopAppInfo>

Creates a new DesktopAppInfo based on a desktop file ID.

A desktop file ID is the basename of the desktop file, including the .desktop extension. GIO is looking for a desktop file with this name in the applications subdirectories of the XDG data directories (i.e. the directories specified in the XDG_DATA_HOME and XDG_DATA_DIRS environment variables). GIO also supports the prefix-to-subdirectory mapping that is described in the Menu Spec (i.e. a desktop ID of kde-foo.desktop will match /usr/share/applications/kde/foo.desktop).

desktop_id

the desktop file ID

Returns

a new DesktopAppInfo, or NULL if no desktop file with that ID exists.

pub fn from_filename(filename: impl AsRef<Path>) -> Option<DesktopAppInfo>

Creates a new DesktopAppInfo.

filename

the path of a desktop file, in the GLib filename encoding

Returns

a new DesktopAppInfo or NULL on error.

pub fn from_keyfile(key_file: &KeyFile) -> Option<DesktopAppInfo>

Creates a new DesktopAppInfo.

key_file

an opened [type@GLib.KeyFile]

Returns

a new DesktopAppInfo or NULL on error.

pub fn action_name(&self, action_name: &str) -> GString

Gets the user-visible display name of the ‘additional application actions’ specified by @action_name.

This corresponds to the Name key within the keyfile group for the action.

action_name

the name of the action as from list_actions()

Returns

the locale-specific action name

pub fn boolean(&self, key: &str) -> bool

Looks up a boolean value in the keyfile backing @self.

The @key is looked up in the Desktop Entry group.

key

the key to look up

Returns

the boolean value, or FALSE if the key is not found

pub fn categories(&self) -> Option<GString>

Gets the categories from the desktop file.

Returns

The unparsed Categories key from the desktop file; i.e. no attempt is made to split it by ; or validate it.

pub fn filename(&self) -> Option<PathBuf>

When @self was created from a known filename, return it. In some situations such as a DesktopAppInfo returned from from_keyfile(), this function will return NULL.

Returns

The full path to the file for @self, or NULL if not known.

pub fn generic_name(&self) -> Option<GString>

Gets the generic name from the desktop file.

Returns

The value of the GenericName key

pub fn is_hidden(&self) -> bool

A desktop file is hidden if the Hidden key in it is set to True.

Returns

TRUE if hidden, FALSE otherwise.

pub fn keywords(&self) -> Vec<GString>

Gets the keywords from the desktop file.

Returns

The value of the Keywords key

pub fn locale_string(&self, key: &str) -> Option<GString>

Looks up a localized string value in the keyfile backing @self translated to the current locale.

The @key is looked up in the Desktop Entry group.

key

the key to look up

Returns

a newly allocated string, or NULL if the key is not found

pub fn is_nodisplay(&self) -> bool

Gets the value of the NoDisplay key which helps determine if the application info should be shown in menus. See G_KEY_FILE_DESKTOP_KEY_NO_DISPLAY and AppInfoExt::should_show().

Returns

The value of the NoDisplay key

pub fn shows_in(&self, desktop_env: Option<&str>) -> bool

Checks if the application info should be shown in menus that list available applications for a specific name of the desktop, based on the OnlyShowIn and NotShowIn keys.

@desktop_env should typically be given as NULL, in which case the XDG_CURRENT_DESKTOP environment variable is consulted. If you want to override the default mechanism then you may specify @desktop_env, but this is not recommended.

Note that AppInfoExt::should_show() for @self will include this check (with NULL for @desktop_env) as well as additional checks.

desktop_env

a string specifying a desktop name

Returns

TRUE if the @self should be shown in @desktop_env according to the OnlyShowIn and NotShowIn keys, FALSE otherwise.

pub fn startup_wm_class(&self) -> Option<GString>

Retrieves the StartupWMClass field from @self. This represents the WM_CLASS property of the main window of the application, if launched through @self.

Returns

the startup WM class, or NULL if none is set in the desktop file.

pub fn string(&self, key: &str) -> Option<GString>

Looks up a string value in the keyfile backing @self.

The @key is looked up in the Desktop Entry group.

key

the key to look up

Returns

a newly allocated string, or NULL if the key is not found

pub fn string_list(&self, key: &str) -> Vec<GString>

Available on crate feature v2_60 only.

Looks up a string list value in the keyfile backing @self.

The @key is looked up in the Desktop Entry group.

key

the key to look up

Returns

a NULL-terminated string array or NULL if the specified key cannot be found. The array should be freed with strfreev().

pub fn has_key(&self, key: &str) -> bool

Returns whether @key exists in the Desktop Entry group of the keyfile backing @self.

key

the key to look up

Returns

TRUE if the @key exists

pub fn launch_action( &self, action_name: &str, launch_context: Option<&impl IsA<AppLaunchContext>>, )

Activates the named application action.

You may only call this function on action names that were returned from list_actions().

Note that if the main entry of the desktop file indicates that the application supports startup notification, and @launch_context is non-NULL, then startup notification will be used when activating the action (and as such, invocation of the action on the receiving side must signal the end of startup notification when it is completed). This is the expected behaviour of applications declaring additional actions, as per the desktop file specification.

As with AppInfoExt::launch() there is no way to detect failures that occur while using this function.

action_name

the name of the action as from list_actions()

launch_context

a AppLaunchContext

pub fn launch_uris_as_manager( &self, uris: &[&str], launch_context: Option<&impl IsA<AppLaunchContext>>, spawn_flags: SpawnFlags, user_setup: Option<Box_<dyn FnOnce() + 'static>>, pid_callback: Option<&mut dyn FnMut(&DesktopAppInfo, Pid)>, ) -> Result<(), Error>

This function performs the equivalent of AppInfoExt::launch_uris(), but is intended primarily for operating system components that launch applications. Ordinary applications should use AppInfoExt::launch_uris().

If the application is launched via GSpawn, then @spawn_flags, @user_setup and @user_setup_data are used for the call to spawn_async(). Additionally, @pid_callback (with @pid_callback_data) will be called to inform about the PID of the created process. See spawn_async_with_pipes() for information on certain parameter conditions that can enable an optimized posix_spawn() code path to be used.

If application launching occurs via some other mechanism (for example, D-Bus activation) then @spawn_flags, @user_setup, @user_setup_data, @pid_callback and @pid_callback_data are ignored.

uris

List of URIs

launch_context

a AppLaunchContext

spawn_flags

glib::SpawnFlags, used for each process

user_setup

a `callback::GLib::SpawnChildSetupFunc, used once for each process.

user_setup_data

User data for @user_setup

pid_callback

Callback for child processes

pid_callback_data

User data for @callback

Returns

TRUE on successful launch, FALSE otherwise.

pub fn list_actions(&self) -> Vec<GString>

Returns the list of ‘additional application actions’ supported on the desktop file, as per the desktop file specification.

As per the specification, this is the list of actions that are explicitly listed in the Actions key of the Desktop Entry group.

Returns

a list of strings, always non-NULL

pub fn implementations(interface: &str) -> Vec<DesktopAppInfo>

Gets all applications that implement @interface.

An application implements an interface if that interface is listed in the Implements line of the desktop file of the application.

interface

the name of the interface

Returns

a list of DesktopAppInfo objects.