gio
Struct File

pub struct File { /* private fields */ }

GFile is a high level abstraction for manipulating files on a virtual file system. GFiles are lightweight, immutable objects that do no I/O upon creation. It is necessary to understand that GFile objects do not represent files, merely an identifier for a file. All file content I/O is implemented as streaming operations (see InputStream and OutputStream).

To construct a GFile, you can use:

for_path() if you have a path.

for_uri() if you have a URI.

for_commandline_arg() or for_commandline_arg_and_cwd() for a command line argument.

new_tmp() to create a temporary file from a template.

new_tmp_async() to asynchronously create a temporary file.

new_tmp_dir_async() to asynchronously create a temporary directory.

for_parse_name() from a UTF-8 string gotten from FileExt::parse_name().

Gio::File::new_build_filename() or new_build_filenamev() to create a file from path elements.

One way to think of a GFile is as an abstraction of a pathname. For normal files the system pathname is what is stored internally, but as GFiles are extensible it could also be something else that corresponds to a pathname in a userspace implementation of a filesystem.

GFiles make up hierarchies of directories and files that correspond to the files on a filesystem. You can move through the file system with GFile using FileExt::parent() to get an identifier for the parent directory, FileExt::child() to get a child within a directory, and FileExt::resolve_relative_path() to resolve a relative path between two GFiles. There can be multiple hierarchies, so you may not end up at the same root if you repeatedly call FileExt::parent() on two different files.

All GFiles have a basename (get with FileExt::basename()). These names are byte strings that are used to identify the file on the filesystem (relative to its parent directory) and there is no guarantees that they have any particular charset encoding or even make any sense at all. If you want to use filenames in a user interface you should use the display name that you can get by requesting the G_FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME attribute with FileExt::query_info(). This is guaranteed to be in UTF-8 and can be used in a user interface. But always store the real basename or the GFile to use to actually access the file, because there is no way to go from a display name to the actual name.

Using GFile as an identifier has the same weaknesses as using a path in that there may be multiple aliases for the same file. For instance, hard or soft links may cause two different GFiles to refer to the same file. Other possible causes for aliases are: case insensitive filesystems, short and long names on FAT/NTFS, or bind mounts in Linux. If you want to check if two GFiles point to the same file you can query for the G_FILE_ATTRIBUTE_ID_FILE attribute. Note that GFile does some trivial canonicalization of pathnames passed in, so that trivial differences in the path string used at creation (duplicated slashes, slash at end of path, . or .. path segments, etc) does not create different GFiles.

Many GFile operations have both synchronous and asynchronous versions to suit your application. Asynchronous versions of synchronous functions simply have _async() appended to their function names. The asynchronous I/O functions call a callback::Gio::AsyncReadyCallback which is then used to finalize the operation, producing a [AsyncResult][crate::AsyncResult] which is then passed to the function’s matching _finish()` operation.

It is highly recommended to use asynchronous calls when running within a shared main loop, such as in the main thread of an application. This avoids I/O operations blocking other sources on the main loop from being dispatched. Synchronous I/O operations should be performed from worker threads. See the introduction to asynchronous programming section for more.

Some GFile operations almost always take a noticeable amount of time, and so do not have synchronous analogs. Notable cases include:

FileExt::mount_mountable() to mount a mountable file.

FileExt::unmount_mountable_with_operation() to unmount a mountable file.

FileExt::eject_mountable_with_operation() to eject a mountable file.

Entity Tags

One notable feature of GFiles are entity tags, or ‘etags’ for short. Entity tags are somewhat like a more abstract version of the traditional mtime, and can be used to quickly determine if the file has been modified from the version on the file system. See the HTTP 1.1 specification for HTTP ETag headers, which are a very similar concept.

Implements

FileExt, FileExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl File

pub fn new_tmp_async<P: FnOnce(Result<(File, FileIOStream), Error>) + 'static>( tmpl: Option<impl AsRef<Path>>, io_priority: Priority, cancellable: Option<&impl IsA<Cancellable>>, callback: P, )

Available on crate feature v2_74 only.

Asynchronously opens a file in the preferred directory for temporary files (as returned by g_get_tmp_dir()) as g_file_new_tmp().

@tmpl should be a string in the GLib file name encoding containing a sequence of six ‘X’ characters, and containing no directory components. If it is None, a default template is used.

tmpl

Template for the file name, as in g_file_open_tmp(), or None for a default template

io_priority

the I/O priority of the request

cancellable

optional #GCancellable object, None to ignore

callback

a #GAsyncReadyCallback to call when the request is done

pub fn new_tmp_future( tmpl: Option<impl AsRef<Path>>, io_priority: Priority, ) -> Pin<Box<dyn Future<Output = Result<(File, FileIOStream), Error>> + 'static>>

Available on crate feature v2_74 only.

pub fn new_tmp_dir_async<P: FnOnce(Result<File, Error>) + 'static>( tmpl: Option<impl AsRef<Path>>, io_priority: Priority, cancellable: Option<&impl IsA<Cancellable>>, callback: P, )

Available on crate feature v2_74 only.

Asynchronously creates a directory in the preferred directory for temporary files (as returned by g_get_tmp_dir()) as g_dir_make_tmp().

@tmpl should be a string in the GLib file name encoding containing a sequence of six ‘X’ characters, and containing no directory components. If it is None, a default template is used.

tmpl

Template for the file name, as in g_dir_make_tmp(), or None for a default template

io_priority

the I/O priority of the request

cancellable

optional #GCancellable object, None to ignore

callback

a #GAsyncReadyCallback to call when the request is done

pub fn new_tmp_dir_future( tmpl: Option<impl AsRef<Path>>, io_priority: Priority, ) -> Pin<Box<dyn Future<Output = Result<File, Error>> + 'static>>

Available on crate feature v2_74 only.

impl File

pub const NONE: Option<&'static File> = None

pub fn new_build_filenamev(args: &[&Path]) -> File

Available on crate feature v2_78 only.

Constructs a #GFile from a vector of elements using the correct separator for filenames.

Using this function is equivalent to calling g_build_filenamev(), followed by g_file_new_for_path() on the result.

args

None-terminated array of strings containing the path elements.

Returns

a new #GFile

pub fn for_commandline_arg(arg: impl AsRef<OsStr>) -> File

Creates a #GFile with the given argument from the command line. The value of @arg can be either a URI, an absolute path or a relative path resolved relative to the current working directory. This operation never fails, but the returned object might not support any I/O operation if @arg points to a malformed path.

Note that on Windows, this function expects its argument to be in UTF-8 – not the system code page. This means that you should not use this function with string from argv as it is passed to main(). g_win32_get_command_line() will return a UTF-8 version of the commandline. #GApplication also uses UTF-8 but g_application_command_line_create_file_for_arg() may be more useful for you there. It is also always possible to use this function with #GOptionContext arguments of type glib::OptionArg::Filename.

arg

a command line string

Returns

a new #GFile. Free the returned object with g_object_unref().

pub fn for_commandline_arg_and_cwd( arg: impl AsRef<OsStr>, cwd: impl AsRef<Path>, ) -> File

Creates a #GFile with the given argument from the command line.

This function is similar to g_file_new_for_commandline_arg() except that it allows for passing the current working directory as an argument instead of using the current working directory of the process.

This is useful if the commandline argument was given in a context other than the invocation of the current process.

See also g_application_command_line_create_file_for_arg().

arg

a command line string

cwd

the current working directory of the commandline

Returns

a new #GFile

pub fn for_path(path: impl AsRef<Path>) -> File

Constructs a #GFile for a given path. This operation never fails, but the returned object might not support any I/O operation if @path is malformed.

path

a string containing a relative or absolute path. The string must be encoded in the glib filename encoding.

Returns

a new #GFile for the given @path. Free the returned object with g_object_unref().

pub fn for_uri(uri: &str) -> File

Constructs a #GFile for a given URI. This operation never fails, but the returned object might not support any I/O operation if @uri is malformed or if the uri type is not supported.

uri

a UTF-8 string containing a URI

Returns

a new #GFile for the given @uri. Free the returned object with g_object_unref().

pub fn new_tmp( tmpl: Option<impl AsRef<Path>>, ) -> Result<(File, FileIOStream), Error>

Opens a file in the preferred directory for temporary files (as returned by g_get_tmp_dir()) and returns a #GFile and #GFileIOStream pointing to it.

@tmpl should be a string in the GLib file name encoding containing a sequence of six ‘X’ characters, and containing no directory components. If it is None, a default template is used.

Unlike the other #GFile constructors, this will return None if a temporary file could not be created.

tmpl

Template for the file name, as in g_file_open_tmp(), or None for a default template

Returns

a new #GFile. Free the returned object with g_object_unref().

iostream

on return, a #GFileIOStream for the created file

pub fn for_parse_name(parse_name: &str) -> File

Constructs a #GFile with the given @parse_name (i.e. something given by g_file_get_parse_name()). This operation never fails, but the returned object might not support any I/O operation if the @parse_name cannot be parsed.

parse_name

a file name or path to be parsed

Returns

a new #GFile.