Struct gio::FileEnumerator

pub struct FileEnumerator(_);

FileEnumerator allows you to operate on a set of GFiles, returning a FileInfo structure for each file enumerated (e.g. FileExt::enumerate_children() will return a FileEnumerator for each of the children within a directory).

To get the next file’s information from a FileEnumerator, use FileEnumeratorExt::next_file() or its asynchronous version, FileEnumeratorExt::next_files_async(). Note that the asynchronous version will return a list of GFileInfos, whereas the synchronous will only return the next file in the enumerator.

The ordering of returned files is unspecified for non-Unix platforms; for more information, see g_dir_read_name(). On Unix, when operating on local files, returned files will be sorted by inode number. Effectively you can assume that the ordering of returned files will be stable between successive calls (and applications) assuming the directory is unchanged.

If your application needs a specific ordering, such as by name or modification time, you will have to implement that in your application code.

To close a FileEnumerator, use FileEnumeratorExt::close(), or its asynchronous version, FileEnumeratorExt::close_async(). Once a FileEnumerator is closed, no further actions may be performed on it, and it should be freed with g_object_unref().

Implements

FileEnumeratorExt, glib::ObjectExt