gio
Struct Settings

pub struct Settings { /* private fields */ }

The GSettings class provides a convenient API for storing and retrieving application settings.

Reads and writes can be considered to be non-blocking. Reading settings with GSettings is typically extremely fast: on approximately the same order of magnitude (but slower than) a GLib::HashTable lookup. Writing settings is also extremely fast in terms of time to return to your application, but can be extremely expensive for other threads and other processes. Many settings backends (including dconf) have lazy initialisation which means in the common case of the user using their computer without modifying any settings a lot of work can be avoided. For dconf, the D-Bus service doesn’t even need to be started in this case. For this reason, you should only ever modify GSettings keys in response to explicit user action. Particular care should be paid to ensure that modifications are not made during startup — for example, when setting the initial value of preferences widgets. The built-in SettingsExtManual::bind() functionality is careful not to write settings in response to notify signals as a result of modifications that it makes to widgets.

When creating a GSettings instance, you have to specify a schema that describes the keys in your settings and their types and default values, as well as some other information.

Normally, a schema has a fixed path that determines where the settings are stored in the conceptual global tree of settings. However, schemas can also be ‘relocatable’, i.e. not equipped with a fixed path. This is useful e.g. when the schema describes an ‘account’, and you want to be able to store a arbitrary number of accounts.

Paths must start with and end with a forward slash character (/) and must not contain two sequential slash characters. Paths should be chosen based on a domain name associated with the program or library to which the settings belong. Examples of paths are /org/gtk/settings/file-chooser/ and /ca/desrt/dconf-editor/. Paths should not start with /apps/, /desktop/ or /system/ as they often did in GConf.

Unlike other configuration systems (like GConf), GSettings does not restrict keys to basic types like strings and numbers. GSettings stores values as glib::Variant, and allows any [type@GLib.VariantType] for keys. Key names are restricted to lowercase characters, numbers and -. Furthermore, the names must begin with a lowercase character, must not end with a -, and must not contain consecutive dashes.

Similar to GConf, the default values in GSettings schemas can be localized, but the localized values are stored in gettext catalogs and looked up with the domain that is specified in the gettext-domain attribute of the <schemalist> or <schema> elements and the category that is specified in the l10n attribute of the <default> element. The string which is translated includes all text in the <default> element, including any surrounding quotation marks.

The l10n attribute must be set to messages or time, and sets the locale category for translation. The messages category should be used by default; use time for translatable date or time formats. A translation comment can be added as an XML comment immediately above the <default> element — it is recommended to add these comments to aid translators understand the meaning and implications of the default value. An optional translation context attribute can be set on the <default> element to disambiguate multiple defaults which use the same string.

For example:

 <!-- Translators: A list of words which are not allowed to be typed, in
      GVariant serialization syntax.
      See: https://developer.gnome.org/glib/stable/gvariant-text.html -->
 <default l10n='messages' context='Banned words'>['bad', 'words']</default>

Translations of default values must remain syntactically valid serialized glib::Variants (e.g. retaining any surrounding quotation marks) or runtime errors will occur.

GSettings uses schemas in a compact binary form that is created by the glib-compile-schemas utility. The input is a schema description in an XML format.

A DTD for the gschema XML format can be found here: gschema.dtd

The glib-compile-schemas tool expects schema files to have the extension .gschema.xml.

At runtime, schemas are identified by their ID (as specified in the id attribute of the <schema> element). The convention for schema IDs is to use a dotted name, similar in style to a D-Bus bus name, e.g. org.gnome.SessionManager. In particular, if the settings are for a specific service that owns a D-Bus bus name, the D-Bus bus name and schema ID should match. For schemas which deal with settings not associated with one named application, the ID should not use StudlyCaps, e.g. org.gnome.font-rendering.

In addition to glib::Variant types, keys can have types that have enumerated types. These can be described by a <choice>, <enum> or <flags> element, as seen in the second example below. The underlying type of such a key is string, but you can use SettingsExt::enum_(), SettingsExt::set_enum(), SettingsExt::flags(), SettingsExt::set_flags() access the numeric values corresponding to the string value of enum and flags keys.

An example for default value:

<schemalist>
  <schema id="org.gtk.Test" path="/org/gtk/Test/" gettext-domain="test">

    <key name="greeting" type="s">
      <default l10n="messages">"Hello, earthlings"</default>
      <summary>A greeting</summary>
      <description>
        Greeting of the invading martians
      </description>
    </key>

    <key name="box" type="(ii)">
      <default>(20,30)</default>
    </key>

    <key name="empty-string" type="s">
      <default>""</default>
      <summary>Empty strings have to be provided in GVariant form</summary>
    </key>

  </schema>
</schemalist>

An example for ranges, choices and enumerated types:

<schemalist>

  <enum id="org.gtk.Test.myenum">
    <value nick="first" value="1"/>
    <value nick="second" value="2"/>
  </enum>

  <flags id="org.gtk.Test.myflags">
    <value nick="flag1" value="1"/>
    <value nick="flag2" value="2"/>
    <value nick="flag3" value="4"/>
  </flags>

  <schema id="org.gtk.Test">

    <key name="key-with-range" type="i">
      <range min="1" max="100"/>
      <default>10</default>
    </key>

    <key name="key-with-choices" type="s">
      <choices>
        <choice value='Elisabeth'/>
        <choice value='Annabeth'/>
        <choice value='Joe'/>
      </choices>
      <aliases>
        <alias value='Anna' target='Annabeth'/>
        <alias value='Beth' target='Elisabeth'/>
      </aliases>
      <default>'Joe'</default>
    </key>

    <key name='enumerated-key' enum='org.gtk.Test.myenum'>
      <default>'first'</default>
    </key>

    <key name='flags-key' flags='org.gtk.Test.myflags'>
      <default>["flag1","flag2"]</default>
    </key>
  </schema>
</schemalist>

Vendor overrides

Default values are defined in the schemas that get installed by an application. Sometimes, it is necessary for a vendor or distributor to adjust these defaults. Since patching the XML source for the schema is inconvenient and error-prone, glib-compile-schemas reads so-called ‘vendor override’ files. These are keyfiles in the same directory as the XML schema sources which can override default values. The schema ID serves as the group name in the key file, and the values are expected in serialized glib::Variant form, as in the following example:

[org.gtk.Example]
key1='string'
key2=1.5

glib-compile-schemas expects schema files to have the extension .gschema.override.

Binding

A very convenient feature of GSettings lets you bind glib::Object properties directly to settings, using SettingsExtManual::bind(). Once a glib::Object property has been bound to a setting, changes on either side are automatically propagated to the other side. GSettings handles details like mapping between glib::Object and glib::Variant types, and preventing infinite cycles.

This makes it very easy to hook up a preferences dialog to the underlying settings. To make this even more convenient, GSettings looks for a boolean property with the name sensitivity and automatically binds it to the writability of the bound setting. If this ‘magic’ gets in the way, it can be suppressed with the G_SETTINGS_BIND_NO_SENSITIVITY flag.

Relocatable schemas

A relocatable schema is one with no path attribute specified on its <schema> element. By using with_path(), a GSettings object can be instantiated for a relocatable schema, assigning a path to the instance. Paths passed to with_path() will typically be constructed dynamically from a constant prefix plus some form of instance identifier; but they must still be valid GSettings paths. Paths could also be constant and used with a globally installed schema originating from a dependency library.

For example, a relocatable schema could be used to store geometry information for different windows in an application. If the schema ID was org.foo.MyApp.Window, it could be instantiated for paths /org/foo/MyApp/main/, /org/foo/MyApp/document-1/, /org/foo/MyApp/document-2/, etc. If any of the paths are well-known they can be specified as <child> elements in the parent schema, e.g.:

<schema id="org.foo.MyApp" path="/org/foo/MyApp/">
  <child name="main" schema="org.foo.MyApp.Window"/>
</schema>

Build system integration

Meson

GSettings is natively supported by Meson’s GNOME module.

You can install the schemas as any other data file:

install_data(
  'org.foo.MyApp.gschema.xml',
  install_dir: get_option('datadir') / 'glib-2.0/schemas',
)

You can use gnome.post_install() function to compile the schemas on installation:

gnome = import('gnome')
gnome.post_install(
  glib_compile_schemas: true,
)

If an enumerated type defined in a C header file is to be used in a GSettings schema, it can either be defined manually using an <enum> element in the schema XML, or it can be extracted automatically from the C header. This approach is preferred, as it ensures the two representations are always synchronised. To do so, you will need to use the gnome.mkenums() function with the following templates:

schemas_enums = gnome.mkenums('org.foo.MyApp.enums.xml',
  comments: '<!-- @comment@ -->',
  fhead: '<schemalist>',
  vhead: '  <@type@ id="org.foo.MyApp.@EnumName@">',
  vprod: '    <value nick="@valuenick@" value="@valuenum@"/>',
  vtail: '  </@type@>',
  ftail: '</schemalist>',
  sources: enum_sources,
  install_header: true,
  install_dir: get_option('datadir') / 'glib-2.0/schemas',
)

It is recommended to validate your schemas as part of the test suite for your application:

test('validate-schema',
  find_program('glib-compile-schemas'),
  args: ['--strict', '--dry-run', meson.current_source_dir()],
)

If your application allows running uninstalled, you should also use the gnome.compile_schemas() function to compile the schemas in the current build directory:

gnome.compile_schemas()

Autotools

GSettings comes with autotools integration to simplify compiling and installing schemas. To add GSettings support to an application, add the following to your configure.ac:

GLIB_GSETTINGS

In the appropriate Makefile.am, use the following snippet to compile and install the named schema:

gsettings_SCHEMAS = org.foo.MyApp.gschema.xml
EXTRA_DIST = $(gsettings_SCHEMAS)

@GSETTINGS_RULES@

If an enumerated type defined in a C header file is to be used in a GSettings schema, it can either be defined manually using an <enum> element in the schema XML, or it can be extracted automatically from the C header. This approach is preferred, as it ensures the two representations are always synchronised. To do so, add the following to the relevant Makefile.am:

gsettings_ENUM_NAMESPACE = org.foo.MyApp
gsettings_ENUM_FILES = my-app-enums.h my-app-misc.h

gsettings_ENUM_NAMESPACE specifies the schema namespace for the enum files, which are specified in gsettings_ENUM_FILES. This will generate a org.foo.MyApp.enums.xml file containing the extracted enums, which will be automatically included in the schema compilation, install and uninstall rules. It should not be committed to version control or included in EXTRA_DIST.

Localization

No changes are needed to the build system to mark a schema XML file for translation. Assuming it sets the gettext-domain attribute, a schema may be marked for translation by adding it to POTFILES.in, assuming gettext 0.19 or newer is in use (the preferred method for translation):

data/org.foo.MyApp.gschema.xml

Alternatively, if intltool 0.50.1 is in use:

[type: gettext/gsettings]data/org.foo.MyApp.gschema.xml

GSettings will use gettext to look up translations for the <summary> and <description> elements, and also any <default> elements which have a l10n attribute set.

Translations must not be included in the .gschema.xml file by the build system, for example by using a rule to generate the XML file from a template.

Properties

backend

The name of the context that the settings are stored in.

Readable | Writeable | Construct Only

delay-apply

Whether the #GSettings object is in ‘delay-apply’ mode. See g_settings_delay() for details.

Readable

has-unapplied

If this property is true, the #GSettings object has outstanding changes that will be applied when g_settings_apply() is called.

Readable

path

The path within the backend where the settings are stored.

Readable | Writeable | Construct Only

schema

The name of the schema that describes the types of keys for this #GSettings object.

The type of this property is not #GSettingsSchema. #GSettingsSchema has only existed since version 2.32 and unfortunately this name was used in previous versions to refer to the schema ID rather than the schema itself. Take care to use the ‘settings-schema’ property if you wish to pass in a #GSettingsSchema.

Readable | Writeable | Construct Only

schema-id

The name of the schema that describes the types of keys for this #GSettings object.

Readable | Writeable | Construct Only

settings-schema

The #GSettingsSchema describing the types of keys for this #GSettings object.

Ideally, this property would be called ‘schema’. #GSettingsSchema has only existed since version 2.32, however, and before then the ‘schema’ property was used to refer to the ID of the schema rather than the schema itself. Take care.

Readable | Writeable | Construct Only

Signals

change-event

The “change-event” signal is emitted once per change event that affects this settings object. You should connect to this signal only if you are interested in viewing groups of changes before they are split out into multiple emissions of the “changed” signal. For most use cases it is more appropriate to use the “changed” signal.

In the event that the change event applies to one or more specified keys, @keys will be an array of #GQuark of length @n_keys. In the event that the change event applies to the #GSettings object as a whole (ie: potentially every key has been changed) then @keys will be None and @n_keys will be 0.

The default handler for this signal invokes the “changed” signal for each affected key. If any other connected handler returns true then this default functionality will be suppressed.

changed

The “changed” signal is emitted when a key has potentially changed. You should call one of the g_settings_get() calls to check the new value.

This signal supports detailed connections. You can connect to the detailed signal “changed::x” in order to only receive callbacks when key “x” changes.

Note that @settings only emits this signal if you have read @key at least once while a signal handler was already connected for @key.

Detailed

writable-change-event

The “writable-change-event” signal is emitted once per writability change event that affects this settings object. You should connect to this signal if you are interested in viewing groups of changes before they are split out into multiple emissions of the “writable-changed” signal. For most use cases it is more appropriate to use the “writable-changed” signal.

In the event that the writability change applies only to a single key, @key will be set to the #GQuark for that key. In the event that the writability change affects the entire settings object, @key will be 0.

The default handler for this signal invokes the “writable-changed” and “changed” signals for each affected key. This is done because changes in writability might also imply changes in value (if for example, a new mandatory setting is introduced). If any other connected handler returns true then this default functionality will be suppressed.

writable-changed

The “writable-changed” signal is emitted when the writability of a key has potentially changed. You should call g_settings_is_writable() in order to determine the new status.

This signal supports detailed connections. You can connect to the detailed signal “writable-changed::x” in order to only receive callbacks when the writability of “x” changes.

Detailed

Implements

SettingsExt, [trait@glib::ObjectExt], SettingsExtManual

GLib type: GObject with reference counted clone semantics.

Implementations

impl Settings

pub const NONE: Option<&'static Settings> = None

pub fn new(schema_id: &str) -> Settings

Creates a new #GSettings object with the schema specified by @schema_id.

It is an error for the schema to not exist: schemas are an essential part of a program, as they provide type information. If schemas need to be dynamically loaded (for example, from an optional runtime dependency), g_settings_schema_source_lookup() can be used to test for their existence before loading them.

Signals on the newly created #GSettings object will be dispatched via the thread-default #GMainContext in effect at the time of the call to g_settings_new(). The new #GSettings will hold a reference on the context. See g_main_context_push_thread_default().

schema_id

the id of the schema

Returns

a new #GSettings object

pub fn new_full( schema: &SettingsSchema, backend: Option<&impl IsA<SettingsBackend>>, path: Option<&str>, ) -> Settings

Creates a new #GSettings object with a given schema, backend and path.

It should be extremely rare that you ever want to use this function. It is made available for advanced use-cases (such as plugin systems that want to provide access to schemas loaded from custom locations, etc).

At the most basic level, a #GSettings object is a pure composition of 4 things: a #GSettingsSchema, a #GSettingsBackend, a path within that backend, and a #GMainContext to which signals are dispatched.

This constructor therefore gives you full control over constructing #GSettings instances. The first 3 parameters are given directly as @schema, @backend and @path, and the main context is taken from the thread-default (as per g_settings_new()).

If @backend is None then the default backend is used.

If @path is None then the path from the schema is used. It is an error if @path is None and the schema has no path of its own or if @path is non-None and not equal to the path that the schema does have.

schema

a #GSettingsSchema

backend

a #GSettingsBackend

path

the path to use

Returns

a new #GSettings object

pub fn with_backend( schema_id: &str, backend: &impl IsA<SettingsBackend>, ) -> Settings

Creates a new #GSettings object with the schema specified by @schema_id and a given #GSettingsBackend.

Creating a #GSettings object with a different backend allows accessing settings from a database other than the usual one. For example, it may make sense to pass a backend corresponding to the “defaults” settings database on the system to get a settings object that modifies the system default settings instead of the settings for this user.

schema_id

the id of the schema

backend

the #GSettingsBackend to use

Returns

a new #GSettings object

pub fn with_backend_and_path( schema_id: &str, backend: &impl IsA<SettingsBackend>, path: &str, ) -> Settings

Creates a new #GSettings object with the schema specified by @schema_id and a given #GSettingsBackend and path.

This is a mix of g_settings_new_with_backend() and g_settings_new_with_path().

schema_id

the id of the schema

backend

the #GSettingsBackend to use

path

the path to use

Returns

a new #GSettings object

pub fn with_path(schema_id: &str, path: &str) -> Settings

Creates a new #GSettings object with the relocatable schema specified by @schema_id and a given path.

You only need to do this if you want to directly create a settings object with a schema that doesn’t have a specified path of its own. That’s quite rare.

It is a programmer error to call this function for a schema that has an explicitly specified path.

It is a programmer error if @path is not a valid path. A valid path begins and ends with ‘/’ and does not contain two consecutive ‘/’ characters.

schema_id

the id of the schema

path

the path to use

Returns

a new #GSettings object

pub fn sync()

Ensures that all pending operations are complete for the default backend.

Writes made to a #GSettings are handled asynchronously. For this reason, it is very unlikely that the changes have it to disk by the time g_settings_set() returns.

This call will block until all of the writes have made it to the backend. Since the mainloop is not running, no change notifications will be dispatched during this call (but some may be queued by the time the call is done).

pub fn unbind(object: &impl IsA<Object>, property: &str)

Removes an existing binding for @property on @object.

Note that bindings are automatically removed when the object is finalized, so it is rarely necessary to call this function.

object

the object

property

the property whose binding is removed