Struct gio::Subprocess

pub struct Subprocess(_, _);

Subprocess allows the creation of and interaction with child processes.

Processes can be communicated with using standard GIO-style APIs (ie: InputStream, OutputStream). There are GIO-style APIs to wait for process termination (ie: cancellable and with an asynchronous variant).

There is an API to force a process to terminate, as well as a race-free API for sending UNIX signals to a subprocess.

One major advantage that GIO brings over the core GLib library is comprehensive API for asynchronous I/O, such OutputStreamExt::splice_async. This makes GSubprocess significantly more powerful and flexible than equivalent APIs in some other languages such as the subprocess.py included with Python. For example, using Subprocess one could create two child processes, reading standard output from the first, processing it, and writing to the input stream of the second, all without blocking the main loop.

A powerful Subprocess::communicate API is provided similar to the communicate() method of subprocess.py. This enables very easy interaction with a subprocess that has been opened with pipes.

Subprocess defaults to tight control over the file descriptors open in the child process, avoiding dangling-fd issues that are caused by a simple fork/exec. The only open file descriptors in the spawned process are ones that were explicitly specified by the Subprocess API (unless SubprocessFlags::InheritFds was specified).

Subprocess will quickly reap all child processes as they exit, avoiding "zombie processes" remaining around for long periods of time. Subprocess::wait can be used to wait for this to happen, but it will happen even without the call being explicitly made.

As a matter of principle, Subprocess has no API that accepts shell-style space-separated strings. It will, however, match the typical shell behaviour of searching the PATH for executables that do not contain a directory separator in their name.

Subprocess attempts to have a very simple API for most uses (ie: spawning a subprocess with arguments and support for most typical kinds of input and output redirection). See Subprocess::new. The SubprocessLauncher API is provided for more complicated cases (advanced types of redirection, environment variable manipulation, change of working directory, child setup functions, etc).

A typical use of Subprocess will involve calling Subprocess::new, followed by Subprocess::wait_async or Subprocess::wait. After the process exits, the status can be checked using functions such as Subprocess::get_if_exited (which are similar to the familiar WIFEXITED-style POSIX macros).

Implements

glib::object::ObjectExt

Methods

impl Subprocess

pub fn communicate_utf8_async<R: FnOnce(Result<(GString, GString), Error>) + Send + 'static, C: IsA<Cancellable>>(
&self,
stdin_buf: Option<String>,
cancellable: Option<&C>,
callback: R
)

Asynchronous version of Subprocess::communicate_utf8. Complete invocation with Subprocess::communicate_utf8_finish.

stdin_buf

Input data, or None

cancellable

Cancellable

callback

Callback

user_data

User data

pub fn communicate_utf8_async_future(
&self,
stdin_buf: Option<String>
) -> Pin<Box<dyn Future<Output = Result<(GString, GString), Error>> + 'static>>

impl Subprocess

pub fn newv(
argv: &[&OsStr],
flags: SubprocessFlags
) -> Result<Subprocess, Error>

Create a new process with the given flags and argument list.

The argument list is expected to be None-terminated.

argv

commandline arguments for the subprocess

flags

flags that define the behaviour of the subprocess

Returns

A newly created Subprocess, or None on error (and error will be set)

pub fn communicate<P: IsA<Cancellable>>(
&self,
stdin_buf: Option<&Bytes>,
cancellable: Option<&P>
) -> Result<(Option<Bytes>, Option<Bytes>), Error>

Communicate with the subprocess until it terminates, and all input and output has been completed.

If stdin_buf is given, the subprocess must have been created with SubprocessFlags::StdinPipe. The given data is fed to the stdin of the subprocess and the pipe is closed (ie: EOF).

At the same time (as not to cause blocking when dealing with large amounts of data), if SubprocessFlags::StdoutPipe or SubprocessFlags::StderrPipe were used, reads from those streams. The data that was read is returned in stdout and/or the stderr.

If the subprocess was created with SubprocessFlags::StdoutPipe, stdout_buf will contain the data read from stdout. Otherwise, for subprocesses not created with SubprocessFlags::StdoutPipe, stdout_buf will be set to None. Similar provisions apply to stderr_buf and SubprocessFlags::StderrPipe.

As usual, any output variable may be given as None to ignore it.

If you desire the stdout and stderr data to be interleaved, create the subprocess with SubprocessFlags::StdoutPipe and SubprocessFlags::StderrMerge. The merged result will be returned in stdout_buf and stderr_buf will be set to None.

In case of any error (including cancellation), false will be returned with error set. Some or all of the stdin data may have been written. Any stdout or stderr data that has been read will be discarded. None of the out variables (aside from error) will have been set to anything in particular and should not be inspected.

In the case that true is returned, the subprocess has exited and the exit status inspection APIs (eg: Subprocess::get_if_exited, Subprocess::get_exit_status) may be used.

You should not attempt to use any of the subprocess pipes after starting this function, since they may be left in strange states, even if the operation was cancelled. You should especially not attempt to interact with the pipes while the operation is in progress (either from another thread or if using the asynchronous version).

stdin_buf

data to send to the stdin of the subprocess, or None

cancellable

a Cancellable

stdout_buf

data read from the subprocess stdout

stderr_buf

data read from the subprocess stderr

Returns

true if successful

pub fn communicate_async<P: IsA<Cancellable>, Q: FnOnce(Result<(Bytes, Bytes), Error>) + Send + 'static>(
&self,
stdin_buf: Option<&Bytes>,
cancellable: Option<&P>,
callback: Q
)

Asynchronous version of Subprocess::communicate. Complete invocation with Subprocess::communicate_finish.

stdin_buf

Input data, or None

cancellable

Cancellable

callback

Callback

user_data

User data

pub fn communicate_async_future(
&self,
stdin_buf: Option<&Bytes>
) -> Pin<Box_<dyn Future<Output = Result<(Bytes, Bytes), Error>> + 'static>>

pub fn communicate_utf8<P: IsA<Cancellable>>(
&self,
stdin_buf: Option<&str>,
cancellable: Option<&P>
) -> Result<(Option<GString>, Option<GString>), Error>

Like Subprocess::communicate, but validates the output of the process as UTF-8, and returns it as a regular NUL terminated string.

On error, stdout_buf and stderr_buf will be set to undefined values and should not be used.

stdin_buf

data to send to the stdin of the subprocess, or None

cancellable

a Cancellable

stdout_buf

data read from the subprocess stdout

stderr_buf

data read from the subprocess stderr

pub fn force_exit(&self)

Use an operating-system specific method to attempt an immediate, forceful termination of the process. There is no mechanism to determine whether or not the request itself was successful; however, you can use Subprocess::wait to monitor the status of the process after calling this function.

On Unix, this function sends SIGKILL.

pub fn get_exit_status(&self) -> i32

Check the exit status of the subprocess, given that it exited normally. This is the value passed to the exit system call or the return value from main.

This is equivalent to the system WEXITSTATUS macro.

It is an error to call this function before Subprocess::wait and unless Subprocess::get_if_exited returned true.

Returns

the exit status

pub fn get_identifier(&self) -> Option<GString>

On UNIX, returns the process ID as a decimal string. On Windows, returns the result of GetProcessId() also as a string. If the subprocess has terminated, this will return None.

Returns

the subprocess identifier, or None if the subprocess has terminated

pub fn get_if_exited(&self) -> bool

Check if the given subprocess exited normally (ie: by way of exit or return from main).

This is equivalent to the system WIFEXITED macro.

It is an error to call this function before Subprocess::wait has returned.

Returns

true if the case of a normal exit

pub fn get_if_signaled(&self) -> bool

Check if the given subprocess terminated in response to a signal.

This is equivalent to the system WIFSIGNALED macro.

It is an error to call this function before Subprocess::wait has returned.

Returns

true if the case of termination due to a signal

pub fn get_status(&self) -> i32

Gets the raw status code of the process, as from waitpid.

This value has no particular meaning, but it can be used with the macros defined by the system headers such as WIFEXITED. It can also be used with g_spawn_check_exit_status.

It is more likely that you want to use Subprocess::get_if_exited followed by Subprocess::get_exit_status.

It is an error to call this function before Subprocess::wait has returned.

Returns

the (meaningless) waitpid exit status from the kernel

pub fn get_stderr_pipe(&self) -> Option<InputStream>

Gets the InputStream from which to read the stderr output of self.

The process must have been created with SubprocessFlags::StderrPipe.

Returns

the stderr pipe

pub fn get_stdin_pipe(&self) -> Option<OutputStream>

Gets the OutputStream that you can write to in order to give data to the stdin of self.

The process must have been created with SubprocessFlags::StdinPipe.

Returns

the stdout pipe

pub fn get_stdout_pipe(&self) -> Option<InputStream>

Gets the InputStream from which to read the stdout output of self.

The process must have been created with SubprocessFlags::StdoutPipe.

Returns

the stdout pipe

pub fn get_successful(&self) -> bool

Checks if the process was "successful". A process is considered successful if it exited cleanly with an exit status of 0, either by way of the exit system call or return from main.

It is an error to call this function before Subprocess::wait has returned.

Returns

true if the process exited cleanly with a exit status of 0

pub fn get_term_sig(&self) -> i32

Get the signal number that caused the subprocess to terminate, given that it terminated due to a signal.

This is equivalent to the system WTERMSIG macro.

It is an error to call this function before Subprocess::wait and unless Subprocess::get_if_signaled returned true.

Returns

the signal causing termination

pub fn send_signal(&self, signal_num: i32)

Sends the UNIX signal signal_num to the subprocess, if it is still running.

This API is race-free. If the subprocess has terminated, it will not be signalled.

This API is not available on Windows.

signal_num

the signal number to send

pub fn wait<P: IsA<Cancellable>>(
&self,
cancellable: Option<&P>
) -> Result<(), Error>

Synchronously wait for the subprocess to terminate.

After the process terminates you can query its exit status with functions such as Subprocess::get_if_exited and Subprocess::get_exit_status.

This function does not fail in the case of the subprocess having abnormal termination. See Subprocess::wait_check for that.

Cancelling cancellable doesn't kill the subprocess. Call Subprocess::force_exit if it is desirable.

cancellable

a Cancellable

Returns

true on success, false if cancellable was cancelled

pub fn wait_async<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
cancellable: Option<&P>,
callback: Q
)

Wait for the subprocess to terminate.

This is the asynchronous version of Subprocess::wait.

cancellable

a Cancellable, or None

callback

a GAsyncReadyCallback to call when the operation is complete

user_data

user_data for callback

pub fn wait_async_future(
&self
) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

pub fn wait_check<P: IsA<Cancellable>>(
&self,
cancellable: Option<&P>
) -> Result<(), Error>

Combines Subprocess::wait with g_spawn_check_exit_status.

cancellable

a Cancellable

Returns

true on success, false if process exited abnormally, or cancellable was cancelled

pub fn wait_check_async<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
cancellable: Option<&P>,
callback: Q
)

Combines Subprocess::wait_async with g_spawn_check_exit_status.

This is the asynchronous version of Subprocess::wait_check.

cancellable

a Cancellable, or None

callback

a GAsyncReadyCallback to call when the operation is complete

user_data

user_data for callback

pub fn wait_check_async_future(
&self
) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>