Struct gio::Task

pub struct Task(_);

A crate::Task represents and manages a cancellable “task”.

Asynchronous operations

The most common usage of crate::Task is as a crate::AsyncResult, to manage data during an asynchronous operation. You call Self::new() in the “start” method, followed by [Self::set_task_data()] and the like if you need to keep some additional data associated with the task, and then pass the task object around through your asynchronous operation. Eventually, you will call a method such as [Self::return_pointer()] or Self::return_error(), which will save the value you give it and then invoke the task’s callback function in the [thread-default main context][g-main-context-push-thread-default] where it was created (waiting until the next iteration of the main loop first, if necessary). The caller will pass the crate::Task back to the operation’s finish function (as a crate::AsyncResult), and you can use [Self::propagate_pointer()] or the like to extract the return value.

Here is an example for using GTask as a GAsyncResult:

    typedef struct {
      CakeFrostingType frosting;
      char *message;
    } DecorationData;

    static void
    decoration_data_free (DecorationData *decoration)
    {
      g_free (decoration->message);
      g_slice_free (DecorationData, decoration);
    }

    static void
    baked_cb (Cake     *cake,
              gpointer  user_data)
    {
      GTask *task = user_data;
      DecorationData *decoration = g_task_get_task_data (task);
      GError *error = NULL;

      if (cake == NULL)
        {
          g_task_return_new_error (task, BAKER_ERROR, BAKER_ERROR_NO_FLOUR,
                                   "Go to the supermarket");
          g_object_unref (task);
          return;
        }

      if (!cake_decorate (cake, decoration->frosting, decoration->message, &error))
        {
          g_object_unref (cake);
          // g_task_return_error() takes ownership of error
          g_task_return_error (task, error);
          g_object_unref (task);
          return;
        }

      g_task_return_pointer (task, cake, g_object_unref);
      g_object_unref (task);
    }

    void
    baker_bake_cake_async (Baker               *self,
                           guint                radius,
                           CakeFlavor           flavor,
                           CakeFrostingType     frosting,
                           const char          *message,
                           GCancellable        *cancellable,
                           GAsyncReadyCallback  callback,
                           gpointer             user_data)
    {
      GTask *task;
      DecorationData *decoration;
      Cake  *cake;

      task = g_task_new (self, cancellable, callback, user_data);
      if (radius < 3)
        {
          g_task_return_new_error (task, BAKER_ERROR, BAKER_ERROR_TOO_SMALL,
                                   "%ucm radius cakes are silly",
                                   radius);
          g_object_unref (task);
          return;
        }

      cake = _baker_get_cached_cake (self, radius, flavor, frosting, message);
      if (cake != NULL)
        {
          // _baker_get_cached_cake() returns a reffed cake
          g_task_return_pointer (task, cake, g_object_unref);
          g_object_unref (task);
          return;
        }

      decoration = g_slice_new (DecorationData);
      decoration->frosting = frosting;
      decoration->message = g_strdup (message);
      g_task_set_task_data (task, decoration, (GDestroyNotify) decoration_data_free);

      _baker_begin_cake (self, radius, flavor, cancellable, baked_cb, task);
    }

    Cake *
    baker_bake_cake_finish (Baker         *self,
                            GAsyncResult  *result,
                            GError       **error)
    {
      g_return_val_if_fail (g_task_is_valid (result, self), NULL);

      return g_task_propagate_pointer (G_TASK (result), error);
    }

Chained asynchronous operations

crate::Task also tries to simplify asynchronous operations that internally chain together several smaller asynchronous operations. [Self::get_cancellable()], [Self::get_context()], and [Self::get_priority()] allow you to get back the task’s crate::Cancellable, crate::glib::MainContext, and [I/O priority][io-priority] when starting a new subtask, so you don’t have to keep track of them yourself. [Self::attach_source()] simplifies the case of waiting for a source to fire (automatically using the correct crate::glib::MainContext and priority).

Here is an example for chained asynchronous operations:

    typedef struct {
      Cake *cake;
      CakeFrostingType frosting;
      char *message;
    } BakingData;

    static void
    decoration_data_free (BakingData *bd)
    {
      if (bd->cake)
        g_object_unref (bd->cake);
      g_free (bd->message);
      g_slice_free (BakingData, bd);
    }

    static void
    decorated_cb (Cake         *cake,
                  GAsyncResult *result,
                  gpointer      user_data)
    {
      GTask *task = user_data;
      GError *error = NULL;

      if (!cake_decorate_finish (cake, result, &error))
        {
          g_object_unref (cake);
          g_task_return_error (task, error);
          g_object_unref (task);
          return;
        }

      // baking_data_free() will drop its ref on the cake, so we have to
      // take another here to give to the caller.
      g_task_return_pointer (task, g_object_ref (cake), g_object_unref);
      g_object_unref (task);
    }

    static gboolean
    decorator_ready (gpointer user_data)
    {
      GTask *task = user_data;
      BakingData *bd = g_task_get_task_data (task);

      cake_decorate_async (bd->cake, bd->frosting, bd->message,
                           g_task_get_cancellable (task),
                           decorated_cb, task);

      return G_SOURCE_REMOVE;
    }

    static void
    baked_cb (Cake     *cake,
              gpointer  user_data)
    {
      GTask *task = user_data;
      BakingData *bd = g_task_get_task_data (task);
      GError *error = NULL;

      if (cake == NULL)
        {
          g_task_return_new_error (task, BAKER_ERROR, BAKER_ERROR_NO_FLOUR,
                                   "Go to the supermarket");
          g_object_unref (task);
          return;
        }

      bd->cake = cake;

      // Bail out now if the user has already cancelled
      if (g_task_return_error_if_cancelled (task))
        {
          g_object_unref (task);
          return;
        }

      if (cake_decorator_available (cake))
        decorator_ready (task);
      else
        {
          GSource *source;

          source = cake_decorator_wait_source_new (cake);
          // Attach @source to @task's GMainContext and have it call
          // decorator_ready() when it is ready.
          g_task_attach_source (task, source, decorator_ready);
          g_source_unref (source);
        }
    }

    void
    baker_bake_cake_async (Baker               *self,
                           guint                radius,
                           CakeFlavor           flavor,
                           CakeFrostingType     frosting,
                           const char          *message,
                           gint                 priority,
                           GCancellable        *cancellable,
                           GAsyncReadyCallback  callback,
                           gpointer             user_data)
    {
      GTask *task;
      BakingData *bd;

      task = g_task_new (self, cancellable, callback, user_data);
      g_task_set_priority (task, priority);

      bd = g_slice_new0 (BakingData);
      bd->frosting = frosting;
      bd->message = g_strdup (message);
      g_task_set_task_data (task, bd, (GDestroyNotify) baking_data_free);

      _baker_begin_cake (self, radius, flavor, cancellable, baked_cb, task);
    }

    Cake *
    baker_bake_cake_finish (Baker         *self,
                            GAsyncResult  *result,
                            GError       **error)
    {
      g_return_val_if_fail (g_task_is_valid (result, self), NULL);

      return g_task_propagate_pointer (G_TASK (result), error);
    }

Asynchronous operations from synchronous ones

You can use [Self::run_in_thread()] to turn a synchronous operation into an asynchronous one, by running it in a thread. When it completes, the result will be dispatched to the [thread-default main context][g-main-context-push-thread-default] where the crate::Task was created.

Running a task in a thread:

    typedef struct {
      guint radius;
      CakeFlavor flavor;
      CakeFrostingType frosting;
      char *message;
    } CakeData;

    static void
    cake_data_free (CakeData *cake_data)
    {
      g_free (cake_data->message);
      g_slice_free (CakeData, cake_data);
    }

    static void
    bake_cake_thread (GTask         *task,
                      gpointer       source_object,
                      gpointer       task_data,
                      GCancellable  *cancellable)
    {
      Baker *self = source_object;
      CakeData *cake_data = task_data;
      Cake *cake;
      GError *error = NULL;

      cake = bake_cake (baker, cake_data->radius, cake_data->flavor,
                        cake_data->frosting, cake_data->message,
                        cancellable, &error);
      if (cake)
        g_task_return_pointer (task, cake, g_object_unref);
      else
        g_task_return_error (task, error);
    }

    void
    baker_bake_cake_async (Baker               *self,
                           guint                radius,
                           CakeFlavor           flavor,
                           CakeFrostingType     frosting,
                           const char          *message,
                           GCancellable        *cancellable,
                           GAsyncReadyCallback  callback,
                           gpointer             user_data)
    {
      CakeData *cake_data;
      GTask *task;

      cake_data = g_slice_new (CakeData);
      cake_data->radius = radius;
      cake_data->flavor = flavor;
      cake_data->frosting = frosting;
      cake_data->message = g_strdup (message);
      task = g_task_new (self, cancellable, callback, user_data);
      g_task_set_task_data (task, cake_data, (GDestroyNotify) cake_data_free);
      g_task_run_in_thread (task, bake_cake_thread);
      g_object_unref (task);
    }

    Cake *
    baker_bake_cake_finish (Baker         *self,
                            GAsyncResult  *result,
                            GError       **error)
    {
      g_return_val_if_fail (g_task_is_valid (result, self), NULL);

      return g_task_propagate_pointer (G_TASK (result), error);
    }

Adding cancellability to uncancellable tasks

Finally, [Self::run_in_thread()] and [Self::run_in_thread_sync()] can be used to turn an uncancellable operation into a cancellable one. If you call Self::set_return_on_cancel(), passing true, then if the task’s crate::Cancellable is cancelled, it will return control back to the caller immediately, while allowing the task thread to continue running in the background (and simply discarding its result when it finally does finish). Provided that the task thread is careful about how it uses locks and other externally-visible resources, this allows you to make “GLib-friendly” asynchronous and cancellable synchronous variants of blocking APIs.

Cancelling a task:

    static void
    bake_cake_thread (GTask         *task,
                      gpointer       source_object,
                      gpointer       task_data,
                      GCancellable  *cancellable)
    {
      Baker *self = source_object;
      CakeData *cake_data = task_data;
      Cake *cake;
      GError *error = NULL;

      cake = bake_cake (baker, cake_data->radius, cake_data->flavor,
                        cake_data->frosting, cake_data->message,
                        &error);
      if (error)
        {
          g_task_return_error (task, error);
          return;
        }

      // If the task has already been cancelled, then we don't want to add
      // the cake to the cake cache. Likewise, we don't  want to have the
      // task get cancelled in the middle of updating the cache.
      // g_task_set_return_on_cancel() will return %TRUE here if it managed
      // to disable return-on-cancel, or %FALSE if the task was cancelled
      // before it could.
      if (g_task_set_return_on_cancel (task, FALSE))
        {
          // If the caller cancels at this point, their
          // GAsyncReadyCallback won't be invoked until we return,
          // so we don't have to worry that this code will run at
          // the same time as that code does. But if there were
          // other functions that might look at the cake cache,
          // then we'd probably need a GMutex here as well.
          baker_add_cake_to_cache (baker, cake);
          g_task_return_pointer (task, cake, g_object_unref);
        }
    }

    void
    baker_bake_cake_async (Baker               *self,
                           guint                radius,
                           CakeFlavor           flavor,
                           CakeFrostingType     frosting,
                           const char          *message,
                           GCancellable        *cancellable,
                           GAsyncReadyCallback  callback,
                           gpointer             user_data)
    {
      CakeData *cake_data;
      GTask *task;

      cake_data = g_slice_new (CakeData);

      ...

      task = g_task_new (self, cancellable, callback, user_data);
      g_task_set_task_data (task, cake_data, (GDestroyNotify) cake_data_free);
      g_task_set_return_on_cancel (task, TRUE);
      g_task_run_in_thread (task, bake_cake_thread);
    }

    Cake *
    baker_bake_cake_sync (Baker               *self,
                          guint                radius,
                          CakeFlavor           flavor,
                          CakeFrostingType     frosting,
                          const char          *message,
                          GCancellable        *cancellable,
                          GError             **error)
    {
      CakeData *cake_data;
      GTask *task;
      Cake *cake;

      cake_data = g_slice_new (CakeData);

      ...

      task = g_task_new (self, cancellable, NULL, NULL);
      g_task_set_task_data (task, cake_data, (GDestroyNotify) cake_data_free);
      g_task_set_return_on_cancel (task, TRUE);
      g_task_run_in_thread_sync (task, bake_cake_thread);

      cake = g_task_propagate_pointer (task, error);
      g_object_unref (task);
      return cake;
    }

Porting from GSimpleAsyncResult

crate::Task’s API attempts to be simpler than [crate::SimpleAsyncResult]’s in several ways:

You can save task-specific data with [Self::set_task_data()], and retrieve it later with [Self::get_task_data()]. This replaces the abuse of [crate::SimpleAsyncResult::set_op_res_gpointer()] for the same purpose with [crate::SimpleAsyncResult].

In addition to the task data, crate::Task also keeps track of the [priority][io-priority], crate::Cancellable, and crate::glib::MainContext associated with the task, so tasks that consist of a chain of simpler asynchronous operations will have easy access to those values when starting each sub-task.

Self::return_error_if_cancelled() provides simplified handling for cancellation. In addition, cancellation overrides any other crate::Task return value by default, like [crate::SimpleAsyncResult] does when [crate::SimpleAsyncResult::set_check_cancellable()] is called. (You can use Self::set_check_cancellable() to turn off that behavior.) On the other hand, [Self::run_in_thread()] guarantees that it will always run your task_func, even if the task’s crate::Cancellable is already cancelled before the task gets a chance to run; you can start your task_func with a Self::return_error_if_cancelled() check if you need the old behavior.

The “return” methods (eg, [Self::return_pointer()]) automatically cause the task to be “completed” as well, and there is no need to worry about the “complete” vs “complete in idle” distinction. (crate::Task automatically figures out whether the task’s callback can be invoked directly, or if it needs to be sent to another crate::glib::MainContext, or delayed until the next iteration of the current crate::glib::MainContext.)

The “finish” functions for crate::Task based operations are generally much simpler than [crate::SimpleAsyncResult] ones, normally consisting of only a single call to [Self::propagate_pointer()] or the like. Since [Self::propagate_pointer()] “steals” the return value from the crate::Task, it is not necessary to juggle pointers around to prevent it from being freed twice.

With [crate::SimpleAsyncResult], it was common to call [crate::SimpleAsyncResult::propagate_error()] from the _finish() wrapper function, and have virtual method implementations only deal with successful returns. This behavior is deprecated, because it makes it difficult for a subclass to chain to a parent class’s async methods. Instead, the wrapper function should just be a simple wrapper, and the virtual method should call an appropriate g_task_propagate_ function. Note that wrapper methods can now use [crate::AsyncResult::legacy_propagate_error()] to do old-style [crate::SimpleAsyncResult] error-returning behavior, and [crate::AsyncResult::is_tagged()] to check if a result is tagged as having come from the _async() wrapper function (for “short-circuit” results, such as when passing 0 to [crate::InputStream::read_async()]).

Implements

glib::object::ObjectExt, crate::prelude::AsyncResultExt

Implementations

impl Task

pub fn cancellable(&self) -> Cancellable

Gets self’s crate::Cancellable

Returns

self’s crate::Cancellable

pub fn is_check_cancellable(&self) -> bool

Gets self’s check-cancellable flag. See Self::set_check_cancellable() for more details.

pub fn is_completed(&self) -> bool

Gets the value of [crate::Task:completed]. This changes from false to true after the task’s callback is invoked, and will return false if called from inside the callback.

Returns

true if the task has completed, false otherwise.

pub fn context(&self) -> MainContext

Gets the crate::glib::MainContext that self will return its result in (that is, the context that was the [thread-default main context][g-main-context-push-thread-default] at the point when self was created).

This will always return a non-None value, even if the task’s context is the default crate::glib::MainContext.

Returns

self’s crate::glib::MainContext

pub fn name(&self) -> Option<GString>

This is supported on crate feature v2_60 only.

Gets self’s name. See Self::set_name().

Returns

self’s name, or None

pub fn is_return_on_cancel(&self) -> bool

Gets self’s return-on-cancel flag. See Self::set_return_on_cancel() for more details.

pub fn had_error(&self) -> bool

Tests if self resulted in an error.

Returns

true if the task resulted in an error, false otherwise.

pub fn return_error_if_cancelled(&self) -> bool

Checks if self’s crate::Cancellable has been cancelled, and if so, sets self’s error accordingly and completes the task (see [Self::return_pointer()] for more discussion of exactly what this means).

Returns

true if self has been cancelled, false if not

pub fn set_check_cancellable(&self, check_cancellable: bool)

Sets or clears self’s check-cancellable flag. If this is true (the default), then [Self::propagate_pointer()], etc, and Self::had_error() will check the task’s crate::Cancellable first, and if it has been cancelled, then they will consider the task to have returned an “Operation was cancelled” error (crate::IOErrorEnum::Cancelled), regardless of any other error or return value the task may have had.

If check_cancellable is false, then the crate::Task will not check the cancellable itself, and it is up to self’s owner to do this (eg, via Self::return_error_if_cancelled()).

If you are using Self::set_return_on_cancel() as well, then you must leave check-cancellable set true.

check_cancellable

whether crate::Task will check the state of its crate::Cancellable for you.

pub fn set_name(&self, name: Option<&str>)

This is supported on crate feature v2_60 only.

Sets self’s name, used in debugging and profiling. The name defaults to None.

The task name should describe in a human readable way what the task does. For example, ‘Open file’ or ‘Connect to network host’. It is used to set the name of the crate::glib::Source used for idle completion of the task.

This function may only be called before the self is first used in a thread other than the one it was constructed in.

name

a human readable name for the task, or None to unset it

pub fn set_return_on_cancel(&self, return_on_cancel: bool) -> bool

Sets or clears self’s return-on-cancel flag. This is only meaningful for tasks run via [Self::run_in_thread()] or [Self::run_in_thread_sync()].

If return_on_cancel is true, then cancelling self’s crate::Cancellable will immediately cause it to return, as though the task’s GTaskThreadFunc had called Self::return_error_if_cancelled() and then returned.

This allows you to create a cancellable wrapper around an uninterruptible function. The GTaskThreadFunc just needs to be careful that it does not modify any externally-visible state after it has been cancelled. To do that, the thread should call Self::set_return_on_cancel() again to (atomically) set return-on-cancel false before making externally-visible changes; if the task gets cancelled before the return-on-cancel flag could be changed, Self::set_return_on_cancel() will indicate this by returning false.

You can disable and re-enable this flag multiple times if you wish. If the task’s crate::Cancellable is cancelled while return-on-cancel is false, then calling Self::set_return_on_cancel() to set it true again will cause the task to be cancelled at that point.

If the task’s crate::Cancellable is already cancelled before you call [Self::run_in_thread()]/[Self::run_in_thread_sync()], then the GTaskThreadFunc will still be run (for consistency), but the task will also be completed right away.

return_on_cancel

whether the task returns automatically when it is cancelled.

Returns

true if self’s return-on-cancel flag was changed to match return_on_cancel. false if self has already been cancelled.

pub fn is_valid<P: IsA<AsyncResult>, Q: IsA<Object>>(
result: &P,
source_object: Option<&Q>
) -> bool

Checks that result is a crate::Task, and that source_object is its source object (or that source_object is None and result has no source object). This can be used in g_return_if_fail() checks.

result

A crate::AsyncResult

source_object

the source object expected to be associated with the task

Returns

true if result and source_object are valid, false if not

pub fn connect_completed_notify<F: Fn(&Task) + 'static>(
&self,
f: F
) -> SignalHandlerId

impl Task

pub fn new<P: IsA<Cancellable>, Q: FnOnce(&AsyncResult, Option<&Object>) + 'static>(
source_object: Option<&Object>,
cancellable: Option<&P>,
callback: Q
) -> Task

Creates a crate::Task acting on source_object, which will eventually be used to invoke callback in the current [thread-default main context][g-main-context-push-thread-default].

Call this in the “start” method of your asynchronous method, and pass the crate::Task around throughout the asynchronous operation. You can use [Self::set_task_data()] to attach task-specific data to the object, which you can retrieve later via [Self::get_task_data()].

By default, if cancellable is cancelled, then the return value of the task will always be crate::IOErrorEnum::Cancelled, even if the task had already completed before the cancellation. This allows for simplified handling in cases where cancellation may imply that other objects that the task depends on have been destroyed. If you do not want this behavior, you can use Self::set_check_cancellable() to change it.

source_object

the crate::glib::object::Object that owns this task, or None.

cancellable

optional crate::Cancellable object, None to ignore.

callback

a GAsyncReadyCallback.

callback_data

user data passed to callback.

Returns

a crate::Task.

pub fn return_error(&self, error: Error)

Sets self’s result to error (which self assumes ownership of) and completes the task (see [Self::return_pointer()] for more discussion of exactly what this means).

Note that since the task takes ownership of error, and since the task may be completed before returning from Self::return_error(), you cannot assume that error is still valid after calling this. Call [crate::glib::Error::copy()] on the error if you need to keep a local copy as well.

See also [Self::return_new_error()].

error

the crate::glib::Error result of a task function.

pub fn priority(&self) -> Priority

Gets self’s priority

Returns

self’s priority

pub fn set_priority(&self, priority: Priority)

Sets self’s priority. If you do not call this, it will default to G_PRIORITY_DEFAULT.

This will affect the priority of GSources created with [Self::attach_source()] and the scheduling of tasks run in threads, and can also be explicitly retrieved later via [Self::get_priority()].

priority

the [priority][io-priority] of the request

pub fn return_value(&self, result: &Value)

Sets self’s result to result (by copying it) and completes the task.

If result is None then a [crate::glib::object::Value] of type G_TYPE_POINTER with a value of None will be used for the result.

This is a very generic low-level method intended primarily for use by language bindings; for C code, [Self::return_pointer()] and the like will normally be much easier to use.

result

the [crate::glib::object::Value] result of a task function

pub fn propagate_value(&self) -> Result<Value, Error>

Gets the result of self as a [crate::glib::object::Value], and transfers ownership of that value to the caller. As with Self::return_value(), this is a generic low-level method; [Self::propagate_pointer()] and the like will usually be more useful for C code.

If the task resulted in an error, or was cancelled, then this will instead set error and return false.

Since this method transfers ownership of the return value (or error) to the caller, you may only call it once.

value

return location for the [crate::glib::object::Value]

Returns

true if self succeeded, false on error.