Struct gio::Task

pub struct Task(_);

Implementations

impl Task

pub fn get_cancellable(&self) -> Option<Cancellable>

Gets self's Cancellable

Returns

self's Cancellable

pub fn get_check_cancellable(&self) -> bool

Gets self's check-cancellable flag. See Task::set_check_cancellable for more details.

pub fn get_completed(&self) -> bool

This is supported on crate feature v2_44 only.

Gets the value of Task:completed. This changes from false to true after the task’s callback is invoked, and will return false if called from inside the callback.

Feature: v2_44

Returns

true if the task has completed, false otherwise.

pub fn get_context(&self) -> Option<MainContext>

Gets the glib::MainContext that self will return its result in (that is, the context that was the [thread-default main context][g-main-context-push-thread-default] at the point when self was created).

This will always return a non-None value, even if the task's context is the default glib::MainContext.

Returns

self's glib::MainContext

pub fn get_name(&self) -> Option<GString>

This is supported on crate feature v2_60 only.

Gets self’s name. See Task::set_name.

Feature: v2_60

Returns

self’s name, or None

pub fn get_return_on_cancel(&self) -> bool

Gets self's return-on-cancel flag. See Task::set_return_on_cancel for more details.

pub fn had_error(&self) -> bool

Tests if self resulted in an error.

Returns

true if the task resulted in an error, false otherwise.

pub fn return_error_if_cancelled(&self) -> bool

Checks if self's Cancellable has been cancelled, and if so, sets self's error accordingly and completes the task (see Task::return_pointer for more discussion of exactly what this means).

Returns

true if self has been cancelled, false if not

pub fn set_check_cancellable(&self, check_cancellable: bool)

Sets or clears self's check-cancellable flag. If this is true (the default), then Task::propagate_pointer, etc, and Task::had_error will check the task's Cancellable first, and if it has been cancelled, then they will consider the task to have returned an "Operation was cancelled" error (IOErrorEnum::Cancelled), regardless of any other error or return value the task may have had.

If check_cancellable is false, then the Task will not check the cancellable itself, and it is up to self's owner to do this (eg, via Task::return_error_if_cancelled).

If you are using Task::set_return_on_cancel as well, then you must leave check-cancellable set true.

check_cancellable

whether Task will check the state of its Cancellable for you.

pub fn set_name(&self, name: Option<&str>)

This is supported on crate feature v2_60 only.

Sets self’s name, used in debugging and profiling. The name defaults to None.

The task name should describe in a human readable way what the task does. For example, ‘Open file’ or ‘Connect to network host’. It is used to set the name of the glib::Source used for idle completion of the task.

This function may only be called before the self is first used in a thread other than the one it was constructed in.

Feature: v2_60

name

a human readable name for the task, or None to unset it

pub fn set_return_on_cancel(&self, return_on_cancel: bool) -> bool

Sets or clears self's return-on-cancel flag. This is only meaningful for tasks run via Task::run_in_thread or Task::run_in_thread_sync.

If return_on_cancel is true, then cancelling self's Cancellable will immediately cause it to return, as though the task's GTaskThreadFunc had called Task::return_error_if_cancelled and then returned.

This allows you to create a cancellable wrapper around an uninterruptible function. The GTaskThreadFunc just needs to be careful that it does not modify any externally-visible state after it has been cancelled. To do that, the thread should call Task::set_return_on_cancel again to (atomically) set return-on-cancel false before making externally-visible changes; if the task gets cancelled before the return-on-cancel flag could be changed, Task::set_return_on_cancel will indicate this by returning false.

You can disable and re-enable this flag multiple times if you wish. If the task's Cancellable is cancelled while return-on-cancel is false, then calling Task::set_return_on_cancel to set it true again will cause the task to be cancelled at that point.

If the task's Cancellable is already cancelled before you call Task::run_in_thread/Task::run_in_thread_sync, then the GTaskThreadFunc will still be run (for consistency), but the task will also be completed right away.

return_on_cancel

whether the task returns automatically when it is cancelled.

Returns

true if self's return-on-cancel flag was changed to match return_on_cancel. false if self has already been cancelled.

pub fn is_valid<P: IsA<AsyncResult>, Q: IsA<Object>>(
result: &P,
source_object: Option<&Q>
) -> bool

Checks that result is a Task, and that source_object is its source object (or that source_object is None and result has no source object). This can be used in g_return_if_fail checks.

result

A AsyncResult

source_object

the source object expected to be associated with the task

Returns

true if result and source_object are valid, false if not

pub fn connect_property_completed_notify<F: Fn(&Task) + 'static>(
&self,
f: F
) -> SignalHandlerId

This is supported on crate feature v2_44 only.

impl Task

pub fn new<P: IsA<Cancellable>, Q: FnOnce(&AsyncResult, Option<&Object>) + 'static>(
source_object: Option<&Object>,
cancellable: Option<&P>,
callback: Q
) -> Task

Creates a Task acting on source_object, which will eventually be used to invoke callback in the current [thread-default main context][g-main-context-push-thread-default].

Call this in the "start" method of your asynchronous method, and pass the Task around throughout the asynchronous operation. You can use Task::set_task_data to attach task-specific data to the object, which you can retrieve later via Task::get_task_data.

By default, if cancellable is cancelled, then the return value of the task will always be IOErrorEnum::Cancelled, even if the task had already completed before the cancellation. This allows for simplified handling in cases where cancellation may imply that other objects that the task depends on have been destroyed. If you do not want this behavior, you can use Task::set_check_cancellable to change it.

source_object

the gobject::Object that owns this task, or None.

cancellable

optional Cancellable object, None to ignore.

callback

a GAsyncReadyCallback.

callback_data

user data passed to callback.

Returns

a Task.

pub fn return_error(&self, error: Error)

Sets self's result to error (which self assumes ownership of) and completes the task (see Task::return_pointer for more discussion of exactly what this means).

Note that since the task takes ownership of error, and since the task may be completed before returning from Task::return_error, you cannot assume that error is still valid after calling this. Call glib::Error::copy on the error if you need to keep a local copy as well.

See also Task::return_new_error.

error

the glib::Error result of a task function.

pub fn get_priority(&self) -> Priority

Gets self's priority

Returns

self's priority

pub fn set_priority(&self, priority: Priority)

Sets self's priority. If you do not call this, it will default to G_PRIORITY_DEFAULT.

This will affect the priority of GSources created with Task::attach_source and the scheduling of tasks run in threads, and can also be explicitly retrieved later via Task::get_priority.

priority

the [priority][io-priority] of the request

pub fn return_value(&self, result: &Value)

Sets self's result to result (by copying it) and completes the task.

If result is None then a gobject::Value of type G_TYPE_POINTER with a value of None will be used for the result.

This is a very generic low-level method intended primarily for use by language bindings; for C code, Task::return_pointer and the like will normally be much easier to use.

Feature: v2_64

result

the gobject::Value result of a task function

pub fn propagate_value(&self) -> Result<Value, Error>

Gets the result of self as a gobject::Value, and transfers ownership of that value to the caller. As with Task::return_value, this is a generic low-level method; Task::propagate_pointer and the like will usually be more useful for C code.

If the task resulted in an error, or was cancelled, then this will instead set error and return false.

Since this method transfers ownership of the return value (or error) to the caller, you may only call it once.

Feature: v2_64

value

return location for the gobject::Value

Returns

true if self succeeded, false on error.