Trait gio::TlsCertificateExt

pub trait TlsCertificateExt {
    fn get_issuer(&self) -> Option<TlsCertificate>;

    fn is_same(&self, cert_two: &TlsCertificate) -> bool;

    fn verify<'a, 'b, P: IsA<SocketConnectable> + 'a, Q: Into<Option<&'a P>>, R: Into<Option<&'b TlsCertificate>>>(
        &self, 
        identity: Q, 
        trusted_ca: R
    ) -> TlsCertificateFlags;

    fn get_property_certificate_pem(&self) -> Option<String>;

    fn connect_property_certificate_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_certificate_pem_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_issuer_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_private_key_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_private_key_pem_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all TlsCertificate methods.

Implementors

TlsCertificate

Required Methods

fn get_issuer(&self) -> Option<TlsCertificate>

Gets the TlsCertificate representing self's issuer, if known

Returns

The certificate of self's issuer, or None if self is self-signed or signed with an unknown certificate.

fn is_same(&self, cert_two: &TlsCertificate) -> bool

Check if two TlsCertificate objects represent the same certificate. The raw DER byte data of the two certificates are checked for equality. This has the effect that two certificates may compare equal even if their TlsCertificate:issuer, TlsCertificate:private-key, or TlsCertificate:private-key-pem properties differ.

Feature: v2_34

cert_two

second certificate to compare

Returns

whether the same or not

fn verify<'a, 'b, P: IsA<SocketConnectable> + 'a, Q: Into<Option<&'a P>>, R: Into<Option<&'b TlsCertificate>>>(
&self,
identity: Q,
trusted_ca: R
) -> TlsCertificateFlags

This verifies self and returns a set of TlsCertificateFlags indicating any problems found with it. This can be used to verify a certificate outside the context of making a connection, or to check a certificate against a CA that is not part of the system CA database.

If identity is not None, self's name(s) will be compared against it, and TlsCertificateFlags::BadIdentity will be set in the return value if it does not match. If identity is None, that bit will never be set in the return value.

If trusted_ca is not None, then self (or one of the certificates in its chain) must be signed by it, or else TlsCertificateFlags::UnknownCa will be set in the return value. If trusted_ca is None, that bit will never be set in the return value.

(All other TlsCertificateFlags values will always be set or unset as appropriate.)

identity

the expected peer identity

trusted_ca

the certificate of a trusted authority

Returns

the appropriate TlsCertificateFlags

fn get_property_certificate_pem(&self) -> Option<String>

fn connect_property_certificate_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_certificate_pem_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_issuer_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_private_key_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_private_key_pem_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId